refactor(i18n): split overrides monolith into per-locale files (#285) (#301)
Break the 3261-line src/core/i18n/messages/overrides.ts into one file
per locale under overrides/, with an index that assembles
MESSAGE_OVERRIDES. Folds the cross-cutting REVIEW_GATE_TRANSLATIONS map
and its 14 spreads into each locale's own override object.

Pure restructure: each locale's final override catalog is byte-identical
to before (verified by deep-equal snapshot), so getMessages output is
unchanged.
