--- src/core/i18n/messages/overrides.ts
@@ -1,3261 +1,0 @@
-import type { SupportedLocale } from '@/core/i18n/locales'
-import type { MessageCatalog } from '@/core/i18n/messages/types'
-
-const REVIEW_GATE_TRANSLATIONS: Partial<
-  Record<Exclude<SupportedLocale, 'en'>, Partial<MessageCatalog>>
-> = {
-  es: {
-    'common.confirm': 'Confirmar',
-    'discover.filters': 'Filtros',
-    'discover.filtersActive': 'Filtros ({0})',
-    'discover.moreActions': 'Más acciones',
-    'discover.decadesLabel': 'Décadas',
-    'analytics.title': 'Analíticas',
-    'dashboard.title': 'Panel',
-    'discover.title': 'Descubrir',
-    'setup.title': 'Configuración',
-    'todaysPick.discoverSimilar': 'Descubrir similares',
-    'libraryHealth.fix': 'Corregir',
-    'libraryHealth.fixing': 'Corrigiendo...',
-    'libraryHealth.showLess': 'Mostrar menos',
-    'libraryHealth.showAll': 'Mostrar todo {0}',
-    'libraryHealth.fixHint.missingMetadata': 'Actualiza metadatos de artistas desde Lidarr',
-    'libraryHealth.fixHint.unmonitored':
-      'Marca todos los artistas listados como monitorizados en Lidarr',
-    'libraryHealth.fixHint.missingAlbums': 'Actualiza álbumes en Lidarr para encontrar faltantes',
-    'libraryHealth.fixHint.genreGaps': 'Actualiza metadatos en Lidarr para obtener géneros',
-    'libraryHealth.fixHint.imageGaps': 'Vuelve a buscar imágenes de artistas',
-    'libraryHealth.fixHint.missingWikidata': 'Obtiene biografías y enlaces desde Wikidata',
-  },
-  fr: {
-    'common.confirm': 'Confirmer',
-    'discover.filters': 'Filtres',
-    'discover.filtersActive': 'Filtres ({0})',
-    'discover.moreActions': "Plus d'actions",
-    'discover.decadesLabel': 'Décennies',
-    'analytics.title': 'Analyses',
-    'dashboard.title': 'Tableau de bord',
-    'discover.title': 'Découvrir',
-    'setup.title': 'Configuration',
-    'todaysPick.discoverSimilar': 'Découvrir des similaires',
-    'libraryHealth.fix': 'Corriger',
-    'libraryHealth.fixing': 'Correction...',
-    'libraryHealth.showLess': 'Afficher moins',
-    'libraryHealth.showAll': 'Tout afficher {0}',
-    'libraryHealth.fixHint.missingMetadata': 'Actualise les métadonnées des artistes depuis Lidarr',
-    'libraryHealth.fixHint.unmonitored':
-      'Marque tous les artistes listés comme surveillés dans Lidarr',
-    'libraryHealth.fixHint.missingAlbums':
-      'Actualise les albums dans Lidarr pour trouver les manques',
-    'libraryHealth.fixHint.genreGaps': 'Actualise les métadonnées Lidarr pour récupérer les genres',
-    'libraryHealth.fixHint.imageGaps': "Relance la recherche d'images d'artistes",
-    'libraryHealth.fixHint.missingWikidata': 'Récupère les biographies et liens depuis Wikidata',
-  },
-  de: {
-    'common.confirm': 'Bestätigen',
-    'discover.filters': 'Filter',
-    'discover.filtersActive': 'Filter ({0})',
-    'discover.moreActions': 'Weitere Aktionen',
-    'discover.decadesLabel': 'Jahrzehnte',
-    'analytics.title': 'Analysen',
-    'dashboard.title': 'Übersicht',
-    'discover.title': 'Entdecken',
-    'setup.title': 'Einrichtung',
-    'todaysPick.discoverSimilar': 'Ähnliche entdecken',
-    'libraryHealth.fix': 'Beheben',
-    'libraryHealth.fixing': 'Wird behoben...',
-    'libraryHealth.showLess': 'Weniger anzeigen',
-    'libraryHealth.showAll': 'Alle {0} anzeigen',
-    'libraryHealth.fixHint.missingMetadata': 'Aktualisiert Künstler-Metadaten aus Lidarr',
-    'libraryHealth.fixHint.unmonitored': 'Setzt alle aufgeführten Künstler in Lidarr auf überwacht',
-    'libraryHealth.fixHint.missingAlbums': 'Aktualisiert Alben in Lidarr, um fehlende zu finden',
-    'libraryHealth.fixHint.genreGaps': 'Aktualisiert Lidarr-Metadaten, um Genres abzurufen',
-    'libraryHealth.fixHint.imageGaps': 'Sucht Künstlerbilder erneut',
-    'libraryHealth.fixHint.missingWikidata': 'Holt Biografien und Links aus Wikidata',
-  },
-  'pt-BR': {
-    'common.confirm': 'Confirmar',
-    'discover.filters': 'Filtros',
-    'discover.filtersActive': 'Filtros ({0})',
-    'discover.moreActions': 'Mais ações',
-    'discover.decadesLabel': 'Décadas',
-    'analytics.title': 'Análises',
-    'dashboard.title': 'Painel',
-    'discover.title': 'Descobrir',
-    'setup.title': 'Configuração',
-    'todaysPick.discoverSimilar': 'Descobrir similares',
-    'libraryHealth.fix': 'Corrigir',
-    'libraryHealth.fixing': 'Corrigindo...',
-    'libraryHealth.showLess': 'Mostrar menos',
-    'libraryHealth.showAll': 'Mostrar todos {0}',
-    'libraryHealth.fixHint.missingMetadata': 'Atualiza metadados de artistas pelo Lidarr',
-    'libraryHealth.fixHint.unmonitored':
-      'Define todos os artistas listados como monitorados no Lidarr',
-    'libraryHealth.fixHint.missingAlbums': 'Atualiza álbuns no Lidarr para encontrar ausentes',
-    'libraryHealth.fixHint.genreGaps': 'Atualiza metadados do Lidarr para buscar gêneros',
-    'libraryHealth.fixHint.imageGaps': 'Busca imagens de artistas novamente',
-    'libraryHealth.fixHint.missingWikidata': 'Busca biografias e links no Wikidata',
-  },
-  it: {
-    'common.confirm': 'Conferma',
-    'discover.filters': 'Filtri',
-    'discover.filtersActive': 'Filtri ({0})',
-    'discover.moreActions': 'Altre azioni',
-    'discover.decadesLabel': 'Decenni',
-    'analytics.title': 'Analisi',
-    'dashboard.title': 'Cruscotto',
-    'discover.title': 'Scopri',
-    'setup.title': 'Configurazione',
-    'todaysPick.discoverSimilar': 'Scopri simili',
-    'libraryHealth.fix': 'Correggi',
-    'libraryHealth.fixing': 'Correzione...',
-    'libraryHealth.showLess': 'Mostra meno',
-    'libraryHealth.showAll': 'Mostra tutti {0}',
-    'libraryHealth.fixHint.missingMetadata': 'Aggiorna i metadati degli artisti da Lidarr',
-    'libraryHealth.fixHint.unmonitored':
-      'Imposta tutti gli artisti elencati come monitorati in Lidarr',
-    'libraryHealth.fixHint.missingAlbums':
-      'Aggiorna gli album in Lidarr per trovare quelli mancanti',
-    'libraryHealth.fixHint.genreGaps': 'Aggiorna i metadati Lidarr per recuperare i generi',
-    'libraryHealth.fixHint.imageGaps': 'Cerca di nuovo le immagini degli artisti',
-    'libraryHealth.fixHint.missingWikidata': 'Recupera biografie e link da Wikidata',
-  },
-  nl: {
-    'common.confirm': 'Bevestigen',
-    'discover.filters': 'Filteropties',
-    'discover.filtersActive': 'Filteropties ({0})',
-    'discover.moreActions': 'Meer acties',
-    'discover.decadesLabel': 'Decennia',
-    'analytics.title': 'Analyse',
-    'dashboard.title': 'Overzicht',
-    'discover.title': 'Ontdekken',
-    'setup.title': 'Instellen',
-    'todaysPick.discoverSimilar': 'Vergelijkbare ontdekken',
-    'libraryHealth.fix': 'Repareren',
-    'libraryHealth.fixing': 'Repareren...',
-    'libraryHealth.showLess': 'Minder tonen',
-    'libraryHealth.showAll': 'Alle {0} tonen',
-    'libraryHealth.fixHint.missingMetadata': 'Vernieuwt artiestmetadata vanuit Lidarr',
-    'libraryHealth.fixHint.unmonitored': 'Zet alle vermelde artiesten in Lidarr op gemonitord',
-    'libraryHealth.fixHint.missingAlbums': 'Vernieuwt albums in Lidarr om ontbrekende te vinden',
-    'libraryHealth.fixHint.genreGaps': 'Vernieuwt Lidarr-metadata om genres op te halen',
-    'libraryHealth.fixHint.imageGaps': 'Zoekt artiestafbeeldingen opnieuw',
-    'libraryHealth.fixHint.missingWikidata': 'Haalt biografieën en links op uit Wikidata',
-  },
-  ro: {
-    'common.confirm': 'Confirmă',
-    'discover.filters': 'Filtre',
-    'discover.filtersActive': 'Filtre ({0})',
-    'discover.moreActions': 'Mai multe acțiuni',
-    'discover.decadesLabel': 'Decenii',
-    'analytics.title': 'Analize',
-    'dashboard.title': 'Tablou de bord',
-    'discover.title': 'Descoperire',
-    'setup.title': 'Configurare',
-    'todaysPick.discoverSimilar': 'Descoperă similari',
-    'libraryHealth.fix': 'Remediază',
-    'libraryHealth.fixing': 'Se remediază...',
-    'libraryHealth.showLess': 'Afișează mai puțin',
-    'libraryHealth.showAll': 'Afișează toate {0}',
-    'libraryHealth.fixHint.missingMetadata': 'Actualizează metadatele artiștilor din Lidarr',
-    'libraryHealth.fixHint.unmonitored': 'Setează toți artiștii listați ca monitorizați în Lidarr',
-    'libraryHealth.fixHint.missingAlbums':
-      'Actualizează albumele în Lidarr pentru a găsi lipsurile',
-    'libraryHealth.fixHint.genreGaps': 'Actualizează metadatele Lidarr pentru genuri',
-    'libraryHealth.fixHint.imageGaps': 'Caută din nou imagini pentru artiști',
-    'libraryHealth.fixHint.missingWikidata': 'Preia biografii și linkuri din Wikidata',
-  },
-  pl: {
-    'common.confirm': 'Potwierdź',
-    'discover.filters': 'Filtry',
-    'discover.filtersActive': 'Filtry ({0})',
-    'discover.moreActions': 'Więcej działań',
-    'discover.decadesLabel': 'Dekady',
-    'analytics.title': 'Analityka',
-    'dashboard.title': 'Panel',
-    'discover.title': 'Odkrywaj',
-    'setup.title': 'Konfiguracja',
-    'todaysPick.discoverSimilar': 'Odkryj podobnych',
-    'libraryHealth.fix': 'Napraw',
-    'libraryHealth.fixing': 'Naprawianie...',
-    'libraryHealth.showLess': 'Pokaż mniej',
-    'libraryHealth.showAll': 'Pokaż wszystkie {0}',
-    'libraryHealth.fixHint.missingMetadata': 'Odświeża metadane artystów z Lidarr',
-    'libraryHealth.fixHint.unmonitored':
-      'Ustawia wszystkich wymienionych artystów jako monitorowanych w Lidarr',
-    'libraryHealth.fixHint.missingAlbums': 'Odświeża albumy w Lidarr, aby znaleźć brakujące',
-    'libraryHealth.fixHint.genreGaps': 'Odświeża metadane Lidarr, aby pobrać gatunki',
-    'libraryHealth.fixHint.imageGaps': 'Ponownie wyszukuje obrazy artystów',
-    'libraryHealth.fixHint.missingWikidata': 'Pobiera biografie i linki z Wikidata',
-  },
-  tr: {
-    'common.confirm': 'Onayla',
-    'discover.filters': 'Filtreler',
-    'discover.filtersActive': 'Filtreler ({0})',
-    'discover.moreActions': 'Daha fazla işlem',
-    'discover.decadesLabel': 'On yıllar',
-    'analytics.title': 'Analizler',
-    'dashboard.title': 'Pano',
-    'discover.title': 'Keşfet',
-    'setup.title': 'Kurulum',
-    'todaysPick.discoverSimilar': 'Benzerleri keşfet',
-    'libraryHealth.fix': 'Düzelt',
-    'libraryHealth.fixing': 'Düzeltiliyor...',
-    'libraryHealth.showLess': 'Daha az göster',
-    'libraryHealth.showAll': 'Tümünü göster {0}',
-    'libraryHealth.fixHint.missingMetadata': 'Lidarr üzerinden sanatçı meta verilerini yeniler',
-    'libraryHealth.fixHint.unmonitored': 'Listelenen tüm sanatçıları Lidarr içinde izleniyor yapar',
-    'libraryHealth.fixHint.missingAlbums': 'Eksikleri bulmak için Lidarr albümlerini yeniler',
-    'libraryHealth.fixHint.genreGaps': 'Türleri almak için Lidarr meta verilerini yeniler',
-    'libraryHealth.fixHint.imageGaps': 'Sanatçı görsellerini yeniden arar',
-    'libraryHealth.fixHint.missingWikidata': 'Wikidata üzerinden biyografi ve bağlantıları alır',
-  },
-  uk: {
-    'common.confirm': 'Підтвердити',
-    'discover.filters': 'Фільтри',
-    'discover.filtersActive': 'Фільтри ({0})',
-    'discover.moreActions': 'Більше дій',
-    'discover.decadesLabel': 'Десятиліття',
-    'analytics.title': 'Аналітика',
-    'dashboard.title': 'Панель',
-    'discover.title': 'Відкриття',
-    'setup.title': 'Налаштування',
-    'todaysPick.discoverSimilar': 'Знайти схожих',
-    'libraryHealth.fix': 'Виправити',
-    'libraryHealth.fixing': 'Виправлення...',
-    'libraryHealth.showLess': 'Показати менше',
-    'libraryHealth.showAll': 'Показати всі {0}',
-    'libraryHealth.fixHint.missingMetadata': 'Оновлює метадані артистів із Lidarr',
-    'libraryHealth.fixHint.unmonitored':
-      'Позначає всіх перелічених артистів як відстежуваних у Lidarr',
-    'libraryHealth.fixHint.missingAlbums': 'Оновлює альбоми в Lidarr, щоб знайти відсутні',
-    'libraryHealth.fixHint.genreGaps': 'Оновлює метадані Lidarr, щоб отримати жанри',
-    'libraryHealth.fixHint.imageGaps': 'Повторно шукає зображення артистів',
-    'libraryHealth.fixHint.missingWikidata': 'Отримує біографії та посилання з Wikidata',
-  },
-  ru: {
-    'common.confirm': 'Подтвердить',
-    'discover.filters': 'Фильтры',
-    'discover.filtersActive': 'Фильтры ({0})',
-    'discover.moreActions': 'Больше действий',
-    'discover.decadesLabel': 'Десятилетия',
-    'analytics.title': 'Аналитика',
-    'dashboard.title': 'Панель',
-    'discover.title': 'Поиск',
-    'setup.title': 'Настройка',
-    'todaysPick.discoverSimilar': 'Найти похожих',
-    'libraryHealth.fix': 'Исправить',
-    'libraryHealth.fixing': 'Исправление...',
-    'libraryHealth.showLess': 'Показать меньше',
-    'libraryHealth.showAll': 'Показать все {0}',
-    'libraryHealth.fixHint.missingMetadata': 'Обновляет метаданные артистов из Lidarr',
-    'libraryHealth.fixHint.unmonitored':
-      'Помечает всех перечисленных артистов как отслеживаемых в Lidarr',
-    'libraryHealth.fixHint.missingAlbums': 'Обновляет альбомы в Lidarr, чтобы найти отсутствующие',
-    'libraryHealth.fixHint.genreGaps': 'Обновляет метаданные Lidarr для получения жанров',
-    'libraryHealth.fixHint.imageGaps': 'Повторно ищет изображения артистов',
-    'libraryHealth.fixHint.missingWikidata': 'Получает биографии и ссылки из Wikidata',
-  },
-  ja: {
-    'common.confirm': '確認',
-    'discover.filters': 'フィルター',
-    'discover.filtersActive': 'フィルター ({0})',
-    'discover.moreActions': 'その他の操作',
-    'discover.decadesLabel': '年代',
-    'analytics.title': '分析',
-    'dashboard.title': 'ダッシュボード',
-    'discover.title': '発見',
-    'setup.title': 'セットアップ',
-    'todaysPick.discoverSimilar': '似たアーティストを探す',
-    'libraryHealth.fix': '修正',
-    'libraryHealth.fixing': '修正中...',
-    'libraryHealth.showLess': '少なく表示',
-    'libraryHealth.showAll': '{0} 件すべて表示',
-    'libraryHealth.fixHint.missingMetadata': 'Lidarr からアーティストのメタデータを更新します',
-    'libraryHealth.fixHint.unmonitored': '一覧のすべてのアーティストを Lidarr で監視対象にします',
-    'libraryHealth.fixHint.missingAlbums': '不足を見つけるため Lidarr のアルバムを更新します',
-    'libraryHealth.fixHint.genreGaps': 'ジャンルを取得するため Lidarr メタデータを更新します',
-    'libraryHealth.fixHint.imageGaps': 'アーティスト画像を再検索します',
-    'libraryHealth.fixHint.missingWikidata': 'Wikidata から略歴とリンクを取得します',
-  },
-  ko: {
-    'common.confirm': '확인',
-    'discover.filters': '필터',
-    'discover.filtersActive': '필터 ({0})',
-    'discover.moreActions': '추가 작업',
-    'discover.decadesLabel': '연대',
-    'analytics.title': '분석',
-    'dashboard.title': '대시보드',
-    'discover.title': '발견',
-    'setup.title': '설정',
-    'todaysPick.discoverSimilar': '비슷한 아티스트 찾기',
-    'libraryHealth.fix': '수정',
-    'libraryHealth.fixing': '수정 중...',
-    'libraryHealth.showLess': '간단히 보기',
-    'libraryHealth.showAll': '{0}개 모두 보기',
-    'libraryHealth.fixHint.missingMetadata': 'Lidarr에서 아티스트 메타데이터를 새로 고칩니다',
-    'libraryHealth.fixHint.unmonitored':
-      '나열된 모든 아티스트를 Lidarr에서 모니터링 대상으로 설정합니다',
-    'libraryHealth.fixHint.missingAlbums': '누락 항목을 찾도록 Lidarr 앨범을 새로 고칩니다',
-    'libraryHealth.fixHint.genreGaps': '장르 태그를 가져오도록 Lidarr 메타데이터를 새로 고칩니다',
-    'libraryHealth.fixHint.imageGaps': '아티스트 이미지를 다시 검색합니다',
-    'libraryHealth.fixHint.missingWikidata': 'Wikidata에서 약력과 링크를 가져옵니다',
-  },
-  'zh-CN': {
-    'common.confirm': '确认',
-    'discover.filters': '筛选',
-    'discover.filtersActive': '筛选 ({0})',
-    'discover.moreActions': '更多操作',
-    'discover.decadesLabel': '年代',
-    'analytics.title': '分析',
-    'dashboard.title': '仪表盘',
-    'discover.title': '发现',
-    'setup.title': '设置',
-    'todaysPick.discoverSimilar': '发现相似艺人',
-    'libraryHealth.fix': '修复',
-    'libraryHealth.fixing': '修复中...',
-    'libraryHealth.showLess': '收起',
-    'libraryHealth.showAll': '显示全部 {0}',
-    'libraryHealth.fixHint.missingMetadata': '从 Lidarr 刷新艺人元数据',
-    'libraryHealth.fixHint.unmonitored': '将所有列出的艺人在 Lidarr 中设为监控',
-    'libraryHealth.fixHint.missingAlbums': '刷新 Lidarr 专辑以查找缺失项',
-    'libraryHealth.fixHint.genreGaps': '刷新 Lidarr 元数据以获取流派标签',
-    'libraryHealth.fixHint.imageGaps': '重新查找艺人图片',
-    'libraryHealth.fixHint.missingWikidata': '从 Wikidata 获取简介和链接',
-  },
-}
-
-export const MESSAGE_OVERRIDES: Partial<Record<SupportedLocale, Partial<MessageCatalog>>> = {
-  es: {
-    ...REVIEW_GATE_TRANSLATIONS.es,
-    'common.unauthorized': 'No autorizado',
-    'common.adminAccessRequired': 'Se requiere acceso de administrador',
-    'common.current': 'Actual',
-    'common.custom': 'Personalizado',
-    'common.enabled': 'Activado',
-    'common.importing': 'Importando...',
-    'subscriptionForm.dialogCreate': 'Crear suscripción',
-    'subscriptionForm.dialogEdit': 'Editar suscripción',
-    'subscriptionForm.titleCreate': 'Crear suscripción',
-    'subscriptionForm.titleEdit': 'Editar suscripción',
-    'subscriptionForm.name': 'Nombre',
-    'subscriptionForm.namePlaceholder': 'Nuevo Shoegaze Weekly',
-    'subscriptionForm.nameRequired': 'El nombre es obligatorio',
-    'subscriptionForm.source': 'Fuente',
-    'subscriptionForm.sources': 'Fuentes',
-    'subscriptionForm.sourceType': 'Tipo de fuente',
-    'subscriptionForm.sourceTypeGenre': 'Género',
-    'subscriptionForm.sourceTypeGenreDesc': 'Descubre artistas dentro de un género concreto',
-    'subscriptionForm.sourceTypeSimilar': 'Similares',
-    'subscriptionForm.sourceTypeSimilarDesc': 'Encuentra artistas parecidos a tus favoritos',
-    'subscriptionForm.sourceTypeDiscoveryModeDesc':
-      'Usa un modo de descubrimiento para generar artistas',
-    'subscriptionForm.discoveryMode': 'Modo de descubrimiento',
-    'subscriptionForm.selectDiscoveryMode':
-      'Selecciona un modo de descubrimiento para configurarlo.',
-    'subscriptionForm.discoveryModeSettingsRequired':
-      'La configuración del modo de descubrimiento es obligatoria',
-    'subscriptionForm.genre': 'Género',
-    'subscriptionForm.genrePlaceholder': 'shoegaze',
-    'subscriptionForm.genreRequired': 'El género es obligatorio',
-    'subscriptionForm.seedArtists': 'Artistas semilla',
-    'subscriptionForm.seedArtistsPlaceholder': 'Radiohead, Portishead, Massive Attack',
-    'subscriptionForm.seedArtistsHelp': 'Nombres de artistas separados por comas',
-    'subscriptionForm.seedArtistsRequired': 'Se requiere al menos un artista semilla',
-    'subscriptionForm.configureInSettings': 'Configuralo en Ajustes > Conexiones',
-    'subscriptionForm.notConfigured': 'sin configurar',
-    'subscriptionForm.selectAtLeastOneSource': 'Selecciona al menos una fuente',
-    'subscriptionForm.sourceLockedHelp':
-      'El tipo de fuente y la configuración no se pueden cambiar para esta suscripción.',
-    'subscriptionForm.maxArtistsPerRun': 'Máximo de artistas por ejecución',
-    'subscriptionForm.scoringPreset': 'Preajuste de puntuación',
-    'subscriptionForm.scoringPresetDefault': 'Predeterminado',
-    'subscriptionForm.scoringPresetGenre': 'Optimizado para género',
-    'cronPicker.preset': 'Preajuste',
-    'cronPicker.dailyMidnight': 'Diario (medianoche)',
-    'cronPicker.twiceAWeek': 'Dos veces por semana (lun + jue)',
-    'cronPicker.everyTwoWeeks': 'Cada 2 semanas (1 y 15)',
-    'cronPicker.monthlyFirst': 'Mensual (día 1)',
-    'cronPicker.standardHint': 'Cron estándar: minuto hora día-del-mes mes día-de-la-semana',
-    'libraryHealth.syncNow': 'Sincronizar ahora',
-    'libraryHealth.snapshotTitle': 'Última instantánea',
-    'libraryHealth.lastSynced': 'Última sincronización',
-    'libraryHealth.neverSynced': 'Nunca se ha sincronizado.',
-    'libraryHealth.autoSyncEvery': 'Sincronización automática cada {0} h',
-    'jobHistory.librarySync': 'Sincronización de biblioteca',
-    'settings.tabs.systemHealth': 'Salud del sistema',
-    'settings.targetUpdated': 'Destino actualizado',
-    'settings.targetUpdateFailed': 'No se pudo actualizar el destino',
-    'settings.editTarget': 'Editar destino',
-    'settings.saveChanges': 'Guardar cambios',
-    'settings.linkedTarget': 'Destino vinculado:',
-    'errorBoundary.title': 'Algo salió mal',
-    'errorBoundary.retry': 'Reintentar',
-    'errorBoundary.home': 'Ir al inicio',
-    'connectionSuggestions.title': 'Sugerencias para {0}',
-    'connectionSuggestions.close': 'Cerrar sugerencias',
-    'connectionSuggestions.gotIt': 'Entendido',
-    'connectionSuggestions.skipForNow': 'Omitir por ahora',
-    'connectionSuggestions.spotifyDiscoverWeekly': 'Configurar una suscripción a Discover Weekly',
-    'connectionSuggestions.spotifyReleaseRadar': 'Activar una suscripción a Release Radar',
-    'connectionSuggestions.lastfmTopGenreTags': 'Suscribirse a tus etiquetas de género principales',
-    'connectionSuggestions.listenbrainzWeeklyJams': 'Activar la suscripción Weekly Jams',
-    'connectionSuggestions.listenbrainzWeeklyExploration': 'Activar Weekly Exploration',
-    'connectionSuggestions.weeklyDigestPlaylist': 'Crear una lista semanal Digarr Digest',
-    'keyboardShortcuts.title': 'Atajos de teclado',
-    'keyboardShortcuts.close': 'Cerrar atajos',
-    'keyboardShortcuts.nextCard': 'Siguiente tarjeta',
-    'keyboardShortcuts.previousCard': 'Tarjeta anterior',
-    'keyboardShortcuts.approveFocusedCard': 'Aprobar la tarjeta activa',
-    'keyboardShortcuts.rejectFocusedCard': 'Rechazar la tarjeta activa',
-    'keyboardShortcuts.toggleDetailView': 'Abrir / cerrar la vista de detalle',
-    'keyboardShortcuts.toggleShortcuts': 'Mostrar / ocultar atajos',
-    'keyboardShortcuts.previousCardStack': 'Tarjeta anterior (vista apilada)',
-    'keyboardShortcuts.nextCardStack': 'Siguiente tarjeta (vista apilada)',
-    'keyboardShortcuts.disabledWhileTyping':
-      'Los atajos se desactivan cuando hay un campo de texto enfocado.',
-    'admin.restoreSummary': 'Se restauraron {0} filas en {1} tablas',
-    'admin.restoreSummaryForce':
-      'Se restauraron {0} filas. Vuelve a introducir las credenciales de los campos cifrados.',
-    'admin.encryptionMismatchMessage': 'La copia se creó con una clave de cifrado distinta.',
-    'admin.encryptionMismatchAffected': 'Estos campos tendrán que volver a introducirse: {0}',
-    'admin.encryptionMismatchFallback':
-      'Algunos campos cifrados podrían requerir que vuelvas a introducirlos.',
-    'subscriptions.spotifyNotConnected': 'Spotify no está conectado',
-    'playlistForm.dialogCreate': 'Crear lista',
-    'playlistForm.dialogEdit': 'Editar lista',
-    'playlistForm.titleCreate': 'Crear lista de reproducción',
-    'playlistForm.titleEdit': 'Editar lista de reproducción',
-    'playlistForm.namePlaceholder': 'Mi mezcla semanal',
-    'playlistForm.strategy': 'Estrategia',
-    'playlistForm.strategyWeeklyDigestDesc': 'Selección curada de tus descubrimientos recientes',
-    'playlistForm.strategyGenreFocusDesc': 'Profundiza en un género concreto',
-    'playlistForm.strategyMoodMixDesc': 'Canciones que encajan con un estado de ánimo concreto',
-    'playlistForm.strategyRediscoverDesc': 'Recupera artistas que quizás habias olvidado',
-    'playlistForm.mood': 'Estado de ánimo',
-    'playlistForm.moodPlaceholder': 'conduciendo de noche',
-    'playlistForm.trackCount': 'Número de canciones',
-    'playlistForm.scheduleAutomaticGeneration': 'Programar generación automática',
-    'playlistForm.nameRequired': 'El nombre es obligatorio',
-    'playlistForm.genreRequired': 'El género es obligatorio para la estrategia Enfoque de género',
-    'playlistForm.moodRequired':
-      'El estado de ánimo es obligatorio para la estrategia Mezcla de estados de ánimo',
-    'searchResultCard.addToQueue': 'Añadir a la cola de descubrimiento',
-    'searchResultCard.queuing': 'Añadiendo...',
-    'searchResultCard.queuedSuccess': '{0} se ha añadido a la cola de descubrimiento',
-    'searchResultCard.queuedFailure': 'No se pudo añadir {0} a la cola',
-    'searchResultCard.popularity': '{0}{1} de popularidad',
-    'subscriptionPresets.getStarted': 'Empieza con un preajuste',
-    'subscriptionPresets.intro':
-      'Elige un preajuste para crear suscripciones automáticamente o monta una configuración personalizada.',
-    'subscriptionPresets.casualTitle': 'Tranquilo',
-    'subscriptionPresets.casualDescription': 'Recomendaciones semanales con poco ruido',
-    'subscriptionPresets.deepDiverTitle': 'Explorador profundo',
-    'subscriptionPresets.deepDiverDescription':
-      'Descubrimiento amplio con varias fuentes y actualizaciones diarias',
-    'subscriptionPresets.discoveryOnlyTitle': 'Solo descubrimiento',
-    'subscriptionPresets.discoveryOnlyDescription': 'Solo pipeline, sin fuentes externas',
-    'subscriptionPresets.customTitle': 'Personalizado',
-    'subscriptionPresets.customDescription': 'Construye tu propia combinación de suscripciones',
-    'subscriptionPresets.oneConnectedService': '1 servicio conectado',
-    'subscriptionPresets.twoConnectedServices': '2 servicios conectados',
-    'subscriptionPresets.requires': 'Requiere {0}',
-    'subscriptionPresets.setUp': 'Configurar',
-    'subscriptionPresets.applying': 'Aplicando...',
-    'subscriptionPresets.casualApplied': 'Se ha aplicado el preajuste Casual',
-    'subscriptionPresets.deepDiverApplied': 'Se ha aplicado el preajuste Explorador profundo',
-    'subscriptionPresets.applyFailed': 'No se pudo aplicar el preajuste',
-    'subscriptionPresets.aiOnlyEnabled': 'Se ha activado el modo solo IA',
-    'subscriptionPresets.aiOnlyEnableFailed': 'No se pudo activar el modo solo IA',
-    'common.manual': 'manual',
-    'jobHistory.pipeline': 'Proceso de descubrimiento',
-    'jobHistory.playlistType': 'Lista de reproducción',
-    'jobHistory.errorLabel': 'Fallo:',
-    'libraryReconciliation.albumsCouldNotBeMatched':
-      'álbumes que no se pudieron asociar automáticamente con MusicBrainz.',
-    'libraryReconciliation.artistsCouldNotBeMatched':
-      'artistas que no se pudieron asociar automáticamente con MusicBrainz.',
-    'playlist.total': 'en total',
-    'common.total': 'Total acumulado',
-    'settings.capDiscovery': 'Descubrimiento',
-    'settings.capImport': 'Importación',
-    'settings.capLibrarySync': 'Sincronización de biblioteca',
-    'settings.capPlaylistExport': 'Exportación de listas',
-    'settings.capService': 'Servicio musical',
-    'settings.capSubscriptions': 'Suscripciones',
-    'systemHealth.pipeline': 'Proceso de descubrimiento',
-    'systemHealth.librarySync': 'Sincronización de biblioteca',
-    'search.experimental': 'En pruebas',
-    'discoveryMode.field.feed': 'Origen del feed',
-    'settings.totalWeight': 'peso total:',
-    'settings.librarySyncInterval': 'Sincronización de biblioteca',
-    'settings.librarySyncIntervalHours': 'Intervalo de sincronización (horas)',
-    'settings.librarySyncIntervalHelp':
-      'Controla con qué frecuencia Digarr actualiza los datos de sincronización de la biblioteca y vuelve a ejecutar las comprobaciones automáticamente.',
-    'settings.aiProviderTitle': 'Proveedor de IA',
-    'settings.webhookTitle': 'Integración webhook',
-    'settings.plexToken': 'Token de Plex',
-    'settings.serviceConnectedToast': '{0} conectado',
-    'settings.serviceUnreachableToast': 'No se pudo acceder a {0}',
-    'settings.serviceSettingsSavedToast': 'Se han guardado los ajustes de {0}',
-    'settings.serviceSettingsFailedToast': 'No se pudieron guardar los ajustes de {0}',
-    'settings.spotifyAuthorizationFailed': 'No se pudo iniciar la autorización de Spotify',
-    'settings.spotifyDisconnectFailed': 'No se pudo desconectar Spotify',
-    'settings.spotifyLikedSongsFailed': 'No se pudo iniciar la importación de Spotify Liked Songs',
-    'settings.spotifyPlaylistFailed': 'No se pudo iniciar la importación de la lista',
-    'settings.deezerFavoritesFailed': 'No se pudo iniciar la importación de favoritos de Deezer',
-    'settings.deezerFollowedFailed':
-      'No se pudo iniciar la importación de artistas seguidos en Deezer',
-    'settings.deezerDisconnectFailed': 'No se pudo desconectar Deezer',
-    'settings.deezerAuthorizationFailed': 'No se pudo iniciar la autorización de Deezer',
-    'settings.reverseProxyConfigHint':
-      'Usa PROXY_AUTH_ENABLED y PROXY_AUTH_TRUSTED_PROXIES; tu proxy debe enviar X-Forwarded-User.',
-    'service.error': 'Fallo',
-    'admin.autoFixInProgress': 'Corrección automática en curso... ({0}/{1} corregidos)',
-    'integration.playlist': 'Lista de reproducción',
-    'auth.registerDescription': 'Crea la primera cuenta para empezar.',
-    'albumPicker.title': 'Seleccionar álbumes',
-    'albumPicker.loadFailed': 'No se pudieron cargar los álbumes',
-    'albumPicker.selectAll': 'Todos',
-    'albumPicker.selectNone': 'Ninguno',
-    'albumPicker.selectedCount': '{0} seleccionados',
-    'albumPicker.aiPick': 'Elección de IA',
-    'albumPicker.monitor': 'Monitorizar',
-    'albumPicker.monitorSingular': 'Monitorizar {0} álbum',
-    'albumPicker.monitorPlural': 'Monitorizar {0} álbumes',
-    'approveDialog.title': 'Ajustes de Lidarr para este artista',
-    'approveDialog.loadFailed':
-      'No se pudieron cargar los perfiles de Lidarr. Revisa tu conexión con Lidarr.',
-    'approveDialog.qualityProfile': 'Perfil de calidad',
-    'approveDialog.metadataProfile': 'Perfil de metadatos',
-    'approveDialog.rootFolder': 'Carpeta raíz',
-    'cardStack.counter': '{0} de {1}',
-    'cardStack.swipeHint': 'Desliza a la derecha para aprobar, a la izquierda para rechazar',
-    'serviceCard.testConnection': 'Probar conexión',
-    'preview.noPreviewAvailable': 'No hay vista previa disponible para este artista',
-    'preview.playbackBlocked': 'Reproducción bloqueada por el navegador; inténtalo de nuevo',
-    'errors.auth.invalidCredentials': 'Credenciales no válidas',
-    'errors.auth.notAuthenticated': 'No autenticado',
-    'errors.auth.passwordIncorrect': 'La contraseña actual es incorrecta',
-    'errors.auth.usernameTaken': 'El nombre de usuario ya está en uso',
-    'errors.target.notFound': 'Destino no encontrado',
-    'errors.pipeline.alreadyRunning': 'Ya hay una ejecución del pipeline en curso',
-    'errors.recommendation.notFound': 'Recomendación no encontrada',
-    'errors.subscription.notFound': 'Suscripción no encontrada',
-    'errors.playlist.notFound': 'Lista de reproducción no encontrada',
-  },
-  fr: {
-    ...REVIEW_GATE_TRANSLATIONS.fr,
-    'app.themeMode': 'Thème',
-    'nav.genres': 'Genres musicaux',
-    'setup.mode': 'Mode choisi',
-    'setup.modelSuggestionGeminiFlash': 'Gemini 2.5 Flash (version stable)',
-    'common.unauthorized': 'Non autorise',
-    'common.adminAccessRequired': 'Acces administrateur requis',
-    'common.importing': 'Importation...',
-    'common.current': 'Actuel',
-    'common.custom': 'Personnalise',
-    'common.enabled': 'Active',
-    'common.date': 'Jour',
-    'common.total': 'Total cumule',
-    'genres.genrePlural': 'styles',
-    'genres.genreSingular': 'style',
-    'jobHistory.pipeline': 'Flux de découverte',
-    'jobHistory.librarySync': 'Synchronisation de la bibliothèque',
-    'jobHistory.playlistType': 'Liste de lecture',
-    'libraryReconciliation.page': 'Page de résultats',
-    'playlistForm.dialogCreate': 'Créer une playlist',
-    'playlistForm.dialogEdit': 'Modifier la playlist',
-    'playlistForm.titleCreate': 'Créer une playlist',
-    'playlistForm.titleEdit': 'Modifier la playlist',
-    'playlistForm.namePlaceholder': 'Mon mix hebdomadaire',
-    'playlistForm.strategy': 'Strategie',
-    'playlistForm.strategyWeeklyDigestDesc':
-      'Selection soignee à partir de vos découvertes recentes',
-    'playlistForm.strategyGenreFocusDesc': 'Explorer en profondeur un genre precis',
-    'playlistForm.strategyMoodMixDesc': 'Titres adaptes a une ambiance precise',
-    'playlistForm.strategyRediscoverDesc':
-      'Retrouver des artistes que vous aviez peut-etre oublies',
-    'playlistForm.mood': 'Ambiance',
-    'playlistForm.moodPlaceholder': 'trajet de nuit',
-    'playlistForm.trackCount': 'Nombre de titres',
-    'playlistForm.scheduleAutomaticGeneration': 'Planifier la generation automatique',
-    'playlistForm.nameRequired': 'Le nom est obligatoire',
-    'playlistForm.genreRequired': 'Le genre est obligatoire pour la strategie Focus genre',
-    'playlistForm.moodRequired': "L'ambiance est obligatoire pour la strategie Mix d'ambiance",
-    'subscriptionForm.dialogCreate': 'Créer un abonnement',
-    'subscriptionForm.dialogEdit': "Modifier l'abonnement",
-    'subscriptionForm.titleCreate': 'Créer un abonnement',
-    'subscriptionForm.titleEdit': "Modifier l'abonnement",
-    'subscriptionForm.name': 'Nom',
-    'subscriptionForm.namePlaceholder': 'Nouveau Shoegaze Weekly',
-    'subscriptionForm.nameRequired': 'Le nom est obligatoire',
-    'subscriptionForm.source': 'Origine',
-    'subscriptionForm.sources': 'Origines',
-    'subscriptionForm.sourceType': 'Type de source',
-    'subscriptionForm.sourceTypeGenre': 'Style musical',
-    'subscriptionForm.sourceTypeGenreDesc': 'Découvrir des artistes dans un genre precis',
-    'subscriptionForm.sourceTypeSimilar': 'Similaires',
-    'subscriptionForm.sourceTypeSimilarDesc': 'Trouver des artistes proches de vos favoris',
-    'subscriptionForm.sourceTypeDiscoveryModeDesc':
-      'Utiliser un mode de découverte pour générer des artistes',
-    'subscriptionForm.discoveryMode': 'Mode de découverte',
-    'subscriptionForm.selectDiscoveryMode': 'Selectionnez un mode de découverte a configurer.',
-    'subscriptionForm.discoveryModeSettingsRequired':
-      'Les réglages du mode de découverte sont obligatoires',
-    'subscriptionForm.genre': 'Style musical',
-    'subscriptionForm.genrePlaceholder': 'shoegaze',
-    'subscriptionForm.genreRequired': 'Le genre est obligatoire',
-    'subscriptionForm.seedArtists': 'Artistes de depart',
-    'subscriptionForm.seedArtistsPlaceholder': 'Radiohead, Portishead, Massive Attack',
-    'subscriptionForm.seedArtistsHelp': "Noms d'artistes separes par des virgules",
-    'subscriptionForm.seedArtistsRequired': 'Au moins un artiste de depart est requis',
-    'subscriptionForm.configureInSettings': 'A configurer dans Réglages > Connexions',
-    'subscriptionForm.notConfigured': 'non configuré',
-    'subscriptionForm.selectAtLeastOneSource': 'Selectionnez au moins une source',
-    'subscriptionForm.sourceLockedHelp':
-      'Le type de source et sa configuration ne peuvent pas etre modifies pour cet abonnement.',
-    'subscriptionForm.maxArtistsPerRun': 'Nombre maximal d artistes par execution',
-    'subscriptionForm.scoringPreset': 'Preset de score',
-    'subscriptionForm.scoringPresetDefault': 'Par defaut',
-    'subscriptionForm.scoringPresetGenre': 'Optimise pour le genre',
-    'subscriptions.spotifyNotConnected': 'Spotify n est pas connecte',
-    'subscriptionPresets.getStarted': 'Commencer avec un preset',
-    'subscriptionPresets.intro':
-      'Choisissez un preset pour créer des abonnements automatiquement ou composez votre propre configuration.',
-    'subscriptionPresets.casualTitle': 'Detendu',
-    'subscriptionPresets.casualDescription': 'Recommandations hebdomadaires avec peu de bruit',
-    'subscriptionPresets.deepDiverTitle': 'Exploration poussee',
-    'subscriptionPresets.deepDiverDescription':
-      'Découverte large avec plusieurs sources et des mises a jour quotidiennes',
-    'subscriptionPresets.discoveryOnlyTitle': 'Découverte uniquement',
-    'subscriptionPresets.discoveryOnlyDescription': 'Pipeline seul, sans flux externes',
-    'subscriptionPresets.customTitle': 'Personnalise',
-    'subscriptionPresets.customDescription': 'Composez votre propre combinaison d abonnements',
-    'subscriptionPresets.oneConnectedService': '1 service connecte',
-    'subscriptionPresets.twoConnectedServices': '2 services connectes',
-    'subscriptionPresets.requires': 'Necessite {0}',
-    'subscriptionPresets.setUp': 'Configurer',
-    'subscriptionPresets.applying': 'Application...',
-    'subscriptionPresets.casualApplied': 'Le preset Detendu a été applique',
-    'subscriptionPresets.deepDiverApplied': 'Le preset Exploration poussee a été applique',
-    'subscriptionPresets.applyFailed': "Impossible d'appliquer le preset",
-    'subscriptionPresets.aiOnlyEnabled': 'Le mode IA uniquement a été active',
-    'subscriptionPresets.aiOnlyEnableFailed': "Impossible d'activer le mode IA uniquement",
-    'cronPicker.preset': 'Reglage type',
-    'cronPicker.dailyMidnight': 'Tous les jours (minuit)',
-    'cronPicker.twiceAWeek': 'Deux fois par semaine (lun + jeu)',
-    'cronPicker.everyTwoWeeks': 'Toutes les 2 semaines (1 et 15)',
-    'cronPicker.monthlyFirst': 'Chaque mois (jour 1)',
-    'cronPicker.standardHint': 'Cron standard : minute heure jour-du-mois mois jour-de-la-semaine',
-    'recommendation.sourceGenre': 'Style musical',
-    'libraryHealth.syncNow': 'Synchroniser maintenant',
-    'libraryHealth.snapshotTitle': 'Dernier instantane',
-    'libraryHealth.lastSynced': 'Dernière synchronisation',
-    'libraryHealth.neverSynced': 'Jamais synchronise.',
-    'libraryHealth.autoSyncEvery': 'Synchronisation automatique toutes les {0} h',
-    'systemHealth.pipeline': 'Flux de découverte',
-    'systemHealth.librarySync': 'Sync bibliothèque',
-    'systemHealth.sources': 'Connexions',
-    'discoveryMode.labels.label': 'Libelles',
-    'discoveryMode.field.tags': 'Etiquettes',
-    'libraryStats.albums': 'Albums musicaux',
-    'errorBoundary.title': 'Un problème est survenu',
-    'errorBoundary.retry': 'Réessayer',
-    'errorBoundary.home': "Retour à l'accueil",
-    'connectionSuggestions.title': 'Suggestions pour {0}',
-    'connectionSuggestions.close': 'Fermer les suggestions',
-    'connectionSuggestions.gotIt': 'Compris',
-    'connectionSuggestions.skipForNow': 'Ignorer pour le moment',
-    'connectionSuggestions.spotifyDiscoverWeekly': 'Configurer un abonnement Discover Weekly',
-    'connectionSuggestions.spotifyReleaseRadar': "Activer l'abonnement Release Radar",
-    'connectionSuggestions.lastfmTopGenreTags': 'Vous abonner a vos principaux tags de genre',
-    'connectionSuggestions.listenbrainzWeeklyJams': "Activer l'abonnement Weekly Jams",
-    'connectionSuggestions.listenbrainzWeeklyExploration': 'Activer Weekly Exploration',
-    'connectionSuggestions.weeklyDigestPlaylist': 'Créer une playlist Digarr Digest hebdomadaire',
-    'keyboardShortcuts.title': 'Raccourcis clavier',
-    'keyboardShortcuts.close': 'Fermer les raccourcis',
-    'keyboardShortcuts.nextCard': 'Carte suivante',
-    'keyboardShortcuts.previousCard': 'Carte precedente',
-    'keyboardShortcuts.approveFocusedCard': 'Approuver la carte active',
-    'keyboardShortcuts.rejectFocusedCard': 'Rejeter la carte active',
-    'keyboardShortcuts.toggleDetailView': 'Ouvrir / fermer la vue detaillee',
-    'keyboardShortcuts.toggleShortcuts': 'Afficher / masquer les raccourcis',
-    'keyboardShortcuts.previousCardStack': 'Carte precedente (vue pile)',
-    'keyboardShortcuts.nextCardStack': 'Carte suivante (vue pile)',
-    'keyboardShortcuts.disabledWhileTyping':
-      'Les raccourcis sont desactives lorsqu un champ texte est actif.',
-    'admin.restoreSummary': 'Restauration de {0} lignes dans {1} tables',
-    'admin.restoreSummaryForce':
-      'Restauration de {0} lignes. Ressaisissez les identifiants des champs chiffres.',
-    'admin.encryptionMismatchMessage':
-      'La sauvegarde a été créée avec une clé de chiffrement differente.',
-    'admin.encryptionMismatchAffected': 'Ces champs devront etre ressaisis : {0}',
-    'admin.encryptionMismatchFallback':
-      'Certains champs chiffres devront peut-etre etre ressaisis.',
-    'admin.autoFixInProgress': 'Correction automatique en cours... ({0}/{1} corriges)',
-    'settings.capDiscovery': 'Découverte',
-    'settings.capImport': 'Importation',
-    'settings.capLibrarySync': 'Sync bibliothèque',
-    'settings.capPlaylistExport': 'Export de playlist',
-    'settings.capService': 'Connecteur',
-    'settings.capSubscriptions': 'Abonnements',
-    'settings.tabs.administration': 'Gestion',
-    'settings.tabs.systemHealth': 'Etat du systeme',
-    'settings.targetUpdated': 'Destination mise a jour',
-    'settings.targetUpdateFailed': 'Échec de la mise a jour de la destination',
-    'settings.editTarget': 'Modifier la destination',
-    'settings.saveChanges': 'Enregistrer les changements',
-    'settings.linkedTarget': 'Destination liee :',
-    'settings.type': 'Categorie',
-    'settings.consensus': 'Consensus global',
-    'settings.librarySyncInterval': 'Sync bibliothèque',
-    'settings.librarySyncIntervalHours': 'Intervalle de sync (heures)',
-    'settings.librarySyncIntervalHelp':
-      'Controle la frequence a laquelle Digarr actualise les donnees de sync bibliothèque et relance les verifications.',
-    'settings.oidcSso': 'OIDC / SSO',
-    'settings.aiProviderTitle': 'Fournisseur IA',
-    'settings.webhookTitle': 'Canal webhook',
-    'settings.plexToken': 'Jeton Plex',
-    'settings.serviceConnectedToast': '{0} connecte',
-    'settings.serviceUnreachableToast': '{0} est inaccessible',
-    'settings.serviceSettingsSavedToast': 'Réglages de {0} enregistres',
-    'settings.serviceSettingsFailedToast': 'Impossible d enregistrer les réglages de {0}',
-    'settings.spotifyAuthorizationFailed': "Impossible de lancer l'autorisation Spotify",
-    'settings.spotifyDisconnectFailed': 'Impossible de deconnecter Spotify',
-    'settings.spotifyLikedSongsFailed': "Impossible de lancer l'import des Titres J aimes Spotify",
-    'settings.spotifyPlaylistFailed': "Impossible de lancer l'import de playlist",
-    'settings.deezerFavoritesFailed': "Impossible de lancer l'import des favoris Deezer",
-    'settings.deezerFollowedFailed': "Impossible de lancer l'import des artistes suivis Deezer",
-    'settings.deezerDisconnectFailed': 'Impossible de deconnecter Deezer',
-    'settings.deezerAuthorizationFailed': "Impossible de lancer l'autorisation Deezer",
-    'settings.reverseProxyConfigHint':
-      'Utilisez PROXY_AUTH_ENABLED et PROXY_AUTH_TRUSTED_PROXIES ; votre proxy doit envoyer X-Forwarded-User.',
-    'searchResultCard.addToQueue': 'Ajouter a la file de découverte',
-    'searchResultCard.queuing': 'Ajout...',
-    'searchResultCard.queuedSuccess': '{0} a été ajouté a la file de découverte',
-    'searchResultCard.queuedFailure': "Impossible d'ajouter {0} a la file",
-    'searchResultCard.popularity': 'Popularite : {0}{1}',
-    'integration.playlist': 'Liste de lecture',
-    'analytics.source.consensus': 'Consensus global',
-    'albumPicker.title': 'Sélectionner les albums',
-    'albumPicker.loadFailed': 'Échec du chargement des albums',
-    'albumPicker.selectAll': 'Tous',
-    'albumPicker.selectNone': 'Aucun',
-    'albumPicker.selectedCount': '{0} sélectionnés',
-    'albumPicker.aiPick': 'Choix de IA',
-    'albumPicker.monitor': 'Surveiller',
-    'albumPicker.monitorSingular': 'Surveiller {0} album',
-    'albumPicker.monitorPlural': 'Surveiller {0} albums',
-    'approveDialog.title': 'Paramètres Lidarr pour cet artiste',
-    'approveDialog.loadFailed':
-      'Échec du chargement des profils Lidarr. Vérifiez votre connexion Lidarr.',
-    'approveDialog.qualityProfile': 'Profil de qualité',
-    'approveDialog.metadataProfile': 'Profil de métadonnées',
-    'approveDialog.rootFolder': 'Dossier racine',
-    'cardStack.counter': '{0} sur {1}',
-    'cardStack.swipeHint': 'Glissez à droite pour approuver, à gauche pour rejeter',
-    'serviceCard.testConnection': 'Tester la connexion',
-    'preview.noPreviewAvailable': 'Aucun aperçu disponible pour cet artiste',
-    'preview.playbackBlocked': 'Lecture bloquée par le navigateur ; essayez de cliquer à nouveau',
-    'errors.auth.invalidCredentials': 'Identifiants non valides',
-    'errors.auth.notAuthenticated': 'Non authentifié',
-    'errors.auth.passwordIncorrect': 'Le mot de passe actuel est incorrect',
-    'errors.auth.usernameTaken': "Nom d'utilisateur déjà pris",
-    'errors.target.notFound': 'Destination introuvable',
-    'errors.pipeline.alreadyRunning': 'Une exécution du pipeline est déjà en cours',
-    'errors.recommendation.notFound': 'Recommandation introuvable',
-    'errors.subscription.notFound': 'Abonnement introuvable',
-    'errors.playlist.notFound': 'Liste de lecture introuvable',
-  },
-  de: {
-    ...REVIEW_GATE_TRANSLATIONS.de,
-    'app.themeModeSystem': 'Vom System',
-    'app.themeGroupStreaming': 'Streaming-Dienste',
-    'nav.genres': 'Musikstile',
-    'common.unauthorized': 'Nicht autorisiert',
-    'common.adminAccessRequired': 'Adminzugriff erforderlich',
-    'common.importing': 'Import wird ausgeführt...',
-    'common.optional': 'nach Bedarf',
-    'common.current': 'Aktuell',
-    'common.custom': 'Benutzerdefiniert',
-    'common.enabled': 'Aktiviert',
-    'common.status': 'Zustand',
-    'genreDetail.tabLibrary': 'In Ihrer Bibliothek',
-    'jobHistory.pipeline': 'Entdeckungslauf',
-    'jobHistory.librarySync': 'Bibliothekssynchronisierung',
-    'jobHistory.playlistType': 'Wiedergabeliste',
-    'playlistForm.dialogCreate': 'Playlist erstellen',
-    'playlistForm.dialogEdit': 'Playlist bearbeiten',
-    'playlistForm.titleCreate': 'Playlist erstellen',
-    'playlistForm.titleEdit': 'Playlist bearbeiten',
-    'playlistForm.namePlaceholder': 'Mein Wochenmix',
-    'playlistForm.strategy': 'Strategie',
-    'playlistForm.strategyWeeklyDigestDesc': 'Kuratiere Auswahl aus Ihren aktuellen Entdeckungen',
-    'playlistForm.strategyGenreFocusDesc': 'Ein bestimmtes Genre gezielt tiefer erkunden',
-    'playlistForm.strategyMoodMixDesc': 'Titel, die zu einer bestimmten Stimmung passen',
-    'playlistForm.strategyRediscoverDesc':
-      'Künstler wiederfinden, die Sie vielleicht vergessen haben',
-    'playlistForm.mood': 'Stimmung',
-    'playlistForm.moodPlaceholder': 'nächtliche Fahrt',
-    'playlistForm.trackCount': 'Anzahl der Titel',
-    'playlistForm.scheduleAutomaticGeneration': 'Automatische Generierung planen',
-    'playlistForm.nameRequired': 'Name ist erforderlich',
-    'playlistForm.genreRequired': 'Das Genre ist für die Strategie Genre-Fokus erforderlich',
-    'playlistForm.moodRequired': 'Die Stimmung ist fuer die Strategie Mood Mix erforderlich',
-    'subscriptionForm.dialogCreate': 'Abo erstellen',
-    'subscriptionForm.dialogEdit': 'Abo bearbeiten',
-    'subscriptionForm.titleCreate': 'Abo erstellen',
-    'subscriptionForm.titleEdit': 'Abo bearbeiten',
-    'subscriptionForm.name': 'Bezeichnung',
-    'subscriptionForm.namePlaceholder': 'Neues Shoegaze Weekly',
-    'subscriptionForm.nameRequired': 'Name ist erforderlich',
-    'subscriptionForm.source': 'Quelle',
-    'subscriptionForm.sources': 'Quellen',
-    'subscriptionForm.sourceType': 'Quellentyp',
-    'subscriptionForm.sourceTypeGenre': 'Musikgenre',
-    'subscriptionForm.sourceTypeGenreDesc': 'Neue Künstler in einem bestimmten Genre entdecken',
-    'subscriptionForm.sourceTypeSimilar': 'Ähnliche',
-    'subscriptionForm.sourceTypeSimilarDesc': 'Künstler finden, die Ihren Favoriten ähneln',
-    'subscriptionForm.sourceTypeDiscoveryModeDesc':
-      'Einen Entdeckungsmodus verwenden, um Künstler zu erzeugen',
-    'subscriptionForm.discoveryMode': 'Entdeckungsmodus',
-    'subscriptionForm.selectDiscoveryMode':
-      'Wählen Sie einen Entdeckungsmodus zur Konfiguration aus.',
-    'subscriptionForm.discoveryModeSettingsRequired':
-      'Die Einstellungen fuer den Entdeckungsmodus sind erforderlich',
-    'subscriptionForm.genre': 'Musikgenre',
-    'subscriptionForm.genrePlaceholder': 'shoegaze',
-    'subscriptionForm.genreRequired': 'Genre ist erforderlich',
-    'subscriptionForm.seedArtists': 'Startkünstler',
-    'subscriptionForm.seedArtistsPlaceholder': 'Radiohead, Portishead, Massive Attack',
-    'subscriptionForm.seedArtistsHelp': 'Künstlernamen durch Kommas getrennt',
-    'subscriptionForm.seedArtistsRequired': 'Mindestens ein Startkünstler ist erforderlich',
-    'subscriptionForm.configureInSettings': 'Unter Einstellungen > Verbindungen konfigurieren',
-    'subscriptionForm.notConfigured': 'nicht konfiguriert',
-    'subscriptionForm.selectAtLeastOneSource': 'Wählen Sie mindestens eine Quelle aus',
-    'subscriptionForm.sourceLockedHelp':
-      'Quellentyp und Konfiguration können für dieses Abo nicht geändert werden.',
-    'subscriptionForm.maxArtistsPerRun': 'Maximale Künstler pro Lauf',
-    'subscriptionForm.scoringPreset': 'Bewertungs-Preset',
-    'subscriptionForm.scoringPresetDefault': 'Standard',
-    'subscriptionForm.scoringPresetGenre': 'Für Genres optimiert',
-    'subscriptions.spotifyNotConnected': 'Spotify ist nicht verbunden',
-    'subscriptionPresets.getStarted': 'Mit einem Preset starten',
-    'subscriptionPresets.intro':
-      'Wählen Sie ein Preset, um Abos automatisch anzulegen, oder bauen Sie Ihre eigene Konfiguration.',
-    'subscriptionPresets.casualTitle': 'Locker',
-    'subscriptionPresets.casualDescription': 'Wöchentliche Empfehlungen mit wenig Rauschen',
-    'subscriptionPresets.deepDiverTitle': 'Tiefgang',
-    'subscriptionPresets.deepDiverDescription':
-      'Breite Entdeckung mit mehreren Quellen und täglichen Aktualisierungen',
-    'subscriptionPresets.discoveryOnlyTitle': 'Nur entdecken',
-    'subscriptionPresets.discoveryOnlyDescription': 'Nur Pipeline, keine externen Feeds',
-    'subscriptionPresets.customTitle': 'Individuell',
-    'subscriptionPresets.customDescription': 'Stellen Sie Ihre eigene Abo-Kombination zusammen',
-    'subscriptionPresets.oneConnectedService': '1 verbundener Dienst',
-    'subscriptionPresets.twoConnectedServices': '2 verbundene Dienste',
-    'subscriptionPresets.requires': 'Erfordert {0}',
-    'subscriptionPresets.setUp': 'Einrichten',
-    'subscriptionPresets.applying': 'Wird angewendet...',
-    'subscriptionPresets.casualApplied': 'Das Preset Locker wurde angewendet',
-    'subscriptionPresets.deepDiverApplied': 'Das Preset Tiefgang wurde angewendet',
-    'subscriptionPresets.applyFailed': 'Preset konnte nicht angewendet werden',
-    'subscriptionPresets.aiOnlyEnabled': 'Nur-KI-Modus wurde aktiviert',
-    'subscriptionPresets.aiOnlyEnableFailed': 'Nur-KI-Modus konnte nicht aktiviert werden',
-    'cronPicker.preset': 'Zeitplanvorlage',
-    'cronPicker.dailyMidnight': 'Täglich (Mitternacht)',
-    'cronPicker.twiceAWeek': 'Zweimal pro Woche (Mo + Do)',
-    'cronPicker.everyTwoWeeks': 'Alle 2 Wochen (1. + 15.)',
-    'cronPicker.monthlyFirst': 'Monatlich (Tag 1)',
-    'cronPicker.standardHint': 'Standard-Cron: Minute Stunde Tag-des-Monats Monat Wochentag',
-    'recommendation.sourceGenre': 'Musikgenre',
-    'recommendation.streaming': 'Streaming-Dienst',
-    'libraryHealth.syncNow': 'Jetzt synchronisieren',
-    'libraryHealth.snapshotTitle': 'Letzte Momentaufnahme',
-    'libraryHealth.lastSynced': 'Zuletzt synchronisiert',
-    'libraryHealth.neverSynced': 'Noch nie synchronisiert.',
-    'libraryHealth.autoSyncEvery': 'Automatische Synchronisierung alle {0} Std.',
-    'settings.capImport': 'Importe',
-    'systemHealth.pipeline': 'Entdeckungslauf',
-    'systemHealth.playlists': 'Wiedergabelisten',
-    'systemHealth.librarySync': 'Bibliothekssync',
-    'systemHealth.inMinutes': 'in {0} Min.',
-    'systemHealth.inHours': 'in {0} Std.',
-    'discoveryMode.labels.label': 'Bezeichnungen',
-    'discoveryMode.field.feed': 'Feed-Quelle',
-    'discoveryMode.field.limit': 'Obergrenze',
-    'discoveryMode.field.tags': 'Schlagwörter',
-    'discoveryMode.option.weeklyJams': 'Weekly Jams',
-    'settings.tabs.administration': 'Verwaltung',
-    'settings.tabs.systemHealth': 'Systemzustand',
-    'settings.name': 'Bezeichnung',
-    'settings.targetTypeEmbyPlaylist': 'Emby-Playlist',
-    'settings.limits': 'Grenzwerte',
-    'settings.librarySyncInterval': 'Bibliothekssync',
-    'settings.librarySyncIntervalHours': 'Sync-Intervall (Stunden)',
-    'settings.librarySyncIntervalHelp':
-      'Steuert, wie oft Digarr die Bibliothekssync-Daten aktualisiert und Bibliotheksprüfungen erneut startet.',
-    'settings.app': 'Anwendung',
-    'settings.oidcSso': 'OIDC / SSO',
-    'settings.fieldApiKeyOptional': 'API-Schlüssel (optional)',
-    'settings.fieldClientId': 'Client-ID',
-    'settings.fieldClientSecret': 'Client-Secret',
-    'settings.fieldScopes': 'Berechtigungen',
-    'settings.aiProviderTitle': 'KI-Anbieter',
-    'settings.webhookTitle': 'Webhook-Integration',
-    'settings.plexToken': 'Plex-Token',
-    'settings.serviceConnectedToast': '{0} verbunden',
-    'settings.serviceUnreachableToast': '{0} ist nicht erreichbar',
-    'settings.serviceSettingsSavedToast': '{0}-Einstellungen gespeichert',
-    'settings.serviceSettingsFailedToast': '{0}-Einstellungen konnten nicht gespeichert werden',
-    'settings.spotifyAuthorizationFailed': 'Spotify-Autorisierung konnte nicht gestartet werden',
-    'settings.spotifyDisconnectFailed': 'Spotify konnte nicht getrennt werden',
-    'settings.spotifyLikedSongsFailed':
-      'Import der Spotify-Lieblingssongs konnte nicht gestartet werden',
-    'settings.spotifyPlaylistFailed': 'Playlist-Import konnte nicht gestartet werden',
-    'settings.deezerFavoritesFailed': 'Import der Deezer-Favoriten konnte nicht gestartet werden',
-    'settings.deezerFollowedFailed':
-      'Import gefolgter Deezer-Künstler konnte nicht gestartet werden',
-    'settings.deezerDisconnectFailed': 'Deezer konnte nicht getrennt werden',
-    'settings.deezerAuthorizationFailed': 'Deezer-Autorisierung konnte nicht gestartet werden',
-    'settings.reverseProxyConfigHint':
-      'Verwenden Sie PROXY_AUTH_ENABLED und PROXY_AUTH_TRUSTED_PROXIES; Ihr Proxy muss X-Forwarded-User senden.',
-    'searchResultCard.addToQueue': 'Zur Entdeckungswarteschlange hinzufügen',
-    'searchResultCard.queuing': 'Wird hinzugefügt...',
-    'searchResultCard.queuedSuccess': '{0} wurde zur Entdeckungswarteschlange hinzugefügt',
-    'searchResultCard.queuedFailure': '{0} konnte nicht zur Warteschlange hinzugefügt werden',
-    'searchResultCard.popularity': 'Popularität {0}{1}',
-    'admin.autoFixInProgress': 'Automatische Korrektur läuft... ({0}/{1} korrigiert)',
-    'errorBoundary.title': 'Etwas ist schiefgelaufen',
-    'errorBoundary.retry': 'Erneut versuchen',
-    'errorBoundary.home': 'Zur Startseite',
-    'connectionSuggestions.title': 'Vorschläge für {0}',
-    'connectionSuggestions.close': 'Vorschläge schließen',
-    'connectionSuggestions.gotIt': 'Verstanden',
-    'connectionSuggestions.skipForNow': 'Später',
-    'connectionSuggestions.spotifyDiscoverWeekly': 'Ein Discover-Weekly-Abo einrichten',
-    'connectionSuggestions.spotifyReleaseRadar': 'Ein Release-Radar-Abo aktivieren',
-    'connectionSuggestions.lastfmTopGenreTags': 'Ihre wichtigsten Genre-Tags abonnieren',
-    'connectionSuggestions.listenbrainzWeeklyJams': 'Ein Weekly-Jams-Abo aktivieren',
-    'connectionSuggestions.listenbrainzWeeklyExploration': 'Weekly Exploration aktivieren',
-    'connectionSuggestions.weeklyDigestPlaylist':
-      'Eine wöchentliche Digarr-Digest-Playlist erstellen',
-    'keyboardShortcuts.title': 'Tastenkombinationen',
-    'keyboardShortcuts.close': 'Tastenkombinationen schließen',
-    'keyboardShortcuts.nextCard': 'Weiter zur nächsten Karte',
-    'keyboardShortcuts.previousCard': 'Zur vorherigen Karte',
-    'keyboardShortcuts.approveFocusedCard': 'Aktive Karte bestätigen',
-    'keyboardShortcuts.rejectFocusedCard': 'Aktive Karte ablehnen',
-    'keyboardShortcuts.toggleDetailView': 'Detailansicht öffnen / schließen',
-    'keyboardShortcuts.toggleShortcuts': 'Tastenkombinationen ein-/ausblenden',
-    'keyboardShortcuts.previousCardStack': 'Vorherige Karte (Stapelansicht)',
-    'keyboardShortcuts.nextCardStack': 'Nächste Karte (Stapelansicht)',
-    'keyboardShortcuts.disabledWhileTyping':
-      'Tastenkombinationen sind deaktiviert, wenn ein Textfeld fokussiert ist.',
-    'admin.restoreSummary': '{0} Zeilen in {1} Tabellen wiederhergestellt',
-    'admin.restoreSummaryForce':
-      '{0} Zeilen wiederhergestellt. Geben Sie Anmeldedaten für verschlüsselte Felder erneut ein.',
-    'admin.encryptionMismatchMessage':
-      'Das Backup wurde mit einem anderen Verschlüsselungsschlüssel erstellt.',
-    'admin.encryptionMismatchAffected': 'Diese Felder müssen erneut eingegeben werden: {0}',
-    'admin.encryptionMismatchFallback':
-      'Einige verschlüsselte Felder müssen möglicherweise erneut eingegeben werden.',
-    'integration.playlist': 'Wiedergabeliste',
-    'settings.yourAccount': 'Ihr Konto',
-    'settings.lidarrPreferences': 'Ihre Lidarr-Einstellungen',
-    'settings.targetUpdated': 'Ziel aktualisiert',
-    'settings.targetUpdateFailed': 'Ziel konnte nicht aktualisiert werden',
-    'settings.editTarget': 'Ziel bearbeiten',
-    'settings.saveChanges': 'Änderungen speichern',
-    'settings.linkedTarget': 'Verknüpftes Ziel:',
-    'albumPicker.title': 'Alben auswählen',
-    'albumPicker.loadFailed': 'Alben konnten nicht geladen werden',
-    'albumPicker.selectAll': 'Alle',
-    'albumPicker.selectNone': 'Keine',
-    'albumPicker.selectedCount': '{0} ausgewählt',
-    'albumPicker.aiPick': 'KI-Auswahl',
-    'albumPicker.monitor': 'Überwachen',
-    'albumPicker.monitorSingular': '{0} Album überwachen',
-    'albumPicker.monitorPlural': '{0} Alben überwachen',
-    'approveDialog.title': 'Lidarr-Einstellungen für diesen Künstler',
-    'approveDialog.loadFailed':
-      'Lidarr-Profile konnten nicht geladen werden. Prüfen Sie Ihre Lidarr-Verbindung.',
-    'approveDialog.qualityProfile': 'Qualitätsprofil',
-    'approveDialog.metadataProfile': 'Metadatenprofil',
-    'approveDialog.rootFolder': 'Stammordner',
-    'cardStack.counter': '{0} von {1}',
-    'cardStack.swipeHint': 'Nach rechts wischen zum Bestätigen, nach links zum Ablehnen',
-    'serviceCard.testConnection': 'Verbindung testen',
-    'preview.noPreviewAvailable': 'Keine Vorschau für diesen Künstler verfügbar',
-    'preview.playbackBlocked': 'Wiedergabe vom Browser blockiert - bitte erneut klicken',
-    'errors.auth.invalidCredentials': 'Ungültige Anmeldedaten',
-    'errors.auth.notAuthenticated': 'Nicht authentifiziert',
-    'errors.auth.passwordIncorrect': 'Aktuelles Passwort ist falsch',
-    'errors.auth.usernameTaken': 'Benutzername bereits vergeben',
-    'errors.target.notFound': 'Ziel nicht gefunden',
-    'errors.pipeline.alreadyRunning': 'Eine Pipeline-Ausführung läuft bereits',
-    'errors.recommendation.notFound': 'Empfehlung nicht gefunden',
-    'errors.subscription.notFound': 'Abonnement nicht gefunden',
-    'errors.playlist.notFound': 'Wiedergabeliste nicht gefunden',
-  },
-  it: {
-    ...REVIEW_GATE_TRANSLATIONS.it,
-    'common.unauthorized': 'Non autorizzato',
-    'common.adminAccessRequired': 'Accesso amministratore richiesto',
-    'common.importing': 'Importazione in corso...',
-    'common.current': 'Attuale',
-    'common.custom': 'Personalizzato',
-    'common.enabled': 'Attivato',
-    'jobHistory.pipeline': 'Pipeline di scoperta',
-    'jobHistory.librarySync': 'Sincronizzazione libreria',
-    'jobHistory.playlistType': 'Lista di riproduzione',
-    'libraryReconciliation.albumsCouldNotBeMatched':
-      'album che non e stato possibile associare automaticamente a MusicBrainz.',
-    'libraryReconciliation.artistsCouldNotBeMatched':
-      'artisti che non e stato possibile associare automaticamente a MusicBrainz.',
-    'playlists.playlistSingular': 'lista di riproduzione',
-    'subscriptionForm.dialogCreate': 'Crea abbonamento',
-    'subscriptionForm.dialogEdit': 'Modifica abbonamento',
-    'subscriptionForm.titleCreate': 'Crea abbonamento',
-    'subscriptionForm.titleEdit': 'Modifica abbonamento',
-    'subscriptionForm.name': 'Nome',
-    'subscriptionForm.namePlaceholder': 'Nuovo Shoegaze Weekly',
-    'subscriptionForm.nameRequired': 'Il nome e obbligatorio',
-    'subscriptionForm.source': 'Origine',
-    'subscriptionForm.sources': 'Origini',
-    'subscriptionForm.sourceType': 'Tipo di sorgente',
-    'subscriptionForm.sourceTypeGenre': 'Genere',
-    'subscriptionForm.sourceTypeGenreDesc': 'Scopri artisti in un genere specifico',
-    'subscriptionForm.sourceTypeSimilar': 'Simili',
-    'subscriptionForm.sourceTypeSimilarDesc': 'Trova artisti simili ai tuoi preferiti',
-    'subscriptionForm.sourceTypeDiscoveryModeDesc':
-      'Usa una modalita di scoperta per generare artisti',
-    'subscriptionForm.discoveryMode': 'Modalita di scoperta',
-    'subscriptionForm.selectDiscoveryMode': 'Seleziona una modalita di scoperta per configurarla.',
-    'subscriptionForm.discoveryModeSettingsRequired':
-      'Le impostazioni della modalita di scoperta sono obbligatorie',
-    'subscriptionForm.genre': 'Genere',
-    'subscriptionForm.genrePlaceholder': 'shoegaze',
-    'subscriptionForm.genreRequired': 'Il genere e obbligatorio',
-    'subscriptionForm.seedArtists': 'Artisti seed',
-    'subscriptionForm.seedArtistsPlaceholder': 'Radiohead, Portishead, Massive Attack',
-    'subscriptionForm.seedArtistsHelp': 'Nomi di artisti separati da virgole',
-    'subscriptionForm.seedArtistsRequired': 'E richiesto almeno un artista seed',
-    'subscriptionForm.configureInSettings': 'Configura in Impostazioni > Connessioni',
-    'subscriptionForm.notConfigured': 'non configurato',
-    'subscriptionForm.selectAtLeastOneSource': 'Seleziona almeno una sorgente',
-    'subscriptionForm.sourceLockedHelp':
-      'Il tipo di sorgente e la configurazione non possono essere modificati per questo abbonamento.',
-    'subscriptionForm.maxArtistsPerRun': 'Artisti massimi per esecuzione',
-    'subscriptionForm.scoringPreset': 'Preset di punteggio',
-    'subscriptionForm.scoringPresetDefault': 'Predefinito',
-    'subscriptionForm.scoringPresetGenre': 'Ottimizzato per genere',
-    'subscriptions.spotifyNotConnected': 'Spotify non e connesso',
-    'subscriptionPresets.getStarted': 'Inizia da un preset',
-    'subscriptionPresets.intro':
-      'Scegli un preset per creare abbonamenti automaticamente oppure costruisci una configurazione personalizzata.',
-    'subscriptionPresets.casualTitle': 'Leggero',
-    'subscriptionPresets.casualDescription': 'Suggerimenti settimanali con poco rumore',
-    'subscriptionPresets.deepDiverTitle': 'Esplorazione profonda',
-    'subscriptionPresets.deepDiverDescription':
-      'Scoperta ampia da più fonti con aggiornamenti quotidiani',
-    'subscriptionPresets.discoveryOnlyTitle': 'Solo scoperta',
-    'subscriptionPresets.discoveryOnlyDescription': 'Solo pipeline, senza feed esterni',
-    'subscriptionPresets.customTitle': 'Personalizzato',
-    'subscriptionPresets.customDescription': 'Crea la tua combinazione di abbonamenti',
-    'subscriptionPresets.oneConnectedService': '1 servizio connesso',
-    'subscriptionPresets.twoConnectedServices': '2 servizi connessi',
-    'subscriptionPresets.requires': 'Richiede {0}',
-    'subscriptionPresets.setUp': 'Configura',
-    'subscriptionPresets.applying': 'Applicazione in corso...',
-    'subscriptionPresets.casualApplied': 'Preset Leggero applicato',
-    'subscriptionPresets.deepDiverApplied': 'Preset Esplorazione profonda applicato',
-    'subscriptionPresets.applyFailed': 'Impossibile applicare il preset',
-    'subscriptionPresets.aiOnlyEnabled': 'Modalita solo AI attivata',
-    'subscriptionPresets.aiOnlyEnableFailed': 'Impossibile attivare la modalita solo AI',
-    'cronPicker.preset': 'Preimpostazione',
-    'cronPicker.dailyMidnight': 'Ogni giorno (mezzanotte)',
-    'cronPicker.twiceAWeek': 'Due volte a settimana (lun + gio)',
-    'cronPicker.everyTwoWeeks': 'Ogni 2 settimane (1 + 15)',
-    'cronPicker.monthlyFirst': 'Mensile (giorno 1)',
-    'cronPicker.standardHint':
-      'Cron standard: minuto ora giorno-del-mese mese giorno-della-settimana',
-    'recommendation.streaming': 'Servizio di streaming',
-    'libraryHealth.syncNow': 'Sincronizza ora',
-    'libraryHealth.snapshotTitle': 'Ultima istantanea',
-    'libraryHealth.lastSynced': 'Ultima sincronizzazione',
-    'libraryHealth.neverSynced': 'Mai sincronizzato.',
-    'libraryHealth.autoSyncEvery': 'Sincronizzazione automatica ogni {0} h',
-    'systemHealth.pipeline': 'Pipeline di scoperta',
-    'systemHealth.librarySync': 'Sync libreria',
-    'discoveryMode.field.feed': 'Feed di origine',
-    'errorBoundary.title': 'Qualcosa e andato storto',
-    'errorBoundary.retry': 'Riprova',
-    'errorBoundary.home': 'Vai alla home',
-    'connectionSuggestions.title': 'Suggerimenti per {0}',
-    'connectionSuggestions.close': 'Chiudi suggerimenti',
-    'connectionSuggestions.gotIt': 'Capito',
-    'connectionSuggestions.skipForNow': 'Salta per ora',
-    'connectionSuggestions.spotifyDiscoverWeekly': 'Configura un abbonamento Discover Weekly',
-    'connectionSuggestions.spotifyReleaseRadar': 'Attiva l abbonamento Release Radar',
-    'connectionSuggestions.lastfmTopGenreTags': 'Abbonati ai tuoi tag di genere principali',
-    'connectionSuggestions.listenbrainzWeeklyJams': 'Attiva l abbonamento Weekly Jams',
-    'connectionSuggestions.listenbrainzWeeklyExploration': 'Attiva Weekly Exploration',
-    'connectionSuggestions.weeklyDigestPlaylist': 'Crea una playlist Digarr Digest settimanale',
-    'keyboardShortcuts.title': 'Scorciatoie da tastiera',
-    'keyboardShortcuts.close': 'Chiudi scorciatoie',
-    'keyboardShortcuts.nextCard': 'Scheda successiva',
-    'keyboardShortcuts.previousCard': 'Scheda precedente',
-    'keyboardShortcuts.approveFocusedCard': 'Approva la scheda attiva',
-    'keyboardShortcuts.rejectFocusedCard': 'Rifiuta la scheda attiva',
-    'keyboardShortcuts.toggleDetailView': 'Apri / chiudi la vista dettagliata',
-    'keyboardShortcuts.toggleShortcuts': 'Mostra / nascondi le scorciatoie',
-    'keyboardShortcuts.previousCardStack': 'Scheda precedente (vista pila)',
-    'keyboardShortcuts.nextCardStack': 'Scheda successiva (vista pila)',
-    'keyboardShortcuts.disabledWhileTyping':
-      'Le scorciatoie sono disattivate quando un campo di testo e attivo.',
-    'admin.restoreSummary': 'Ripristinate {0} righe in {1} tabelle',
-    'admin.restoreSummaryForce':
-      'Ripristinate {0} righe. Reinserisci le credenziali per i campi cifrati.',
-    'admin.encryptionMismatchMessage':
-      'Il backup e stato creato con una chiave di crittografia diversa.',
-    'admin.encryptionMismatchAffected': 'Questi campi dovranno essere reinseriti: {0}',
-    'admin.encryptionMismatchFallback': 'Alcuni campi cifrati potrebbero dover essere reinseriti.',
-    'admin.autoFixInProgress': 'Correzione automatica in corso... ({0}/{1} corretti)',
-    'settings.capDiscovery': 'Scoperta',
-    'settings.capImport': 'Importazione',
-    'settings.capLibrarySync': 'Sync libreria',
-    'settings.capPlaylistExport': 'Esportazione playlist',
-    'settings.capService': 'Servizio',
-    'settings.capSubscriptions': 'Abbonamenti',
-    'settings.tabs.account': 'Account utente',
-    'settings.tabs.systemHealth': 'Stato del sistema',
-    'settings.targetUpdated': 'Destinazione aggiornata',
-    'settings.targetUpdateFailed': 'Impossibile aggiornare la destinazione',
-    'settings.editTarget': 'Modifica destinazione',
-    'settings.saveChanges': 'Salva modifiche',
-    'settings.linkedTarget': 'Destinazione collegata:',
-    'settings.librarySyncInterval': 'Sync libreria',
-    'settings.librarySyncIntervalHours': 'Intervallo di sync (ore)',
-    'settings.librarySyncIntervalHelp':
-      'Controlla con quale frequenza Digarr aggiorna i dati di sincronizzazione della libreria e rilancia i controlli automaticamente.',
-    'settings.app': 'Applicazione',
-    'settings.oidcSso': 'OIDC / SSO',
-    'settings.fieldProvider': 'Fornitore',
-    'settings.aiProviderTitle': 'Provider AI',
-    'settings.webhookTitle': 'Integrazione webhook',
-    'settings.plexToken': 'Token Plex',
-    'settings.serviceConnectedToast': '{0} connesso',
-    'settings.serviceUnreachableToast': '{0} non raggiungibile',
-    'settings.serviceSettingsSavedToast': 'Impostazioni di {0} salvate',
-    'settings.serviceSettingsFailedToast': 'Impossibile salvare le impostazioni di {0}',
-    'settings.spotifyAuthorizationFailed': 'Impossibile avviare l autorizzazione Spotify',
-    'settings.spotifyDisconnectFailed': 'Impossibile disconnettere Spotify',
-    'settings.spotifyLikedSongsFailed':
-      'Impossibile avviare l import dei brani preferiti di Spotify',
-    'settings.spotifyPlaylistFailed': 'Impossibile avviare l import della playlist',
-    'settings.deezerFavoritesFailed': 'Impossibile avviare l import dei preferiti Deezer',
-    'settings.deezerFollowedFailed': 'Impossibile avviare l import degli artisti seguiti su Deezer',
-    'settings.deezerDisconnectFailed': 'Impossibile disconnettere Deezer',
-    'settings.deezerAuthorizationFailed': 'Impossibile avviare l autorizzazione Deezer',
-    'settings.reverseProxyConfigHint':
-      'Usa PROXY_AUTH_ENABLED e PROXY_AUTH_TRUSTED_PROXIES; il proxy deve inviare X-Forwarded-User.',
-    'searchResultCard.addToQueue': 'Aggiungi alla coda di scoperta',
-    'searchResultCard.queuing': 'Aggiunta in corso...',
-    'searchResultCard.queuedSuccess': '{0} e stato aggiunto alla coda di scoperta',
-    'searchResultCard.queuedFailure': 'Impossibile aggiungere {0} alla coda',
-    'searchResultCard.popularity': 'Popolarita {0}{1}',
-    'integration.playlist': 'Lista di riproduzione',
-    'app.themeGroupStreaming': 'Servizi streaming',
-    'playlistForm.dialogCreate': 'Crea playlist',
-    'playlistForm.dialogEdit': 'Modifica playlist',
-    'playlistForm.titleCreate': 'Crea playlist',
-    'playlistForm.titleEdit': 'Modifica playlist',
-    'playlistForm.namePlaceholder': 'Il mio mix settimanale',
-    'playlistForm.strategy': 'Strategia',
-    'playlistForm.strategyWeeklyDigestDesc': 'Selezione curata dalle tue scoperte recenti',
-    'playlistForm.strategyGenreFocusDesc': 'Approfondisci un genere specifico',
-    'playlistForm.strategyMoodMixDesc': 'Brani adatti a un atmosfera precisa',
-    'playlistForm.strategyRediscoverDesc':
-      'Riporta in primo piano artisti che potresti aver dimenticato',
-    'playlistForm.mood': 'Atmosfera',
-    'playlistForm.moodPlaceholder': 'guida notturna',
-    'playlistForm.trackCount': 'Numero di brani',
-    'playlistForm.scheduleAutomaticGeneration': 'Pianifica la generazione automatica',
-    'playlistForm.nameRequired': 'Il nome e obbligatorio',
-    'playlistForm.genreRequired': 'Il genere e obbligatorio per la strategia Focus sul genere',
-    'playlistForm.moodRequired': 'L atmosfera e obbligatoria per la strategia Mood Mix',
-    'albumPicker.title': 'Seleziona album',
-    'albumPicker.loadFailed': 'Impossibile caricare gli album',
-    'albumPicker.selectAll': 'Tutti',
-    'albumPicker.selectNone': 'Nessuno',
-    'albumPicker.selectedCount': '{0} selezionati',
-    'albumPicker.aiPick': 'Scelta AI',
-    'albumPicker.monitor': 'Monitora',
-    'albumPicker.monitorSingular': 'Monitora {0} album',
-    'albumPicker.monitorPlural': 'Monitora {0} album',
-    'approveDialog.title': 'Impostazioni Lidarr per questo artista',
-    'approveDialog.loadFailed':
-      'Impossibile caricare i profili Lidarr. Controlla la connessione a Lidarr.',
-    'approveDialog.qualityProfile': 'Profilo qualità',
-    'approveDialog.metadataProfile': 'Profilo metadati',
-    'approveDialog.rootFolder': 'Cartella radice',
-    'cardStack.counter': '{0} di {1}',
-    'cardStack.swipeHint': 'Scorri a destra per approvare, a sinistra per rifiutare',
-    'serviceCard.testConnection': 'Testa connessione',
-    'preview.noPreviewAvailable': 'Nessuna anteprima disponibile per questo artista',
-    'preview.playbackBlocked': 'Riproduzione bloccata dal browser; prova a cliccare di nuovo',
-    'errors.auth.invalidCredentials': 'Credenziali non valide',
-    'errors.auth.notAuthenticated': 'Non autenticato',
-    'errors.auth.passwordIncorrect': 'La password attuale non e corretta',
-    'errors.auth.usernameTaken': 'Nome utente gia in uso',
-    'errors.target.notFound': 'Destinazione non trovata',
-    'errors.pipeline.alreadyRunning': 'Una esecuzione della pipeline e gia in corso',
-    'errors.recommendation.notFound': 'Raccomandazione non trovata',
-    'errors.subscription.notFound': 'Abbonamento non trovato',
-    'errors.playlist.notFound': 'Playlist non trovata',
-  },
-  nl: {
-    ...REVIEW_GATE_TRANSLATIONS.nl,
-    'nav.dashboard': 'Overzicht',
-    'setup.model': 'Modelkeuze',
-    'common.importing': 'Importeren...',
-    'common.current': 'Huidig',
-    'common.custom': 'Aangepast',
-    'common.enabled': 'Ingeschakeld',
-    'common.unauthorized': 'Niet gemachtigd',
-    'common.adminAccessRequired': 'Beheerderstoegang vereist',
-    'common.status': 'Toestand',
-    'genres.genrePlural': 'muziekgenres',
-    'genres.genreSingular': 'muziekgenre',
-    'jobHistory.pipeline': 'Ontdekkingsrun',
-    'jobHistory.librarySync': 'Bibliotheeksynchronisatie',
-    'jobHistory.playlistType': 'Afspeellijst',
-    'jobHistory.metadataLabel': 'Metagegevens:',
-    'playlistForm.dialogCreate': 'Afspeellijst maken',
-    'playlistForm.dialogEdit': 'Afspeellijst bewerken',
-    'playlistForm.titleCreate': 'Afspeellijst maken',
-    'playlistForm.titleEdit': 'Afspeellijst bewerken',
-    'playlistForm.namePlaceholder': 'Mijn wekelijkse mix',
-    'playlistForm.strategy': 'Samenstelling',
-    'playlistForm.strategyWeeklyDigestDesc': 'Een zorgvuldige selectie uit je recente ontdekkingen',
-    'playlistForm.strategyGenreFocusDesc': 'Duik diep in één specifiek genre',
-    'playlistForm.strategyMoodMixDesc': 'Tracks die bij een bepaalde sfeer passen',
-    'playlistForm.strategyRediscoverDesc': 'Haal artiesten terug die je misschien was vergeten',
-    'playlistForm.mood': 'Sfeer',
-    'playlistForm.moodPlaceholder': 'nachtelijke rit',
-    'playlistForm.trackCount': 'Aantal tracks',
-    'playlistForm.scheduleAutomaticGeneration': 'Automatisch genereren plannen',
-    'playlistForm.nameRequired': 'Naam is verplicht',
-    'playlistForm.genreRequired': 'Genre is verplicht voor de strategie Genrefocus',
-    'playlistForm.moodRequired': 'Sfeer is verplicht voor de strategie Moodmix',
-    'subscriptionForm.dialogCreate': 'Abonnement maken',
-    'subscriptionForm.dialogEdit': 'Abonnement bewerken',
-    'subscriptionForm.titleCreate': 'Abonnement maken',
-    'subscriptionForm.titleEdit': 'Abonnement bewerken',
-    'subscriptionForm.name': 'Naam',
-    'subscriptionForm.namePlaceholder': 'Nieuwe Shoegaze Weekly',
-    'subscriptionForm.nameRequired': 'Naam is verplicht',
-    'subscriptionForm.source': 'Bron',
-    'subscriptionForm.sources': 'Bronnen',
-    'subscriptionForm.sourceType': 'Brontype',
-    'subscriptionForm.sourceTypeGenre': 'Muziekgenre',
-    'subscriptionForm.sourceTypeGenreDesc': 'Ontdek artiesten binnen een specifiek genre',
-    'subscriptionForm.sourceTypeSimilar': 'Vergelijkbare artiesten',
-    'subscriptionForm.sourceTypeSimilarDesc': 'Vind artiesten die op je favorieten lijken',
-    'subscriptionForm.sourceTypeDiscoveryModeDesc':
-      'Gebruik een ontdekkingsmodus om artiesten te genereren',
-    'subscriptionForm.discoveryMode': 'Ontdekkingsmodus',
-    'subscriptionForm.selectDiscoveryMode': 'Kies een ontdekkingsmodus om in te stellen.',
-    'subscriptionForm.discoveryModeSettingsRequired':
-      'De instellingen voor de ontdekkingsmodus zijn verplicht',
-    'subscriptionForm.genre': 'Muziekgenre',
-    'subscriptionForm.genrePlaceholder': 'shoegaze',
-    'subscriptionForm.seedArtistsPlaceholder': 'Radiohead, Portishead, Massive Attack',
-    'subscriptionForm.genreRequired': 'Genre is verplicht',
-    'subscriptionForm.seedArtists': 'Startartiesten',
-    'subscriptionForm.seedArtistsHelp': 'Artiestnamen gescheiden door komma s',
-    'subscriptionForm.seedArtistsRequired': 'Minstens één startartiest is verplicht',
-    'subscriptionForm.configureInSettings': 'Stel dit in via Instellingen > Verbindingen',
-    'subscriptionForm.notConfigured': 'niet ingesteld',
-    'subscriptionForm.selectAtLeastOneSource': 'Selecteer minstens één bron',
-    'subscriptionForm.sourceLockedHelp':
-      'Het brontype en de configuratie kunnen voor dit abonnement niet worden gewijzigd.',
-    'subscriptionForm.maxArtistsPerRun': 'Maximaal aantal artiesten per uitvoering',
-    'subscriptionForm.scoringPreset': 'Scoreprofiel',
-    'subscriptionForm.scoringPresetDefault': 'Standaard',
-    'subscriptionForm.scoringPresetGenre': 'Geoptimaliseerd voor genres',
-    'subscriptions.spotifyNotConnected': 'Spotify is niet gekoppeld',
-    'subscriptionPresets.getStarted': 'Begin met een voorinstelling',
-    'subscriptionPresets.intro':
-      'Kies een voorinstelling om automatisch abonnementen aan te maken of stel zelf een configuratie samen.',
-    'subscriptionPresets.casualTitle': 'Ontspannen',
-    'subscriptionPresets.casualDescription': 'Wekelijkse aanbevelingen met weinig ruis',
-    'subscriptionPresets.deepDiverTitle': 'Diepgraver',
-    'subscriptionPresets.deepDiverDescription':
-      'Brede ontdekking uit meerdere bronnen met dagelijkse updates',
-    'subscriptionPresets.discoveryOnlyTitle': 'Alleen ontdekken',
-    'subscriptionPresets.discoveryOnlyDescription': 'Alleen de pijplijn, zonder externe feeds',
-    'subscriptionPresets.customTitle': 'Aangepast',
-    'subscriptionPresets.customDescription': 'Stel je eigen combinatie van abonnementen samen',
-    'subscriptionPresets.oneConnectedService': '1 verbonden dienst',
-    'subscriptionPresets.twoConnectedServices': '2 verbonden diensten',
-    'subscriptionPresets.requires': 'Vereist {0}',
-    'subscriptionPresets.setUp': 'Instellen',
-    'subscriptionPresets.applying': 'Bezig met toepassen...',
-    'subscriptionPresets.casualApplied': 'De voorinstelling Ontspannen is toegepast',
-    'subscriptionPresets.deepDiverApplied': 'De voorinstelling Diepgraver is toegepast',
-    'subscriptionPresets.applyFailed': 'De voorinstelling kon niet worden toegepast',
-    'subscriptionPresets.aiOnlyEnabled': 'Alleen-AI-modus is ingeschakeld',
-    'subscriptionPresets.aiOnlyEnableFailed': 'Alleen-AI-modus kon niet worden ingeschakeld',
-    'cronPicker.preset': 'Voorinstelling',
-    'cronPicker.dailyMidnight': 'Dagelijks (middernacht)',
-    'cronPicker.twiceAWeek': 'Twee keer per week (ma + do)',
-    'cronPicker.everyTwoWeeks': 'Om de 2 weken (1 en 15)',
-    'cronPicker.monthlyFirst': 'Maandelijks (dag 1)',
-    'cronPicker.standardHint': 'Standaard cron: minuut uur dag-van-de-maand maand dag-van-de-week',
-    'recommendation.sourceGenre': 'Muziekgenre',
-    'libraryHealth.syncNow': 'Nu synchroniseren',
-    'libraryHealth.snapshotTitle': 'Laatste momentopname',
-    'libraryHealth.lastSynced': 'Laatst gesynchroniseerd',
-    'libraryHealth.neverSynced': 'Nog niet gesynchroniseerd.',
-    'libraryHealth.autoSyncEvery': 'Automatisch synchroniseren elke {0} uur',
-    'settings.capImport': 'Importeren',
-    'systemHealth.pipeline': 'Ontdekkingsrun',
-    'systemHealth.librarySync': 'Bibliotheeksynchronisatie',
-    'discoveryMode.labels.label': 'Etiketten',
-    'discoveryMode.field.feed': 'Feedbron',
-    'discoveryMode.field.tags': 'Trefwoorden',
-    'discoveryMode.option.weeklyJams': 'Weekly Jams',
-    'libraryStats.albums': 'Albumreleases',
-    'settings.tabs.account': 'Gebruikersaccount',
-    'settings.tabs.systemHealth': 'Systeemstatus',
-    'settings.targetUpdated': 'Doel is bijgewerkt',
-    'settings.targetUpdateFailed': 'Doel bijwerken is mislukt',
-    'settings.editTarget': 'Doel bewerken',
-    'settings.saveChanges': 'Wijzigingen opslaan',
-    'settings.linkedTarget': 'Gekoppeld doel:',
-    'settings.type': 'Soort',
-    'settings.consensus': 'Consensusscore',
-    'settings.librarySyncInterval': 'Bibliotheeksynchronisatie',
-    'settings.librarySyncIntervalHours': 'Synchronisatie-interval (uren)',
-    'settings.librarySyncIntervalHelp':
-      'Bepaalt hoe vaak Digarr synchronisatiegegevens van de bibliotheek ververst en gezondheidscontroles opnieuw uitvoert.',
-    'settings.app': 'Applicatie',
-    'settings.oidcSso': 'OIDC / SSO',
-    'settings.fieldModel': 'Modelkeuze',
-    'settings.aiProviderTitle': 'AI-provider',
-    'settings.webhookTitle': 'Webhook-koppeling',
-    'settings.plexToken': 'Plex-token',
-    'settings.serviceConnectedToast': '{0} gekoppeld',
-    'settings.serviceUnreachableToast': '{0} is niet bereikbaar',
-    'settings.serviceSettingsSavedToast': 'Instellingen voor {0} zijn opgeslagen',
-    'settings.serviceSettingsFailedToast': 'Instellingen voor {0} konden niet worden opgeslagen',
-    'settings.spotifyAuthorizationFailed': 'Spotify-autorisatie kon niet worden gestart',
-    'settings.spotifyDisconnectFailed': 'Spotify kon niet worden losgekoppeld',
-    'settings.spotifyLikedSongsFailed':
-      'Het importeren van Spotify-liked songs kon niet worden gestart',
-    'settings.spotifyPlaylistFailed': 'Het importeren van de afspeellijst kon niet worden gestart',
-    'settings.deezerFavoritesFailed':
-      'Het importeren van Deezer-favorieten kon niet worden gestart',
-    'settings.deezerFollowedFailed':
-      'Het importeren van gevolgde Deezer-artiesten kon niet worden gestart',
-    'settings.deezerDisconnectFailed': 'Deezer kon niet worden losgekoppeld',
-    'settings.deezerAuthorizationFailed': 'Deezer-autorisatie kon niet worden gestart',
-    'settings.reverseProxyConfigHint':
-      'Gebruik PROXY_AUTH_ENABLED en PROXY_AUTH_TRUSTED_PROXIES; je proxy moet X-Forwarded-User meesturen.',
-    'searchResultCard.addToQueue': 'Aan ontdekkingswachtrij toevoegen',
-    'searchResultCard.queuing': 'Toevoegen...',
-    'searchResultCard.queuedSuccess': '{0} is toegevoegd aan de ontdekkingswachtrij',
-    'searchResultCard.queuedFailure': '{0} kon niet aan de wachtrij worden toegevoegd',
-    'searchResultCard.popularity': 'Populariteit {0}{1}',
-    'admin.restoreSummary': '{0} rijen hersteld in {1} tabellen',
-    'admin.restoreSummaryForce':
-      '{0} rijen hersteld. Voer de gegevens voor versleutelde velden opnieuw in.',
-    'admin.encryptionMismatchMessage':
-      'Deze back-up is gemaakt met een andere versleutelingssleutel.',
-    'admin.encryptionMismatchAffected': 'Deze velden moeten opnieuw worden ingevuld: {0}',
-    'admin.encryptionMismatchFallback':
-      'Sommige versleutelde velden moeten mogelijk opnieuw worden ingevuld.',
-    'admin.autoFixInProgress': 'Automatische correctie bezig... ({0}/{1} hersteld)',
-    'errorBoundary.title': 'Er is iets misgegaan',
-    'errorBoundary.retry': 'Opnieuw proberen',
-    'errorBoundary.home': 'Naar startpagina',
-    'connectionSuggestions.title': 'Suggesties voor {0}',
-    'connectionSuggestions.close': 'Suggesties sluiten',
-    'connectionSuggestions.gotIt': 'Begrepen',
-    'connectionSuggestions.skipForNow': 'Nu overslaan',
-    'connectionSuggestions.spotifyDiscoverWeekly': 'Een Discover Weekly-abonnement instellen',
-    'connectionSuggestions.spotifyReleaseRadar': 'Een Release Radar-abonnement inschakelen',
-    'connectionSuggestions.lastfmTopGenreTags': 'Je belangrijkste genrelabels volgen',
-    'connectionSuggestions.listenbrainzWeeklyJams': 'Het Weekly Jams-abonnement inschakelen',
-    'connectionSuggestions.listenbrainzWeeklyExploration': 'Weekly Exploration inschakelen',
-    'connectionSuggestions.weeklyDigestPlaylist': 'Een wekelijkse Digarr Digest-afspeellijst maken',
-    'keyboardShortcuts.title': 'Sneltoetsen',
-    'keyboardShortcuts.close': 'Sneltoetsen sluiten',
-    'keyboardShortcuts.nextCard': 'Volgende kaart',
-    'keyboardShortcuts.previousCard': 'Vorige kaart',
-    'keyboardShortcuts.approveFocusedCard': 'Actieve kaart goedkeuren',
-    'keyboardShortcuts.rejectFocusedCard': 'Actieve kaart afwijzen',
-    'keyboardShortcuts.toggleDetailView': 'Detailweergave openen / sluiten',
-    'keyboardShortcuts.toggleShortcuts': 'Sneltoetsen tonen / verbergen',
-    'keyboardShortcuts.previousCardStack': 'Vorige kaart (stapelweergave)',
-    'keyboardShortcuts.nextCardStack': 'Volgende kaart (stapelweergave)',
-    'keyboardShortcuts.disabledWhileTyping':
-      'Sneltoetsen zijn uitgeschakeld zolang een tekstveld focus heeft.',
-    'settings.fieldClientId': 'Client-ID',
-    'settings.fieldClientSecret': 'Clientgeheim',
-    'settings.fieldScopes': 'Reikwijdten',
-    'analytics.source.consensus': 'Overeenstemming',
-    'albumPicker.title': 'Albums selecteren',
-    'albumPicker.loadFailed': 'Laden van albums mislukt',
-    'albumPicker.selectAll': 'Alle',
-    'albumPicker.selectNone': 'Geen',
-    'albumPicker.selectedCount': '{0} geselecteerd',
-    'albumPicker.aiPick': 'AI-keuze',
-    'albumPicker.monitor': 'Volgen',
-    'albumPicker.monitorSingular': '{0} album volgen',
-    'albumPicker.monitorPlural': '{0} albums volgen',
-    'approveDialog.title': 'Lidarr-instellingen voor deze artiest',
-    'approveDialog.loadFailed':
-      'Laden van Lidarr-profielen mislukt. Controleer je Lidarr-verbinding.',
-    'approveDialog.qualityProfile': 'Kwaliteitsprofiel',
-    'approveDialog.metadataProfile': 'Metadataprofiel',
-    'approveDialog.rootFolder': 'Hoofdmap',
-    'cardStack.counter': '{0} van {1}',
-    'cardStack.swipeHint': 'Veeg naar rechts om goed te keuren, naar links om af te wijzen',
-    'serviceCard.testConnection': 'Verbinding testen',
-    'preview.noPreviewAvailable': 'Geen voorbeeld beschikbaar voor deze artiest',
-    'preview.playbackBlocked': 'Afspelen geblokkeerd door browser - probeer opnieuw te klikken',
-    'errors.auth.invalidCredentials': 'Ongeldige inloggegevens',
-    'errors.auth.notAuthenticated': 'Niet aangemeld',
-    'errors.auth.passwordIncorrect': 'Huidig wachtwoord is onjuist',
-    'errors.auth.usernameTaken': 'Gebruikersnaam al in gebruik',
-    'errors.target.notFound': 'Doel niet gevonden',
-    'errors.pipeline.alreadyRunning': 'Er is al een pipeline-uitvoering bezig',
-    'errors.recommendation.notFound': 'Aanbeveling niet gevonden',
-    'errors.subscription.notFound': 'Abonnement niet gevonden',
-    'errors.playlist.notFound': 'Afspeellijst niet gevonden',
-  },
-  ro: {
-    ...REVIEW_GATE_TRANSLATIONS.ro,
-    'app.themeGroupEditor': 'Editare',
-    'app.themeGroupStreaming': 'Servicii de streaming',
-    'setup.lidarrUrl': 'Adresa Lidarr',
-    'common.importing': 'Importare...',
-    'common.manual': 'manual',
-    'common.current': 'Curent',
-    'common.custom': 'Personalizat',
-    'common.enabled': 'Activat',
-    'common.unauthorized': 'Neautorizat',
-    'common.adminAccessRequired': 'Este necesar accesul de administrator',
-    'common.total': 'În total',
-    'genreDetail.artistSingular': 'muzician',
-    'jobHistory.pipeline': 'Rulare de descoperire',
-    'jobHistory.librarySync': 'Sincronizarea bibliotecii',
-    'jobHistory.playlistType': 'Listă de redare',
-    'playlist.total': 'în total',
-    'playlistForm.dialogCreate': 'Creează lista de redare',
-    'playlistForm.dialogEdit': 'Editează lista de redare',
-    'playlistForm.titleCreate': 'Creează lista de redare',
-    'playlistForm.titleEdit': 'Editează lista de redare',
-    'playlistForm.namePlaceholder': 'Mixul meu săptămânal',
-    'playlistForm.strategy': 'Strategie',
-    'playlistForm.strategyWeeklyDigestDesc': 'Selecție atentă din descoperirile tale recente',
-    'playlistForm.strategyGenreFocusDesc': 'Explorează în profunzime un anumit gen',
-    'playlistForm.strategyMoodMixDesc': 'Piese potrivite pentru o anumită stare',
-    'playlistForm.strategyRediscoverDesc': 'Adu înapoi artiști pe care poate i-ai uitat',
-    'playlistForm.mood': 'Stare',
-    'playlistForm.moodPlaceholder': 'drum de noapte',
-    'playlistForm.trackCount': 'Număr de piese',
-    'playlistForm.scheduleAutomaticGeneration': 'Planifică generarea automată',
-    'playlistForm.nameRequired': 'Numele este obligatoriu',
-    'playlistForm.genreRequired': 'Genul este obligatoriu pentru strategia Focus pe gen',
-    'playlistForm.moodRequired': 'Starea este obligatorie pentru strategia Mood Mix',
-    'subscriptionForm.dialogCreate': 'Creează abonament',
-    'subscriptionForm.dialogEdit': 'Editează abonamentul',
-    'subscriptionForm.titleCreate': 'Creează abonament',
-    'subscriptionForm.titleEdit': 'Editează abonamentul',
-    'subscriptionForm.name': 'Nume',
-    'subscriptionForm.namePlaceholder': 'Shoegaze Weekly nou',
-    'subscriptionForm.nameRequired': 'Numele este obligatoriu',
-    'subscriptionForm.source': 'Sursă',
-    'subscriptionForm.sources': 'Surse',
-    'subscriptionForm.sourceType': 'Tip de sursă',
-    'subscriptionForm.sourceTypeGenre': 'Gen muzical',
-    'subscriptionForm.sourceTypeGenreDesc': 'Descoperă artiști într-un gen anume',
-    'subscriptionForm.sourceTypeSimilar': 'Artiști similari',
-    'subscriptionForm.sourceTypeSimilarDesc': 'Găsește artiști asemănători cu favoriții tăi',
-    'subscriptionForm.sourceTypeDiscoveryModeDesc':
-      'Folosește un mod de descoperire pentru a genera artiști',
-    'subscriptionForm.discoveryMode': 'Mod de descoperire',
-    'subscriptionForm.selectDiscoveryMode':
-      'Alege un mod de descoperire pe care să îl configurezi.',
-    'subscriptionForm.discoveryModeSettingsRequired':
-      'Setările pentru modul de descoperire sunt obligatorii',
-    'subscriptionForm.genre': 'Gen muzical',
-    'subscriptionForm.genrePlaceholder': 'shoegaze',
-    'subscriptionForm.seedArtistsPlaceholder': 'Radiohead, Portishead, Massive Attack',
-    'subscriptionForm.genreRequired': 'Genul este obligatoriu',
-    'subscriptionForm.seedArtists': 'Artiști de pornire',
-    'subscriptionForm.seedArtistsHelp': 'Nume de artiști separate prin virgulă',
-    'subscriptionForm.seedArtistsRequired': 'Este necesar cel puțin un artist de pornire',
-    'subscriptionForm.configureInSettings': 'Configurează în Setări > Conexiuni',
-    'subscriptionForm.notConfigured': 'neconfigurat',
-    'subscriptionForm.selectAtLeastOneSource': 'Selectează cel puțin o sursă',
-    'subscriptionForm.sourceLockedHelp':
-      'Tipul sursei și configurația nu pot fi schimbate pentru acest abonament.',
-    'subscriptionForm.maxArtistsPerRun': 'Număr maxim de artiști pe rulare',
-    'subscriptionForm.scoringPreset': 'Preset de scor',
-    'subscriptionForm.scoringPresetDefault': 'Implicit',
-    'subscriptionForm.scoringPresetGenre': 'Optimizat pentru gen',
-    'subscriptions.spotifyNotConnected': 'Spotify nu este conectat',
-    'subscriptionPresets.getStarted': 'Începe cu un preset',
-    'subscriptionPresets.intro':
-      'Alege un preset pentru a crea abonamente automat sau construiește-ți propria configurație.',
-    'subscriptionPresets.casualTitle': 'Lejer',
-    'subscriptionPresets.casualDescription': 'Recomandări săptămânale cu puțin zgomot',
-    'subscriptionPresets.deepDiverTitle': 'Explorare profundă',
-    'subscriptionPresets.deepDiverDescription':
-      'Descoperire amplă din mai multe surse și actualizări zilnice',
-    'subscriptionPresets.discoveryOnlyTitle': 'Doar descoperire',
-    'subscriptionPresets.discoveryOnlyDescription': 'Doar pipeline, fără feeduri externe',
-    'subscriptionPresets.customTitle': 'Personalizat',
-    'subscriptionPresets.customDescription': 'Construiește-ți propria combinație de abonamente',
-    'subscriptionPresets.oneConnectedService': '1 serviciu conectat',
-    'subscriptionPresets.twoConnectedServices': '2 servicii conectate',
-    'subscriptionPresets.requires': 'Necesită {0}',
-    'subscriptionPresets.setUp': 'Configurează',
-    'subscriptionPresets.applying': 'Se aplică...',
-    'subscriptionPresets.casualApplied': 'Presetul Lejer a fost aplicat',
-    'subscriptionPresets.deepDiverApplied': 'Presetul Explorare profundă a fost aplicat',
-    'subscriptionPresets.applyFailed': 'Presetul nu a putut fi aplicat',
-    'subscriptionPresets.aiOnlyEnabled': 'Modul doar AI a fost activat',
-    'subscriptionPresets.aiOnlyEnableFailed': 'Modul doar AI nu a putut fi activat',
-    'cronPicker.preset': 'Programare predefinită',
-    'cronPicker.dailyMidnight': 'Zilnic (miezul nopții)',
-    'cronPicker.twiceAWeek': 'De două ori pe săptămână (lun + joi)',
-    'cronPicker.everyTwoWeeks': 'La fiecare 2 săptămâni (1 și 15)',
-    'cronPicker.monthlyFirst': 'Lunar (ziua 1)',
-    'cronPicker.standardHint': 'Cron standard: minut oră zi-din-lună lună zi-din-săptămână',
-    'recommendation.sourceSimilar': 'Asemănător',
-    'recommendation.streaming': 'Serviciu de streaming',
-    'libraryHealth.syncNow': 'Sincronizează acum',
-    'libraryHealth.snapshotTitle': 'Ultimul instantaneu',
-    'libraryHealth.lastSynced': 'Ultima sincronizare',
-    'libraryHealth.neverSynced': 'Nu a fost sincronizat niciodată.',
-    'libraryHealth.autoSyncEvery': 'Sincronizare automată la fiecare {0} h',
-    'settings.capDiscovery': 'Descoperire',
-    'settings.capImport': 'Importare',
-    'settings.capLibrarySync': 'Sync bibliotecă',
-    'settings.capPlaylistExport': 'Export listă',
-    'settings.capService': 'Serviciu',
-    'settings.capSubscriptions': 'Abonamente',
-    'systemHealth.pipeline': 'Rulare de descoperire',
-    'systemHealth.librarySync': 'Sincronizarea bibliotecii',
-    'search.experimental': 'În test',
-    'discoveryMode.field.artist': 'Nume artist',
-    'settings.tabs.systemHealth': 'Starea sistemului',
-    'settings.targetUpdated': 'Ținta a fost actualizată',
-    'settings.targetUpdateFailed': 'Actualizarea țintei a eșuat',
-    'settings.editTarget': 'Editează ținta',
-    'settings.saveChanges': 'Salvează modificările',
-    'settings.linkedTarget': 'Țintă asociată:',
-    'settings.totalWeight': 'în total:',
-    'settings.librarySyncInterval': 'Sync bibliotecă',
-    'settings.librarySyncIntervalHours': 'Interval de sincronizare (ore)',
-    'settings.librarySyncIntervalHelp':
-      'Controlează cât de des Digarr reîmprospătează datele de sincronizare ale bibliotecii și rerulează verificările automat.',
-    'settings.oidcSso': 'OIDC / SSO',
-    'settings.fieldModel': 'Model AI',
-    'settings.aiProviderTitle': 'Furnizor AI',
-    'settings.webhookTitle': 'Integrare webhook',
-    'settings.plexToken': 'Token Plex',
-    'settings.serviceConnectedToast': '{0} conectat',
-    'settings.serviceUnreachableToast': '{0} nu poate fi accesat',
-    'settings.serviceSettingsSavedToast': 'Setările pentru {0} au fost salvate',
-    'settings.serviceSettingsFailedToast': 'Setările pentru {0} nu au putut fi salvate',
-    'settings.spotifyAuthorizationFailed': 'Pornirea autorizării Spotify a eșuat',
-    'settings.spotifyDisconnectFailed': 'Deconectarea Spotify a eșuat',
-    'settings.spotifyLikedSongsFailed': 'Pornirea importului Spotify Liked Songs a eșuat',
-    'settings.spotifyPlaylistFailed': 'Pornirea importului playlistului a eșuat',
-    'settings.deezerFavoritesFailed': 'Pornirea importului favorite Deezer a eșuat',
-    'settings.deezerFollowedFailed': 'Pornirea importului artiștilor urmăriți pe Deezer a eșuat',
-    'settings.deezerDisconnectFailed': 'Deconectarea Deezer a eșuat',
-    'settings.deezerAuthorizationFailed': 'Pornirea autorizării Deezer a eșuat',
-    'settings.reverseProxyConfigHint':
-      'Folosește PROXY_AUTH_ENABLED și PROXY_AUTH_TRUSTED_PROXIES; proxy-ul tău trebuie să trimită X-Forwarded-User.',
-    'searchResultCard.addToQueue': 'Adaugă în coada de descoperire',
-    'searchResultCard.queuing': 'Se adaugă...',
-    'searchResultCard.queuedSuccess': '{0} a fost adăugat în coada de descoperire',
-    'searchResultCard.queuedFailure': '{0} nu a putut fi adăugat în coadă',
-    'searchResultCard.popularity': 'Popularitate {0}{1}',
-    'genres.artistCountSingular': 'artist',
-    'admin.restoreSummary': '{0} rânduri restaurate în {1} tabele',
-    'admin.restoreSummaryForce':
-      '{0} rânduri restaurate. Reintrodu credențialele pentru câmpurile criptate.',
-    'admin.encryptionMismatchMessage':
-      'Această copie de siguranță a fost creată cu o altă cheie de criptare.',
-    'admin.encryptionMismatchAffected': 'Aceste câmpuri trebuie completate din nou: {0}',
-    'admin.encryptionMismatchFallback': 'Unele câmpuri criptate pot necesita completare din nou.',
-    'admin.autoFixInProgress': 'Corectare automată în curs... ({0}/{1} corectate)',
-    'integration.playlist': 'Listă de redare',
-    'errorBoundary.title': 'A apărut o problemă',
-    'errorBoundary.retry': 'Încearcă din nou',
-    'errorBoundary.home': 'Mergi la început',
-    'connectionSuggestions.title': 'Sugestii pentru {0}',
-    'connectionSuggestions.close': 'Închide sugestiile',
-    'connectionSuggestions.gotIt': 'Am înțeles',
-    'connectionSuggestions.skipForNow': 'Sari peste momentan',
-    'connectionSuggestions.spotifyDiscoverWeekly': 'Configurează un abonament Discover Weekly',
-    'connectionSuggestions.spotifyReleaseRadar': 'Activează un abonament Release Radar',
-    'connectionSuggestions.lastfmTopGenreTags': 'Abonează-te la principalele tale etichete de gen',
-    'connectionSuggestions.listenbrainzWeeklyJams': 'Activează abonamentul Weekly Jams',
-    'connectionSuggestions.listenbrainzWeeklyExploration': 'Activează Weekly Exploration',
-    'connectionSuggestions.weeklyDigestPlaylist': 'Creează o listă săptămânală Digarr Digest',
-    'keyboardShortcuts.title': 'Scurtături de tastatură',
-    'keyboardShortcuts.close': 'Închide scurtăturile',
-    'keyboardShortcuts.nextCard': 'Cardul următor',
-    'keyboardShortcuts.previousCard': 'Cardul anterior',
-    'keyboardShortcuts.approveFocusedCard': 'Aprobă cardul activ',
-    'keyboardShortcuts.rejectFocusedCard': 'Respinge cardul activ',
-    'keyboardShortcuts.toggleDetailView': 'Deschide / închide vizualizarea detaliată',
-    'keyboardShortcuts.toggleShortcuts': 'Arată / ascunde scurtăturile',
-    'keyboardShortcuts.previousCardStack': 'Cardul anterior (vizualizare stivă)',
-    'keyboardShortcuts.nextCardStack': 'Cardul următor (vizualizare stivă)',
-    'keyboardShortcuts.disabledWhileTyping':
-      'Scurtăturile sunt dezactivate cât timp un câmp text are focus.',
-    'setup.model': 'Model AI',
-    'albumPicker.title': 'Selectează albume',
-    'albumPicker.loadFailed': 'Încărcarea albumelor a eșuat',
-    'albumPicker.selectAll': 'Toate',
-    'albumPicker.selectNone': 'Niciunul',
-    'albumPicker.selectedCount': '{0} selectate',
-    'albumPicker.aiPick': 'Alegere AI',
-    'albumPicker.monitor': 'Monitorizează',
-    'albumPicker.monitorSingular': 'Monitorizează {0} album',
-    'albumPicker.monitorPlural': 'Monitorizează {0} albume',
-    'approveDialog.title': 'Setări Lidarr pentru acest artist',
-    'approveDialog.loadFailed':
-      'Încărcarea profilurilor Lidarr a eșuat. Verifică conexiunea la Lidarr.',
-    'approveDialog.qualityProfile': 'Profil de calitate',
-    'approveDialog.metadataProfile': 'Profil de metadate',
-    'approveDialog.rootFolder': 'Dosar rădăcină',
-    'cardStack.counter': '{0} din {1}',
-    'cardStack.swipeHint': 'Glisează dreapta pentru a aproba, stânga pentru a respinge',
-    'serviceCard.testConnection': 'Testează conexiunea',
-    'preview.noPreviewAvailable': 'Nu există previzualizare pentru acest artist',
-    'preview.playbackBlocked': 'Redare blocată de browser - încearcă să dai clic din nou',
-    'errors.auth.invalidCredentials': 'Credențiale nevalide',
-    'errors.auth.notAuthenticated': 'Neautentificat',
-    'errors.auth.passwordIncorrect': 'Parola actuală este incorectă',
-    'errors.auth.usernameTaken': 'Numele de utilizator este deja folosit',
-    'errors.target.notFound': 'Destinația nu a fost găsită',
-    'errors.pipeline.alreadyRunning': 'O execuție a pipeline-ului este deja în desfășurare',
-    'errors.recommendation.notFound': 'Recomandarea nu a fost găsită',
-    'errors.subscription.notFound': 'Abonamentul nu a fost găsit',
-    'errors.playlist.notFound': 'Lista de redare nu a fost găsită',
-  },
-  tr: {
-    ...REVIEW_GATE_TRANSLATIONS.tr,
-    'common.importing': 'İçe aktarılıyor...',
-    'common.current': 'Geçerli',
-    'common.custom': 'Özel',
-    'common.enabled': 'Etkin',
-    'common.unauthorized': 'Yetkisiz',
-    'common.adminAccessRequired': 'Yönetici erişimi gerekiyor',
-    'jobHistory.pipeline': 'Keşif çalıştırması',
-    'jobHistory.librarySync': 'Kütüphane eşitleme',
-    'playlistForm.dialogCreate': 'Çalma listesi oluştur',
-    'playlistForm.dialogEdit': 'Çalma listesini düzenle',
-    'playlistForm.titleCreate': 'Çalma listesi oluştur',
-    'playlistForm.titleEdit': 'Çalma listesini düzenle',
-    'playlistForm.namePlaceholder': 'Haftalık karışımım',
-    'playlistForm.strategy': 'Strateji',
-    'playlistForm.strategyWeeklyDigestDesc': 'Son keşiflerinden özenle seçilmiş parçalar',
-    'playlistForm.strategyGenreFocusDesc': 'Belirli bir türe derinlemesine dal',
-    'playlistForm.strategyMoodMixDesc': 'Belirli bir havaya uyan parçalar',
-    'playlistForm.strategyRediscoverDesc': 'Unutmuş olabileceğin sanatçıları yeniden ortaya çıkar',
-    'playlistForm.mood': 'Ruh hali',
-    'playlistForm.moodPlaceholder': 'gece yolculuğu',
-    'playlistForm.trackCount': 'Parça sayısı',
-    'playlistForm.scheduleAutomaticGeneration': 'Otomatik oluşturmayı zamanla',
-    'playlistForm.nameRequired': 'Ad zorunludur',
-    'playlistForm.genreRequired': 'Tür Odağı stratejisi için tür zorunludur',
-    'playlistForm.moodRequired': 'Mood Mix stratejisi için ruh hali zorunludur',
-    'subscriptionForm.dialogCreate': 'Abonelik oluştur',
-    'subscriptionForm.dialogEdit': 'Aboneliği düzenle',
-    'subscriptionForm.titleCreate': 'Abonelik oluştur',
-    'subscriptionForm.titleEdit': 'Aboneliği düzenle',
-    'subscriptionForm.name': 'Ad',
-    'subscriptionForm.namePlaceholder': 'Yeni Shoegaze Weekly',
-    'subscriptionForm.nameRequired': 'Ad zorunludur',
-    'subscriptionForm.source': 'Kaynak',
-    'subscriptionForm.sources': 'Kaynaklar',
-    'subscriptionForm.sourceType': 'Kaynak türü',
-    'subscriptionForm.sourceTypeGenre': 'Müzik türü',
-    'subscriptionForm.sourceTypeGenreDesc': 'Belirli bir tür içinde sanatçılar keşfet',
-    'subscriptionForm.sourceTypeSimilar': 'Benzer sanatçılar',
-    'subscriptionForm.sourceTypeSimilarDesc': 'Favorilerine benzeyen sanatçıları bul',
-    'subscriptionForm.sourceTypeDiscoveryModeDesc': 'Sanatçı üretmek için bir keşif modu kullan',
-    'subscriptionForm.discoveryMode': 'Keşif modu',
-    'subscriptionForm.selectDiscoveryMode': 'Yapılandırmak için bir keşif modu seç.',
-    'subscriptionForm.discoveryModeSettingsRequired': 'Keşif modu ayarları zorunludur',
-    'subscriptionForm.genre': 'Müzik türü',
-    'subscriptionForm.genrePlaceholder': 'shoegaze',
-    'subscriptionForm.seedArtistsPlaceholder': 'Radiohead, Portishead, Massive Attack',
-    'subscriptionForm.genreRequired': 'Tür zorunludur',
-    'subscriptionForm.seedArtists': 'Başlangıç sanatçıları',
-    'subscriptionForm.seedArtistsHelp': 'Virgülle ayrılmış sanatçı adları',
-    'subscriptionForm.seedArtistsRequired': 'En az bir başlangıç sanatçısı gereklidir',
-    'subscriptionForm.configureInSettings': 'Ayarlar > Bağlantılar bölümünden yapılandır',
-    'subscriptionForm.notConfigured': 'yapılandırılmadı',
-    'subscriptionForm.selectAtLeastOneSource': 'En az bir kaynak seç',
-    'subscriptionForm.sourceLockedHelp':
-      'Bu abonelik için kaynak türü ve yapılandırma değiştirilemez.',
-    'subscriptionForm.maxArtistsPerRun': 'Çalıştırma başına en fazla sanatçı',
-    'subscriptionForm.scoringPreset': 'Puanlama ön ayarı',
-    'subscriptionForm.scoringPresetDefault': 'Varsayılan',
-    'subscriptionForm.scoringPresetGenre': 'Türe göre optimize edildi',
-    'subscriptions.spotifyNotConnected': 'Spotify bağlı değil',
-    'subscriptionPresets.getStarted': 'Bir ön ayarla başla',
-    'subscriptionPresets.intro':
-      'Abonelikleri otomatik oluşturmak için bir ön ayar seç ya da kendi yapılandırmanı kur.',
-    'subscriptionPresets.casualTitle': 'Rahat',
-    'subscriptionPresets.casualDescription': 'Az gürültülü haftalık öneriler',
-    'subscriptionPresets.deepDiverTitle': 'Derin keşif',
-    'subscriptionPresets.deepDiverDescription':
-      'Birden fazla kaynaktan geniş keşif ve günlük güncellemeler',
-    'subscriptionPresets.discoveryOnlyTitle': 'Sadece keşif',
-    'subscriptionPresets.discoveryOnlyDescription': 'Yalnızca pipeline, harici akış yok',
-    'subscriptionPresets.customTitle': 'Özel',
-    'subscriptionPresets.customDescription': 'Kendi abonelik karışımını oluştur',
-    'subscriptionPresets.oneConnectedService': '1 bağlı hizmet',
-    'subscriptionPresets.twoConnectedServices': '2 bağlı hizmet',
-    'subscriptionPresets.requires': '{0} gerekir',
-    'subscriptionPresets.setUp': 'Kur',
-    'subscriptionPresets.applying': 'Uygulanıyor...',
-    'subscriptionPresets.casualApplied': 'Rahat ön ayarı uygulandı',
-    'subscriptionPresets.deepDiverApplied': 'Derin keşif ön ayarı uygulandı',
-    'subscriptionPresets.applyFailed': 'Ön ayar uygulanamadı',
-    'subscriptionPresets.aiOnlyEnabled': 'Yalnızca AI modu etkinleştirildi',
-    'subscriptionPresets.aiOnlyEnableFailed': 'Yalnızca AI modu etkinleştirilemedi',
-    'cronPicker.preset': 'Ön ayar',
-    'cronPicker.dailyMidnight': 'Her gün (gece yarısı)',
-    'cronPicker.twiceAWeek': 'Haftada iki kez (Pzt + Per)',
-    'cronPicker.everyTwoWeeks': 'Her 2 haftada bir (1 ve 15)',
-    'cronPicker.monthlyFirst': 'Aylık (1. gün)',
-    'cronPicker.standardHint': 'Standart cron: dakika saat ayın-günü ay haftanın-günü',
-    'libraryHealth.syncNow': 'Şimdi eşitle',
-    'libraryHealth.snapshotTitle': 'Son anlık görüntü',
-    'libraryHealth.lastSynced': 'Son eşitleme',
-    'libraryHealth.neverSynced': 'Henüz hiç eşitlenmedi.',
-    'libraryHealth.autoSyncEvery': 'Her {0} saatte otomatik eşitleme',
-    'systemHealth.pipeline': 'Keşif çalıştırması',
-    'systemHealth.librarySync': 'Kütüphane eşitleme',
-    'discoveryMode.field.feed': 'Besleme kaynağı',
-    'settings.tabs.systemHealth': 'Sistem durumu',
-    'settings.targetUpdated': 'Hedef güncellendi',
-    'settings.targetUpdateFailed': 'Hedef güncellenemedi',
-    'settings.editTarget': 'Hedefi düzenle',
-    'settings.saveChanges': 'Değişiklikleri kaydet',
-    'settings.linkedTarget': 'Bağlı hedef:',
-    'settings.librarySyncInterval': 'Kütüphane eşitleme',
-    'settings.librarySyncIntervalHours': 'Eşitleme aralığı (saat)',
-    'settings.librarySyncIntervalHelp':
-      'Digarr ın kütüphane eşitleme verilerini ne sıklıkla yenileyeceğini ve denetimleri otomatik olarak yeniden çalıştıracağını belirler.',
-    'settings.oidcSso': 'OIDC / SSO',
-    'settings.fieldModel': 'Model seçimi',
-    'settings.fieldClientId': 'İstemci kimliği',
-    'settings.fieldClientSecret': 'İstemci gizlisi',
-    'settings.fieldWebhookUrl': 'Webhook adresi',
-    'settings.aiProviderTitle': 'AI sağlayıcısı',
-    'settings.webhookTitle': 'Webhook entegrasyonu',
-    'settings.plexToken': 'Plex belirteci',
-    'settings.serviceConnectedToast': '{0} bağlandı',
-    'settings.serviceUnreachableToast': '{0} erişilemiyor',
-    'settings.serviceSettingsSavedToast': '{0} ayarları kaydedildi',
-    'settings.serviceSettingsFailedToast': '{0} ayarları kaydedilemedi',
-    'settings.spotifyAuthorizationFailed': 'Spotify yetkilendirmesi başlatılamadı',
-    'settings.spotifyDisconnectFailed': 'Spotify bağlantısı kesilemedi',
-    'settings.spotifyLikedSongsFailed': 'Spotify Liked Songs içe aktarma başlatılamadı',
-    'settings.spotifyPlaylistFailed': 'Çalma listesi içe aktarma başlatılamadı',
-    'settings.deezerFavoritesFailed': 'Deezer favorileri içe aktarma başlatılamadı',
-    'settings.deezerFollowedFailed': 'Takip edilen Deezer sanatçıları içe aktarma başlatılamadı',
-    'settings.deezerDisconnectFailed': 'Deezer bağlantısı kesilemedi',
-    'settings.deezerAuthorizationFailed': 'Deezer yetkilendirmesi başlatılamadı',
-    'settings.reverseProxyConfigHint':
-      'PROXY_AUTH_ENABLED ve PROXY_AUTH_TRUSTED_PROXIES kullanın; proxy nizin X-Forwarded-User göndermesi gerekir.',
-    'searchResultCard.addToQueue': 'Keşif kuyruğuna ekle',
-    'searchResultCard.queuing': 'Ekleniyor...',
-    'searchResultCard.queuedSuccess': '{0} keşif kuyruğuna eklendi',
-    'searchResultCard.queuedFailure': '{0} kuyruğa eklenemedi',
-    'searchResultCard.popularity': 'Popülerlik {0}{1}',
-    'admin.restoreSummary': '{1} tabloda {0} satır geri yüklendi',
-    'admin.restoreSummaryForce':
-      '{0} satır geri yüklendi. Şifrelenmiş alanlar için kimlik bilgilerini yeniden girin.',
-    'admin.encryptionMismatchMessage': 'Bu yedek farklı bir şifreleme anahtarıyla oluşturulmuş.',
-    'admin.encryptionMismatchAffected': 'Bu alanların yeniden girilmesi gerekecek: {0}',
-    'admin.encryptionMismatchFallback': 'Bazı şifrelenmiş alanların yeniden girilmesi gerekebilir.',
-    'admin.autoFixInProgress': 'Otomatik düzeltme sürüyor... ({0}/{1} düzeltildi)',
-    'errorBoundary.title': 'Bir sorun oluştu',
-    'errorBoundary.retry': 'Tekrar dene',
-    'errorBoundary.home': 'Ana sayfaya git',
-    'connectionSuggestions.title': '{0} için öneriler',
-    'connectionSuggestions.close': 'Önerileri kapat',
-    'connectionSuggestions.gotIt': 'Tamam',
-    'connectionSuggestions.skipForNow': 'Şimdilik geç',
-    'connectionSuggestions.spotifyDiscoverWeekly': 'Bir Discover Weekly aboneliği ayarla',
-    'connectionSuggestions.spotifyReleaseRadar': 'Bir Release Radar aboneliğini etkinleştir',
-    'connectionSuggestions.lastfmTopGenreTags': 'En önemli tür etiketlerine abone ol',
-    'connectionSuggestions.listenbrainzWeeklyJams': 'Weekly Jams aboneliğini etkinleştir',
-    'connectionSuggestions.listenbrainzWeeklyExploration': 'Weekly Exploration ı etkinleştir',
-    'connectionSuggestions.weeklyDigestPlaylist':
-      'Haftalık bir Digarr Digest çalma listesi oluştur',
-    'keyboardShortcuts.title': 'Klavye kısayolları',
-    'keyboardShortcuts.close': 'Kısayolları kapat',
-    'keyboardShortcuts.nextCard': 'Sonraki kart',
-    'keyboardShortcuts.previousCard': 'Önceki kart',
-    'keyboardShortcuts.approveFocusedCard': 'Etkin kartı onayla',
-    'keyboardShortcuts.rejectFocusedCard': 'Etkin kartı reddet',
-    'keyboardShortcuts.toggleDetailView': 'Detay görünümünü aç / kapat',
-    'keyboardShortcuts.toggleShortcuts': 'Kısayolları göster / gizle',
-    'keyboardShortcuts.previousCardStack': 'Önceki kart (yığın görünümü)',
-    'keyboardShortcuts.nextCardStack': 'Sonraki kart (yığın görünümü)',
-    'keyboardShortcuts.disabledWhileTyping':
-      'Bir metin alanı odaktayken kısayollar devre dışı kalır.',
-    'albumPicker.title': 'Albümleri seç',
-    'albumPicker.loadFailed': 'Albümler yüklenemedi',
-    'albumPicker.selectAll': 'Tümü',
-    'albumPicker.selectNone': 'Hiçbiri',
-    'albumPicker.selectedCount': '{0} seçili',
-    'albumPicker.aiPick': 'AI seçimi',
-    'albumPicker.monitor': 'İzle',
-    'albumPicker.monitorSingular': '{0} albümü izle',
-    'albumPicker.monitorPlural': '{0} albümü izle',
-    'approveDialog.title': 'Bu sanatçı için Lidarr ayarları',
-    'approveDialog.loadFailed': 'Lidarr profilleri yüklenemedi. Lidarr bağlantınızı kontrol edin.',
-    'approveDialog.qualityProfile': 'Kalite Profili',
-    'approveDialog.metadataProfile': 'Meta Veri Profili',
-    'approveDialog.rootFolder': 'Kök Klasör',
-    'cardStack.counter': '{0} / {1}',
-    'cardStack.swipeHint': 'Onaylamak için sağa, reddetmek için sola kaydırın',
-    'serviceCard.testConnection': 'Bağlantıyı Test Et',
-    'preview.noPreviewAvailable': 'Bu sanatçı için önizleme yok',
-    'preview.playbackBlocked': 'Oynatma tarayıcı tarafından engellendi - tekrar tıklamayı deneyin',
-    'errors.auth.invalidCredentials': 'Geçersiz kimlik bilgileri',
-    'errors.auth.notAuthenticated': 'Kimlik doğrulanmadı',
-    'errors.auth.passwordIncorrect': 'Mevcut parola yanlış',
-    'errors.auth.usernameTaken': 'Kullanıcı adı zaten alınmış',
-    'errors.target.notFound': 'Hedef bulunamadı',
-    'errors.pipeline.alreadyRunning': 'Zaten bir pipeline çalışması devam ediyor',
-    'errors.recommendation.notFound': 'Öneri bulunamadı',
-    'errors.subscription.notFound': 'Abonelik bulunamadı',
-    'errors.playlist.notFound': 'Oynatma listesi bulunamadı',
-  },
-  uk: {
-    ...REVIEW_GATE_TRANSLATIONS.uk,
-    'setup.embyUrl': 'Адреса Emby',
-    'common.importing': 'Імпорт...',
-    'common.current': 'Поточний',
-    'common.custom': 'Власний',
-    'common.enabled': 'Увімкнено',
-    'common.unauthorized': 'Немає авторизації',
-    'common.adminAccessRequired': 'Потрібен доступ адміністратора',
-    'jobHistory.pipeline': 'Запуск пошуку',
-    'jobHistory.librarySync': 'Синхронізація бібліотеки',
-    'playlistForm.dialogCreate': 'Створити плейлист',
-    'playlistForm.dialogEdit': 'Редагувати плейлист',
-    'playlistForm.titleCreate': 'Створити плейлист',
-    'playlistForm.titleEdit': 'Редагувати плейлист',
-    'playlistForm.namePlaceholder': 'Мій тижневий мікс',
-    'playlistForm.strategy': 'Стратегія',
-    'playlistForm.strategyWeeklyDigestDesc': 'Добірка з ваших недавніх відкриттів',
-    'playlistForm.strategyGenreFocusDesc': 'Глибоке занурення в конкретний жанр',
-    'playlistForm.strategyMoodMixDesc': 'Треки під певний настрій',
-    'playlistForm.strategyRediscoverDesc': 'Поверніть артистів, про яких могли забути',
-    'playlistForm.mood': 'Настрій',
-    'playlistForm.moodPlaceholder': 'нічна поїздка',
-    'playlistForm.trackCount': 'Кількість треків',
-    'playlistForm.scheduleAutomaticGeneration': 'Запланувати автоматичне створення',
-    'playlistForm.nameRequired': 'Назва обов язкова',
-    'playlistForm.genreRequired': 'Жанр обов язковий для стратегії Фокус на жанрі',
-    'playlistForm.moodRequired': 'Настрій обов язковий для стратегії Mood Mix',
-    'subscriptionForm.dialogCreate': 'Створити підписку',
-    'subscriptionForm.dialogEdit': 'Редагувати підписку',
-    'subscriptionForm.titleCreate': 'Створити підписку',
-    'subscriptionForm.titleEdit': 'Редагувати підписку',
-    'subscriptionForm.name': 'Назва',
-    'subscriptionForm.namePlaceholder': 'Нова Shoegaze Weekly',
-    'subscriptionForm.nameRequired': 'Назва обов язкова',
-    'subscriptionForm.source': 'Джерело',
-    'subscriptionForm.sources': 'Джерела',
-    'subscriptionForm.sourceType': 'Тип джерела',
-    'subscriptionForm.sourceTypeGenre': 'Музичний жанр',
-    'subscriptionForm.sourceTypeGenreDesc': 'Відкривайте артистів у певному жанрі',
-    'subscriptionForm.sourceTypeSimilar': 'Схожі артисти',
-    'subscriptionForm.sourceTypeSimilarDesc': 'Знайдіть артистів, схожих на ваших улюблених',
-    'subscriptionForm.sourceTypeDiscoveryModeDesc':
-      'Використовуйте режим пошуку, щоб генерувати артистів',
-    'subscriptionForm.discoveryMode': 'Режим пошуку',
-    'subscriptionForm.selectDiscoveryMode': 'Виберіть режим пошуку для налаштування.',
-    'subscriptionForm.discoveryModeSettingsRequired': 'Налаштування режиму пошуку є обов язковими',
-    'subscriptionForm.genre': 'Музичний жанр',
-    'subscriptionForm.genrePlaceholder': 'shoegaze',
-    'subscriptionForm.seedArtistsPlaceholder': 'Radiohead, Portishead, Massive Attack',
-    'subscriptionForm.genreRequired': 'Жанр обов язковий',
-    'subscriptionForm.seedArtists': 'Початкові артисти',
-    'subscriptionForm.seedArtistsHelp': 'Імена артистів, розділені комами',
-    'subscriptionForm.seedArtistsRequired': 'Потрібен щонайменше один початковий артист',
-    'subscriptionForm.configureInSettings': 'Налаштуйте в Розділі Налаштування > Підключення',
-    'subscriptionForm.notConfigured': 'не налаштовано',
-    'subscriptionForm.selectAtLeastOneSource': 'Виберіть принаймні одне джерело',
-    'subscriptionForm.sourceLockedHelp':
-      'Тип джерела та конфігурацію не можна змінити для цієї підписки.',
-    'subscriptionForm.maxArtistsPerRun': 'Максимум артистів за запуск',
-    'subscriptionForm.scoringPreset': 'Шаблон оцінювання',
-    'subscriptionForm.scoringPresetDefault': 'Типовий',
-    'subscriptionForm.scoringPresetGenre': 'Оптимізовано для жанру',
-    'subscriptions.spotifyNotConnected': 'Spotify не підключено',
-    'subscriptionPresets.getStarted': 'Почніть із пресета',
-    'subscriptionPresets.intro':
-      'Виберіть пресет, щоб автоматично створювати підписки, або зберіть власну конфігурацію.',
-    'subscriptionPresets.casualTitle': 'Легкий режим',
-    'subscriptionPresets.casualDescription': 'Щотижневі рекомендації з мінімумом шуму',
-    'subscriptionPresets.deepDiverTitle': 'Глибоке занурення',
-    'subscriptionPresets.deepDiverDescription':
-      'Широке відкриття з кількох джерел і щоденними оновленнями',
-    'subscriptionPresets.discoveryOnlyTitle': 'Лише пошук',
-    'subscriptionPresets.discoveryOnlyDescription': 'Лише pipeline, без зовнішніх стрічок',
-    'subscriptionPresets.customTitle': 'Власний',
-    'subscriptionPresets.customDescription': 'Створіть власну комбінацію підписок',
-    'subscriptionPresets.oneConnectedService': '1 підключений сервіс',
-    'subscriptionPresets.twoConnectedServices': '2 підключені сервіси',
-    'subscriptionPresets.requires': 'Потрібно: {0}',
-    'subscriptionPresets.setUp': 'Налаштувати',
-    'subscriptionPresets.applying': 'Застосування...',
-    'subscriptionPresets.casualApplied': 'Пресет Легкий режим застосовано',
-    'subscriptionPresets.deepDiverApplied': 'Пресет Глибоке занурення застосовано',
-    'subscriptionPresets.applyFailed': 'Не вдалося застосувати пресет',
-    'subscriptionPresets.aiOnlyEnabled': 'Режим лише AI увімкнено',
-    'subscriptionPresets.aiOnlyEnableFailed': 'Не вдалося увімкнути режим лише AI',
-    'cronPicker.preset': 'Пресет',
-    'cronPicker.dailyMidnight': 'Щодня (опівночі)',
-    'cronPicker.twiceAWeek': 'Двічі на тиждень (пн + чт)',
-    'cronPicker.everyTwoWeeks': 'Кожні 2 тижні (1 і 15)',
-    'cronPicker.monthlyFirst': 'Щомісяця (день 1)',
-    'cronPicker.standardHint': 'Стандартний cron: хвилина година день-місяця місяць день-тижня',
-    'libraryHealth.syncNow': 'Синхронізувати зараз',
-    'libraryHealth.snapshotTitle': 'Останній знімок',
-    'libraryHealth.lastSynced': 'Остання синхронізація',
-    'libraryHealth.neverSynced': 'Ще не синхронізовано.',
-    'libraryHealth.autoSyncEvery': 'Автосинхронізація кожні {0} год',
-    'settings.capDiscovery': 'Пошук',
-    'settings.capImport': 'Імпорт',
-    'settings.capLibrarySync': 'Синхронізація бібліотеки',
-    'settings.capPlaylistExport': 'Експорт плейлистів',
-    'settings.capService': 'Сервіс',
-    'settings.capSubscriptions': 'Підписки',
-    'settings.connectDeezer': 'Підключити Deezer',
-    'settings.deezerConnectHelp':
-      'Підключіть свій обліковий запис Deezer, щоб імпортувати улюблене, артистів, на яких ви підписані, і дані плейлистів.',
-    'settings.importFavorites': 'Імпортувати улюблене',
-    'settings.importFollowed': 'Імпортувати підписки',
-    'settings.integrationCapabilities': 'Можливості інтеграції',
-    'systemHealth.librarySync': 'Синхронізація бібліотеки',
-    'settings.tabs.systemHealth': 'Стан системи',
-    'settings.targetUpdated': 'Ціль оновлено',
-    'settings.targetUpdateFailed': 'Не вдалося оновити ціль',
-    'settings.editTarget': 'Редагувати ціль',
-    'settings.saveChanges': 'Зберегти зміни',
-    'settings.linkedTarget': 'Пов язана ціль:',
-    'settings.librarySyncInterval': 'Синхронізація бібліотеки',
-    'settings.librarySyncIntervalHours': 'Інтервал синхронізації (години)',
-    'settings.librarySyncIntervalHelp':
-      'Визначає, як часто Digarr оновлює дані синхронізації бібліотеки та автоматично повторно запускає перевірки.',
-    'settings.oidcSso': 'OIDC / SSO',
-    'settings.fieldClientId': 'ID клієнта',
-    'settings.fieldClientSecret': 'Секрет клієнта',
-    'settings.aiProviderTitle': 'AI-провайдер',
-    'settings.webhookTitle': 'Webhook-інтеграція',
-    'settings.plexToken': 'Токен Plex',
-    'settings.serviceConnectedToast': '{0} підключено',
-    'settings.serviceUnreachableToast': 'Не вдалося дістатися до {0}',
-    'settings.serviceSettingsSavedToast': 'Налаштування {0} збережено',
-    'settings.serviceSettingsFailedToast': 'Не вдалося зберегти налаштування {0}',
-    'settings.spotifyAuthorizationFailed': 'Не вдалося запустити авторизацію Spotify',
-    'settings.spotifyDisconnectFailed': 'Не вдалося відключити Spotify',
-    'settings.spotifyLikedSongsFailed': 'Не вдалося запустити імпорт Spotify Liked Songs',
-    'settings.spotifyPlaylistFailed': 'Не вдалося запустити імпорт плейлиста',
-    'settings.deezerFavoritesFailed': 'Не вдалося запустити імпорт Deezer Favorites',
-    'settings.deezerFollowedFailed': 'Не вдалося запустити імпорт Deezer Followed',
-    'settings.deezerDisconnectFailed': 'Не вдалося відключити Deezer',
-    'settings.deezerAuthorizationFailed': 'Не вдалося запустити авторизацію Deezer',
-    'settings.reverseProxyConfigHint':
-      'Використовуйте PROXY_AUTH_ENABLED і PROXY_AUTH_TRUSTED_PROXIES; ваш proxy має надсилати X-Forwarded-User.',
-    'searchResultCard.addToQueue': 'Додати до черги пошуку',
-    'searchResultCard.queuing': 'Додається...',
-    'searchResultCard.queuedSuccess': '{0} додано до черги пошуку',
-    'searchResultCard.queuedFailure': '{0} не вдалося додати до черги',
-    'searchResultCard.popularity': 'Популярність {0}{1}',
-    'admin.restoreSummary': '{0} рядків відновлено у {1} таблицях',
-    'admin.restoreSummaryForce':
-      '{0} рядків відновлено. Повторно введіть облікові дані для зашифрованих полів.',
-    'admin.encryptionMismatchMessage': 'Цю резервну копію створено з іншим ключем шифрування.',
-    'admin.encryptionMismatchAffected': 'Ці поля потрібно буде ввести знову: {0}',
-    'admin.encryptionMismatchFallback':
-      'Деякі зашифровані поля можуть вимагати повторного введення.',
-    'admin.autoFixInProgress': 'Автовиправлення триває... ({0}/{1} виправлено)',
-    'errorBoundary.title': 'Сталася проблема',
-    'errorBoundary.retry': 'Спробувати знову',
-    'errorBoundary.home': 'Перейти на головну',
-    'connectionSuggestions.title': 'Пропозиції для {0}',
-    'connectionSuggestions.close': 'Закрити пропозиції',
-    'connectionSuggestions.gotIt': 'Зрозуміло',
-    'connectionSuggestions.skipForNow': 'Пропустити поки що',
-    'connectionSuggestions.spotifyDiscoverWeekly': 'Налаштувати підписку Discover Weekly',
-    'connectionSuggestions.spotifyReleaseRadar': 'Увімкнути підписку Release Radar',
-    'connectionSuggestions.lastfmTopGenreTags': 'Підписатися на ваші основні теги жанрів',
-    'connectionSuggestions.listenbrainzWeeklyJams': 'Увімкнути підписку Weekly Jams',
-    'connectionSuggestions.listenbrainzWeeklyExploration': 'Увімкнути Weekly Exploration',
-    'connectionSuggestions.weeklyDigestPlaylist': 'Створити щотижневий плейлист Digarr Digest',
-    'keyboardShortcuts.title': 'Клавіатурні скорочення',
-    'keyboardShortcuts.close': 'Закрити скорочення',
-    'keyboardShortcuts.nextCard': 'Наступна картка',
-    'keyboardShortcuts.previousCard': 'Попередня картка',
-    'keyboardShortcuts.approveFocusedCard': 'Схвалити активну картку',
-    'keyboardShortcuts.rejectFocusedCard': 'Відхилити активну картку',
-    'keyboardShortcuts.toggleDetailView': 'Відкрити / закрити детальний перегляд',
-    'keyboardShortcuts.toggleShortcuts': 'Показати / приховати скорочення',
-    'keyboardShortcuts.previousCardStack': 'Попередня картка (вигляд стосу)',
-    'keyboardShortcuts.nextCardStack': 'Наступна картка (вигляд стосу)',
-    'keyboardShortcuts.disabledWhileTyping': 'Скорочення вимикаються, коли у фокусі текстове поле.',
-    'albumPicker.title': 'Вибрати альбоми',
-    'albumPicker.loadFailed': 'Не вдалося завантажити альбоми',
-    'albumPicker.selectAll': 'Усі',
-    'albumPicker.selectNone': 'Жодного',
-    'albumPicker.selectedCount': 'Вибрано: {0}',
-    'albumPicker.aiPick': 'Вибір AI',
-    'albumPicker.monitor': 'Відстежувати',
-    'albumPicker.monitorSingular': 'Відстежувати {0} альбом',
-    'albumPicker.monitorPlural': 'Відстежувати {0} альбомів',
-    'approveDialog.title': 'Налаштування Lidarr для цього виконавця',
-    'approveDialog.loadFailed':
-      'Не вдалося завантажити профілі Lidarr. Перевірте підключення до Lidarr.',
-    'approveDialog.qualityProfile': 'Профіль якості',
-    'approveDialog.metadataProfile': 'Профіль метаданих',
-    'approveDialog.rootFolder': 'Коренева папка',
-    'cardStack.counter': '{0} з {1}',
-    'cardStack.swipeHint': 'Проведіть праворуч, щоб схвалити, ліворуч - щоб відхилити',
-    'serviceCard.testConnection': 'Перевірити зʼєднання',
-    'preview.noPreviewAvailable': 'Попередній перегляд для цього виконавця недоступний',
-    'preview.playbackBlocked': 'Відтворення заблоковано браузером - спробуйте натиснути ще раз',
-    'errors.auth.invalidCredentials': 'Недійсні облікові дані',
-    'errors.auth.notAuthenticated': 'Не автентифіковано',
-    'errors.auth.passwordIncorrect': 'Поточний пароль неправильний',
-    'errors.auth.usernameTaken': 'Імʼя користувача вже зайняте',
-    'errors.target.notFound': 'Ціль не знайдено',
-    'errors.pipeline.alreadyRunning': 'Запуск конвеєра вже триває',
-    'errors.recommendation.notFound': 'Рекомендацію не знайдено',
-    'errors.subscription.notFound': 'Підписку не знайдено',
-    'errors.playlist.notFound': 'Список відтворення не знайдено',
-  },
-  pl: {
-    ...REVIEW_GATE_TRANSLATIONS.pl,
-    'common.unauthorized': 'Brak autoryzacji',
-    'common.adminAccessRequired': 'Wymagany jest dostep administratora',
-    'common.current': 'Biezacy',
-    'common.custom': 'Niestandardowy',
-    'common.enabled': 'Wlaczone',
-    'common.importing': 'Importowanie...',
-    'libraryHealth.syncNow': 'Synchronizuj teraz',
-    'libraryHealth.snapshotTitle': 'Najnowsza migawka',
-    'libraryHealth.lastSynced': 'Ostatnia synchronizacja',
-    'libraryHealth.neverSynced': 'Nigdy nie synchronizowano.',
-    'libraryHealth.autoSyncEvery': 'Automatyczna synchronizacja co {0} h',
-    'jobHistory.librarySync': 'Synchronizacja biblioteki',
-    'errorBoundary.title': 'Cos poszlo nie tak',
-    'errorBoundary.retry': 'Sprobuj ponownie',
-    'errorBoundary.home': 'Przejdz do strony glownej',
-    'connectionSuggestions.title': 'Sugestie dla {0}',
-    'connectionSuggestions.close': 'Zamknij sugestie',
-    'connectionSuggestions.gotIt': 'Rozumiem',
-    'connectionSuggestions.skipForNow': 'Pomin na razie',
-    'connectionSuggestions.spotifyDiscoverWeekly': 'Skonfiguruj subskrypcje Discover Weekly',
-    'connectionSuggestions.spotifyReleaseRadar': 'Wlacz subskrypcje Release Radar',
-    'connectionSuggestions.lastfmTopGenreTags': 'Subskrybuj swoje najwazniejsze tagi gatunkowe',
-    'connectionSuggestions.listenbrainzWeeklyJams': 'Wlacz subskrypcje Weekly Jams',
-    'connectionSuggestions.listenbrainzWeeklyExploration': 'Wlacz Weekly Exploration',
-    'connectionSuggestions.weeklyDigestPlaylist': 'Utworz cotygodniowa playliste Digarr Digest',
-    'keyboardShortcuts.title': 'Skroty klawiaturowe',
-    'keyboardShortcuts.close': 'Zamknij skroty',
-    'keyboardShortcuts.nextCard': 'Nastepna karta',
-    'keyboardShortcuts.previousCard': 'Poprzednia karta',
-    'keyboardShortcuts.approveFocusedCard': 'Zatwierdz aktywna karte',
-    'keyboardShortcuts.rejectFocusedCard': 'Odrzuc aktywna karte',
-    'keyboardShortcuts.toggleDetailView': 'Otworz / zamknij widok szczegolow',
-    'keyboardShortcuts.toggleShortcuts': 'Pokaz / ukryj skroty',
-    'keyboardShortcuts.previousCardStack': 'Poprzednia karta (widok stosu)',
-    'keyboardShortcuts.nextCardStack': 'Nastepna karta (widok stosu)',
-    'keyboardShortcuts.disabledWhileTyping': 'Skroty sa wylaczone, gdy aktywne jest pole tekstowe.',
-    'admin.restoreSummary': 'Przywrocono {0} wierszy w {1} tabelach',
-    'admin.restoreSummaryForce':
-      'Przywrocono {0} wierszy. Wpisz ponownie dane logowania dla zaszyfrowanych pol.',
-    'admin.encryptionMismatchMessage':
-      'Kopia zostala utworzona przy uzyciu innego klucza szyfrowania.',
-    'admin.encryptionMismatchAffected': 'Te pola trzeba bedzie wpisac ponownie: {0}',
-    'admin.encryptionMismatchFallback':
-      'Niektore zaszyfrowane pola moga wymagac ponownego wpisania.',
-    'subscriptions.spotifyNotConnected': 'Spotify nie jest polaczone',
-    'playlistForm.dialogCreate': 'Utworz playliste',
-    'playlistForm.dialogEdit': 'Edytuj playliste',
-    'playlistForm.titleCreate': 'Utworz playliste',
-    'playlistForm.titleEdit': 'Edytuj playliste',
-    'playlistForm.namePlaceholder': 'Moj tygodniowy miks',
-    'playlistForm.strategy': 'Strategia',
-    'playlistForm.strategyWeeklyDigestDesc': 'Wybor utworow z ostatnich odkryc',
-    'playlistForm.strategyGenreFocusDesc': 'Glebokie zanurzenie w wybrany gatunek',
-    'playlistForm.strategyMoodMixDesc': 'Utwory pasujace do wybranego nastroju',
-    'playlistForm.strategyRediscoverDesc': 'Przypomnij sobie artystów, o ktorych mogles zapomniec',
-    'playlistForm.mood': 'Nastroj',
-    'playlistForm.moodPlaceholder': 'nocna jazda autem',
-    'playlistForm.trackCount': 'Liczba utworow',
-    'playlistForm.scheduleAutomaticGeneration': 'Zaplanuj automatyczne generowanie',
-    'playlistForm.nameRequired': 'Nazwa jest wymagana',
-    'playlistForm.genreRequired': 'Gatunek jest wymagany dla strategii Fokus na gatunek',
-    'playlistForm.moodRequired': 'Nastroj jest wymagany dla strategii Mood Mix',
-    'searchResultCard.addToQueue': 'Dodaj do kolejki odkryc',
-    'searchResultCard.queuing': 'Dodawanie...',
-    'searchResultCard.queuedSuccess': '{0} dodano do kolejki odkryc',
-    'searchResultCard.queuedFailure': 'Nie udalo się dodac {0} do kolejki',
-    'searchResultCard.popularity': '{0}{1} popularnosci',
-    'subscriptionPresets.getStarted': 'Zacznij od gotowego zestawu',
-    'subscriptionPresets.intro':
-      'Wybierz zestaw, aby automatycznie utworzyc subskrypcje, albo zbuduj własna konfiguracje.',
-    'subscriptionPresets.casualTitle': 'Na luzie',
-    'subscriptionPresets.casualDescription': 'Cotygodniowe rekomendacje z niewielka iloscia szumu',
-    'subscriptionPresets.deepDiverTitle': 'Gleboki nurek',
-    'subscriptionPresets.deepDiverDescription':
-      'Szerokie odkrywanie z wielu zrodel i codziennymi aktualizacjami',
-    'subscriptionPresets.discoveryOnlyTitle': 'Tylko odkrywanie',
-    'subscriptionPresets.discoveryOnlyDescription': 'Tylko pipeline, bez zewnetrznych feedow',
-    'subscriptionPresets.customTitle': 'Wlasny',
-    'subscriptionPresets.customDescription': 'Zbuduj własna mieszanke subskrypcji',
-    'subscriptionPresets.oneConnectedService': '1 podlaczona usluga',
-    'subscriptionPresets.twoConnectedServices': '2 podlaczone uslugi',
-    'subscriptionPresets.requires': 'Wymaga: {0}',
-    'subscriptionPresets.setUp': 'Skonfiguruj',
-    'subscriptionPresets.applying': 'Stosowanie...',
-    'subscriptionPresets.casualApplied': 'Zastosowano zestaw Na luzie',
-    'subscriptionPresets.deepDiverApplied': 'Zastosowano zestaw Gleboki nurek',
-    'subscriptionPresets.applyFailed': 'Nie udalo się zastosowac zestawu',
-    'subscriptionPresets.aiOnlyEnabled': 'Wlaczono tryb tylko AI',
-    'subscriptionPresets.aiOnlyEnableFailed': 'Nie udalo się włączyć trybu tylko AI',
-    'jobHistory.pipeline': 'Proces odkrywania',
-    'common.status': 'Stan',
-    'subscriptionForm.dialogCreate': 'Utwórz subskrypcję',
-    'subscriptionForm.dialogEdit': 'Edytuj subskrypcję',
-    'subscriptionForm.titleCreate': 'Utwórz subskrypcję',
-    'subscriptionForm.titleEdit': 'Edytuj subskrypcję',
-    'subscriptionForm.name': 'Nazwa',
-    'subscriptionForm.namePlaceholder': 'Nowe Shoegaze Weekly',
-    'subscriptionForm.nameRequired': 'Nazwa jest wymagana',
-    'subscriptionForm.source': 'Źródło',
-    'subscriptionForm.sources': 'Źródła',
-    'subscriptionForm.sourceType': 'Typ źródła',
-    'subscriptionForm.sourceTypeGenre': 'Gatunek muzyczny',
-    'subscriptionForm.sourceTypeGenreDesc': 'Odkrywaj artystów w określonym gatunku',
-    'subscriptionForm.sourceTypeSimilar': 'Podobni artyści',
-    'subscriptionForm.sourceTypeSimilarDesc': 'Znajdź artystów podobnych do swoich ulubionych',
-    'subscriptionForm.sourceTypeDiscoveryModeDesc': 'Użyj trybu odkrywania, aby generować artystów',
-    'subscriptionForm.discoveryMode': 'Tryb odkrywania',
-    'subscriptionForm.selectDiscoveryMode': 'Wybierz tryb odkrywania do skonfigurowania.',
-    'subscriptionForm.discoveryModeSettingsRequired': 'Ustawienia trybu odkrywania są wymagane',
-    'subscriptionForm.genre': 'Gatunek muzyczny',
-    'subscriptionForm.genrePlaceholder': 'shoegaze',
-    'subscriptionForm.seedArtistsPlaceholder': 'Radiohead, Portishead, Massive Attack',
-    'subscriptionForm.genreRequired': 'Gatunek jest wymagany',
-    'subscriptionForm.seedArtists': 'Artyści startowi',
-    'subscriptionForm.seedArtistsHelp': 'Nazwy artystów oddzielone przecinkami',
-    'subscriptionForm.seedArtistsRequired': 'Wymagany jest co najmniej jeden artysta startowy',
-    'subscriptionForm.configureInSettings': 'Skonfiguruj w Ustawienia > Połączenia',
-    'subscriptionForm.notConfigured': 'nie skonfigurowano',
-    'subscriptionForm.selectAtLeastOneSource': 'Wybierz co najmniej jedno źródło',
-    'subscriptionForm.sourceLockedHelp':
-      'Typ źródła i konfiguracja nie mogą zostać zmienione dla tej subskrypcji.',
-    'subscriptionForm.maxArtistsPerRun': 'Maksymalna liczba artystów na uruchomienie',
-    'subscriptionForm.scoringPreset': 'Preset punktacji',
-    'subscriptionForm.scoringPresetDefault': 'Domyślny',
-    'subscriptionForm.scoringPresetGenre': 'Zoptymalizowany pod gatunki',
-    'cronPicker.preset': 'Ustawienie harmonogramu',
-    'cronPicker.dailyMidnight': 'Codziennie (północ)',
-    'cronPicker.twiceAWeek': 'Dwa razy w tygodniu (pon + czw)',
-    'cronPicker.everyTwoWeeks': 'Co 2 tygodnie (1 i 15)',
-    'cronPicker.monthlyFirst': 'Co miesiąc (1. dzień)',
-    'cronPicker.standardHint':
-      'Standardowy cron: minuta godzina dzień-miesiąca miesiąc dzień-tygodnia',
-    'settings.capImport': 'Importowanie',
-    'systemHealth.pipeline': 'Proces odkrywania',
-    'systemHealth.librarySync': 'Synchronizacja biblioteki',
-    'discoveryMode.field.feed': 'Źródło feedu',
-    'discoveryMode.field.limit': 'Maksymalna liczba',
-    'settings.tabs.systemHealth': 'Stan systemu',
-    'settings.targetUpdated': 'Cel został zaktualizowany',
-    'settings.targetUpdateFailed': 'Nie udało się zaktualizować celu',
-    'settings.editTarget': 'Edytuj cel',
-    'settings.saveChanges': 'Zapisz zmiany',
-    'settings.linkedTarget': 'Powiązany cel:',
-    'settings.librarySyncInterval': 'Synchronizacja biblioteki',
-    'settings.librarySyncIntervalHours': 'Interwał synchronizacji (godziny)',
-    'settings.librarySyncIntervalHelp':
-      'Określa, jak często Digarr odświeża dane synchronizacji biblioteki i automatycznie ponawia kontrole.',
-    'settings.fieldModel': 'Model AI',
-    'settings.fieldClientId': 'Identyfikator klienta',
-    'settings.fieldClientSecret': 'Sekret klienta',
-    'settings.aiProviderTitle': 'Dostawca AI',
-    'settings.webhookTitle': 'Integracja webhook',
-    'settings.plexToken': 'Token Plex',
-    'settings.serviceConnectedToast': '{0} połączono',
-    'settings.serviceUnreachableToast': 'Nie udało się połączyć z {0}',
-    'settings.serviceSettingsSavedToast': 'Ustawienia {0} zostały zapisane',
-    'settings.serviceSettingsFailedToast': 'Nie udało się zapisać ustawień {0}',
-    'settings.spotifyAuthorizationFailed': 'Nie udało się rozpocząć autoryzacji Spotify',
-    'settings.spotifyDisconnectFailed': 'Nie udało się odłączyć Spotify',
-    'settings.spotifyLikedSongsFailed': 'Nie udało się rozpocząć importu Spotify Liked Songs',
-    'settings.spotifyPlaylistFailed': 'Nie udało się rozpocząć importu playlisty',
-    'settings.deezerFavoritesFailed': 'Nie udało się rozpocząć importu ulubionych z Deezer',
-    'settings.deezerFollowedFailed':
-      'Nie udało się rozpocząć importu obserwowanych artystów z Deezer',
-    'settings.deezerDisconnectFailed': 'Nie udało się odłączyć Deezer',
-    'settings.deezerAuthorizationFailed': 'Nie udało się rozpocząć autoryzacji Deezer',
-    'settings.reverseProxyConfigHint':
-      'Użyj PROXY_AUTH_ENABLED oraz PROXY_AUTH_TRUSTED_PROXIES; proxy musi wysyłać X-Forwarded-User.',
-    'admin.autoFixInProgress': 'Automatyczna naprawa trwa... ({0}/{1} naprawiono)',
-    'albumPicker.title': 'Wybierz albumy',
-    'albumPicker.loadFailed': 'Nie udało się wczytać albumów',
-    'albumPicker.selectAll': 'Wszystkie',
-    'albumPicker.selectNone': 'Żadne',
-    'albumPicker.selectedCount': 'Wybrano: {0}',
-    'albumPicker.aiPick': 'Wybór AI',
-    'albumPicker.monitor': 'Monitoruj',
-    'albumPicker.monitorSingular': 'Monitoruj {0} album',
-    'albumPicker.monitorPlural': 'Monitoruj albumy: {0}',
-    'approveDialog.title': 'Ustawienia Lidarr dla tego wykonawcy',
-    'approveDialog.loadFailed':
-      'Nie udało się wczytać profili Lidarr. Sprawdź połączenie z Lidarr.',
-    'approveDialog.qualityProfile': 'Profil jakości',
-    'approveDialog.metadataProfile': 'Profil metadanych',
-    'approveDialog.rootFolder': 'Folder główny',
-    'cardStack.counter': '{0} z {1}',
-    'cardStack.swipeHint': 'Przesuń w prawo, aby zatwierdzić, w lewo, aby odrzucić',
-    'serviceCard.testConnection': 'Testuj połączenie',
-    'preview.noPreviewAvailable': 'Brak dostępnego podglądu dla tego wykonawcy',
-    'preview.playbackBlocked':
-      'Odtwarzanie zablokowane przez przeglądarkę - spróbuj kliknąć ponownie',
-    'errors.auth.invalidCredentials': 'Nieprawidłowe dane logowania',
-    'errors.auth.notAuthenticated': 'Nie uwierzytelniono',
-    'errors.auth.passwordIncorrect': 'Bieżące hasło jest nieprawidłowe',
-    'errors.auth.usernameTaken': 'Nazwa użytkownika jest już zajęta',
-    'errors.target.notFound': 'Nie znaleziono celu',
-    'errors.pipeline.alreadyRunning': 'Uruchomienie pipeline już trwa',
-    'errors.recommendation.notFound': 'Nie znaleziono rekomendacji',
-    'errors.subscription.notFound': 'Nie znaleziono subskrypcji',
-    'errors.playlist.notFound': 'Nie znaleziono playlisty',
-  },
-  ja: {
-    ...REVIEW_GATE_TRANSLATIONS.ja,
-    'common.unauthorized': '認証されていません',
-    'common.adminAccessRequired': '管理者権限が必要です',
-    'common.enabled': '有効',
-    'common.importing': 'インポート中...',
-    'settings.tabs.systemHealth': 'システムの健全性',
-    'settings.targetUpdated': 'ターゲットを更新しました',
-    'settings.targetUpdateFailed': 'ターゲットの更新に失敗しました',
-    'settings.editTarget': 'ターゲットを編集',
-    'settings.saveChanges': '変更を保存',
-    'settings.linkedTarget': '関連付けられたターゲット:',
-    'libraryHealth.syncNow': '今すぐ同期',
-    'libraryHealth.snapshotTitle': '最新のスナップショット',
-    'libraryHealth.lastSynced': '最終同期',
-    'libraryHealth.neverSynced': 'まだ同期されていません。',
-    'libraryHealth.autoSyncEvery': '{0}時間ごとに自動同期',
-    'jobHistory.librarySync': 'ライブラリ同期',
-    'errorBoundary.title': '問題が発生しました',
-    'errorBoundary.retry': '再試行',
-    'errorBoundary.home': 'ホームへ戻る',
-    'connectionSuggestions.title': '{0} の提案',
-    'connectionSuggestions.close': '提案を閉じる',
-    'connectionSuggestions.gotIt': '了解',
-    'connectionSuggestions.skipForNow': '今はスキップ',
-    'connectionSuggestions.spotifyDiscoverWeekly': 'Discover Weekly のサブスクリプションを設定',
-    'connectionSuggestions.spotifyReleaseRadar': 'Release Radar のサブスクリプションを有効化',
-    'connectionSuggestions.lastfmTopGenreTags': 'よく聴くジャンルタグのサブスクリプションを作成',
-    'connectionSuggestions.listenbrainzWeeklyJams': 'Weekly Jams のサブスクリプションを有効化',
-    'connectionSuggestions.listenbrainzWeeklyExploration': 'Weekly Exploration を有効化',
-    'connectionSuggestions.weeklyDigestPlaylist': '毎週の Digarr Digest プレイリストを作成',
-    'keyboardShortcuts.title': 'キーボードショートカット',
-    'keyboardShortcuts.close': 'ショートカットを閉じる',
-    'keyboardShortcuts.nextCard': '次のカード',
-    'keyboardShortcuts.previousCard': '前のカード',
-    'keyboardShortcuts.approveFocusedCard': '選択中のカードを承認',
-    'keyboardShortcuts.rejectFocusedCard': '選択中のカードを拒否',
-    'keyboardShortcuts.toggleDetailView': '詳細表示を開く / 閉じる',
-    'keyboardShortcuts.toggleShortcuts': 'ショートカットを表示 / 非表示',
-    'keyboardShortcuts.previousCardStack': '前のカード（スタック表示）',
-    'keyboardShortcuts.nextCardStack': '次のカード（スタック表示）',
-    'keyboardShortcuts.disabledWhileTyping':
-      'テキスト入力欄にフォーカスがある間はショートカットが無効になります。',
-    'admin.restoreSummary': '{1} テーブルで {0} 行を復元しました',
-    'admin.restoreSummaryForce':
-      '{0} 行を復元しました。暗号化された項目の認証情報を再入力してください。',
-    'admin.encryptionMismatchMessage': 'このバックアップは別の暗号化キーで作成されています。',
-    'admin.encryptionMismatchAffected': '次の項目は再入力が必要です: {0}',
-    'admin.encryptionMismatchFallback':
-      '一部の暗号化された項目は再入力が必要になる場合があります。',
-    'subscriptions.spotifyNotConnected': 'Spotify が接続されていません',
-    'playlistForm.dialogCreate': 'プレイリストを作成',
-    'playlistForm.dialogEdit': 'プレイリストを編集',
-    'playlistForm.titleCreate': 'プレイリストを作成',
-    'playlistForm.titleEdit': 'プレイリストを編集',
-    'playlistForm.namePlaceholder': 'マイ・ウィークリー・ミックス',
-    'playlistForm.strategy': '戦略',
-    'playlistForm.strategyWeeklyDigestDesc': '最近の発見から厳選した曲',
-    'playlistForm.strategyGenreFocusDesc': '特定のジャンルを深掘り',
-    'playlistForm.strategyMoodMixDesc': '特定のムードに合う曲を集める',
-    'playlistForm.strategyRediscoverDesc': '忘れていたかもしれないアーティストを掘り起こす',
-    'playlistForm.mood': 'ムード',
-    'playlistForm.moodPlaceholder': '夜のドライブ',
-    'playlistForm.trackCount': '曲数',
-    'playlistForm.scheduleAutomaticGeneration': '自動生成をスケジュールする',
-    'playlistForm.nameRequired': '名前は必須です',
-    'playlistForm.genreRequired': 'ジャンルフォーカス戦略ではジャンルが必須です',
-    'playlistForm.moodRequired': 'ムードミックス戦略ではムードが必須です',
-    'searchResultCard.addToQueue': '発見キューに追加',
-    'searchResultCard.queuing': '追加中...',
-    'searchResultCard.queuedSuccess': '{0} を発見キューに追加しました',
-    'searchResultCard.queuedFailure': '{0} をキューに追加できませんでした',
-    'searchResultCard.popularity': '人気 {0}{1}',
-    'subscriptionPresets.getStarted': 'プリセットから始める',
-    'subscriptionPresets.intro':
-      'プリセットを選んでサブスクリプションを自動作成するか、独自の構成を作成してください。',
-    'subscriptionPresets.casualTitle': 'カジュアル',
-    'subscriptionPresets.casualDescription': 'ノイズの少ない毎週のおすすめ',
-    'subscriptionPresets.deepDiverTitle': 'ディープダイバー',
-    'subscriptionPresets.deepDiverDescription': '複数ソースを使った幅広い発見と毎日の更新',
-    'subscriptionPresets.discoveryOnlyTitle': '発見のみ',
-    'subscriptionPresets.discoveryOnlyDescription': 'パイプラインのみ、外部フィードなし',
-    'subscriptionPresets.customTitle': 'カスタム',
-    'subscriptionPresets.customDescription': '自分だけのサブスクリプション構成を作る',
-    'subscriptionPresets.oneConnectedService': '1 つの接続済みサービス',
-    'subscriptionPresets.twoConnectedServices': '2 つの接続済みサービス',
-    'subscriptionPresets.requires': '{0} が必要です',
-    'subscriptionPresets.setUp': '設定する',
-    'subscriptionPresets.applying': '適用中...',
-    'subscriptionPresets.casualApplied': 'カジュアルのプリセットを適用しました',
-    'subscriptionPresets.deepDiverApplied': 'ディープダイバーのプリセットを適用しました',
-    'subscriptionPresets.applyFailed': 'プリセットを適用できませんでした',
-    'subscriptionPresets.aiOnlyEnabled': 'AI のみモードを有効にしました',
-    'subscriptionPresets.aiOnlyEnableFailed': 'AI のみモードを有効にできませんでした',
-    'common.current': '現在の設定',
-    'common.custom': 'カスタム設定',
-    'subscriptionForm.dialogCreate': 'サブスクリプションを作成',
-    'subscriptionForm.dialogEdit': 'サブスクリプションを編集',
-    'subscriptionForm.titleCreate': 'サブスクリプションを作成',
-    'subscriptionForm.titleEdit': 'サブスクリプションを編集',
-    'subscriptionForm.name': '名前',
-    'subscriptionForm.namePlaceholder': '新しい Shoegaze Weekly',
-    'subscriptionForm.nameRequired': '名前は必須です',
-    'subscriptionForm.source': 'ソース',
-    'subscriptionForm.sources': 'ソース',
-    'subscriptionForm.sourceType': 'ソース種別',
-    'subscriptionForm.sourceTypeGenre': '音楽ジャンル',
-    'subscriptionForm.sourceTypeGenreDesc': '特定のジャンルからアーティストを見つけます',
-    'subscriptionForm.sourceTypeSimilar': '似たアーティスト',
-    'subscriptionForm.sourceTypeSimilarDesc': 'お気に入りに似たアーティストを探します',
-    'subscriptionForm.sourceTypeDiscoveryModeDesc':
-      'ディスカバリーモードでアーティストを生成します',
-    'subscriptionForm.discoveryMode': 'ディスカバリーモード',
-    'subscriptionForm.selectDiscoveryMode': '設定するディスカバリーモードを選択してください。',
-    'subscriptionForm.discoveryModeSettingsRequired': 'ディスカバリーモードの設定は必須です',
-    'subscriptionForm.genre': '音楽ジャンル',
-    'subscriptionForm.genrePlaceholder': 'shoegaze',
-    'subscriptionForm.seedArtistsPlaceholder': 'Radiohead, Portishead, Massive Attack',
-    'subscriptionForm.genreRequired': 'ジャンルは必須です',
-    'subscriptionForm.seedArtists': '起点アーティスト',
-    'subscriptionForm.seedArtistsHelp': 'アーティスト名をカンマ区切りで入力してください',
-    'subscriptionForm.seedArtistsRequired': '少なくとも 1 人の起点アーティストが必要です',
-    'subscriptionForm.configureInSettings': '設定 > 接続 で設定してください',
-    'subscriptionForm.notConfigured': '未設定',
-    'subscriptionForm.selectAtLeastOneSource': '少なくとも 1 つのソースを選択してください',
-    'subscriptionForm.sourceLockedHelp':
-      'このサブスクリプションではソース種別と設定を変更できません。',
-    'subscriptionForm.maxArtistsPerRun': '1 回あたりの最大アーティスト数',
-    'subscriptionForm.scoringPreset': 'スコアプリセット',
-    'subscriptionForm.scoringPresetDefault': '標準',
-    'subscriptionForm.scoringPresetGenre': 'ジャンル重視',
-    'cronPicker.preset': 'スケジュール設定',
-    'cronPicker.dailyMidnight': '毎日（深夜0時）',
-    'cronPicker.twiceAWeek': '週2回（月 + 木）',
-    'cronPicker.everyTwoWeeks': '2週間ごと（1日と15日）',
-    'cronPicker.monthlyFirst': '毎月（1日）',
-    'cronPicker.standardHint': '標準 cron: 分 時 日 月 曜日',
-    'systemHealth.librarySync': 'ライブラリ同期',
-    'settings.targetTypeEmbyPlaylist': 'Emby プレイリスト',
-    'settings.librarySyncInterval': 'ライブラリ同期',
-    'settings.librarySyncIntervalHours': '同期間隔（時間）',
-    'settings.librarySyncIntervalHelp':
-      'Digarr がライブラリ同期データを更新し、チェックを自動で再実行する間隔を設定します。',
-    'settings.fieldClientId': 'クライアント ID',
-    'settings.fieldClientSecret': 'クライアントシークレット',
-    'settings.fieldWebhookUrl': 'Webhook 送信先 URL',
-    'settings.aiProviderTitle': 'AI プロバイダー',
-    'settings.webhookTitle': 'Webhook 連携',
-    'settings.plexToken': 'Plex トークン',
-    'settings.serviceConnectedToast': '{0} を接続しました',
-    'settings.serviceUnreachableToast': '{0} に接続できませんでした',
-    'settings.serviceSettingsSavedToast': '{0} の設定を保存しました',
-    'settings.serviceSettingsFailedToast': '{0} の設定を保存できませんでした',
-    'settings.spotifyAuthorizationFailed': 'Spotify の認可を開始できませんでした',
-    'settings.spotifyDisconnectFailed': 'Spotify を切断できませんでした',
-    'settings.spotifyLikedSongsFailed': 'Spotify の「いいね」曲の取り込みを開始できませんでした',
-    'settings.spotifyPlaylistFailed': 'プレイリストの取り込みを開始できませんでした',
-    'settings.deezerFavoritesFailed': 'Deezer のお気に入り取り込みを開始できませんでした',
-    'settings.deezerFollowedFailed':
-      'Deezer のフォロー中アーティスト取り込みを開始できませんでした',
-    'settings.deezerDisconnectFailed': 'Deezer を切断できませんでした',
-    'settings.deezerAuthorizationFailed': 'Deezer の認可を開始できませんでした',
-    'settings.reverseProxyConfigHint':
-      'PROXY_AUTH_ENABLED と PROXY_AUTH_TRUSTED_PROXIES を使い、プロキシは X-Forwarded-User を送信してください。',
-    'admin.autoFixInProgress': '自動修正を実行中...（{0}/{1} 件修正済み）',
-    'albumPicker.title': 'アルバムを選択',
-    'albumPicker.loadFailed': 'アルバムの読み込みに失敗しました',
-    'albumPicker.selectAll': 'すべて',
-    'albumPicker.selectNone': 'なし',
-    'albumPicker.selectedCount': '{0} 件選択中',
-    'albumPicker.aiPick': 'AI のおすすめ',
-    'albumPicker.monitor': '監視',
-    'albumPicker.monitorSingular': '{0} 枚のアルバムを監視',
-    'albumPicker.monitorPlural': '{0} 枚のアルバムを監視',
-    'approveDialog.title': 'このアーティストの Lidarr 設定',
-    'approveDialog.loadFailed':
-      'Lidarr プロファイルの読み込みに失敗しました。Lidarr の接続を確認してください。',
-    'approveDialog.qualityProfile': '品質プロファイル',
-    'approveDialog.metadataProfile': 'メタデータプロファイル',
-    'approveDialog.rootFolder': 'ルートフォルダ',
-    'cardStack.counter': '{1} 件中 {0} 件目',
-    'cardStack.swipeHint': '右にスワイプで承認、左にスワイプで拒否',
-    'serviceCard.testConnection': '接続テスト',
-    'preview.noPreviewAvailable': 'このアーティストのプレビューはありません',
-    'preview.playbackBlocked':
-      'ブラウザによって再生がブロックされました。もう一度クリックしてください',
-    'errors.auth.invalidCredentials': '認証情報が無効です',
-    'errors.auth.notAuthenticated': '認証されていません',
-    'errors.auth.passwordIncorrect': '現在のパスワードが正しくありません',
-    'errors.auth.usernameTaken': 'そのユーザー名はすでに使われています',
-    'errors.target.notFound': 'ターゲットが見つかりません',
-    'errors.pipeline.alreadyRunning': 'パイプラインの実行はすでに進行中です',
-    'errors.recommendation.notFound': 'おすすめが見つかりません',
-    'errors.subscription.notFound': 'サブスクリプションが見つかりません',
-    'errors.playlist.notFound': 'プレイリストが見つかりません',
-  },
-  'pt-BR': {
-    ...REVIEW_GATE_TRANSLATIONS['pt-BR'],
-    'common.importing': 'Importando...',
-    'common.current': 'Atual',
-    'common.custom': 'Personalizado',
-    'common.enabled': 'Ativado',
-    'common.unauthorized': 'Não autorizado',
-    'common.adminAccessRequired': 'Acesso de administrador obrigatório',
-    'jobHistory.pipeline': 'Fluxo de descoberta',
-    'jobHistory.librarySync': 'Sincronização da biblioteca',
-    'jobHistory.playlists': 'Listas de reprodução',
-    'playlist.total': 'no total',
-    'playlistForm.dialogCreate': 'Criar playlist',
-    'playlistForm.dialogEdit': 'Editar playlist',
-    'playlistForm.titleCreate': 'Criar Playlist',
-    'playlistForm.titleEdit': 'Editar Playlist',
-    'playlistForm.namePlaceholder': 'Meu Mix Semanal',
-    'playlistForm.strategy': 'Estratégia',
-    'playlistForm.strategyWeeklyDigestDesc': 'Seleção curada das suas descobertas recentes',
-    'playlistForm.strategyGenreFocusDesc': 'Mergulho profundo em um gênero específico',
-    'playlistForm.strategyMoodMixDesc': 'Faixas que combinam com um clima específico',
-    'playlistForm.strategyRediscoverDesc': 'Traga de volta artistas que você pode ter esquecido',
-    'playlistForm.mood': 'Clima',
-    'playlistForm.moodPlaceholder': 'dirigindo tarde da noite',
-    'playlistForm.trackCount': 'Número de faixas',
-    'playlistForm.scheduleAutomaticGeneration': 'Agendar geração automática',
-    'playlistForm.nameRequired': 'O nome é obrigatório',
-    'playlistForm.genreRequired': 'O gênero é obrigatório para a estratégia Foco em gênero',
-    'playlistForm.moodRequired': 'O clima é obrigatório para a estratégia Mix de clima',
-    'subscriptionForm.dialogCreate': 'Criar assinatura',
-    'subscriptionForm.dialogEdit': 'Editar assinatura',
-    'subscriptionForm.titleCreate': 'Criar assinatura',
-    'subscriptionForm.titleEdit': 'Editar assinatura',
-    'subscriptionForm.name': 'Nome',
-    'subscriptionForm.namePlaceholder': 'Nova Shoegaze Weekly',
-    'subscriptionForm.nameRequired': 'O nome é obrigatório',
-    'subscriptionForm.source': 'Fonte',
-    'subscriptionForm.sources': 'Fontes',
-    'subscriptionForm.sourceType': 'Tipo de fonte',
-    'subscriptionForm.sourceTypeGenre': 'Gênero',
-    'subscriptionForm.sourceTypeGenreDesc': 'Descubra artistas em um gênero específico',
-    'subscriptionForm.sourceTypeSimilar': 'Semelhantes',
-    'subscriptionForm.sourceTypeSimilarDesc': 'Encontre artistas parecidos com os seus favoritos',
-    'subscriptionForm.sourceTypeDiscoveryModeDesc': 'Use um modo de descoberta para gerar artistas',
-    'subscriptionForm.discoveryMode': 'Modo de descoberta',
-    'subscriptionForm.selectDiscoveryMode': 'Selecione um modo de descoberta para configurá-lo.',
-    'subscriptionForm.discoveryModeSettingsRequired':
-      'As configurações do modo de descoberta são obrigatórias',
-    'subscriptionForm.genre': 'Gênero',
-    'subscriptionForm.genrePlaceholder': 'shoegaze',
-    'subscriptionForm.genreRequired': 'O gênero é obrigatório',
-    'subscriptionForm.seedArtists': 'Artistas semente',
-    'subscriptionForm.seedArtistsPlaceholder': 'Radiohead, Portishead, Massive Attack',
-    'subscriptionForm.seedArtistsHelp': 'Nomes de artistas separados por vírgulas',
-    'subscriptionForm.seedArtistsRequired': 'É necessário pelo menos um artista semente',
-    'subscriptionForm.configureInSettings': 'Configure em Configurações > Conexões',
-    'subscriptionForm.notConfigured': 'não configurado',
-    'subscriptionForm.selectAtLeastOneSource': 'Selecione pelo menos uma fonte',
-    'subscriptionForm.sourceLockedHelp':
-      'O tipo de fonte e a configuração não podem ser alterados para esta assinatura.',
-    'subscriptionForm.maxArtistsPerRun': 'Máximo de artistas por execução',
-    'subscriptionForm.scoringPreset': 'Preset de pontuação',
-    'subscriptionForm.scoringPresetDefault': 'Padrão',
-    'subscriptionForm.scoringPresetGenre': 'Otimizado para gênero',
-    'subscriptions.spotifyNotConnected': 'O Spotify não está conectado',
-    'subscriptionPresets.getStarted': 'Comece com um preset',
-    'subscriptionPresets.intro':
-      'Escolha um preset para criar assinaturas automaticamente ou monte uma configuração personalizada.',
-    'subscriptionPresets.casualTitle': 'Leve',
-    'subscriptionPresets.casualDescription': 'Recomendações semanais com pouco ruído',
-    'subscriptionPresets.deepDiverTitle': 'Mergulho profundo',
-    'subscriptionPresets.deepDiverDescription':
-      'Descoberta ampla com várias fontes e atualizações diárias',
-    'subscriptionPresets.discoveryOnlyTitle': 'Somente descoberta',
-    'subscriptionPresets.discoveryOnlyDescription': 'Apenas pipeline, sem feeds externos',
-    'subscriptionPresets.customTitle': 'Personalizado',
-    'subscriptionPresets.customDescription': 'Monte sua própria combinação de assinaturas',
-    'subscriptionPresets.oneConnectedService': '1 serviço conectado',
-    'subscriptionPresets.twoConnectedServices': '2 serviços conectados',
-    'subscriptionPresets.requires': 'Requer {0}',
-    'subscriptionPresets.setUp': 'Configurar',
-    'subscriptionPresets.applying': 'Aplicando...',
-    'subscriptionPresets.casualApplied': 'Preset Leve aplicado',
-    'subscriptionPresets.deepDiverApplied': 'Preset Mergulho profundo aplicado',
-    'subscriptionPresets.applyFailed': 'Falha ao aplicar o preset',
-    'subscriptionPresets.aiOnlyEnabled': 'Modo somente IA ativado',
-    'subscriptionPresets.aiOnlyEnableFailed': 'Falha ao ativar o modo somente IA',
-    'cronPicker.preset': 'Predefinição',
-    'cronPicker.dailyMidnight': 'Diário (meia-noite)',
-    'cronPicker.twiceAWeek': 'Duas vezes por semana (seg + qui)',
-    'cronPicker.everyTwoWeeks': 'A cada 2 semanas (1 e 15)',
-    'cronPicker.monthlyFirst': 'Mensal (dia 1)',
-    'cronPicker.standardHint': 'Cron padrão: minuto hora dia-do-mês mês dia-da-semana',
-    'libraryHealth.syncNow': 'Sincronizar agora',
-    'libraryHealth.snapshotTitle': 'Última captura',
-    'libraryHealth.lastSynced': 'Última sincronização',
-    'libraryHealth.neverSynced': 'Nunca sincronizado.',
-    'libraryHealth.autoSyncEvery': 'Sincronização automática a cada {0} h',
-    'systemHealth.pipeline': 'Fluxo de descoberta',
-    'systemHealth.librarySync': 'Sincronização da biblioteca',
-    'settings.tabs.systemHealth': 'Saúde do sistema',
-    'settings.targetUpdated': 'Destino atualizado',
-    'settings.targetUpdateFailed': 'Falha ao atualizar o destino',
-    'settings.editTarget': 'Editar destino',
-    'settings.saveChanges': 'Salvar alterações',
-    'settings.linkedTarget': 'Destino vinculado:',
-    'settings.librarySyncInterval': 'Sincronização da biblioteca',
-    'settings.librarySyncIntervalHours': 'Intervalo de sincronização (horas)',
-    'settings.librarySyncIntervalHelp':
-      'Controla com que frequência o Digarr atualiza os dados de sincronização da biblioteca e executa novamente as verificações automaticamente.',
-    'settings.oidcSso': 'OIDC e SSO',
-    'settings.aiProviderTitle': 'Provedor de IA',
-    'settings.webhookTitle': 'Integração webhook',
-    'settings.plexToken': 'Token do Plex',
-    'settings.serviceConnectedToast': '{0} conectado',
-    'settings.serviceUnreachableToast': '{0} está inacessível',
-    'settings.serviceSettingsSavedToast': 'Configurações de {0} salvas',
-    'settings.serviceSettingsFailedToast': 'Falha ao salvar as configurações de {0}',
-    'settings.spotifyAuthorizationFailed': 'Falha ao iniciar a autorização do Spotify',
-    'settings.spotifyDisconnectFailed': 'Falha ao desconectar o Spotify',
-    'settings.spotifyLikedSongsFailed': 'Falha ao iniciar a importação das curtidas do Spotify',
-    'settings.spotifyPlaylistFailed': 'Falha ao importar playlist do Spotify',
-    'settings.deezerFavoritesFailed': 'Falha ao iniciar a importação dos favoritos do Deezer',
-    'settings.deezerFollowedFailed':
-      'Falha ao iniciar a importação dos artistas seguidos do Deezer',
-    'settings.deezerDisconnectFailed': 'Falha ao desconectar o Deezer',
-    'settings.deezerAuthorizationFailed': 'Falha ao iniciar a autorização do Deezer',
-    'settings.reverseProxyConfigHint':
-      'Use PROXY_AUTH_ENABLED e PROXY_AUTH_TRUSTED_PROXIES; seu proxy deve enviar X-Forwarded-User.',
-    'searchResultCard.addToQueue': 'Adicionar à fila de descoberta',
-    'searchResultCard.queuing': 'Adicionando...',
-    'searchResultCard.queuedSuccess': '{0} foi adicionado à fila de descoberta',
-    'searchResultCard.queuedFailure': 'Falha ao adicionar {0} à fila',
-    'searchResultCard.popularity': '{0}{1} de popularidade',
-    'admin.restoreSummary': '{0} linhas restauradas em {1} tabelas',
-    'admin.restoreSummaryForce':
-      '{0} linhas restauradas. Digite novamente as credenciais dos campos criptografados.',
-    'admin.encryptionMismatchMessage':
-      'O backup foi criado com uma chave de criptografia diferente.',
-    'admin.encryptionMismatchAffected': 'Estes campos precisarão ser preenchidos novamente: {0}',
-    'admin.encryptionMismatchFallback':
-      'Alguns campos criptografados podem precisar ser preenchidos novamente.',
-    'admin.autoFixInProgress': 'Correção automática em andamento... ({0}/{1} corrigidos)',
-    'errorBoundary.title': 'Algo deu errado',
-    'errorBoundary.retry': 'Tentar novamente',
-    'errorBoundary.home': 'Ir para o início',
-    'connectionSuggestions.title': 'Sugestões para {0}',
-    'connectionSuggestions.close': 'Fechar sugestões',
-    'connectionSuggestions.gotIt': 'Entendi',
-    'connectionSuggestions.skipForNow': 'Pular por enquanto',
-    'connectionSuggestions.spotifyDiscoverWeekly': 'Configurar uma assinatura do Discover Weekly',
-    'connectionSuggestions.spotifyReleaseRadar': 'Ativar uma assinatura do Release Radar',
-    'connectionSuggestions.lastfmTopGenreTags': 'Assinar suas principais tags de gênero',
-    'connectionSuggestions.listenbrainzWeeklyJams': 'Ativar a assinatura Weekly Jams',
-    'connectionSuggestions.listenbrainzWeeklyExploration': 'Ativar Weekly Exploration',
-    'connectionSuggestions.weeklyDigestPlaylist': 'Criar uma playlist semanal do Digarr Digest',
-    'keyboardShortcuts.title': 'Atalhos de teclado',
-    'keyboardShortcuts.close': 'Fechar atalhos',
-    'keyboardShortcuts.nextCard': 'Próximo cartão',
-    'keyboardShortcuts.previousCard': 'Cartão anterior',
-    'keyboardShortcuts.approveFocusedCard': 'Aprovar o cartão em foco',
-    'keyboardShortcuts.rejectFocusedCard': 'Rejeitar o cartão em foco',
-    'keyboardShortcuts.toggleDetailView': 'Abrir / fechar a visualização de detalhes',
-    'keyboardShortcuts.toggleShortcuts': 'Mostrar / ocultar atalhos',
-    'keyboardShortcuts.previousCardStack': 'Cartão anterior (visão em pilha)',
-    'keyboardShortcuts.nextCardStack': 'Próximo cartão (visão em pilha)',
-    'keyboardShortcuts.disabledWhileTyping':
-      'Os atalhos são desativados quando um campo de texto está em foco.',
-    'app.themeGroupEditor': 'Ferramentas do editor',
-    'common.manual': 'manual',
-    'jobHistory.playlistType': 'Lista de reprodução',
-    'common.status': 'Status atual',
-    'common.total': 'Total acumulado',
-    'settings.capDiscovery': 'Descoberta',
-    'settings.capImport': 'Importação',
-    'settings.capLibrarySync': 'Sincronização da biblioteca',
-    'settings.capPlaylistExport': 'Exportação de playlists',
-    'settings.capService': 'Serviço conectado',
-    'settings.capSubscriptions': 'Assinaturas',
-    'search.experimental': 'Em teste',
-    'discoveryMode.field.feed': 'Origem do feed',
-    'discoveryMode.field.tags': 'Etiquetas',
-    'settings.totalWeight': 'peso total:',
-    'integration.playlist': 'Lista de reprodução',
-    'albumPicker.title': 'Selecionar álbuns',
-    'albumPicker.loadFailed': 'Falha ao carregar álbuns',
-    'albumPicker.selectAll': 'Todos',
-    'albumPicker.selectNone': 'Nenhum',
-    'albumPicker.selectedCount': '{0} selecionados',
-    'albumPicker.aiPick': 'Escolha da IA',
-    'albumPicker.monitor': 'Monitorar',
-    'albumPicker.monitorSingular': 'Monitorar {0} álbum',
-    'albumPicker.monitorPlural': 'Monitorar {0} álbuns',
-    'approveDialog.title': 'Configurações do Lidarr para este artista',
-    'approveDialog.loadFailed':
-      'Falha ao carregar os perfis do Lidarr. Verifique sua conexão com o Lidarr.',
-    'approveDialog.qualityProfile': 'Perfil de qualidade',
-    'approveDialog.metadataProfile': 'Perfil de metadados',
-    'approveDialog.rootFolder': 'Pasta raiz',
-    'cardStack.counter': '{0} de {1}',
-    'cardStack.swipeHint': 'Deslize para a direita para aprovar, para a esquerda para rejeitar',
-    'serviceCard.testConnection': 'Testar conexão',
-    'preview.noPreviewAvailable': 'Sem prévia disponível para este artista',
-    'preview.playbackBlocked': 'Reprodução bloqueada pelo navegador - tente clicar novamente',
-    'errors.auth.invalidCredentials': 'Credenciais inválidas',
-    'errors.auth.notAuthenticated': 'Não autenticado',
-    'errors.auth.passwordIncorrect': 'A senha atual está incorreta',
-    'errors.auth.usernameTaken': 'Nome de usuário já em uso',
-    'errors.target.notFound': 'Destino não encontrado',
-    'errors.pipeline.alreadyRunning': 'Já há uma execução do pipeline em andamento',
-    'errors.recommendation.notFound': 'Recomendação não encontrada',
-    'errors.subscription.notFound': 'Assinatura não encontrada',
-    'errors.playlist.notFound': 'Playlist não encontrada',
-  },
-  ru: {
-    ...REVIEW_GATE_TRANSLATIONS.ru,
-    'common.importing': 'Импорт...',
-    'common.current': 'Текущий',
-    'common.custom': 'Пользовательский',
-    'common.enabled': 'Включено',
-    'common.unauthorized': 'Нет авторизации',
-    'common.adminAccessRequired': 'Требуется доступ администратора',
-    'jobHistory.pipeline': 'Пайплайн',
-    'jobHistory.librarySync': 'Синхронизация библиотеки',
-    'playlistForm.dialogCreate': 'Создать плейлист',
-    'playlistForm.dialogEdit': 'Изменить плейлист',
-    'playlistForm.titleCreate': 'Создать плейлист',
-    'playlistForm.titleEdit': 'Изменить плейлист',
-    'playlistForm.namePlaceholder': 'Мой недельный микс',
-    'playlistForm.strategy': 'Стратегия',
-    'playlistForm.strategyWeeklyDigestDesc': 'Подборка из ваших недавних открытий',
-    'playlistForm.strategyGenreFocusDesc': 'Глубокое погружение в конкретный жанр',
-    'playlistForm.strategyMoodMixDesc': 'Треки под определенное настроение',
-    'playlistForm.strategyRediscoverDesc': 'Верните артистов, о которых могли забыть',
-    'playlistForm.mood': 'Настроение',
-    'playlistForm.moodPlaceholder': 'ночная поездка',
-    'playlistForm.trackCount': 'Количество треков',
-    'playlistForm.scheduleAutomaticGeneration': 'Запланировать автоматическую генерацию',
-    'playlistForm.nameRequired': 'Название обязательно',
-    'playlistForm.genreRequired': 'Жанр обязателен для стратегии Фокус на жанре',
-    'playlistForm.moodRequired': 'Настроение обязательно для стратегии Mood Mix',
-    'subscriptionForm.dialogCreate': 'Создать подписку',
-    'subscriptionForm.dialogEdit': 'Изменить подписку',
-    'subscriptionForm.titleCreate': 'Создать подписку',
-    'subscriptionForm.titleEdit': 'Изменить подписку',
-    'subscriptionForm.name': 'Название',
-    'subscriptionForm.namePlaceholder': 'Новая Shoegaze Weekly',
-    'subscriptionForm.nameRequired': 'Название обязательно',
-    'subscriptionForm.source': 'Источник',
-    'subscriptionForm.sources': 'Источники',
-    'subscriptionForm.sourceType': 'Тип источника',
-    'subscriptionForm.sourceTypeGenre': 'Жанр',
-    'subscriptionForm.sourceTypeGenreDesc': 'Открывайте артистов в определенном жанре',
-    'subscriptionForm.sourceTypeSimilar': 'Похожие',
-    'subscriptionForm.sourceTypeSimilarDesc': 'Находите артистов, похожих на ваших любимых',
-    'subscriptionForm.sourceTypeDiscoveryModeDesc':
-      'Используйте режим открытия для генерации артистов',
-    'subscriptionForm.discoveryMode': 'Режим открытия',
-    'subscriptionForm.selectDiscoveryMode': 'Выберите режим открытия, чтобы настроить его.',
-    'subscriptionForm.discoveryModeSettingsRequired': 'Настройки режима открытия обязательны',
-    'subscriptionForm.genre': 'Жанр',
-    'subscriptionForm.genrePlaceholder': 'shoegaze',
-    'subscriptionForm.genreRequired': 'Жанр обязателен',
-    'subscriptionForm.seedArtists': 'Базовые артисты',
-    'subscriptionForm.seedArtistsPlaceholder': 'Radiohead, Portishead, Massive Attack',
-    'subscriptionForm.seedArtistsHelp': 'Имена артистов через запятую',
-    'subscriptionForm.seedArtistsRequired': 'Нужен хотя бы один базовый артист',
-    'subscriptionForm.configureInSettings': 'Настройте в Настройках > Подключения',
-    'subscriptionForm.notConfigured': 'не настроено',
-    'subscriptionForm.selectAtLeastOneSource': 'Выберите хотя бы один источник',
-    'subscriptionForm.sourceLockedHelp':
-      'Тип источника и конфигурацию нельзя изменить для этой подписки.',
-    'subscriptionForm.maxArtistsPerRun': 'Максимум артистов за запуск',
-    'subscriptionForm.scoringPreset': 'Пресет оценки',
-    'subscriptionForm.scoringPresetDefault': 'По умолчанию',
-    'subscriptionForm.scoringPresetGenre': 'Оптимизировано под жанр',
-    'subscriptions.spotifyNotConnected': 'Spotify не подключен',
-    'subscriptionPresets.getStarted': 'Начните с пресета',
-    'subscriptionPresets.intro':
-      'Выберите пресет, чтобы автоматически создать подписки, или соберите собственную конфигурацию.',
-    'subscriptionPresets.casualTitle': 'Легкий',
-    'subscriptionPresets.casualDescription': 'Еженедельные рекомендации с минимумом шума',
-    'subscriptionPresets.deepDiverTitle': 'Глубокое погружение',
-    'subscriptionPresets.deepDiverDescription':
-      'Широкое открытие из нескольких источников с ежедневными обновлениями',
-    'subscriptionPresets.discoveryOnlyTitle': 'Только открытие',
-    'subscriptionPresets.discoveryOnlyDescription': 'Только пайплайн, без внешних фидов',
-    'subscriptionPresets.customTitle': 'Свой вариант',
-    'subscriptionPresets.customDescription': 'Соберите свою комбинацию подписок',
-    'subscriptionPresets.oneConnectedService': '1 подключенный сервис',
-    'subscriptionPresets.twoConnectedServices': '2 подключенных сервиса',
-    'subscriptionPresets.requires': 'Требуется {0}',
-    'subscriptionPresets.setUp': 'Настроить',
-    'subscriptionPresets.applying': 'Применение...',
-    'subscriptionPresets.casualApplied': 'Пресет Легкий применен',
-    'subscriptionPresets.deepDiverApplied': 'Пресет Глубокое погружение применен',
-    'subscriptionPresets.applyFailed': 'Не удалось применить пресет',
-    'subscriptionPresets.aiOnlyEnabled': 'Режим только AI включен',
-    'subscriptionPresets.aiOnlyEnableFailed': 'Не удалось включить режим только AI',
-    'cronPicker.preset': 'Пресет',
-    'cronPicker.dailyMidnight': 'Ежедневно (полночь)',
-    'cronPicker.twiceAWeek': 'Два раза в неделю (пн + чт)',
-    'cronPicker.everyTwoWeeks': 'Каждые 2 недели (1 и 15)',
-    'cronPicker.monthlyFirst': 'Ежемесячно (1-е число)',
-    'cronPicker.standardHint': 'Стандартный cron: минута час день-месяца месяц день-недели',
-    'libraryHealth.syncNow': 'Синхронизировать сейчас',
-    'libraryHealth.snapshotTitle': 'Последний снимок',
-    'libraryHealth.lastSynced': 'Последняя синхронизация',
-    'libraryHealth.neverSynced': 'Никогда не синхронизировалось.',
-    'libraryHealth.autoSyncEvery': 'Автосинхронизация каждые {0} ч',
-    'systemHealth.librarySync': 'Синхронизация библиотеки',
-    'settings.capDiscovery': 'Открытие',
-    'settings.capImport': 'Импорт',
-    'settings.capLibrarySync': 'Синхронизация библиотеки',
-    'settings.capPlaylistExport': 'Экспорт плейлистов',
-    'settings.capService': 'Сервис',
-    'settings.capSubscriptions': 'Подписки',
-    'settings.connectDeezer': 'Подключить Deezer',
-    'settings.deezerConnectHelp':
-      'Подключите Deezer, чтобы импортировать избранное, подписки и плейлисты.',
-    'settings.importFavorites': 'Импортировать избранное',
-    'settings.importFollowed': 'Импортировать подписки',
-    'settings.integrationCapabilities': 'Возможности интеграций',
-    'settings.tabs.systemHealth': 'Состояние системы',
-    'settings.targetUpdated': 'Цель обновлена',
-    'settings.targetUpdateFailed': 'Не удалось обновить цель',
-    'settings.editTarget': 'Изменить цель',
-    'settings.saveChanges': 'Сохранить изменения',
-    'settings.linkedTarget': 'Связанная цель:',
-    'settings.librarySyncInterval': 'Синхронизация библиотеки',
-    'settings.librarySyncIntervalHours': 'Интервал синхронизации (часы)',
-    'settings.librarySyncIntervalHelp':
-      'Определяет, как часто Digarr обновляет данные синхронизации библиотеки и автоматически заново запускает проверки.',
-    'settings.oidcSso': 'OIDC и SSO',
-    'settings.fieldClientId': 'ID клиента',
-    'settings.fieldClientSecret': 'Секрет клиента',
-    'settings.aiProviderTitle': 'Провайдер AI',
-    'settings.webhookTitle': 'Вебхук',
-    'settings.plexToken': 'Токен Plex',
-    'settings.serviceConnectedToast': '{0} подключен',
-    'settings.serviceUnreachableToast': '{0} недоступен',
-    'settings.serviceSettingsSavedToast': 'Настройки {0} сохранены',
-    'settings.serviceSettingsFailedToast': 'Не удалось сохранить настройки {0}',
-    'settings.spotifyAuthorizationFailed': 'Не удалось начать авторизацию Spotify',
-    'settings.spotifyDisconnectFailed': 'Не удалось отключить Spotify',
-    'settings.spotifyLikedSongsFailed': 'Не удалось запустить импорт понравившихся треков Spotify',
-    'settings.spotifyPlaylistFailed': 'Не удалось импортировать плейлист Spotify',
-    'settings.deezerFavoritesFailed': 'Не удалось запустить импорт избранного Deezer',
-    'settings.deezerFollowedFailed': 'Не удалось запустить импорт подписок Deezer',
-    'settings.deezerDisconnectFailed': 'Не удалось отключить Deezer',
-    'settings.deezerAuthorizationFailed': 'Не удалось начать авторизацию Deezer',
-    'settings.reverseProxyConfigHint':
-      'Используйте PROXY_AUTH_ENABLED и PROXY_AUTH_TRUSTED_PROXIES; ваш прокси должен передавать X-Forwarded-User.',
-    'searchResultCard.addToQueue': 'Добавить в очередь открытий',
-    'searchResultCard.queuing': 'Добавление...',
-    'searchResultCard.queuedSuccess': '{0} добавлен в очередь открытий',
-    'searchResultCard.queuedFailure': 'Не удалось поставить {0} в очередь',
-    'searchResultCard.popularity': '{0}{1} популярности',
-    'admin.restoreSummary': 'Восстановлено {0} строк в {1} таблицах',
-    'admin.restoreSummaryForce':
-      'Восстановлено {0} строк. Повторно введите учетные данные для зашифрованных полей.',
-    'admin.encryptionMismatchMessage': 'Резервная копия создана с другим ключом шифрования.',
-    'admin.encryptionMismatchAffected': 'Эти поля потребуется ввести заново: {0}',
-    'admin.encryptionMismatchFallback':
-      'Некоторые зашифрованные поля, возможно, потребуется ввести заново.',
-    'admin.autoFixInProgress': 'Автоисправление выполняется... ({0}/{1} исправлено)',
-    'errorBoundary.title': 'Что-то пошло не так',
-    'errorBoundary.retry': 'Повторить',
-    'errorBoundary.home': 'На главную',
-    'connectionSuggestions.title': 'Подсказки для {0}',
-    'connectionSuggestions.close': 'Закрыть подсказки',
-    'connectionSuggestions.gotIt': 'Понятно',
-    'connectionSuggestions.skipForNow': 'Пока пропустить',
-    'connectionSuggestions.spotifyDiscoverWeekly': 'Настроить подписку Discover Weekly',
-    'connectionSuggestions.spotifyReleaseRadar': 'Включить подписку Release Radar',
-    'connectionSuggestions.lastfmTopGenreTags': 'Подписаться на ваши основные жанровые теги',
-    'connectionSuggestions.listenbrainzWeeklyJams': 'Включить подписку Weekly Jams',
-    'connectionSuggestions.listenbrainzWeeklyExploration': 'Включить Weekly Exploration',
-    'connectionSuggestions.weeklyDigestPlaylist': 'Создать еженедельный плейлист Digarr Digest',
-    'keyboardShortcuts.title': 'Сочетания клавиш',
-    'keyboardShortcuts.close': 'Закрыть сочетания клавиш',
-    'keyboardShortcuts.nextCard': 'Следующая карточка',
-    'keyboardShortcuts.previousCard': 'Предыдущая карточка',
-    'keyboardShortcuts.approveFocusedCard': 'Одобрить активную карточку',
-    'keyboardShortcuts.rejectFocusedCard': 'Отклонить активную карточку',
-    'keyboardShortcuts.toggleDetailView': 'Открыть / закрыть подробный вид',
-    'keyboardShortcuts.toggleShortcuts': 'Показать / скрыть сочетания клавиш',
-    'keyboardShortcuts.previousCardStack': 'Предыдущая карточка (вид стопкой)',
-    'keyboardShortcuts.nextCardStack': 'Следующая карточка (вид стопкой)',
-    'keyboardShortcuts.disabledWhileTyping':
-      'Сочетания клавиш отключены, когда в фокусе текстовое поле.',
-    'albumPicker.title': 'Выбрать альбомы',
-    'albumPicker.loadFailed': 'Не удалось загрузить альбомы',
-    'albumPicker.selectAll': 'Все',
-    'albumPicker.selectNone': 'Ничего',
-    'albumPicker.selectedCount': 'Выбрано: {0}',
-    'albumPicker.aiPick': 'Выбор AI',
-    'albumPicker.monitor': 'Отслеживать',
-    'albumPicker.monitorSingular': 'Отслеживать {0} альбом',
-    'albumPicker.monitorPlural': 'Отслеживать альбомов: {0}',
-    'approveDialog.title': 'Настройки Lidarr для этого исполнителя',
-    'approveDialog.loadFailed':
-      'Не удалось загрузить профили Lidarr. Проверьте подключение к Lidarr.',
-    'approveDialog.qualityProfile': 'Профиль качества',
-    'approveDialog.metadataProfile': 'Профиль метаданных',
-    'approveDialog.rootFolder': 'Корневая папка',
-    'cardStack.counter': '{0} из {1}',
-    'cardStack.swipeHint': 'Свайп вправо - одобрить, влево - отклонить',
-    'serviceCard.testConnection': 'Проверить подключение',
-    'preview.noPreviewAvailable': 'Предпросмотр для этого исполнителя недоступен',
-    'preview.playbackBlocked':
-      'Воспроизведение заблокировано браузером - попробуйте нажать ещё раз',
-    'errors.auth.invalidCredentials': 'Неверные учётные данные',
-    'errors.auth.notAuthenticated': 'Не аутентифицирован',
-    'errors.auth.passwordIncorrect': 'Текущий пароль неверный',
-    'errors.auth.usernameTaken': 'Имя пользователя уже занято',
-    'errors.target.notFound': 'Цель не найдена',
-    'errors.pipeline.alreadyRunning': 'Запуск пайплайна уже выполняется',
-    'errors.recommendation.notFound': 'Рекомендация не найдена',
-    'errors.subscription.notFound': 'Подписка не найдена',
-    'errors.playlist.notFound': 'Плейлист не найден',
-  },
-  ko: {
-    ...REVIEW_GATE_TRANSLATIONS.ko,
-    'common.importing': '가져오는 중...',
-    'common.current': '현재',
-    'common.custom': '사용자 지정',
-    'common.enabled': '사용됨',
-    'common.unauthorized': '인증되지 않음',
-    'common.adminAccessRequired': '관리자 권한이 필요합니다',
-    'jobHistory.librarySync': '라이브러리 동기화',
-    'playlistForm.dialogCreate': '플레이리스트 만들기',
-    'playlistForm.dialogEdit': '플레이리스트 수정',
-    'playlistForm.titleCreate': '플레이리스트 만들기',
-    'playlistForm.titleEdit': '플레이리스트 수정',
-    'playlistForm.namePlaceholder': '내 주간 믹스',
-    'playlistForm.strategy': '전략',
-    'playlistForm.strategyWeeklyDigestDesc': '최근 발견에서 골라낸 큐레이션 모음',
-    'playlistForm.strategyGenreFocusDesc': '특정 장르를 깊게 파고들기',
-    'playlistForm.strategyMoodMixDesc': '특정 분위기에 맞는 트랙 모음',
-    'playlistForm.strategyRediscoverDesc': '잊고 지냈을지도 모를 아티스트 다시 찾기',
-    'playlistForm.mood': '분위기',
-    'playlistForm.moodPlaceholder': '늦은 밤 드라이브',
-    'playlistForm.trackCount': '트랙 수',
-    'playlistForm.scheduleAutomaticGeneration': '자동 생성을 예약',
-    'playlistForm.nameRequired': '이름은 필수입니다',
-    'playlistForm.genreRequired': '장르 포커스 전략에는 장르가 필요합니다',
-    'playlistForm.moodRequired': '무드 믹스 전략에는 분위기가 필요합니다',
-    'subscriptionForm.dialogCreate': '구독 만들기',
-    'subscriptionForm.dialogEdit': '구독 수정',
-    'subscriptionForm.titleCreate': '구독 만들기',
-    'subscriptionForm.titleEdit': '구독 수정',
-    'subscriptionForm.name': '이름',
-    'subscriptionForm.namePlaceholder': '새로운 Shoegaze Weekly',
-    'subscriptionForm.nameRequired': '이름은 필수입니다',
-    'subscriptionForm.source': '소스',
-    'subscriptionForm.sources': '소스',
-    'subscriptionForm.sourceType': '소스 유형',
-    'subscriptionForm.sourceTypeGenre': '장르',
-    'subscriptionForm.sourceTypeGenreDesc': '특정 장르에서 아티스트를 발견합니다',
-    'subscriptionForm.sourceTypeSimilar': '유사 아티스트',
-    'subscriptionForm.sourceTypeSimilarDesc': '좋아하는 아티스트와 비슷한 아티스트를 찾습니다',
-    'subscriptionForm.sourceTypeDiscoveryModeDesc': '발견 모드를 사용해 아티스트를 생성합니다',
-    'subscriptionForm.discoveryMode': '발견 모드',
-    'subscriptionForm.selectDiscoveryMode': '구성할 발견 모드를 선택하세요.',
-    'subscriptionForm.discoveryModeSettingsRequired': '발견 모드 설정은 필수입니다',
-    'subscriptionForm.genre': '장르',
-    'subscriptionForm.genrePlaceholder': 'shoegaze',
-    'subscriptionForm.genreRequired': '장르는 필수입니다',
-    'subscriptionForm.seedArtists': '시드 아티스트',
-    'subscriptionForm.seedArtistsPlaceholder': 'Radiohead, Portishead, Massive Attack',
-    'subscriptionForm.seedArtistsHelp': '쉼표로 구분한 아티스트 이름',
-    'subscriptionForm.seedArtistsRequired': '최소 한 명의 시드 아티스트가 필요합니다',
-    'subscriptionForm.configureInSettings': '설정 > 연결에서 구성하세요',
-    'subscriptionForm.notConfigured': '구성되지 않음',
-    'subscriptionForm.selectAtLeastOneSource': '최소 한 개의 소스를 선택하세요',
-    'subscriptionForm.sourceLockedHelp': '이 구독에서는 소스 유형과 구성을 변경할 수 없습니다.',
-    'subscriptionForm.maxArtistsPerRun': '실행당 최대 아티스트 수',
-    'subscriptionForm.scoringPreset': '점수 프리셋',
-    'subscriptionForm.scoringPresetDefault': '기본값',
-    'subscriptionForm.scoringPresetGenre': '장르 최적화',
-    'subscriptions.spotifyNotConnected': 'Spotify가 연결되지 않았습니다',
-    'subscriptionPresets.getStarted': '프리셋으로 시작하기',
-    'subscriptionPresets.intro': '프리셋을 선택해 구독을 자동으로 만들거나 직접 구성하세요.',
-    'subscriptionPresets.casualTitle': '가볍게',
-    'subscriptionPresets.casualDescription': '잡음이 적은 주간 추천',
-    'subscriptionPresets.deepDiverTitle': '딥 다이버',
-    'subscriptionPresets.deepDiverDescription': '여러 소스를 활용한 폭넓은 발견과 매일 업데이트',
-    'subscriptionPresets.discoveryOnlyTitle': '발견 전용',
-    'subscriptionPresets.discoveryOnlyDescription': '파이프라인만 사용, 외부 피드 없음',
-    'subscriptionPresets.customTitle': '사용자 지정',
-    'subscriptionPresets.customDescription': '나만의 구독 조합 만들기',
-    'subscriptionPresets.oneConnectedService': '연결된 서비스 1개',
-    'subscriptionPresets.twoConnectedServices': '연결된 서비스 2개',
-    'subscriptionPresets.requires': '{0} 필요',
-    'subscriptionPresets.setUp': '설정',
-    'subscriptionPresets.applying': '적용 중...',
-    'subscriptionPresets.casualApplied': '가볍게 프리셋을 적용했습니다',
-    'subscriptionPresets.deepDiverApplied': '딥 다이버 프리셋을 적용했습니다',
-    'subscriptionPresets.applyFailed': '프리셋을 적용하지 못했습니다',
-    'subscriptionPresets.aiOnlyEnabled': 'AI 전용 모드를 켰습니다',
-    'subscriptionPresets.aiOnlyEnableFailed': 'AI 전용 모드를 켜지 못했습니다',
-    'cronPicker.preset': '프리셋',
-    'cronPicker.dailyMidnight': '매일 (자정)',
-    'cronPicker.twiceAWeek': '주 2회 (월 + 목)',
-    'cronPicker.everyTwoWeeks': '2주마다 (1일 + 15일)',
-    'cronPicker.monthlyFirst': '매월 (1일)',
-    'cronPicker.standardHint': '표준 cron: 분 시 일-월 월 요일',
-    'libraryHealth.syncNow': '지금 동기화',
-    'libraryHealth.snapshotTitle': '최신 스냅샷',
-    'libraryHealth.lastSynced': '마지막 동기화',
-    'libraryHealth.neverSynced': '동기화된 적이 없습니다.',
-    'libraryHealth.autoSyncEvery': '{0}시간마다 자동 동기화',
-    'systemHealth.librarySync': '라이브러리 동기화',
-    'settings.tabs.systemHealth': '시스템 상태',
-    'settings.targetUpdated': '대상을 업데이트했습니다',
-    'settings.targetUpdateFailed': '대상 업데이트에 실패했습니다',
-    'settings.editTarget': '대상 편집',
-    'settings.saveChanges': '변경 사항 저장',
-    'settings.linkedTarget': '연결된 대상:',
-    'settings.targetTypeEmbyPlaylist': 'Emby 플레이리스트',
-    'settings.librarySyncInterval': '라이브러리 동기화',
-    'settings.librarySyncIntervalHours': '동기화 간격(시간)',
-    'settings.librarySyncIntervalHelp':
-      'Digarr가 라이브러리 동기화 데이터를 새로 고치고 검사를 자동으로 다시 실행하는 주기를 제어합니다.',
-    'settings.oidcSso': 'OIDC 및 SSO',
-    'settings.fieldClientId': '클라이언트 ID',
-    'settings.fieldClientSecret': '클라이언트 시크릿',
-    'settings.fieldWebhookUrl': '웹훅 URL',
-    'settings.aiProviderTitle': 'AI 제공자',
-    'settings.webhookTitle': '웹훅',
-    'settings.plexToken': 'Plex 토큰',
-    'settings.serviceConnectedToast': '{0} 연결됨',
-    'settings.serviceUnreachableToast': '{0}에 연결할 수 없습니다',
-    'settings.serviceSettingsSavedToast': '{0} 설정을 저장했습니다',
-    'settings.serviceSettingsFailedToast': '{0} 설정 저장에 실패했습니다',
-    'settings.spotifyAuthorizationFailed': 'Spotify 인증 시작에 실패했습니다',
-    'settings.spotifyDisconnectFailed': 'Spotify 연결 해제에 실패했습니다',
-    'settings.spotifyLikedSongsFailed': 'Spotify 좋아요 가져오기를 시작하지 못했습니다',
-    'settings.spotifyPlaylistFailed': 'Spotify 플레이리스트를 가져오지 못했습니다',
-    'settings.deezerFavoritesFailed': 'Deezer 즐겨찾기 가져오기를 시작하지 못했습니다',
-    'settings.deezerFollowedFailed': 'Deezer 팔로우 아티스트 가져오기를 시작하지 못했습니다',
-    'settings.deezerDisconnectFailed': 'Deezer 연결 해제에 실패했습니다',
-    'settings.deezerAuthorizationFailed': 'Deezer 인증 시작에 실패했습니다',
-    'settings.reverseProxyConfigHint':
-      'PROXY_AUTH_ENABLED 및 PROXY_AUTH_TRUSTED_PROXIES를 사용하세요. 프록시는 X-Forwarded-User를 보내야 합니다.',
-    'searchResultCard.addToQueue': '발견 대기열에 추가',
-    'searchResultCard.queuing': '추가 중...',
-    'searchResultCard.queuedSuccess': '{0}을(를) 발견 대기열에 추가했습니다',
-    'searchResultCard.queuedFailure': '{0}을(를) 대기열에 추가하지 못했습니다',
-    'searchResultCard.popularity': '인기도 {0}{1}',
-    'admin.restoreSummary': '{1}개 테이블에서 {0}개 행을 복원했습니다',
-    'admin.restoreSummaryForce':
-      '{0}개 행을 복원했습니다. 암호화된 필드의 자격 증명을 다시 입력하세요.',
-    'admin.encryptionMismatchMessage': '이 백업은 다른 암호화 키로 생성되었습니다.',
-    'admin.encryptionMismatchAffected': '다음 필드는 다시 입력해야 합니다: {0}',
-    'admin.encryptionMismatchFallback': '일부 암호화된 필드는 다시 입력해야 할 수 있습니다.',
-    'admin.autoFixInProgress': '자동 수정 진행 중... ({0}/{1} 수정됨)',
-    'errorBoundary.title': '문제가 발생했습니다',
-    'errorBoundary.retry': '다시 시도',
-    'errorBoundary.home': '홈으로 이동',
-    'connectionSuggestions.title': '{0}에 대한 제안',
-    'connectionSuggestions.close': '제안 닫기',
-    'connectionSuggestions.gotIt': '알겠습니다',
-    'connectionSuggestions.skipForNow': '지금은 건너뛰기',
-    'connectionSuggestions.spotifyDiscoverWeekly': 'Discover Weekly 구독 설정',
-    'connectionSuggestions.spotifyReleaseRadar': 'Release Radar 구독 사용',
-    'connectionSuggestions.lastfmTopGenreTags': '주요 장르 태그 구독',
-    'connectionSuggestions.listenbrainzWeeklyJams': 'Weekly Jams 구독 사용',
-    'connectionSuggestions.listenbrainzWeeklyExploration': 'Weekly Exploration 사용',
-    'connectionSuggestions.weeklyDigestPlaylist': '주간 Digarr Digest 플레이리스트 만들기',
-    'keyboardShortcuts.title': '키보드 단축키',
-    'keyboardShortcuts.close': '단축키 닫기',
-    'keyboardShortcuts.nextCard': '다음 카드',
-    'keyboardShortcuts.previousCard': '이전 카드',
-    'keyboardShortcuts.approveFocusedCard': '선택된 카드 승인',
-    'keyboardShortcuts.rejectFocusedCard': '선택된 카드 거절',
-    'keyboardShortcuts.toggleDetailView': '상세 보기 열기 / 닫기',
-    'keyboardShortcuts.toggleShortcuts': '단축키 표시 / 숨기기',
-    'keyboardShortcuts.previousCardStack': '이전 카드 (스택 보기)',
-    'keyboardShortcuts.nextCardStack': '다음 카드 (스택 보기)',
-    'keyboardShortcuts.disabledWhileTyping':
-      '텍스트 입력란에 포커스가 있으면 단축키가 비활성화됩니다.',
-    'albumPicker.title': '앨범 선택',
-    'albumPicker.loadFailed': '앨범을 불러오지 못했습니다',
-    'albumPicker.selectAll': '모두',
-    'albumPicker.selectNone': '없음',
-    'albumPicker.selectedCount': '{0}개 선택됨',
-    'albumPicker.aiPick': 'AI 추천',
-    'albumPicker.monitor': '모니터링',
-    'albumPicker.monitorSingular': '{0}개 앨범 모니터링',
-    'albumPicker.monitorPlural': '{0}개 앨범 모니터링',
-    'approveDialog.title': '이 아티스트의 Lidarr 설정',
-    'approveDialog.loadFailed': 'Lidarr 프로필을 불러오지 못했습니다. Lidarr 연결을 확인하세요.',
-    'approveDialog.qualityProfile': '품질 프로필',
-    'approveDialog.metadataProfile': '메타데이터 프로필',
-    'approveDialog.rootFolder': '루트 폴더',
-    'cardStack.counter': '{1} 중 {0}',
-    'cardStack.swipeHint': '오른쪽으로 밀어 승인, 왼쪽으로 밀어 거절',
-    'serviceCard.testConnection': '연결 테스트',
-    'preview.noPreviewAvailable': '이 아티스트에 대한 미리듣기를 사용할 수 없습니다',
-    'preview.playbackBlocked': '브라우저에서 재생이 차단되었습니다. 다시 클릭해 보세요',
-    'errors.auth.invalidCredentials': '잘못된 자격 증명입니다',
-    'errors.auth.notAuthenticated': '인증되지 않았습니다',
-    'errors.auth.passwordIncorrect': '현재 비밀번호가 올바르지 않습니다',
-    'errors.auth.usernameTaken': '이미 사용 중인 사용자 이름입니다',
-    'errors.target.notFound': '대상을 찾을 수 없습니다',
-    'errors.pipeline.alreadyRunning': '파이프라인 실행이 이미 진행 중입니다',
-    'errors.recommendation.notFound': '추천을 찾을 수 없습니다',
-    'errors.subscription.notFound': '구독을 찾을 수 없습니다',
-    'errors.playlist.notFound': '재생목록을 찾을 수 없습니다',
-  },
-  'zh-CN': {
-    ...REVIEW_GATE_TRANSLATIONS['zh-CN'],
-    'common.importing': '导入中...',
-    'common.current': '当前',
-    'common.custom': '自定义',
-    'common.enabled': '已启用',
-    'common.unauthorized': '未授权',
-    'common.adminAccessRequired': '需要管理员权限',
-    'jobHistory.pipeline': '流水线',
-    'jobHistory.librarySync': '资料库同步',
-    'playlistForm.dialogCreate': '创建播放列表',
-    'playlistForm.dialogEdit': '编辑播放列表',
-    'playlistForm.titleCreate': '创建播放列表',
-    'playlistForm.titleEdit': '编辑播放列表',
-    'playlistForm.namePlaceholder': '我的每周混音',
-    'playlistForm.strategy': '策略',
-    'playlistForm.strategyWeeklyDigestDesc': '从你最近的发现中精选整理',
-    'playlistForm.strategyGenreFocusDesc': '深入探索某个特定流派',
-    'playlistForm.strategyMoodMixDesc': '符合特定氛围的曲目集合',
-    'playlistForm.strategyRediscoverDesc': '重新挖掘你可能已经淡忘的艺人',
-    'playlistForm.mood': '氛围',
-    'playlistForm.moodPlaceholder': '深夜开车',
-    'playlistForm.trackCount': '曲目数量',
-    'playlistForm.scheduleAutomaticGeneration': '安排自动生成',
-    'playlistForm.nameRequired': '名称为必填项',
-    'playlistForm.genreRequired': '流派聚焦策略必须填写流派',
-    'playlistForm.moodRequired': 'Mood Mix 策略必须填写氛围',
-    'subscriptionForm.dialogCreate': '创建订阅',
-    'subscriptionForm.dialogEdit': '编辑订阅',
-    'subscriptionForm.titleCreate': '创建订阅',
-    'subscriptionForm.titleEdit': '编辑订阅',
-    'subscriptionForm.name': '名称',
-    'subscriptionForm.namePlaceholder': '新的 Shoegaze Weekly',
-    'subscriptionForm.nameRequired': '名称为必填项',
-    'subscriptionForm.source': '来源',
-    'subscriptionForm.sources': '来源',
-    'subscriptionForm.sourceType': '来源类型',
-    'subscriptionForm.sourceTypeGenre': '流派',
-    'subscriptionForm.sourceTypeGenreDesc': '在特定流派中发现艺人',
-    'subscriptionForm.sourceTypeSimilar': '相似艺人',
-    'subscriptionForm.sourceTypeSimilarDesc': '查找与你喜欢的艺人相似的艺人',
-    'subscriptionForm.sourceTypeDiscoveryModeDesc': '使用发现模式来生成艺人',
-    'subscriptionForm.discoveryMode': '发现模式',
-    'subscriptionForm.selectDiscoveryMode': '选择一个发现模式进行配置。',
-    'subscriptionForm.discoveryModeSettingsRequired': '发现模式设置为必填项',
-    'subscriptionForm.genre': '流派',
-    'subscriptionForm.genrePlaceholder': 'shoegaze',
-    'subscriptionForm.genreRequired': '流派为必填项',
-    'subscriptionForm.seedArtists': '种子艺人',
-    'subscriptionForm.seedArtistsPlaceholder': 'Radiohead, Portishead, Massive Attack',
-    'subscriptionForm.seedArtistsHelp': '用逗号分隔艺人名称',
-    'subscriptionForm.seedArtistsRequired': '至少需要一位种子艺人',
-    'subscriptionForm.configureInSettings': '请在 设置 > 连接 中配置',
-    'subscriptionForm.notConfigured': '未配置',
-    'subscriptionForm.selectAtLeastOneSource': '至少选择一个来源',
-    'subscriptionForm.sourceLockedHelp': '此订阅的来源类型和配置无法更改。',
-    'subscriptionForm.maxArtistsPerRun': '每次运行的最大艺人数',
-    'subscriptionForm.scoringPreset': '评分预设',
-    'subscriptionForm.scoringPresetDefault': '默认',
-    'subscriptionForm.scoringPresetGenre': '流派优化',
-    'subscriptions.spotifyNotConnected': 'Spotify 尚未连接',
-    'subscriptionPresets.getStarted': '从预设开始',
-    'subscriptionPresets.intro': '选择一个预设来自动创建订阅，或自行搭建自定义配置。',
-    'subscriptionPresets.casualTitle': '轻松模式',
-    'subscriptionPresets.casualDescription': '低噪音的每周推荐',
-    'subscriptionPresets.deepDiverTitle': '深度探索',
-    'subscriptionPresets.deepDiverDescription': '多来源广泛发现，并提供每日更新',
-    'subscriptionPresets.discoveryOnlyTitle': '仅发现',
-    'subscriptionPresets.discoveryOnlyDescription': '仅运行流水线，不使用外部订阅源',
-    'subscriptionPresets.customTitle': '自定义',
-    'subscriptionPresets.customDescription': '构建你自己的订阅组合',
-    'subscriptionPresets.oneConnectedService': '1 个已连接服务',
-    'subscriptionPresets.twoConnectedServices': '2 个已连接服务',
-    'subscriptionPresets.requires': '需要 {0}',
-    'subscriptionPresets.setUp': '设置',
-    'subscriptionPresets.applying': '应用中...',
-    'subscriptionPresets.casualApplied': '已应用轻松模式预设',
-    'subscriptionPresets.deepDiverApplied': '已应用深度探索预设',
-    'subscriptionPresets.applyFailed': '无法应用预设',
-    'subscriptionPresets.aiOnlyEnabled': '已启用仅 AI 模式',
-    'subscriptionPresets.aiOnlyEnableFailed': '无法启用仅 AI 模式',
-    'cronPicker.preset': '预设',
-    'cronPicker.dailyMidnight': '每天（午夜）',
-    'cronPicker.twiceAWeek': '每周两次（周一 + 周四）',
-    'cronPicker.everyTwoWeeks': '每两周（1日 + 15日）',
-    'cronPicker.monthlyFirst': '每月（1日）',
-    'cronPicker.standardHint': '标准 cron：分钟 小时 月内日期 月 星期',
-    'libraryHealth.syncNow': '立即同步',
-    'libraryHealth.snapshotTitle': '最新快照',
-    'libraryHealth.lastSynced': '上次同步',
-    'libraryHealth.neverSynced': '从未同步。',
-    'libraryHealth.autoSyncEvery': '每 {0} 小时自动同步',
-    'systemHealth.librarySync': '资料库同步',
-    'settings.tabs.systemHealth': '系统健康',
-    'settings.targetUpdated': '目标已更新',
-    'settings.targetUpdateFailed': '无法更新目标',
-    'settings.editTarget': '编辑目标',
-    'settings.saveChanges': '保存更改',
-    'settings.linkedTarget': '已关联目标：',
-    'settings.librarySyncInterval': '资料库同步',
-    'settings.librarySyncIntervalHours': '同步间隔（小时）',
-    'settings.librarySyncIntervalHelp': '控制 Digarr 刷新资料库同步数据并自动重新运行检查的频率。',
-    'settings.oidcSso': 'OIDC 与 SSO',
-    'settings.fieldClientId': '客户端 ID',
-    'settings.fieldClientSecret': '客户端密钥',
-    'settings.fieldWebhookUrl': 'Webhook 地址',
-    'settings.aiProviderTitle': 'AI 提供商',
-    'settings.webhookTitle': 'Webhook 集成',
-    'settings.plexToken': 'Plex 令牌',
-    'settings.serviceConnectedToast': '{0} 已连接',
-    'settings.serviceUnreachableToast': '无法访问 {0}',
-    'settings.serviceSettingsSavedToast': '{0} 设置已保存',
-    'settings.serviceSettingsFailedToast': '无法保存 {0} 设置',
-    'settings.spotifyAuthorizationFailed': '无法开始 Spotify 授权',
-    'settings.spotifyDisconnectFailed': '无法断开 Spotify 连接',
-    'settings.spotifyLikedSongsFailed': '无法开始导入 Spotify 点赞歌曲',
-    'settings.spotifyPlaylistFailed': '无法导入 Spotify 播放列表',
-    'settings.deezerFavoritesFailed': '无法开始导入 Deezer 收藏',
-    'settings.deezerFollowedFailed': '无法开始导入 Deezer 已关注艺人',
-    'settings.deezerDisconnectFailed': '无法断开 Deezer 连接',
-    'settings.deezerAuthorizationFailed': '无法开始 Deezer 授权',
-    'settings.reverseProxyConfigHint':
-      '使用 PROXY_AUTH_ENABLED 和 PROXY_AUTH_TRUSTED_PROXIES；你的代理必须发送 X-Forwarded-User。',
-    'searchResultCard.addToQueue': '加入发现队列',
-    'searchResultCard.queuing': '加入中...',
-    'searchResultCard.queuedSuccess': '已将 {0} 加入发现队列',
-    'searchResultCard.queuedFailure': '无法将 {0} 加入队列',
-    'searchResultCard.popularity': '热度 {0}{1}',
-    'admin.restoreSummary': '已在 {1} 张表中恢复 {0} 行',
-    'admin.restoreSummaryForce': '已恢复 {0} 行。请重新输入加密字段的凭据。',
-    'admin.encryptionMismatchMessage': '该备份使用了不同的加密密钥创建。',
-    'admin.encryptionMismatchAffected': '这些字段需要重新输入：{0}',
-    'admin.encryptionMismatchFallback': '部分加密字段可能需要重新输入。',
-    'admin.autoFixInProgress': '自动修复进行中...（已修复 {0}/{1}）',
-    'errorBoundary.title': '出了点问题',
-    'errorBoundary.retry': '重试',
-    'errorBoundary.home': '返回首页',
-    'connectionSuggestions.title': '{0} 的建议',
-    'connectionSuggestions.close': '关闭建议',
-    'connectionSuggestions.gotIt': '知道了',
-    'connectionSuggestions.skipForNow': '暂时跳过',
-    'connectionSuggestions.spotifyDiscoverWeekly': '设置 Discover Weekly 订阅',
-    'connectionSuggestions.spotifyReleaseRadar': '启用 Release Radar 订阅',
-    'connectionSuggestions.lastfmTopGenreTags': '订阅你最常用的流派标签',
-    'connectionSuggestions.listenbrainzWeeklyJams': '启用 Weekly Jams 订阅',
-    'connectionSuggestions.listenbrainzWeeklyExploration': '启用 Weekly Exploration',
-    'connectionSuggestions.weeklyDigestPlaylist': '创建每周 Digarr Digest 播放列表',
-    'keyboardShortcuts.title': '键盘快捷键',
-    'keyboardShortcuts.close': '关闭快捷键',
-    'keyboardShortcuts.nextCard': '下一张卡片',
-    'keyboardShortcuts.previousCard': '上一张卡片',
-    'keyboardShortcuts.approveFocusedCard': '批准当前卡片',
-    'keyboardShortcuts.rejectFocusedCard': '拒绝当前卡片',
-    'keyboardShortcuts.toggleDetailView': '打开 / 关闭详情视图',
-    'keyboardShortcuts.toggleShortcuts': '显示 / 隐藏快捷键',
-    'keyboardShortcuts.previousCardStack': '上一张卡片（堆叠视图）',
-    'keyboardShortcuts.nextCardStack': '下一张卡片（堆叠视图）',
-    'keyboardShortcuts.disabledWhileTyping': '当文本输入框处于焦点状态时，快捷键会被禁用。',
-    'albumPicker.title': '选择专辑',
-    'albumPicker.loadFailed': '加载专辑失败',
-    'albumPicker.selectAll': '全部',
-    'albumPicker.selectNone': '无',
-    'albumPicker.selectedCount': '已选择 {0} 个',
-    'albumPicker.aiPick': 'AI 精选',
-    'albumPicker.monitor': '监控',
-    'albumPicker.monitorSingular': '监控 {0} 张专辑',
-    'albumPicker.monitorPlural': '监控 {0} 张专辑',
-    'approveDialog.title': '此艺术家的 Lidarr 设置',
-    'approveDialog.loadFailed': '加载 Lidarr 配置文件失败。请检查 Lidarr 连接。',
-    'approveDialog.qualityProfile': '质量配置文件',
-    'approveDialog.metadataProfile': '元数据配置文件',
-    'approveDialog.rootFolder': '根目录',
-    'cardStack.counter': '第 {0} / 共 {1}',
-    'cardStack.swipeHint': '向右滑动批准，向左滑动拒绝',
-    'serviceCard.testConnection': '测试连接',
-    'preview.noPreviewAvailable': '该艺术家没有可用的预览',
-    'preview.playbackBlocked': '浏览器阻止了播放 - 请再次点击',
-    'errors.auth.invalidCredentials': '凭据无效',
-    'errors.auth.notAuthenticated': '未认证',
-    'errors.auth.passwordIncorrect': '当前密码不正确',
-    'errors.auth.usernameTaken': '用户名已被占用',
-    'errors.target.notFound': '未找到目标',
-    'errors.pipeline.alreadyRunning': '已有流水线正在运行',
-    'errors.recommendation.notFound': '未找到推荐',
-    'errors.subscription.notFound': '未找到订阅',
-    'errors.playlist.notFound': '未找到播放列表',
-  },
-}
