+++ src/core/i18n/messages/overrides/de.ts
@@ -1,0 +1,251 @@
+import type { MessageCatalog } from '../types'
+
+export const deOverrides: Partial<MessageCatalog> = {
+  'common.confirm': 'Bestätigen',
+  'discover.filters': 'Filter',
+  'discover.filtersActive': 'Filter ({0})',
+  'discover.moreActions': 'Weitere Aktionen',
+  'discover.decadesLabel': 'Jahrzehnte',
+  'analytics.title': 'Analysen',
+  'dashboard.title': 'Übersicht',
+  'discover.title': 'Entdecken',
+  'setup.title': 'Einrichtung',
+  'todaysPick.discoverSimilar': 'Ähnliche entdecken',
+  'libraryHealth.fix': 'Beheben',
+  'libraryHealth.fixing': 'Wird behoben...',
+  'libraryHealth.showLess': 'Weniger anzeigen',
+  'libraryHealth.showAll': 'Alle {0} anzeigen',
+  'libraryHealth.fixHint.missingMetadata': 'Aktualisiert Künstler-Metadaten aus Lidarr',
+  'libraryHealth.fixHint.unmonitored': 'Setzt alle aufgeführten Künstler in Lidarr auf überwacht',
+  'libraryHealth.fixHint.missingAlbums': 'Aktualisiert Alben in Lidarr, um fehlende zu finden',
+  'libraryHealth.fixHint.genreGaps': 'Aktualisiert Lidarr-Metadaten, um Genres abzurufen',
+  'libraryHealth.fixHint.imageGaps': 'Sucht Künstlerbilder erneut',
+  'libraryHealth.fixHint.missingWikidata': 'Holt Biografien und Links aus Wikidata',
+  'app.themeModeSystem': 'Vom System',
+  'app.themeGroupStreaming': 'Streaming-Dienste',
+  'nav.genres': 'Musikstile',
+  'common.unauthorized': 'Nicht autorisiert',
+  'common.adminAccessRequired': 'Adminzugriff erforderlich',
+  'common.importing': 'Import wird ausgeführt...',
+  'common.optional': 'nach Bedarf',
+  'common.current': 'Aktuell',
+  'common.custom': 'Benutzerdefiniert',
+  'common.enabled': 'Aktiviert',
+  'common.status': 'Zustand',
+  'genreDetail.tabLibrary': 'In Ihrer Bibliothek',
+  'jobHistory.pipeline': 'Entdeckungslauf',
+  'jobHistory.librarySync': 'Bibliothekssynchronisierung',
+  'jobHistory.playlistType': 'Wiedergabeliste',
+  'playlistForm.dialogCreate': 'Playlist erstellen',
+  'playlistForm.dialogEdit': 'Playlist bearbeiten',
+  'playlistForm.titleCreate': 'Playlist erstellen',
+  'playlistForm.titleEdit': 'Playlist bearbeiten',
+  'playlistForm.namePlaceholder': 'Mein Wochenmix',
+  'playlistForm.strategy': 'Strategie',
+  'playlistForm.strategyWeeklyDigestDesc': 'Kuratiere Auswahl aus Ihren aktuellen Entdeckungen',
+  'playlistForm.strategyGenreFocusDesc': 'Ein bestimmtes Genre gezielt tiefer erkunden',
+  'playlistForm.strategyMoodMixDesc': 'Titel, die zu einer bestimmten Stimmung passen',
+  'playlistForm.strategyRediscoverDesc':
+    'Künstler wiederfinden, die Sie vielleicht vergessen haben',
+  'playlistForm.mood': 'Stimmung',
+  'playlistForm.moodPlaceholder': 'nächtliche Fahrt',
+  'playlistForm.trackCount': 'Anzahl der Titel',
+  'playlistForm.scheduleAutomaticGeneration': 'Automatische Generierung planen',
+  'playlistForm.nameRequired': 'Name ist erforderlich',
+  'playlistForm.genreRequired': 'Das Genre ist für die Strategie Genre-Fokus erforderlich',
+  'playlistForm.moodRequired': 'Die Stimmung ist fuer die Strategie Mood Mix erforderlich',
+  'subscriptionForm.dialogCreate': 'Abo erstellen',
+  'subscriptionForm.dialogEdit': 'Abo bearbeiten',
+  'subscriptionForm.titleCreate': 'Abo erstellen',
+  'subscriptionForm.titleEdit': 'Abo bearbeiten',
+  'subscriptionForm.name': 'Bezeichnung',
+  'subscriptionForm.namePlaceholder': 'Neues Shoegaze Weekly',
+  'subscriptionForm.nameRequired': 'Name ist erforderlich',
+  'subscriptionForm.source': 'Quelle',
+  'subscriptionForm.sources': 'Quellen',
+  'subscriptionForm.sourceType': 'Quellentyp',
+  'subscriptionForm.sourceTypeGenre': 'Musikgenre',
+  'subscriptionForm.sourceTypeGenreDesc': 'Neue Künstler in einem bestimmten Genre entdecken',
+  'subscriptionForm.sourceTypeSimilar': 'Ähnliche',
+  'subscriptionForm.sourceTypeSimilarDesc': 'Künstler finden, die Ihren Favoriten ähneln',
+  'subscriptionForm.sourceTypeDiscoveryModeDesc':
+    'Einen Entdeckungsmodus verwenden, um Künstler zu erzeugen',
+  'subscriptionForm.discoveryMode': 'Entdeckungsmodus',
+  'subscriptionForm.selectDiscoveryMode':
+    'Wählen Sie einen Entdeckungsmodus zur Konfiguration aus.',
+  'subscriptionForm.discoveryModeSettingsRequired':
+    'Die Einstellungen fuer den Entdeckungsmodus sind erforderlich',
+  'subscriptionForm.genre': 'Musikgenre',
+  'subscriptionForm.genrePlaceholder': 'shoegaze',
+  'subscriptionForm.genreRequired': 'Genre ist erforderlich',
+  'subscriptionForm.seedArtists': 'Startkünstler',
+  'subscriptionForm.seedArtistsPlaceholder': 'Radiohead, Portishead, Massive Attack',
+  'subscriptionForm.seedArtistsHelp': 'Künstlernamen durch Kommas getrennt',
+  'subscriptionForm.seedArtistsRequired': 'Mindestens ein Startkünstler ist erforderlich',
+  'subscriptionForm.configureInSettings': 'Unter Einstellungen > Verbindungen konfigurieren',
+  'subscriptionForm.notConfigured': 'nicht konfiguriert',
+  'subscriptionForm.selectAtLeastOneSource': 'Wählen Sie mindestens eine Quelle aus',
+  'subscriptionForm.sourceLockedHelp':
+    'Quellentyp und Konfiguration können für dieses Abo nicht geändert werden.',
+  'subscriptionForm.maxArtistsPerRun': 'Maximale Künstler pro Lauf',
+  'subscriptionForm.scoringPreset': 'Bewertungs-Preset',
+  'subscriptionForm.scoringPresetDefault': 'Standard',
+  'subscriptionForm.scoringPresetGenre': 'Für Genres optimiert',
+  'subscriptions.spotifyNotConnected': 'Spotify ist nicht verbunden',
+  'subscriptionPresets.getStarted': 'Mit einem Preset starten',
+  'subscriptionPresets.intro':
+    'Wählen Sie ein Preset, um Abos automatisch anzulegen, oder bauen Sie Ihre eigene Konfiguration.',
+  'subscriptionPresets.casualTitle': 'Locker',
+  'subscriptionPresets.casualDescription': 'Wöchentliche Empfehlungen mit wenig Rauschen',
+  'subscriptionPresets.deepDiverTitle': 'Tiefgang',
+  'subscriptionPresets.deepDiverDescription':
+    'Breite Entdeckung mit mehreren Quellen und täglichen Aktualisierungen',
+  'subscriptionPresets.discoveryOnlyTitle': 'Nur entdecken',
+  'subscriptionPresets.discoveryOnlyDescription': 'Nur Pipeline, keine externen Feeds',
+  'subscriptionPresets.customTitle': 'Individuell',
+  'subscriptionPresets.customDescription': 'Stellen Sie Ihre eigene Abo-Kombination zusammen',
+  'subscriptionPresets.oneConnectedService': '1 verbundener Dienst',
+  'subscriptionPresets.twoConnectedServices': '2 verbundene Dienste',
+  'subscriptionPresets.requires': 'Erfordert {0}',
+  'subscriptionPresets.setUp': 'Einrichten',
+  'subscriptionPresets.applying': 'Wird angewendet...',
+  'subscriptionPresets.casualApplied': 'Das Preset Locker wurde angewendet',
+  'subscriptionPresets.deepDiverApplied': 'Das Preset Tiefgang wurde angewendet',
+  'subscriptionPresets.applyFailed': 'Preset konnte nicht angewendet werden',
+  'subscriptionPresets.aiOnlyEnabled': 'Nur-KI-Modus wurde aktiviert',
+  'subscriptionPresets.aiOnlyEnableFailed': 'Nur-KI-Modus konnte nicht aktiviert werden',
+  'cronPicker.preset': 'Zeitplanvorlage',
+  'cronPicker.dailyMidnight': 'Täglich (Mitternacht)',
+  'cronPicker.twiceAWeek': 'Zweimal pro Woche (Mo + Do)',
+  'cronPicker.everyTwoWeeks': 'Alle 2 Wochen (1. + 15.)',
+  'cronPicker.monthlyFirst': 'Monatlich (Tag 1)',
+  'cronPicker.standardHint': 'Standard-Cron: Minute Stunde Tag-des-Monats Monat Wochentag',
+  'recommendation.sourceGenre': 'Musikgenre',
+  'recommendation.streaming': 'Streaming-Dienst',
+  'libraryHealth.syncNow': 'Jetzt synchronisieren',
+  'libraryHealth.snapshotTitle': 'Letzte Momentaufnahme',
+  'libraryHealth.lastSynced': 'Zuletzt synchronisiert',
+  'libraryHealth.neverSynced': 'Noch nie synchronisiert.',
+  'libraryHealth.autoSyncEvery': 'Automatische Synchronisierung alle {0} Std.',
+  'settings.capImport': 'Importe',
+  'systemHealth.pipeline': 'Entdeckungslauf',
+  'systemHealth.playlists': 'Wiedergabelisten',
+  'systemHealth.librarySync': 'Bibliothekssync',
+  'systemHealth.inMinutes': 'in {0} Min.',
+  'systemHealth.inHours': 'in {0} Std.',
+  'discoveryMode.labels.label': 'Bezeichnungen',
+  'discoveryMode.field.feed': 'Feed-Quelle',
+  'discoveryMode.field.limit': 'Obergrenze',
+  'discoveryMode.field.tags': 'Schlagwörter',
+  'discoveryMode.option.weeklyJams': 'Weekly Jams',
+  'settings.tabs.administration': 'Verwaltung',
+  'settings.tabs.systemHealth': 'Systemzustand',
+  'settings.name': 'Bezeichnung',
+  'settings.targetTypeEmbyPlaylist': 'Emby-Playlist',
+  'settings.limits': 'Grenzwerte',
+  'settings.librarySyncInterval': 'Bibliothekssync',
+  'settings.librarySyncIntervalHours': 'Sync-Intervall (Stunden)',
+  'settings.librarySyncIntervalHelp':
+    'Steuert, wie oft Digarr die Bibliothekssync-Daten aktualisiert und Bibliotheksprüfungen erneut startet.',
+  'settings.app': 'Anwendung',
+  'settings.oidcSso': 'OIDC / SSO',
+  'settings.fieldApiKeyOptional': 'API-Schlüssel (optional)',
+  'settings.fieldClientId': 'Client-ID',
+  'settings.fieldClientSecret': 'Client-Secret',
+  'settings.fieldScopes': 'Berechtigungen',
+  'settings.aiProviderTitle': 'KI-Anbieter',
+  'settings.webhookTitle': 'Webhook-Integration',
+  'settings.plexToken': 'Plex-Token',
+  'settings.serviceConnectedToast': '{0} verbunden',
+  'settings.serviceUnreachableToast': '{0} ist nicht erreichbar',
+  'settings.serviceSettingsSavedToast': '{0}-Einstellungen gespeichert',
+  'settings.serviceSettingsFailedToast': '{0}-Einstellungen konnten nicht gespeichert werden',
+  'settings.spotifyAuthorizationFailed': 'Spotify-Autorisierung konnte nicht gestartet werden',
+  'settings.spotifyDisconnectFailed': 'Spotify konnte nicht getrennt werden',
+  'settings.spotifyLikedSongsFailed':
+    'Import der Spotify-Lieblingssongs konnte nicht gestartet werden',
+  'settings.spotifyPlaylistFailed': 'Playlist-Import konnte nicht gestartet werden',
+  'settings.deezerFavoritesFailed': 'Import der Deezer-Favoriten konnte nicht gestartet werden',
+  'settings.deezerFollowedFailed': 'Import gefolgter Deezer-Künstler konnte nicht gestartet werden',
+  'settings.deezerDisconnectFailed': 'Deezer konnte nicht getrennt werden',
+  'settings.deezerAuthorizationFailed': 'Deezer-Autorisierung konnte nicht gestartet werden',
+  'settings.reverseProxyConfigHint':
+    'Verwenden Sie PROXY_AUTH_ENABLED und PROXY_AUTH_TRUSTED_PROXIES; Ihr Proxy muss X-Forwarded-User senden.',
+  'searchResultCard.addToQueue': 'Zur Entdeckungswarteschlange hinzufügen',
+  'searchResultCard.queuing': 'Wird hinzugefügt...',
+  'searchResultCard.queuedSuccess': '{0} wurde zur Entdeckungswarteschlange hinzugefügt',
+  'searchResultCard.queuedFailure': '{0} konnte nicht zur Warteschlange hinzugefügt werden',
+  'searchResultCard.popularity': 'Popularität {0}{1}',
+  'admin.autoFixInProgress': 'Automatische Korrektur läuft... ({0}/{1} korrigiert)',
+  'errorBoundary.title': 'Etwas ist schiefgelaufen',
+  'errorBoundary.retry': 'Erneut versuchen',
+  'errorBoundary.home': 'Zur Startseite',
+  'connectionSuggestions.title': 'Vorschläge für {0}',
+  'connectionSuggestions.close': 'Vorschläge schließen',
+  'connectionSuggestions.gotIt': 'Verstanden',
+  'connectionSuggestions.skipForNow': 'Später',
+  'connectionSuggestions.spotifyDiscoverWeekly': 'Ein Discover-Weekly-Abo einrichten',
+  'connectionSuggestions.spotifyReleaseRadar': 'Ein Release-Radar-Abo aktivieren',
+  'connectionSuggestions.lastfmTopGenreTags': 'Ihre wichtigsten Genre-Tags abonnieren',
+  'connectionSuggestions.listenbrainzWeeklyJams': 'Ein Weekly-Jams-Abo aktivieren',
+  'connectionSuggestions.listenbrainzWeeklyExploration': 'Weekly Exploration aktivieren',
+  'connectionSuggestions.weeklyDigestPlaylist':
+    'Eine wöchentliche Digarr-Digest-Playlist erstellen',
+  'keyboardShortcuts.title': 'Tastenkombinationen',
+  'keyboardShortcuts.close': 'Tastenkombinationen schließen',
+  'keyboardShortcuts.nextCard': 'Weiter zur nächsten Karte',
+  'keyboardShortcuts.previousCard': 'Zur vorherigen Karte',
+  'keyboardShortcuts.approveFocusedCard': 'Aktive Karte bestätigen',
+  'keyboardShortcuts.rejectFocusedCard': 'Aktive Karte ablehnen',
+  'keyboardShortcuts.toggleDetailView': 'Detailansicht öffnen / schließen',
+  'keyboardShortcuts.toggleShortcuts': 'Tastenkombinationen ein-/ausblenden',
+  'keyboardShortcuts.previousCardStack': 'Vorherige Karte (Stapelansicht)',
+  'keyboardShortcuts.nextCardStack': 'Nächste Karte (Stapelansicht)',
+  'keyboardShortcuts.disabledWhileTyping':
+    'Tastenkombinationen sind deaktiviert, wenn ein Textfeld fokussiert ist.',
+  'admin.restoreSummary': '{0} Zeilen in {1} Tabellen wiederhergestellt',
+  'admin.restoreSummaryForce':
+    '{0} Zeilen wiederhergestellt. Geben Sie Anmeldedaten für verschlüsselte Felder erneut ein.',
+  'admin.encryptionMismatchMessage':
+    'Das Backup wurde mit einem anderen Verschlüsselungsschlüssel erstellt.',
+  'admin.encryptionMismatchAffected': 'Diese Felder müssen erneut eingegeben werden: {0}',
+  'admin.encryptionMismatchFallback':
+    'Einige verschlüsselte Felder müssen möglicherweise erneut eingegeben werden.',
+  'integration.playlist': 'Wiedergabeliste',
+  'settings.yourAccount': 'Ihr Konto',
+  'settings.lidarrPreferences': 'Ihre Lidarr-Einstellungen',
+  'settings.targetUpdated': 'Ziel aktualisiert',
+  'settings.targetUpdateFailed': 'Ziel konnte nicht aktualisiert werden',
+  'settings.editTarget': 'Ziel bearbeiten',
+  'settings.saveChanges': 'Änderungen speichern',
+  'settings.linkedTarget': 'Verknüpftes Ziel:',
+  'albumPicker.title': 'Alben auswählen',
+  'albumPicker.loadFailed': 'Alben konnten nicht geladen werden',
+  'albumPicker.selectAll': 'Alle',
+  'albumPicker.selectNone': 'Keine',
+  'albumPicker.selectedCount': '{0} ausgewählt',
+  'albumPicker.aiPick': 'KI-Auswahl',
+  'albumPicker.monitor': 'Überwachen',
+  'albumPicker.monitorSingular': '{0} Album überwachen',
+  'albumPicker.monitorPlural': '{0} Alben überwachen',
+  'approveDialog.title': 'Lidarr-Einstellungen für diesen Künstler',
+  'approveDialog.loadFailed':
+    'Lidarr-Profile konnten nicht geladen werden. Prüfen Sie Ihre Lidarr-Verbindung.',
+  'approveDialog.qualityProfile': 'Qualitätsprofil',
+  'approveDialog.metadataProfile': 'Metadatenprofil',
+  'approveDialog.rootFolder': 'Stammordner',
+  'cardStack.counter': '{0} von {1}',
+  'cardStack.swipeHint': 'Nach rechts wischen zum Bestätigen, nach links zum Ablehnen',
+  'serviceCard.testConnection': 'Verbindung testen',
+  'preview.noPreviewAvailable': 'Keine Vorschau für diesen Künstler verfügbar',
+  'preview.playbackBlocked': 'Wiedergabe vom Browser blockiert - bitte erneut klicken',
+  'errors.auth.invalidCredentials': 'Ungültige Anmeldedaten',
+  'errors.auth.notAuthenticated': 'Nicht authentifiziert',
+  'errors.auth.passwordIncorrect': 'Aktuelles Passwort ist falsch',
+  'errors.auth.usernameTaken': 'Benutzername bereits vergeben',
+  'errors.target.notFound': 'Ziel nicht gefunden',
+  'errors.pipeline.alreadyRunning': 'Eine Pipeline-Ausführung läuft bereits',
+  'errors.recommendation.notFound': 'Empfehlung nicht gefunden',
+  'errors.subscription.notFound': 'Abonnement nicht gefunden',
+  'errors.playlist.notFound': 'Wiedergabeliste nicht gefunden',
+}
