+++ src/core/i18n/messages/overrides/es.ts
@@ -1,0 +1,239 @@
+import type { MessageCatalog } from '../types'
+
+export const esOverrides: Partial<MessageCatalog> = {
+  'common.confirm': 'Confirmar',
+  'discover.filters': 'Filtros',
+  'discover.filtersActive': 'Filtros ({0})',
+  'discover.moreActions': 'Más acciones',
+  'discover.decadesLabel': 'Décadas',
+  'analytics.title': 'Analíticas',
+  'dashboard.title': 'Panel',
+  'discover.title': 'Descubrir',
+  'setup.title': 'Configuración',
+  'todaysPick.discoverSimilar': 'Descubrir similares',
+  'libraryHealth.fix': 'Corregir',
+  'libraryHealth.fixing': 'Corrigiendo...',
+  'libraryHealth.showLess': 'Mostrar menos',
+  'libraryHealth.showAll': 'Mostrar todo {0}',
+  'libraryHealth.fixHint.missingMetadata': 'Actualiza metadatos de artistas desde Lidarr',
+  'libraryHealth.fixHint.unmonitored':
+    'Marca todos los artistas listados como monitorizados en Lidarr',
+  'libraryHealth.fixHint.missingAlbums': 'Actualiza álbumes en Lidarr para encontrar faltantes',
+  'libraryHealth.fixHint.genreGaps': 'Actualiza metadatos en Lidarr para obtener géneros',
+  'libraryHealth.fixHint.imageGaps': 'Vuelve a buscar imágenes de artistas',
+  'libraryHealth.fixHint.missingWikidata': 'Obtiene biografías y enlaces desde Wikidata',
+  'common.unauthorized': 'No autorizado',
+  'common.adminAccessRequired': 'Se requiere acceso de administrador',
+  'common.current': 'Actual',
+  'common.custom': 'Personalizado',
+  'common.enabled': 'Activado',
+  'common.importing': 'Importando...',
+  'subscriptionForm.dialogCreate': 'Crear suscripción',
+  'subscriptionForm.dialogEdit': 'Editar suscripción',
+  'subscriptionForm.titleCreate': 'Crear suscripción',
+  'subscriptionForm.titleEdit': 'Editar suscripción',
+  'subscriptionForm.name': 'Nombre',
+  'subscriptionForm.namePlaceholder': 'Nuevo Shoegaze Weekly',
+  'subscriptionForm.nameRequired': 'El nombre es obligatorio',
+  'subscriptionForm.source': 'Fuente',
+  'subscriptionForm.sources': 'Fuentes',
+  'subscriptionForm.sourceType': 'Tipo de fuente',
+  'subscriptionForm.sourceTypeGenre': 'Género',
+  'subscriptionForm.sourceTypeGenreDesc': 'Descubre artistas dentro de un género concreto',
+  'subscriptionForm.sourceTypeSimilar': 'Similares',
+  'subscriptionForm.sourceTypeSimilarDesc': 'Encuentra artistas parecidos a tus favoritos',
+  'subscriptionForm.sourceTypeDiscoveryModeDesc':
+    'Usa un modo de descubrimiento para generar artistas',
+  'subscriptionForm.discoveryMode': 'Modo de descubrimiento',
+  'subscriptionForm.selectDiscoveryMode': 'Selecciona un modo de descubrimiento para configurarlo.',
+  'subscriptionForm.discoveryModeSettingsRequired':
+    'La configuración del modo de descubrimiento es obligatoria',
+  'subscriptionForm.genre': 'Género',
+  'subscriptionForm.genrePlaceholder': 'shoegaze',
+  'subscriptionForm.genreRequired': 'El género es obligatorio',
+  'subscriptionForm.seedArtists': 'Artistas semilla',
+  'subscriptionForm.seedArtistsPlaceholder': 'Radiohead, Portishead, Massive Attack',
+  'subscriptionForm.seedArtistsHelp': 'Nombres de artistas separados por comas',
+  'subscriptionForm.seedArtistsRequired': 'Se requiere al menos un artista semilla',
+  'subscriptionForm.configureInSettings': 'Configuralo en Ajustes > Conexiones',
+  'subscriptionForm.notConfigured': 'sin configurar',
+  'subscriptionForm.selectAtLeastOneSource': 'Selecciona al menos una fuente',
+  'subscriptionForm.sourceLockedHelp':
+    'El tipo de fuente y la configuración no se pueden cambiar para esta suscripción.',
+  'subscriptionForm.maxArtistsPerRun': 'Máximo de artistas por ejecución',
+  'subscriptionForm.scoringPreset': 'Preajuste de puntuación',
+  'subscriptionForm.scoringPresetDefault': 'Predeterminado',
+  'subscriptionForm.scoringPresetGenre': 'Optimizado para género',
+  'cronPicker.preset': 'Preajuste',
+  'cronPicker.dailyMidnight': 'Diario (medianoche)',
+  'cronPicker.twiceAWeek': 'Dos veces por semana (lun + jue)',
+  'cronPicker.everyTwoWeeks': 'Cada 2 semanas (1 y 15)',
+  'cronPicker.monthlyFirst': 'Mensual (día 1)',
+  'cronPicker.standardHint': 'Cron estándar: minuto hora día-del-mes mes día-de-la-semana',
+  'libraryHealth.syncNow': 'Sincronizar ahora',
+  'libraryHealth.snapshotTitle': 'Última instantánea',
+  'libraryHealth.lastSynced': 'Última sincronización',
+  'libraryHealth.neverSynced': 'Nunca se ha sincronizado.',
+  'libraryHealth.autoSyncEvery': 'Sincronización automática cada {0} h',
+  'jobHistory.librarySync': 'Sincronización de biblioteca',
+  'settings.tabs.systemHealth': 'Salud del sistema',
+  'settings.targetUpdated': 'Destino actualizado',
+  'settings.targetUpdateFailed': 'No se pudo actualizar el destino',
+  'settings.editTarget': 'Editar destino',
+  'settings.saveChanges': 'Guardar cambios',
+  'settings.linkedTarget': 'Destino vinculado:',
+  'errorBoundary.title': 'Algo salió mal',
+  'errorBoundary.retry': 'Reintentar',
+  'errorBoundary.home': 'Ir al inicio',
+  'connectionSuggestions.title': 'Sugerencias para {0}',
+  'connectionSuggestions.close': 'Cerrar sugerencias',
+  'connectionSuggestions.gotIt': 'Entendido',
+  'connectionSuggestions.skipForNow': 'Omitir por ahora',
+  'connectionSuggestions.spotifyDiscoverWeekly': 'Configurar una suscripción a Discover Weekly',
+  'connectionSuggestions.spotifyReleaseRadar': 'Activar una suscripción a Release Radar',
+  'connectionSuggestions.lastfmTopGenreTags': 'Suscribirse a tus etiquetas de género principales',
+  'connectionSuggestions.listenbrainzWeeklyJams': 'Activar la suscripción Weekly Jams',
+  'connectionSuggestions.listenbrainzWeeklyExploration': 'Activar Weekly Exploration',
+  'connectionSuggestions.weeklyDigestPlaylist': 'Crear una lista semanal Digarr Digest',
+  'keyboardShortcuts.title': 'Atajos de teclado',
+  'keyboardShortcuts.close': 'Cerrar atajos',
+  'keyboardShortcuts.nextCard': 'Siguiente tarjeta',
+  'keyboardShortcuts.previousCard': 'Tarjeta anterior',
+  'keyboardShortcuts.approveFocusedCard': 'Aprobar la tarjeta activa',
+  'keyboardShortcuts.rejectFocusedCard': 'Rechazar la tarjeta activa',
+  'keyboardShortcuts.toggleDetailView': 'Abrir / cerrar la vista de detalle',
+  'keyboardShortcuts.toggleShortcuts': 'Mostrar / ocultar atajos',
+  'keyboardShortcuts.previousCardStack': 'Tarjeta anterior (vista apilada)',
+  'keyboardShortcuts.nextCardStack': 'Siguiente tarjeta (vista apilada)',
+  'keyboardShortcuts.disabledWhileTyping':
+    'Los atajos se desactivan cuando hay un campo de texto enfocado.',
+  'admin.restoreSummary': 'Se restauraron {0} filas en {1} tablas',
+  'admin.restoreSummaryForce':
+    'Se restauraron {0} filas. Vuelve a introducir las credenciales de los campos cifrados.',
+  'admin.encryptionMismatchMessage': 'La copia se creó con una clave de cifrado distinta.',
+  'admin.encryptionMismatchAffected': 'Estos campos tendrán que volver a introducirse: {0}',
+  'admin.encryptionMismatchFallback':
+    'Algunos campos cifrados podrían requerir que vuelvas a introducirlos.',
+  'subscriptions.spotifyNotConnected': 'Spotify no está conectado',
+  'playlistForm.dialogCreate': 'Crear lista',
+  'playlistForm.dialogEdit': 'Editar lista',
+  'playlistForm.titleCreate': 'Crear lista de reproducción',
+  'playlistForm.titleEdit': 'Editar lista de reproducción',
+  'playlistForm.namePlaceholder': 'Mi mezcla semanal',
+  'playlistForm.strategy': 'Estrategia',
+  'playlistForm.strategyWeeklyDigestDesc': 'Selección curada de tus descubrimientos recientes',
+  'playlistForm.strategyGenreFocusDesc': 'Profundiza en un género concreto',
+  'playlistForm.strategyMoodMixDesc': 'Canciones que encajan con un estado de ánimo concreto',
+  'playlistForm.strategyRediscoverDesc': 'Recupera artistas que quizás habias olvidado',
+  'playlistForm.mood': 'Estado de ánimo',
+  'playlistForm.moodPlaceholder': 'conduciendo de noche',
+  'playlistForm.trackCount': 'Número de canciones',
+  'playlistForm.scheduleAutomaticGeneration': 'Programar generación automática',
+  'playlistForm.nameRequired': 'El nombre es obligatorio',
+  'playlistForm.genreRequired': 'El género es obligatorio para la estrategia Enfoque de género',
+  'playlistForm.moodRequired':
+    'El estado de ánimo es obligatorio para la estrategia Mezcla de estados de ánimo',
+  'searchResultCard.addToQueue': 'Añadir a la cola de descubrimiento',
+  'searchResultCard.queuing': 'Añadiendo...',
+  'searchResultCard.queuedSuccess': '{0} se ha añadido a la cola de descubrimiento',
+  'searchResultCard.queuedFailure': 'No se pudo añadir {0} a la cola',
+  'searchResultCard.popularity': '{0}{1} de popularidad',
+  'subscriptionPresets.getStarted': 'Empieza con un preajuste',
+  'subscriptionPresets.intro':
+    'Elige un preajuste para crear suscripciones automáticamente o monta una configuración personalizada.',
+  'subscriptionPresets.casualTitle': 'Tranquilo',
+  'subscriptionPresets.casualDescription': 'Recomendaciones semanales con poco ruido',
+  'subscriptionPresets.deepDiverTitle': 'Explorador profundo',
+  'subscriptionPresets.deepDiverDescription':
+    'Descubrimiento amplio con varias fuentes y actualizaciones diarias',
+  'subscriptionPresets.discoveryOnlyTitle': 'Solo descubrimiento',
+  'subscriptionPresets.discoveryOnlyDescription': 'Solo pipeline, sin fuentes externas',
+  'subscriptionPresets.customTitle': 'Personalizado',
+  'subscriptionPresets.customDescription': 'Construye tu propia combinación de suscripciones',
+  'subscriptionPresets.oneConnectedService': '1 servicio conectado',
+  'subscriptionPresets.twoConnectedServices': '2 servicios conectados',
+  'subscriptionPresets.requires': 'Requiere {0}',
+  'subscriptionPresets.setUp': 'Configurar',
+  'subscriptionPresets.applying': 'Aplicando...',
+  'subscriptionPresets.casualApplied': 'Se ha aplicado el preajuste Casual',
+  'subscriptionPresets.deepDiverApplied': 'Se ha aplicado el preajuste Explorador profundo',
+  'subscriptionPresets.applyFailed': 'No se pudo aplicar el preajuste',
+  'subscriptionPresets.aiOnlyEnabled': 'Se ha activado el modo solo IA',
+  'subscriptionPresets.aiOnlyEnableFailed': 'No se pudo activar el modo solo IA',
+  'common.manual': 'manual',
+  'jobHistory.pipeline': 'Proceso de descubrimiento',
+  'jobHistory.playlistType': 'Lista de reproducción',
+  'jobHistory.errorLabel': 'Fallo:',
+  'libraryReconciliation.albumsCouldNotBeMatched':
+    'álbumes que no se pudieron asociar automáticamente con MusicBrainz.',
+  'libraryReconciliation.artistsCouldNotBeMatched':
+    'artistas que no se pudieron asociar automáticamente con MusicBrainz.',
+  'playlist.total': 'en total',
+  'common.total': 'Total acumulado',
+  'settings.capDiscovery': 'Descubrimiento',
+  'settings.capImport': 'Importación',
+  'settings.capLibrarySync': 'Sincronización de biblioteca',
+  'settings.capPlaylistExport': 'Exportación de listas',
+  'settings.capService': 'Servicio musical',
+  'settings.capSubscriptions': 'Suscripciones',
+  'systemHealth.pipeline': 'Proceso de descubrimiento',
+  'systemHealth.librarySync': 'Sincronización de biblioteca',
+  'search.experimental': 'En pruebas',
+  'discoveryMode.field.feed': 'Origen del feed',
+  'settings.totalWeight': 'peso total:',
+  'settings.librarySyncInterval': 'Sincronización de biblioteca',
+  'settings.librarySyncIntervalHours': 'Intervalo de sincronización (horas)',
+  'settings.librarySyncIntervalHelp':
+    'Controla con qué frecuencia Digarr actualiza los datos de sincronización de la biblioteca y vuelve a ejecutar las comprobaciones automáticamente.',
+  'settings.aiProviderTitle': 'Proveedor de IA',
+  'settings.webhookTitle': 'Integración webhook',
+  'settings.plexToken': 'Token de Plex',
+  'settings.serviceConnectedToast': '{0} conectado',
+  'settings.serviceUnreachableToast': 'No se pudo acceder a {0}',
+  'settings.serviceSettingsSavedToast': 'Se han guardado los ajustes de {0}',
+  'settings.serviceSettingsFailedToast': 'No se pudieron guardar los ajustes de {0}',
+  'settings.spotifyAuthorizationFailed': 'No se pudo iniciar la autorización de Spotify',
+  'settings.spotifyDisconnectFailed': 'No se pudo desconectar Spotify',
+  'settings.spotifyLikedSongsFailed': 'No se pudo iniciar la importación de Spotify Liked Songs',
+  'settings.spotifyPlaylistFailed': 'No se pudo iniciar la importación de la lista',
+  'settings.deezerFavoritesFailed': 'No se pudo iniciar la importación de favoritos de Deezer',
+  'settings.deezerFollowedFailed':
+    'No se pudo iniciar la importación de artistas seguidos en Deezer',
+  'settings.deezerDisconnectFailed': 'No se pudo desconectar Deezer',
+  'settings.deezerAuthorizationFailed': 'No se pudo iniciar la autorización de Deezer',
+  'settings.reverseProxyConfigHint':
+    'Usa PROXY_AUTH_ENABLED y PROXY_AUTH_TRUSTED_PROXIES; tu proxy debe enviar X-Forwarded-User.',
+  'service.error': 'Fallo',
+  'admin.autoFixInProgress': 'Corrección automática en curso... ({0}/{1} corregidos)',
+  'integration.playlist': 'Lista de reproducción',
+  'auth.registerDescription': 'Crea la primera cuenta para empezar.',
+  'albumPicker.title': 'Seleccionar álbumes',
+  'albumPicker.loadFailed': 'No se pudieron cargar los álbumes',
+  'albumPicker.selectAll': 'Todos',
+  'albumPicker.selectNone': 'Ninguno',
+  'albumPicker.selectedCount': '{0} seleccionados',
+  'albumPicker.aiPick': 'Elección de IA',
+  'albumPicker.monitor': 'Monitorizar',
+  'albumPicker.monitorSingular': 'Monitorizar {0} álbum',
+  'albumPicker.monitorPlural': 'Monitorizar {0} álbumes',
+  'approveDialog.title': 'Ajustes de Lidarr para este artista',
+  'approveDialog.loadFailed':
+    'No se pudieron cargar los perfiles de Lidarr. Revisa tu conexión con Lidarr.',
+  'approveDialog.qualityProfile': 'Perfil de calidad',
+  'approveDialog.metadataProfile': 'Perfil de metadatos',
+  'approveDialog.rootFolder': 'Carpeta raíz',
+  'cardStack.counter': '{0} de {1}',
+  'cardStack.swipeHint': 'Desliza a la derecha para aprobar, a la izquierda para rechazar',
+  'serviceCard.testConnection': 'Probar conexión',
+  'preview.noPreviewAvailable': 'No hay vista previa disponible para este artista',
+  'preview.playbackBlocked': 'Reproducción bloqueada por el navegador; inténtalo de nuevo',
+  'errors.auth.invalidCredentials': 'Credenciales no válidas',
+  'errors.auth.notAuthenticated': 'No autenticado',
+  'errors.auth.passwordIncorrect': 'La contraseña actual es incorrecta',
+  'errors.auth.usernameTaken': 'El nombre de usuario ya está en uso',
+  'errors.target.notFound': 'Destino no encontrado',
+  'errors.pipeline.alreadyRunning': 'Ya hay una ejecución del pipeline en curso',
+  'errors.recommendation.notFound': 'Recomendación no encontrada',
+  'errors.subscription.notFound': 'Suscripción no encontrada',
+  'errors.playlist.notFound': 'Lista de reproducción no encontrada',
+}
