+++ src/core/i18n/messages/overrides/fr.ts
@@ -1,0 +1,244 @@
+import type { MessageCatalog } from '../types'
+
+export const frOverrides: Partial<MessageCatalog> = {
+  'common.confirm': 'Confirmer',
+  'discover.filters': 'Filtres',
+  'discover.filtersActive': 'Filtres ({0})',
+  'discover.moreActions': "Plus d'actions",
+  'discover.decadesLabel': 'Décennies',
+  'analytics.title': 'Analyses',
+  'dashboard.title': 'Tableau de bord',
+  'discover.title': 'Découvrir',
+  'setup.title': 'Configuration',
+  'todaysPick.discoverSimilar': 'Découvrir des similaires',
+  'libraryHealth.fix': 'Corriger',
+  'libraryHealth.fixing': 'Correction...',
+  'libraryHealth.showLess': 'Afficher moins',
+  'libraryHealth.showAll': 'Tout afficher {0}',
+  'libraryHealth.fixHint.missingMetadata': 'Actualise les métadonnées des artistes depuis Lidarr',
+  'libraryHealth.fixHint.unmonitored':
+    'Marque tous les artistes listés comme surveillés dans Lidarr',
+  'libraryHealth.fixHint.missingAlbums':
+    'Actualise les albums dans Lidarr pour trouver les manques',
+  'libraryHealth.fixHint.genreGaps': 'Actualise les métadonnées Lidarr pour récupérer les genres',
+  'libraryHealth.fixHint.imageGaps': "Relance la recherche d'images d'artistes",
+  'libraryHealth.fixHint.missingWikidata': 'Récupère les biographies et liens depuis Wikidata',
+  'app.themeMode': 'Thème',
+  'nav.genres': 'Genres musicaux',
+  'setup.mode': 'Mode choisi',
+  'setup.modelSuggestionGeminiFlash': 'Gemini 2.5 Flash (version stable)',
+  'common.unauthorized': 'Non autorise',
+  'common.adminAccessRequired': 'Acces administrateur requis',
+  'common.importing': 'Importation...',
+  'common.current': 'Actuel',
+  'common.custom': 'Personnalise',
+  'common.enabled': 'Active',
+  'common.date': 'Jour',
+  'common.total': 'Total cumule',
+  'genres.genrePlural': 'styles',
+  'genres.genreSingular': 'style',
+  'jobHistory.pipeline': 'Flux de découverte',
+  'jobHistory.librarySync': 'Synchronisation de la bibliothèque',
+  'jobHistory.playlistType': 'Liste de lecture',
+  'libraryReconciliation.page': 'Page de résultats',
+  'playlistForm.dialogCreate': 'Créer une playlist',
+  'playlistForm.dialogEdit': 'Modifier la playlist',
+  'playlistForm.titleCreate': 'Créer une playlist',
+  'playlistForm.titleEdit': 'Modifier la playlist',
+  'playlistForm.namePlaceholder': 'Mon mix hebdomadaire',
+  'playlistForm.strategy': 'Strategie',
+  'playlistForm.strategyWeeklyDigestDesc': 'Selection soignee à partir de vos découvertes recentes',
+  'playlistForm.strategyGenreFocusDesc': 'Explorer en profondeur un genre precis',
+  'playlistForm.strategyMoodMixDesc': 'Titres adaptes a une ambiance precise',
+  'playlistForm.strategyRediscoverDesc': 'Retrouver des artistes que vous aviez peut-etre oublies',
+  'playlistForm.mood': 'Ambiance',
+  'playlistForm.moodPlaceholder': 'trajet de nuit',
+  'playlistForm.trackCount': 'Nombre de titres',
+  'playlistForm.scheduleAutomaticGeneration': 'Planifier la generation automatique',
+  'playlistForm.nameRequired': 'Le nom est obligatoire',
+  'playlistForm.genreRequired': 'Le genre est obligatoire pour la strategie Focus genre',
+  'playlistForm.moodRequired': "L'ambiance est obligatoire pour la strategie Mix d'ambiance",
+  'subscriptionForm.dialogCreate': 'Créer un abonnement',
+  'subscriptionForm.dialogEdit': "Modifier l'abonnement",
+  'subscriptionForm.titleCreate': 'Créer un abonnement',
+  'subscriptionForm.titleEdit': "Modifier l'abonnement",
+  'subscriptionForm.name': 'Nom',
+  'subscriptionForm.namePlaceholder': 'Nouveau Shoegaze Weekly',
+  'subscriptionForm.nameRequired': 'Le nom est obligatoire',
+  'subscriptionForm.source': 'Origine',
+  'subscriptionForm.sources': 'Origines',
+  'subscriptionForm.sourceType': 'Type de source',
+  'subscriptionForm.sourceTypeGenre': 'Style musical',
+  'subscriptionForm.sourceTypeGenreDesc': 'Découvrir des artistes dans un genre precis',
+  'subscriptionForm.sourceTypeSimilar': 'Similaires',
+  'subscriptionForm.sourceTypeSimilarDesc': 'Trouver des artistes proches de vos favoris',
+  'subscriptionForm.sourceTypeDiscoveryModeDesc':
+    'Utiliser un mode de découverte pour générer des artistes',
+  'subscriptionForm.discoveryMode': 'Mode de découverte',
+  'subscriptionForm.selectDiscoveryMode': 'Selectionnez un mode de découverte a configurer.',
+  'subscriptionForm.discoveryModeSettingsRequired':
+    'Les réglages du mode de découverte sont obligatoires',
+  'subscriptionForm.genre': 'Style musical',
+  'subscriptionForm.genrePlaceholder': 'shoegaze',
+  'subscriptionForm.genreRequired': 'Le genre est obligatoire',
+  'subscriptionForm.seedArtists': 'Artistes de depart',
+  'subscriptionForm.seedArtistsPlaceholder': 'Radiohead, Portishead, Massive Attack',
+  'subscriptionForm.seedArtistsHelp': "Noms d'artistes separes par des virgules",
+  'subscriptionForm.seedArtistsRequired': 'Au moins un artiste de depart est requis',
+  'subscriptionForm.configureInSettings': 'A configurer dans Réglages > Connexions',
+  'subscriptionForm.notConfigured': 'non configuré',
+  'subscriptionForm.selectAtLeastOneSource': 'Selectionnez au moins une source',
+  'subscriptionForm.sourceLockedHelp':
+    'Le type de source et sa configuration ne peuvent pas etre modifies pour cet abonnement.',
+  'subscriptionForm.maxArtistsPerRun': 'Nombre maximal d artistes par execution',
+  'subscriptionForm.scoringPreset': 'Preset de score',
+  'subscriptionForm.scoringPresetDefault': 'Par defaut',
+  'subscriptionForm.scoringPresetGenre': 'Optimise pour le genre',
+  'subscriptions.spotifyNotConnected': 'Spotify n est pas connecte',
+  'subscriptionPresets.getStarted': 'Commencer avec un preset',
+  'subscriptionPresets.intro':
+    'Choisissez un preset pour créer des abonnements automatiquement ou composez votre propre configuration.',
+  'subscriptionPresets.casualTitle': 'Detendu',
+  'subscriptionPresets.casualDescription': 'Recommandations hebdomadaires avec peu de bruit',
+  'subscriptionPresets.deepDiverTitle': 'Exploration poussee',
+  'subscriptionPresets.deepDiverDescription':
+    'Découverte large avec plusieurs sources et des mises a jour quotidiennes',
+  'subscriptionPresets.discoveryOnlyTitle': 'Découverte uniquement',
+  'subscriptionPresets.discoveryOnlyDescription': 'Pipeline seul, sans flux externes',
+  'subscriptionPresets.customTitle': 'Personnalise',
+  'subscriptionPresets.customDescription': 'Composez votre propre combinaison d abonnements',
+  'subscriptionPresets.oneConnectedService': '1 service connecte',
+  'subscriptionPresets.twoConnectedServices': '2 services connectes',
+  'subscriptionPresets.requires': 'Necessite {0}',
+  'subscriptionPresets.setUp': 'Configurer',
+  'subscriptionPresets.applying': 'Application...',
+  'subscriptionPresets.casualApplied': 'Le preset Detendu a été applique',
+  'subscriptionPresets.deepDiverApplied': 'Le preset Exploration poussee a été applique',
+  'subscriptionPresets.applyFailed': "Impossible d'appliquer le preset",
+  'subscriptionPresets.aiOnlyEnabled': 'Le mode IA uniquement a été active',
+  'subscriptionPresets.aiOnlyEnableFailed': "Impossible d'activer le mode IA uniquement",
+  'cronPicker.preset': 'Reglage type',
+  'cronPicker.dailyMidnight': 'Tous les jours (minuit)',
+  'cronPicker.twiceAWeek': 'Deux fois par semaine (lun + jeu)',
+  'cronPicker.everyTwoWeeks': 'Toutes les 2 semaines (1 et 15)',
+  'cronPicker.monthlyFirst': 'Chaque mois (jour 1)',
+  'cronPicker.standardHint': 'Cron standard : minute heure jour-du-mois mois jour-de-la-semaine',
+  'recommendation.sourceGenre': 'Style musical',
+  'libraryHealth.syncNow': 'Synchroniser maintenant',
+  'libraryHealth.snapshotTitle': 'Dernier instantane',
+  'libraryHealth.lastSynced': 'Dernière synchronisation',
+  'libraryHealth.neverSynced': 'Jamais synchronise.',
+  'libraryHealth.autoSyncEvery': 'Synchronisation automatique toutes les {0} h',
+  'systemHealth.pipeline': 'Flux de découverte',
+  'systemHealth.librarySync': 'Sync bibliothèque',
+  'systemHealth.sources': 'Connexions',
+  'discoveryMode.labels.label': 'Libelles',
+  'discoveryMode.field.tags': 'Etiquettes',
+  'libraryStats.albums': 'Albums musicaux',
+  'errorBoundary.title': 'Un problème est survenu',
+  'errorBoundary.retry': 'Réessayer',
+  'errorBoundary.home': "Retour à l'accueil",
+  'connectionSuggestions.title': 'Suggestions pour {0}',
+  'connectionSuggestions.close': 'Fermer les suggestions',
+  'connectionSuggestions.gotIt': 'Compris',
+  'connectionSuggestions.skipForNow': 'Ignorer pour le moment',
+  'connectionSuggestions.spotifyDiscoverWeekly': 'Configurer un abonnement Discover Weekly',
+  'connectionSuggestions.spotifyReleaseRadar': "Activer l'abonnement Release Radar",
+  'connectionSuggestions.lastfmTopGenreTags': 'Vous abonner a vos principaux tags de genre',
+  'connectionSuggestions.listenbrainzWeeklyJams': "Activer l'abonnement Weekly Jams",
+  'connectionSuggestions.listenbrainzWeeklyExploration': 'Activer Weekly Exploration',
+  'connectionSuggestions.weeklyDigestPlaylist': 'Créer une playlist Digarr Digest hebdomadaire',
+  'keyboardShortcuts.title': 'Raccourcis clavier',
+  'keyboardShortcuts.close': 'Fermer les raccourcis',
+  'keyboardShortcuts.nextCard': 'Carte suivante',
+  'keyboardShortcuts.previousCard': 'Carte precedente',
+  'keyboardShortcuts.approveFocusedCard': 'Approuver la carte active',
+  'keyboardShortcuts.rejectFocusedCard': 'Rejeter la carte active',
+  'keyboardShortcuts.toggleDetailView': 'Ouvrir / fermer la vue detaillee',
+  'keyboardShortcuts.toggleShortcuts': 'Afficher / masquer les raccourcis',
+  'keyboardShortcuts.previousCardStack': 'Carte precedente (vue pile)',
+  'keyboardShortcuts.nextCardStack': 'Carte suivante (vue pile)',
+  'keyboardShortcuts.disabledWhileTyping':
+    'Les raccourcis sont desactives lorsqu un champ texte est actif.',
+  'admin.restoreSummary': 'Restauration de {0} lignes dans {1} tables',
+  'admin.restoreSummaryForce':
+    'Restauration de {0} lignes. Ressaisissez les identifiants des champs chiffres.',
+  'admin.encryptionMismatchMessage':
+    'La sauvegarde a été créée avec une clé de chiffrement differente.',
+  'admin.encryptionMismatchAffected': 'Ces champs devront etre ressaisis : {0}',
+  'admin.encryptionMismatchFallback': 'Certains champs chiffres devront peut-etre etre ressaisis.',
+  'admin.autoFixInProgress': 'Correction automatique en cours... ({0}/{1} corriges)',
+  'settings.capDiscovery': 'Découverte',
+  'settings.capImport': 'Importation',
+  'settings.capLibrarySync': 'Sync bibliothèque',
+  'settings.capPlaylistExport': 'Export de playlist',
+  'settings.capService': 'Connecteur',
+  'settings.capSubscriptions': 'Abonnements',
+  'settings.tabs.administration': 'Gestion',
+  'settings.tabs.systemHealth': 'Etat du systeme',
+  'settings.targetUpdated': 'Destination mise a jour',
+  'settings.targetUpdateFailed': 'Échec de la mise a jour de la destination',
+  'settings.editTarget': 'Modifier la destination',
+  'settings.saveChanges': 'Enregistrer les changements',
+  'settings.linkedTarget': 'Destination liee :',
+  'settings.type': 'Categorie',
+  'settings.consensus': 'Consensus global',
+  'settings.librarySyncInterval': 'Sync bibliothèque',
+  'settings.librarySyncIntervalHours': 'Intervalle de sync (heures)',
+  'settings.librarySyncIntervalHelp':
+    'Controle la frequence a laquelle Digarr actualise les donnees de sync bibliothèque et relance les verifications.',
+  'settings.oidcSso': 'OIDC / SSO',
+  'settings.aiProviderTitle': 'Fournisseur IA',
+  'settings.webhookTitle': 'Canal webhook',
+  'settings.plexToken': 'Jeton Plex',
+  'settings.serviceConnectedToast': '{0} connecte',
+  'settings.serviceUnreachableToast': '{0} est inaccessible',
+  'settings.serviceSettingsSavedToast': 'Réglages de {0} enregistres',
+  'settings.serviceSettingsFailedToast': 'Impossible d enregistrer les réglages de {0}',
+  'settings.spotifyAuthorizationFailed': "Impossible de lancer l'autorisation Spotify",
+  'settings.spotifyDisconnectFailed': 'Impossible de deconnecter Spotify',
+  'settings.spotifyLikedSongsFailed': "Impossible de lancer l'import des Titres J aimes Spotify",
+  'settings.spotifyPlaylistFailed': "Impossible de lancer l'import de playlist",
+  'settings.deezerFavoritesFailed': "Impossible de lancer l'import des favoris Deezer",
+  'settings.deezerFollowedFailed': "Impossible de lancer l'import des artistes suivis Deezer",
+  'settings.deezerDisconnectFailed': 'Impossible de deconnecter Deezer',
+  'settings.deezerAuthorizationFailed': "Impossible de lancer l'autorisation Deezer",
+  'settings.reverseProxyConfigHint':
+    'Utilisez PROXY_AUTH_ENABLED et PROXY_AUTH_TRUSTED_PROXIES ; votre proxy doit envoyer X-Forwarded-User.',
+  'searchResultCard.addToQueue': 'Ajouter a la file de découverte',
+  'searchResultCard.queuing': 'Ajout...',
+  'searchResultCard.queuedSuccess': '{0} a été ajouté a la file de découverte',
+  'searchResultCard.queuedFailure': "Impossible d'ajouter {0} a la file",
+  'searchResultCard.popularity': 'Popularite : {0}{1}',
+  'integration.playlist': 'Liste de lecture',
+  'analytics.source.consensus': 'Consensus global',
+  'albumPicker.title': 'Sélectionner les albums',
+  'albumPicker.loadFailed': 'Échec du chargement des albums',
+  'albumPicker.selectAll': 'Tous',
+  'albumPicker.selectNone': 'Aucun',
+  'albumPicker.selectedCount': '{0} sélectionnés',
+  'albumPicker.aiPick': 'Choix de IA',
+  'albumPicker.monitor': 'Surveiller',
+  'albumPicker.monitorSingular': 'Surveiller {0} album',
+  'albumPicker.monitorPlural': 'Surveiller {0} albums',
+  'approveDialog.title': 'Paramètres Lidarr pour cet artiste',
+  'approveDialog.loadFailed':
+    'Échec du chargement des profils Lidarr. Vérifiez votre connexion Lidarr.',
+  'approveDialog.qualityProfile': 'Profil de qualité',
+  'approveDialog.metadataProfile': 'Profil de métadonnées',
+  'approveDialog.rootFolder': 'Dossier racine',
+  'cardStack.counter': '{0} sur {1}',
+  'cardStack.swipeHint': 'Glissez à droite pour approuver, à gauche pour rejeter',
+  'serviceCard.testConnection': 'Tester la connexion',
+  'preview.noPreviewAvailable': 'Aucun aperçu disponible pour cet artiste',
+  'preview.playbackBlocked': 'Lecture bloquée par le navigateur ; essayez de cliquer à nouveau',
+  'errors.auth.invalidCredentials': 'Identifiants non valides',
+  'errors.auth.notAuthenticated': 'Non authentifié',
+  'errors.auth.passwordIncorrect': 'Le mot de passe actuel est incorrect',
+  'errors.auth.usernameTaken': "Nom d'utilisateur déjà pris",
+  'errors.target.notFound': 'Destination introuvable',
+  'errors.pipeline.alreadyRunning': 'Une exécution du pipeline est déjà en cours',
+  'errors.recommendation.notFound': 'Recommandation introuvable',
+  'errors.subscription.notFound': 'Abonnement introuvable',
+  'errors.playlist.notFound': 'Liste de lecture introuvable',
+}
