+++ src/core/i18n/messages/overrides/it.ts
@@ -1,0 +1,238 @@
+import type { MessageCatalog } from '../types'
+
+export const itOverrides: Partial<MessageCatalog> = {
+  'common.confirm': 'Conferma',
+  'discover.filters': 'Filtri',
+  'discover.filtersActive': 'Filtri ({0})',
+  'discover.moreActions': 'Altre azioni',
+  'discover.decadesLabel': 'Decenni',
+  'analytics.title': 'Analisi',
+  'dashboard.title': 'Cruscotto',
+  'discover.title': 'Scopri',
+  'setup.title': 'Configurazione',
+  'todaysPick.discoverSimilar': 'Scopri simili',
+  'libraryHealth.fix': 'Correggi',
+  'libraryHealth.fixing': 'Correzione...',
+  'libraryHealth.showLess': 'Mostra meno',
+  'libraryHealth.showAll': 'Mostra tutti {0}',
+  'libraryHealth.fixHint.missingMetadata': 'Aggiorna i metadati degli artisti da Lidarr',
+  'libraryHealth.fixHint.unmonitored':
+    'Imposta tutti gli artisti elencati come monitorati in Lidarr',
+  'libraryHealth.fixHint.missingAlbums': 'Aggiorna gli album in Lidarr per trovare quelli mancanti',
+  'libraryHealth.fixHint.genreGaps': 'Aggiorna i metadati Lidarr per recuperare i generi',
+  'libraryHealth.fixHint.imageGaps': 'Cerca di nuovo le immagini degli artisti',
+  'libraryHealth.fixHint.missingWikidata': 'Recupera biografie e link da Wikidata',
+  'common.unauthorized': 'Non autorizzato',
+  'common.adminAccessRequired': 'Accesso amministratore richiesto',
+  'common.importing': 'Importazione in corso...',
+  'common.current': 'Attuale',
+  'common.custom': 'Personalizzato',
+  'common.enabled': 'Attivato',
+  'jobHistory.pipeline': 'Pipeline di scoperta',
+  'jobHistory.librarySync': 'Sincronizzazione libreria',
+  'jobHistory.playlistType': 'Lista di riproduzione',
+  'libraryReconciliation.albumsCouldNotBeMatched':
+    'album che non e stato possibile associare automaticamente a MusicBrainz.',
+  'libraryReconciliation.artistsCouldNotBeMatched':
+    'artisti che non e stato possibile associare automaticamente a MusicBrainz.',
+  'playlists.playlistSingular': 'lista di riproduzione',
+  'subscriptionForm.dialogCreate': 'Crea abbonamento',
+  'subscriptionForm.dialogEdit': 'Modifica abbonamento',
+  'subscriptionForm.titleCreate': 'Crea abbonamento',
+  'subscriptionForm.titleEdit': 'Modifica abbonamento',
+  'subscriptionForm.name': 'Nome',
+  'subscriptionForm.namePlaceholder': 'Nuovo Shoegaze Weekly',
+  'subscriptionForm.nameRequired': 'Il nome e obbligatorio',
+  'subscriptionForm.source': 'Origine',
+  'subscriptionForm.sources': 'Origini',
+  'subscriptionForm.sourceType': 'Tipo di sorgente',
+  'subscriptionForm.sourceTypeGenre': 'Genere',
+  'subscriptionForm.sourceTypeGenreDesc': 'Scopri artisti in un genere specifico',
+  'subscriptionForm.sourceTypeSimilar': 'Simili',
+  'subscriptionForm.sourceTypeSimilarDesc': 'Trova artisti simili ai tuoi preferiti',
+  'subscriptionForm.sourceTypeDiscoveryModeDesc':
+    'Usa una modalita di scoperta per generare artisti',
+  'subscriptionForm.discoveryMode': 'Modalita di scoperta',
+  'subscriptionForm.selectDiscoveryMode': 'Seleziona una modalita di scoperta per configurarla.',
+  'subscriptionForm.discoveryModeSettingsRequired':
+    'Le impostazioni della modalita di scoperta sono obbligatorie',
+  'subscriptionForm.genre': 'Genere',
+  'subscriptionForm.genrePlaceholder': 'shoegaze',
+  'subscriptionForm.genreRequired': 'Il genere e obbligatorio',
+  'subscriptionForm.seedArtists': 'Artisti seed',
+  'subscriptionForm.seedArtistsPlaceholder': 'Radiohead, Portishead, Massive Attack',
+  'subscriptionForm.seedArtistsHelp': 'Nomi di artisti separati da virgole',
+  'subscriptionForm.seedArtistsRequired': 'E richiesto almeno un artista seed',
+  'subscriptionForm.configureInSettings': 'Configura in Impostazioni > Connessioni',
+  'subscriptionForm.notConfigured': 'non configurato',
+  'subscriptionForm.selectAtLeastOneSource': 'Seleziona almeno una sorgente',
+  'subscriptionForm.sourceLockedHelp':
+    'Il tipo di sorgente e la configurazione non possono essere modificati per questo abbonamento.',
+  'subscriptionForm.maxArtistsPerRun': 'Artisti massimi per esecuzione',
+  'subscriptionForm.scoringPreset': 'Preset di punteggio',
+  'subscriptionForm.scoringPresetDefault': 'Predefinito',
+  'subscriptionForm.scoringPresetGenre': 'Ottimizzato per genere',
+  'subscriptions.spotifyNotConnected': 'Spotify non e connesso',
+  'subscriptionPresets.getStarted': 'Inizia da un preset',
+  'subscriptionPresets.intro':
+    'Scegli un preset per creare abbonamenti automaticamente oppure costruisci una configurazione personalizzata.',
+  'subscriptionPresets.casualTitle': 'Leggero',
+  'subscriptionPresets.casualDescription': 'Suggerimenti settimanali con poco rumore',
+  'subscriptionPresets.deepDiverTitle': 'Esplorazione profonda',
+  'subscriptionPresets.deepDiverDescription':
+    'Scoperta ampia da più fonti con aggiornamenti quotidiani',
+  'subscriptionPresets.discoveryOnlyTitle': 'Solo scoperta',
+  'subscriptionPresets.discoveryOnlyDescription': 'Solo pipeline, senza feed esterni',
+  'subscriptionPresets.customTitle': 'Personalizzato',
+  'subscriptionPresets.customDescription': 'Crea la tua combinazione di abbonamenti',
+  'subscriptionPresets.oneConnectedService': '1 servizio connesso',
+  'subscriptionPresets.twoConnectedServices': '2 servizi connessi',
+  'subscriptionPresets.requires': 'Richiede {0}',
+  'subscriptionPresets.setUp': 'Configura',
+  'subscriptionPresets.applying': 'Applicazione in corso...',
+  'subscriptionPresets.casualApplied': 'Preset Leggero applicato',
+  'subscriptionPresets.deepDiverApplied': 'Preset Esplorazione profonda applicato',
+  'subscriptionPresets.applyFailed': 'Impossibile applicare il preset',
+  'subscriptionPresets.aiOnlyEnabled': 'Modalita solo AI attivata',
+  'subscriptionPresets.aiOnlyEnableFailed': 'Impossibile attivare la modalita solo AI',
+  'cronPicker.preset': 'Preimpostazione',
+  'cronPicker.dailyMidnight': 'Ogni giorno (mezzanotte)',
+  'cronPicker.twiceAWeek': 'Due volte a settimana (lun + gio)',
+  'cronPicker.everyTwoWeeks': 'Ogni 2 settimane (1 + 15)',
+  'cronPicker.monthlyFirst': 'Mensile (giorno 1)',
+  'cronPicker.standardHint':
+    'Cron standard: minuto ora giorno-del-mese mese giorno-della-settimana',
+  'recommendation.streaming': 'Servizio di streaming',
+  'libraryHealth.syncNow': 'Sincronizza ora',
+  'libraryHealth.snapshotTitle': 'Ultima istantanea',
+  'libraryHealth.lastSynced': 'Ultima sincronizzazione',
+  'libraryHealth.neverSynced': 'Mai sincronizzato.',
+  'libraryHealth.autoSyncEvery': 'Sincronizzazione automatica ogni {0} h',
+  'systemHealth.pipeline': 'Pipeline di scoperta',
+  'systemHealth.librarySync': 'Sync libreria',
+  'discoveryMode.field.feed': 'Feed di origine',
+  'errorBoundary.title': 'Qualcosa e andato storto',
+  'errorBoundary.retry': 'Riprova',
+  'errorBoundary.home': 'Vai alla home',
+  'connectionSuggestions.title': 'Suggerimenti per {0}',
+  'connectionSuggestions.close': 'Chiudi suggerimenti',
+  'connectionSuggestions.gotIt': 'Capito',
+  'connectionSuggestions.skipForNow': 'Salta per ora',
+  'connectionSuggestions.spotifyDiscoverWeekly': 'Configura un abbonamento Discover Weekly',
+  'connectionSuggestions.spotifyReleaseRadar': 'Attiva l abbonamento Release Radar',
+  'connectionSuggestions.lastfmTopGenreTags': 'Abbonati ai tuoi tag di genere principali',
+  'connectionSuggestions.listenbrainzWeeklyJams': 'Attiva l abbonamento Weekly Jams',
+  'connectionSuggestions.listenbrainzWeeklyExploration': 'Attiva Weekly Exploration',
+  'connectionSuggestions.weeklyDigestPlaylist': 'Crea una playlist Digarr Digest settimanale',
+  'keyboardShortcuts.title': 'Scorciatoie da tastiera',
+  'keyboardShortcuts.close': 'Chiudi scorciatoie',
+  'keyboardShortcuts.nextCard': 'Scheda successiva',
+  'keyboardShortcuts.previousCard': 'Scheda precedente',
+  'keyboardShortcuts.approveFocusedCard': 'Approva la scheda attiva',
+  'keyboardShortcuts.rejectFocusedCard': 'Rifiuta la scheda attiva',
+  'keyboardShortcuts.toggleDetailView': 'Apri / chiudi la vista dettagliata',
+  'keyboardShortcuts.toggleShortcuts': 'Mostra / nascondi le scorciatoie',
+  'keyboardShortcuts.previousCardStack': 'Scheda precedente (vista pila)',
+  'keyboardShortcuts.nextCardStack': 'Scheda successiva (vista pila)',
+  'keyboardShortcuts.disabledWhileTyping':
+    'Le scorciatoie sono disattivate quando un campo di testo e attivo.',
+  'admin.restoreSummary': 'Ripristinate {0} righe in {1} tabelle',
+  'admin.restoreSummaryForce':
+    'Ripristinate {0} righe. Reinserisci le credenziali per i campi cifrati.',
+  'admin.encryptionMismatchMessage':
+    'Il backup e stato creato con una chiave di crittografia diversa.',
+  'admin.encryptionMismatchAffected': 'Questi campi dovranno essere reinseriti: {0}',
+  'admin.encryptionMismatchFallback': 'Alcuni campi cifrati potrebbero dover essere reinseriti.',
+  'admin.autoFixInProgress': 'Correzione automatica in corso... ({0}/{1} corretti)',
+  'settings.capDiscovery': 'Scoperta',
+  'settings.capImport': 'Importazione',
+  'settings.capLibrarySync': 'Sync libreria',
+  'settings.capPlaylistExport': 'Esportazione playlist',
+  'settings.capService': 'Servizio',
+  'settings.capSubscriptions': 'Abbonamenti',
+  'settings.tabs.account': 'Account utente',
+  'settings.tabs.systemHealth': 'Stato del sistema',
+  'settings.targetUpdated': 'Destinazione aggiornata',
+  'settings.targetUpdateFailed': 'Impossibile aggiornare la destinazione',
+  'settings.editTarget': 'Modifica destinazione',
+  'settings.saveChanges': 'Salva modifiche',
+  'settings.linkedTarget': 'Destinazione collegata:',
+  'settings.librarySyncInterval': 'Sync libreria',
+  'settings.librarySyncIntervalHours': 'Intervallo di sync (ore)',
+  'settings.librarySyncIntervalHelp':
+    'Controlla con quale frequenza Digarr aggiorna i dati di sincronizzazione della libreria e rilancia i controlli automaticamente.',
+  'settings.app': 'Applicazione',
+  'settings.oidcSso': 'OIDC / SSO',
+  'settings.fieldProvider': 'Fornitore',
+  'settings.aiProviderTitle': 'Provider AI',
+  'settings.webhookTitle': 'Integrazione webhook',
+  'settings.plexToken': 'Token Plex',
+  'settings.serviceConnectedToast': '{0} connesso',
+  'settings.serviceUnreachableToast': '{0} non raggiungibile',
+  'settings.serviceSettingsSavedToast': 'Impostazioni di {0} salvate',
+  'settings.serviceSettingsFailedToast': 'Impossibile salvare le impostazioni di {0}',
+  'settings.spotifyAuthorizationFailed': 'Impossibile avviare l autorizzazione Spotify',
+  'settings.spotifyDisconnectFailed': 'Impossibile disconnettere Spotify',
+  'settings.spotifyLikedSongsFailed': 'Impossibile avviare l import dei brani preferiti di Spotify',
+  'settings.spotifyPlaylistFailed': 'Impossibile avviare l import della playlist',
+  'settings.deezerFavoritesFailed': 'Impossibile avviare l import dei preferiti Deezer',
+  'settings.deezerFollowedFailed': 'Impossibile avviare l import degli artisti seguiti su Deezer',
+  'settings.deezerDisconnectFailed': 'Impossibile disconnettere Deezer',
+  'settings.deezerAuthorizationFailed': 'Impossibile avviare l autorizzazione Deezer',
+  'settings.reverseProxyConfigHint':
+    'Usa PROXY_AUTH_ENABLED e PROXY_AUTH_TRUSTED_PROXIES; il proxy deve inviare X-Forwarded-User.',
+  'searchResultCard.addToQueue': 'Aggiungi alla coda di scoperta',
+  'searchResultCard.queuing': 'Aggiunta in corso...',
+  'searchResultCard.queuedSuccess': '{0} e stato aggiunto alla coda di scoperta',
+  'searchResultCard.queuedFailure': 'Impossibile aggiungere {0} alla coda',
+  'searchResultCard.popularity': 'Popolarita {0}{1}',
+  'integration.playlist': 'Lista di riproduzione',
+  'app.themeGroupStreaming': 'Servizi streaming',
+  'playlistForm.dialogCreate': 'Crea playlist',
+  'playlistForm.dialogEdit': 'Modifica playlist',
+  'playlistForm.titleCreate': 'Crea playlist',
+  'playlistForm.titleEdit': 'Modifica playlist',
+  'playlistForm.namePlaceholder': 'Il mio mix settimanale',
+  'playlistForm.strategy': 'Strategia',
+  'playlistForm.strategyWeeklyDigestDesc': 'Selezione curata dalle tue scoperte recenti',
+  'playlistForm.strategyGenreFocusDesc': 'Approfondisci un genere specifico',
+  'playlistForm.strategyMoodMixDesc': 'Brani adatti a un atmosfera precisa',
+  'playlistForm.strategyRediscoverDesc':
+    'Riporta in primo piano artisti che potresti aver dimenticato',
+  'playlistForm.mood': 'Atmosfera',
+  'playlistForm.moodPlaceholder': 'guida notturna',
+  'playlistForm.trackCount': 'Numero di brani',
+  'playlistForm.scheduleAutomaticGeneration': 'Pianifica la generazione automatica',
+  'playlistForm.nameRequired': 'Il nome e obbligatorio',
+  'playlistForm.genreRequired': 'Il genere e obbligatorio per la strategia Focus sul genere',
+  'playlistForm.moodRequired': 'L atmosfera e obbligatoria per la strategia Mood Mix',
+  'albumPicker.title': 'Seleziona album',
+  'albumPicker.loadFailed': 'Impossibile caricare gli album',
+  'albumPicker.selectAll': 'Tutti',
+  'albumPicker.selectNone': 'Nessuno',
+  'albumPicker.selectedCount': '{0} selezionati',
+  'albumPicker.aiPick': 'Scelta AI',
+  'albumPicker.monitor': 'Monitora',
+  'albumPicker.monitorSingular': 'Monitora {0} album',
+  'albumPicker.monitorPlural': 'Monitora {0} album',
+  'approveDialog.title': 'Impostazioni Lidarr per questo artista',
+  'approveDialog.loadFailed':
+    'Impossibile caricare i profili Lidarr. Controlla la connessione a Lidarr.',
+  'approveDialog.qualityProfile': 'Profilo qualità',
+  'approveDialog.metadataProfile': 'Profilo metadati',
+  'approveDialog.rootFolder': 'Cartella radice',
+  'cardStack.counter': '{0} di {1}',
+  'cardStack.swipeHint': 'Scorri a destra per approvare, a sinistra per rifiutare',
+  'serviceCard.testConnection': 'Testa connessione',
+  'preview.noPreviewAvailable': 'Nessuna anteprima disponibile per questo artista',
+  'preview.playbackBlocked': 'Riproduzione bloccata dal browser; prova a cliccare di nuovo',
+  'errors.auth.invalidCredentials': 'Credenziali non valide',
+  'errors.auth.notAuthenticated': 'Non autenticato',
+  'errors.auth.passwordIncorrect': 'La password attuale non e corretta',
+  'errors.auth.usernameTaken': 'Nome utente gia in uso',
+  'errors.target.notFound': 'Destinazione non trovata',
+  'errors.pipeline.alreadyRunning': 'Una esecuzione della pipeline e gia in corso',
+  'errors.recommendation.notFound': 'Raccomandazione non trovata',
+  'errors.subscription.notFound': 'Abbonamento non trovato',
+  'errors.playlist.notFound': 'Playlist non trovata',
+}
