+++ src/core/i18n/messages/overrides/ja.ts
@@ -1,0 +1,214 @@
+import type { MessageCatalog } from '../types'
+
+export const jaOverrides: Partial<MessageCatalog> = {
+  'common.confirm': '確認',
+  'discover.filters': 'フィルター',
+  'discover.filtersActive': 'フィルター ({0})',
+  'discover.moreActions': 'その他の操作',
+  'discover.decadesLabel': '年代',
+  'analytics.title': '分析',
+  'dashboard.title': 'ダッシュボード',
+  'discover.title': '発見',
+  'setup.title': 'セットアップ',
+  'todaysPick.discoverSimilar': '似たアーティストを探す',
+  'libraryHealth.fix': '修正',
+  'libraryHealth.fixing': '修正中...',
+  'libraryHealth.showLess': '少なく表示',
+  'libraryHealth.showAll': '{0} 件すべて表示',
+  'libraryHealth.fixHint.missingMetadata': 'Lidarr からアーティストのメタデータを更新します',
+  'libraryHealth.fixHint.unmonitored': '一覧のすべてのアーティストを Lidarr で監視対象にします',
+  'libraryHealth.fixHint.missingAlbums': '不足を見つけるため Lidarr のアルバムを更新します',
+  'libraryHealth.fixHint.genreGaps': 'ジャンルを取得するため Lidarr メタデータを更新します',
+  'libraryHealth.fixHint.imageGaps': 'アーティスト画像を再検索します',
+  'libraryHealth.fixHint.missingWikidata': 'Wikidata から略歴とリンクを取得します',
+  'common.unauthorized': '認証されていません',
+  'common.adminAccessRequired': '管理者権限が必要です',
+  'common.enabled': '有効',
+  'common.importing': 'インポート中...',
+  'settings.tabs.systemHealth': 'システムの健全性',
+  'settings.targetUpdated': 'ターゲットを更新しました',
+  'settings.targetUpdateFailed': 'ターゲットの更新に失敗しました',
+  'settings.editTarget': 'ターゲットを編集',
+  'settings.saveChanges': '変更を保存',
+  'settings.linkedTarget': '関連付けられたターゲット:',
+  'libraryHealth.syncNow': '今すぐ同期',
+  'libraryHealth.snapshotTitle': '最新のスナップショット',
+  'libraryHealth.lastSynced': '最終同期',
+  'libraryHealth.neverSynced': 'まだ同期されていません。',
+  'libraryHealth.autoSyncEvery': '{0}時間ごとに自動同期',
+  'jobHistory.librarySync': 'ライブラリ同期',
+  'errorBoundary.title': '問題が発生しました',
+  'errorBoundary.retry': '再試行',
+  'errorBoundary.home': 'ホームへ戻る',
+  'connectionSuggestions.title': '{0} の提案',
+  'connectionSuggestions.close': '提案を閉じる',
+  'connectionSuggestions.gotIt': '了解',
+  'connectionSuggestions.skipForNow': '今はスキップ',
+  'connectionSuggestions.spotifyDiscoverWeekly': 'Discover Weekly のサブスクリプションを設定',
+  'connectionSuggestions.spotifyReleaseRadar': 'Release Radar のサブスクリプションを有効化',
+  'connectionSuggestions.lastfmTopGenreTags': 'よく聴くジャンルタグのサブスクリプションを作成',
+  'connectionSuggestions.listenbrainzWeeklyJams': 'Weekly Jams のサブスクリプションを有効化',
+  'connectionSuggestions.listenbrainzWeeklyExploration': 'Weekly Exploration を有効化',
+  'connectionSuggestions.weeklyDigestPlaylist': '毎週の Digarr Digest プレイリストを作成',
+  'keyboardShortcuts.title': 'キーボードショートカット',
+  'keyboardShortcuts.close': 'ショートカットを閉じる',
+  'keyboardShortcuts.nextCard': '次のカード',
+  'keyboardShortcuts.previousCard': '前のカード',
+  'keyboardShortcuts.approveFocusedCard': '選択中のカードを承認',
+  'keyboardShortcuts.rejectFocusedCard': '選択中のカードを拒否',
+  'keyboardShortcuts.toggleDetailView': '詳細表示を開く / 閉じる',
+  'keyboardShortcuts.toggleShortcuts': 'ショートカットを表示 / 非表示',
+  'keyboardShortcuts.previousCardStack': '前のカード（スタック表示）',
+  'keyboardShortcuts.nextCardStack': '次のカード（スタック表示）',
+  'keyboardShortcuts.disabledWhileTyping':
+    'テキスト入力欄にフォーカスがある間はショートカットが無効になります。',
+  'admin.restoreSummary': '{1} テーブルで {0} 行を復元しました',
+  'admin.restoreSummaryForce':
+    '{0} 行を復元しました。暗号化された項目の認証情報を再入力してください。',
+  'admin.encryptionMismatchMessage': 'このバックアップは別の暗号化キーで作成されています。',
+  'admin.encryptionMismatchAffected': '次の項目は再入力が必要です: {0}',
+  'admin.encryptionMismatchFallback': '一部の暗号化された項目は再入力が必要になる場合があります。',
+  'subscriptions.spotifyNotConnected': 'Spotify が接続されていません',
+  'playlistForm.dialogCreate': 'プレイリストを作成',
+  'playlistForm.dialogEdit': 'プレイリストを編集',
+  'playlistForm.titleCreate': 'プレイリストを作成',
+  'playlistForm.titleEdit': 'プレイリストを編集',
+  'playlistForm.namePlaceholder': 'マイ・ウィークリー・ミックス',
+  'playlistForm.strategy': '戦略',
+  'playlistForm.strategyWeeklyDigestDesc': '最近の発見から厳選した曲',
+  'playlistForm.strategyGenreFocusDesc': '特定のジャンルを深掘り',
+  'playlistForm.strategyMoodMixDesc': '特定のムードに合う曲を集める',
+  'playlistForm.strategyRediscoverDesc': '忘れていたかもしれないアーティストを掘り起こす',
+  'playlistForm.mood': 'ムード',
+  'playlistForm.moodPlaceholder': '夜のドライブ',
+  'playlistForm.trackCount': '曲数',
+  'playlistForm.scheduleAutomaticGeneration': '自動生成をスケジュールする',
+  'playlistForm.nameRequired': '名前は必須です',
+  'playlistForm.genreRequired': 'ジャンルフォーカス戦略ではジャンルが必須です',
+  'playlistForm.moodRequired': 'ムードミックス戦略ではムードが必須です',
+  'searchResultCard.addToQueue': '発見キューに追加',
+  'searchResultCard.queuing': '追加中...',
+  'searchResultCard.queuedSuccess': '{0} を発見キューに追加しました',
+  'searchResultCard.queuedFailure': '{0} をキューに追加できませんでした',
+  'searchResultCard.popularity': '人気 {0}{1}',
+  'subscriptionPresets.getStarted': 'プリセットから始める',
+  'subscriptionPresets.intro':
+    'プリセットを選んでサブスクリプションを自動作成するか、独自の構成を作成してください。',
+  'subscriptionPresets.casualTitle': 'カジュアル',
+  'subscriptionPresets.casualDescription': 'ノイズの少ない毎週のおすすめ',
+  'subscriptionPresets.deepDiverTitle': 'ディープダイバー',
+  'subscriptionPresets.deepDiverDescription': '複数ソースを使った幅広い発見と毎日の更新',
+  'subscriptionPresets.discoveryOnlyTitle': '発見のみ',
+  'subscriptionPresets.discoveryOnlyDescription': 'パイプラインのみ、外部フィードなし',
+  'subscriptionPresets.customTitle': 'カスタム',
+  'subscriptionPresets.customDescription': '自分だけのサブスクリプション構成を作る',
+  'subscriptionPresets.oneConnectedService': '1 つの接続済みサービス',
+  'subscriptionPresets.twoConnectedServices': '2 つの接続済みサービス',
+  'subscriptionPresets.requires': '{0} が必要です',
+  'subscriptionPresets.setUp': '設定する',
+  'subscriptionPresets.applying': '適用中...',
+  'subscriptionPresets.casualApplied': 'カジュアルのプリセットを適用しました',
+  'subscriptionPresets.deepDiverApplied': 'ディープダイバーのプリセットを適用しました',
+  'subscriptionPresets.applyFailed': 'プリセットを適用できませんでした',
+  'subscriptionPresets.aiOnlyEnabled': 'AI のみモードを有効にしました',
+  'subscriptionPresets.aiOnlyEnableFailed': 'AI のみモードを有効にできませんでした',
+  'common.current': '現在の設定',
+  'common.custom': 'カスタム設定',
+  'subscriptionForm.dialogCreate': 'サブスクリプションを作成',
+  'subscriptionForm.dialogEdit': 'サブスクリプションを編集',
+  'subscriptionForm.titleCreate': 'サブスクリプションを作成',
+  'subscriptionForm.titleEdit': 'サブスクリプションを編集',
+  'subscriptionForm.name': '名前',
+  'subscriptionForm.namePlaceholder': '新しい Shoegaze Weekly',
+  'subscriptionForm.nameRequired': '名前は必須です',
+  'subscriptionForm.source': 'ソース',
+  'subscriptionForm.sources': 'ソース',
+  'subscriptionForm.sourceType': 'ソース種別',
+  'subscriptionForm.sourceTypeGenre': '音楽ジャンル',
+  'subscriptionForm.sourceTypeGenreDesc': '特定のジャンルからアーティストを見つけます',
+  'subscriptionForm.sourceTypeSimilar': '似たアーティスト',
+  'subscriptionForm.sourceTypeSimilarDesc': 'お気に入りに似たアーティストを探します',
+  'subscriptionForm.sourceTypeDiscoveryModeDesc': 'ディスカバリーモードでアーティストを生成します',
+  'subscriptionForm.discoveryMode': 'ディスカバリーモード',
+  'subscriptionForm.selectDiscoveryMode': '設定するディスカバリーモードを選択してください。',
+  'subscriptionForm.discoveryModeSettingsRequired': 'ディスカバリーモードの設定は必須です',
+  'subscriptionForm.genre': '音楽ジャンル',
+  'subscriptionForm.genrePlaceholder': 'shoegaze',
+  'subscriptionForm.seedArtistsPlaceholder': 'Radiohead, Portishead, Massive Attack',
+  'subscriptionForm.genreRequired': 'ジャンルは必須です',
+  'subscriptionForm.seedArtists': '起点アーティスト',
+  'subscriptionForm.seedArtistsHelp': 'アーティスト名をカンマ区切りで入力してください',
+  'subscriptionForm.seedArtistsRequired': '少なくとも 1 人の起点アーティストが必要です',
+  'subscriptionForm.configureInSettings': '設定 > 接続 で設定してください',
+  'subscriptionForm.notConfigured': '未設定',
+  'subscriptionForm.selectAtLeastOneSource': '少なくとも 1 つのソースを選択してください',
+  'subscriptionForm.sourceLockedHelp':
+    'このサブスクリプションではソース種別と設定を変更できません。',
+  'subscriptionForm.maxArtistsPerRun': '1 回あたりの最大アーティスト数',
+  'subscriptionForm.scoringPreset': 'スコアプリセット',
+  'subscriptionForm.scoringPresetDefault': '標準',
+  'subscriptionForm.scoringPresetGenre': 'ジャンル重視',
+  'cronPicker.preset': 'スケジュール設定',
+  'cronPicker.dailyMidnight': '毎日（深夜0時）',
+  'cronPicker.twiceAWeek': '週2回（月 + 木）',
+  'cronPicker.everyTwoWeeks': '2週間ごと（1日と15日）',
+  'cronPicker.monthlyFirst': '毎月（1日）',
+  'cronPicker.standardHint': '標準 cron: 分 時 日 月 曜日',
+  'systemHealth.librarySync': 'ライブラリ同期',
+  'settings.targetTypeEmbyPlaylist': 'Emby プレイリスト',
+  'settings.librarySyncInterval': 'ライブラリ同期',
+  'settings.librarySyncIntervalHours': '同期間隔（時間）',
+  'settings.librarySyncIntervalHelp':
+    'Digarr がライブラリ同期データを更新し、チェックを自動で再実行する間隔を設定します。',
+  'settings.fieldClientId': 'クライアント ID',
+  'settings.fieldClientSecret': 'クライアントシークレット',
+  'settings.fieldWebhookUrl': 'Webhook 送信先 URL',
+  'settings.aiProviderTitle': 'AI プロバイダー',
+  'settings.webhookTitle': 'Webhook 連携',
+  'settings.plexToken': 'Plex トークン',
+  'settings.serviceConnectedToast': '{0} を接続しました',
+  'settings.serviceUnreachableToast': '{0} に接続できませんでした',
+  'settings.serviceSettingsSavedToast': '{0} の設定を保存しました',
+  'settings.serviceSettingsFailedToast': '{0} の設定を保存できませんでした',
+  'settings.spotifyAuthorizationFailed': 'Spotify の認可を開始できませんでした',
+  'settings.spotifyDisconnectFailed': 'Spotify を切断できませんでした',
+  'settings.spotifyLikedSongsFailed': 'Spotify の「いいね」曲の取り込みを開始できませんでした',
+  'settings.spotifyPlaylistFailed': 'プレイリストの取り込みを開始できませんでした',
+  'settings.deezerFavoritesFailed': 'Deezer のお気に入り取り込みを開始できませんでした',
+  'settings.deezerFollowedFailed': 'Deezer のフォロー中アーティスト取り込みを開始できませんでした',
+  'settings.deezerDisconnectFailed': 'Deezer を切断できませんでした',
+  'settings.deezerAuthorizationFailed': 'Deezer の認可を開始できませんでした',
+  'settings.reverseProxyConfigHint':
+    'PROXY_AUTH_ENABLED と PROXY_AUTH_TRUSTED_PROXIES を使い、プロキシは X-Forwarded-User を送信してください。',
+  'admin.autoFixInProgress': '自動修正を実行中...（{0}/{1} 件修正済み）',
+  'albumPicker.title': 'アルバムを選択',
+  'albumPicker.loadFailed': 'アルバムの読み込みに失敗しました',
+  'albumPicker.selectAll': 'すべて',
+  'albumPicker.selectNone': 'なし',
+  'albumPicker.selectedCount': '{0} 件選択中',
+  'albumPicker.aiPick': 'AI のおすすめ',
+  'albumPicker.monitor': '監視',
+  'albumPicker.monitorSingular': '{0} 枚のアルバムを監視',
+  'albumPicker.monitorPlural': '{0} 枚のアルバムを監視',
+  'approveDialog.title': 'このアーティストの Lidarr 設定',
+  'approveDialog.loadFailed':
+    'Lidarr プロファイルの読み込みに失敗しました。Lidarr の接続を確認してください。',
+  'approveDialog.qualityProfile': '品質プロファイル',
+  'approveDialog.metadataProfile': 'メタデータプロファイル',
+  'approveDialog.rootFolder': 'ルートフォルダ',
+  'cardStack.counter': '{1} 件中 {0} 件目',
+  'cardStack.swipeHint': '右にスワイプで承認、左にスワイプで拒否',
+  'serviceCard.testConnection': '接続テスト',
+  'preview.noPreviewAvailable': 'このアーティストのプレビューはありません',
+  'preview.playbackBlocked':
+    'ブラウザによって再生がブロックされました。もう一度クリックしてください',
+  'errors.auth.invalidCredentials': '認証情報が無効です',
+  'errors.auth.notAuthenticated': '認証されていません',
+  'errors.auth.passwordIncorrect': '現在のパスワードが正しくありません',
+  'errors.auth.usernameTaken': 'そのユーザー名はすでに使われています',
+  'errors.target.notFound': 'ターゲットが見つかりません',
+  'errors.pipeline.alreadyRunning': 'パイプラインの実行はすでに進行中です',
+  'errors.recommendation.notFound': 'おすすめが見つかりません',
+  'errors.subscription.notFound': 'サブスクリプションが見つかりません',
+  'errors.playlist.notFound': 'プレイリストが見つかりません',
+}
