+++ src/core/i18n/messages/overrides/ko.ts
@@ -1,0 +1,212 @@
+import type { MessageCatalog } from '../types'
+
+export const koOverrides: Partial<MessageCatalog> = {
+  'common.confirm': '확인',
+  'discover.filters': '필터',
+  'discover.filtersActive': '필터 ({0})',
+  'discover.moreActions': '추가 작업',
+  'discover.decadesLabel': '연대',
+  'analytics.title': '분석',
+  'dashboard.title': '대시보드',
+  'discover.title': '발견',
+  'setup.title': '설정',
+  'todaysPick.discoverSimilar': '비슷한 아티스트 찾기',
+  'libraryHealth.fix': '수정',
+  'libraryHealth.fixing': '수정 중...',
+  'libraryHealth.showLess': '간단히 보기',
+  'libraryHealth.showAll': '{0}개 모두 보기',
+  'libraryHealth.fixHint.missingMetadata': 'Lidarr에서 아티스트 메타데이터를 새로 고칩니다',
+  'libraryHealth.fixHint.unmonitored':
+    '나열된 모든 아티스트를 Lidarr에서 모니터링 대상으로 설정합니다',
+  'libraryHealth.fixHint.missingAlbums': '누락 항목을 찾도록 Lidarr 앨범을 새로 고칩니다',
+  'libraryHealth.fixHint.genreGaps': '장르 태그를 가져오도록 Lidarr 메타데이터를 새로 고칩니다',
+  'libraryHealth.fixHint.imageGaps': '아티스트 이미지를 다시 검색합니다',
+  'libraryHealth.fixHint.missingWikidata': 'Wikidata에서 약력과 링크를 가져옵니다',
+  'common.importing': '가져오는 중...',
+  'common.current': '현재',
+  'common.custom': '사용자 지정',
+  'common.enabled': '사용됨',
+  'common.unauthorized': '인증되지 않음',
+  'common.adminAccessRequired': '관리자 권한이 필요합니다',
+  'jobHistory.librarySync': '라이브러리 동기화',
+  'playlistForm.dialogCreate': '플레이리스트 만들기',
+  'playlistForm.dialogEdit': '플레이리스트 수정',
+  'playlistForm.titleCreate': '플레이리스트 만들기',
+  'playlistForm.titleEdit': '플레이리스트 수정',
+  'playlistForm.namePlaceholder': '내 주간 믹스',
+  'playlistForm.strategy': '전략',
+  'playlistForm.strategyWeeklyDigestDesc': '최근 발견에서 골라낸 큐레이션 모음',
+  'playlistForm.strategyGenreFocusDesc': '특정 장르를 깊게 파고들기',
+  'playlistForm.strategyMoodMixDesc': '특정 분위기에 맞는 트랙 모음',
+  'playlistForm.strategyRediscoverDesc': '잊고 지냈을지도 모를 아티스트 다시 찾기',
+  'playlistForm.mood': '분위기',
+  'playlistForm.moodPlaceholder': '늦은 밤 드라이브',
+  'playlistForm.trackCount': '트랙 수',
+  'playlistForm.scheduleAutomaticGeneration': '자동 생성을 예약',
+  'playlistForm.nameRequired': '이름은 필수입니다',
+  'playlistForm.genreRequired': '장르 포커스 전략에는 장르가 필요합니다',
+  'playlistForm.moodRequired': '무드 믹스 전략에는 분위기가 필요합니다',
+  'subscriptionForm.dialogCreate': '구독 만들기',
+  'subscriptionForm.dialogEdit': '구독 수정',
+  'subscriptionForm.titleCreate': '구독 만들기',
+  'subscriptionForm.titleEdit': '구독 수정',
+  'subscriptionForm.name': '이름',
+  'subscriptionForm.namePlaceholder': '새로운 Shoegaze Weekly',
+  'subscriptionForm.nameRequired': '이름은 필수입니다',
+  'subscriptionForm.source': '소스',
+  'subscriptionForm.sources': '소스',
+  'subscriptionForm.sourceType': '소스 유형',
+  'subscriptionForm.sourceTypeGenre': '장르',
+  'subscriptionForm.sourceTypeGenreDesc': '특정 장르에서 아티스트를 발견합니다',
+  'subscriptionForm.sourceTypeSimilar': '유사 아티스트',
+  'subscriptionForm.sourceTypeSimilarDesc': '좋아하는 아티스트와 비슷한 아티스트를 찾습니다',
+  'subscriptionForm.sourceTypeDiscoveryModeDesc': '발견 모드를 사용해 아티스트를 생성합니다',
+  'subscriptionForm.discoveryMode': '발견 모드',
+  'subscriptionForm.selectDiscoveryMode': '구성할 발견 모드를 선택하세요.',
+  'subscriptionForm.discoveryModeSettingsRequired': '발견 모드 설정은 필수입니다',
+  'subscriptionForm.genre': '장르',
+  'subscriptionForm.genrePlaceholder': 'shoegaze',
+  'subscriptionForm.genreRequired': '장르는 필수입니다',
+  'subscriptionForm.seedArtists': '시드 아티스트',
+  'subscriptionForm.seedArtistsPlaceholder': 'Radiohead, Portishead, Massive Attack',
+  'subscriptionForm.seedArtistsHelp': '쉼표로 구분한 아티스트 이름',
+  'subscriptionForm.seedArtistsRequired': '최소 한 명의 시드 아티스트가 필요합니다',
+  'subscriptionForm.configureInSettings': '설정 > 연결에서 구성하세요',
+  'subscriptionForm.notConfigured': '구성되지 않음',
+  'subscriptionForm.selectAtLeastOneSource': '최소 한 개의 소스를 선택하세요',
+  'subscriptionForm.sourceLockedHelp': '이 구독에서는 소스 유형과 구성을 변경할 수 없습니다.',
+  'subscriptionForm.maxArtistsPerRun': '실행당 최대 아티스트 수',
+  'subscriptionForm.scoringPreset': '점수 프리셋',
+  'subscriptionForm.scoringPresetDefault': '기본값',
+  'subscriptionForm.scoringPresetGenre': '장르 최적화',
+  'subscriptions.spotifyNotConnected': 'Spotify가 연결되지 않았습니다',
+  'subscriptionPresets.getStarted': '프리셋으로 시작하기',
+  'subscriptionPresets.intro': '프리셋을 선택해 구독을 자동으로 만들거나 직접 구성하세요.',
+  'subscriptionPresets.casualTitle': '가볍게',
+  'subscriptionPresets.casualDescription': '잡음이 적은 주간 추천',
+  'subscriptionPresets.deepDiverTitle': '딥 다이버',
+  'subscriptionPresets.deepDiverDescription': '여러 소스를 활용한 폭넓은 발견과 매일 업데이트',
+  'subscriptionPresets.discoveryOnlyTitle': '발견 전용',
+  'subscriptionPresets.discoveryOnlyDescription': '파이프라인만 사용, 외부 피드 없음',
+  'subscriptionPresets.customTitle': '사용자 지정',
+  'subscriptionPresets.customDescription': '나만의 구독 조합 만들기',
+  'subscriptionPresets.oneConnectedService': '연결된 서비스 1개',
+  'subscriptionPresets.twoConnectedServices': '연결된 서비스 2개',
+  'subscriptionPresets.requires': '{0} 필요',
+  'subscriptionPresets.setUp': '설정',
+  'subscriptionPresets.applying': '적용 중...',
+  'subscriptionPresets.casualApplied': '가볍게 프리셋을 적용했습니다',
+  'subscriptionPresets.deepDiverApplied': '딥 다이버 프리셋을 적용했습니다',
+  'subscriptionPresets.applyFailed': '프리셋을 적용하지 못했습니다',
+  'subscriptionPresets.aiOnlyEnabled': 'AI 전용 모드를 켰습니다',
+  'subscriptionPresets.aiOnlyEnableFailed': 'AI 전용 모드를 켜지 못했습니다',
+  'cronPicker.preset': '프리셋',
+  'cronPicker.dailyMidnight': '매일 (자정)',
+  'cronPicker.twiceAWeek': '주 2회 (월 + 목)',
+  'cronPicker.everyTwoWeeks': '2주마다 (1일 + 15일)',
+  'cronPicker.monthlyFirst': '매월 (1일)',
+  'cronPicker.standardHint': '표준 cron: 분 시 일-월 월 요일',
+  'libraryHealth.syncNow': '지금 동기화',
+  'libraryHealth.snapshotTitle': '최신 스냅샷',
+  'libraryHealth.lastSynced': '마지막 동기화',
+  'libraryHealth.neverSynced': '동기화된 적이 없습니다.',
+  'libraryHealth.autoSyncEvery': '{0}시간마다 자동 동기화',
+  'systemHealth.librarySync': '라이브러리 동기화',
+  'settings.tabs.systemHealth': '시스템 상태',
+  'settings.targetUpdated': '대상을 업데이트했습니다',
+  'settings.targetUpdateFailed': '대상 업데이트에 실패했습니다',
+  'settings.editTarget': '대상 편집',
+  'settings.saveChanges': '변경 사항 저장',
+  'settings.linkedTarget': '연결된 대상:',
+  'settings.targetTypeEmbyPlaylist': 'Emby 플레이리스트',
+  'settings.librarySyncInterval': '라이브러리 동기화',
+  'settings.librarySyncIntervalHours': '동기화 간격(시간)',
+  'settings.librarySyncIntervalHelp':
+    'Digarr가 라이브러리 동기화 데이터를 새로 고치고 검사를 자동으로 다시 실행하는 주기를 제어합니다.',
+  'settings.oidcSso': 'OIDC 및 SSO',
+  'settings.fieldClientId': '클라이언트 ID',
+  'settings.fieldClientSecret': '클라이언트 시크릿',
+  'settings.fieldWebhookUrl': '웹훅 URL',
+  'settings.aiProviderTitle': 'AI 제공자',
+  'settings.webhookTitle': '웹훅',
+  'settings.plexToken': 'Plex 토큰',
+  'settings.serviceConnectedToast': '{0} 연결됨',
+  'settings.serviceUnreachableToast': '{0}에 연결할 수 없습니다',
+  'settings.serviceSettingsSavedToast': '{0} 설정을 저장했습니다',
+  'settings.serviceSettingsFailedToast': '{0} 설정 저장에 실패했습니다',
+  'settings.spotifyAuthorizationFailed': 'Spotify 인증 시작에 실패했습니다',
+  'settings.spotifyDisconnectFailed': 'Spotify 연결 해제에 실패했습니다',
+  'settings.spotifyLikedSongsFailed': 'Spotify 좋아요 가져오기를 시작하지 못했습니다',
+  'settings.spotifyPlaylistFailed': 'Spotify 플레이리스트를 가져오지 못했습니다',
+  'settings.deezerFavoritesFailed': 'Deezer 즐겨찾기 가져오기를 시작하지 못했습니다',
+  'settings.deezerFollowedFailed': 'Deezer 팔로우 아티스트 가져오기를 시작하지 못했습니다',
+  'settings.deezerDisconnectFailed': 'Deezer 연결 해제에 실패했습니다',
+  'settings.deezerAuthorizationFailed': 'Deezer 인증 시작에 실패했습니다',
+  'settings.reverseProxyConfigHint':
+    'PROXY_AUTH_ENABLED 및 PROXY_AUTH_TRUSTED_PROXIES를 사용하세요. 프록시는 X-Forwarded-User를 보내야 합니다.',
+  'searchResultCard.addToQueue': '발견 대기열에 추가',
+  'searchResultCard.queuing': '추가 중...',
+  'searchResultCard.queuedSuccess': '{0}을(를) 발견 대기열에 추가했습니다',
+  'searchResultCard.queuedFailure': '{0}을(를) 대기열에 추가하지 못했습니다',
+  'searchResultCard.popularity': '인기도 {0}{1}',
+  'admin.restoreSummary': '{1}개 테이블에서 {0}개 행을 복원했습니다',
+  'admin.restoreSummaryForce':
+    '{0}개 행을 복원했습니다. 암호화된 필드의 자격 증명을 다시 입력하세요.',
+  'admin.encryptionMismatchMessage': '이 백업은 다른 암호화 키로 생성되었습니다.',
+  'admin.encryptionMismatchAffected': '다음 필드는 다시 입력해야 합니다: {0}',
+  'admin.encryptionMismatchFallback': '일부 암호화된 필드는 다시 입력해야 할 수 있습니다.',
+  'admin.autoFixInProgress': '자동 수정 진행 중... ({0}/{1} 수정됨)',
+  'errorBoundary.title': '문제가 발생했습니다',
+  'errorBoundary.retry': '다시 시도',
+  'errorBoundary.home': '홈으로 이동',
+  'connectionSuggestions.title': '{0}에 대한 제안',
+  'connectionSuggestions.close': '제안 닫기',
+  'connectionSuggestions.gotIt': '알겠습니다',
+  'connectionSuggestions.skipForNow': '지금은 건너뛰기',
+  'connectionSuggestions.spotifyDiscoverWeekly': 'Discover Weekly 구독 설정',
+  'connectionSuggestions.spotifyReleaseRadar': 'Release Radar 구독 사용',
+  'connectionSuggestions.lastfmTopGenreTags': '주요 장르 태그 구독',
+  'connectionSuggestions.listenbrainzWeeklyJams': 'Weekly Jams 구독 사용',
+  'connectionSuggestions.listenbrainzWeeklyExploration': 'Weekly Exploration 사용',
+  'connectionSuggestions.weeklyDigestPlaylist': '주간 Digarr Digest 플레이리스트 만들기',
+  'keyboardShortcuts.title': '키보드 단축키',
+  'keyboardShortcuts.close': '단축키 닫기',
+  'keyboardShortcuts.nextCard': '다음 카드',
+  'keyboardShortcuts.previousCard': '이전 카드',
+  'keyboardShortcuts.approveFocusedCard': '선택된 카드 승인',
+  'keyboardShortcuts.rejectFocusedCard': '선택된 카드 거절',
+  'keyboardShortcuts.toggleDetailView': '상세 보기 열기 / 닫기',
+  'keyboardShortcuts.toggleShortcuts': '단축키 표시 / 숨기기',
+  'keyboardShortcuts.previousCardStack': '이전 카드 (스택 보기)',
+  'keyboardShortcuts.nextCardStack': '다음 카드 (스택 보기)',
+  'keyboardShortcuts.disabledWhileTyping':
+    '텍스트 입력란에 포커스가 있으면 단축키가 비활성화됩니다.',
+  'albumPicker.title': '앨범 선택',
+  'albumPicker.loadFailed': '앨범을 불러오지 못했습니다',
+  'albumPicker.selectAll': '모두',
+  'albumPicker.selectNone': '없음',
+  'albumPicker.selectedCount': '{0}개 선택됨',
+  'albumPicker.aiPick': 'AI 추천',
+  'albumPicker.monitor': '모니터링',
+  'albumPicker.monitorSingular': '{0}개 앨범 모니터링',
+  'albumPicker.monitorPlural': '{0}개 앨범 모니터링',
+  'approveDialog.title': '이 아티스트의 Lidarr 설정',
+  'approveDialog.loadFailed': 'Lidarr 프로필을 불러오지 못했습니다. Lidarr 연결을 확인하세요.',
+  'approveDialog.qualityProfile': '품질 프로필',
+  'approveDialog.metadataProfile': '메타데이터 프로필',
+  'approveDialog.rootFolder': '루트 폴더',
+  'cardStack.counter': '{1} 중 {0}',
+  'cardStack.swipeHint': '오른쪽으로 밀어 승인, 왼쪽으로 밀어 거절',
+  'serviceCard.testConnection': '연결 테스트',
+  'preview.noPreviewAvailable': '이 아티스트에 대한 미리듣기를 사용할 수 없습니다',
+  'preview.playbackBlocked': '브라우저에서 재생이 차단되었습니다. 다시 클릭해 보세요',
+  'errors.auth.invalidCredentials': '잘못된 자격 증명입니다',
+  'errors.auth.notAuthenticated': '인증되지 않았습니다',
+  'errors.auth.passwordIncorrect': '현재 비밀번호가 올바르지 않습니다',
+  'errors.auth.usernameTaken': '이미 사용 중인 사용자 이름입니다',
+  'errors.target.notFound': '대상을 찾을 수 없습니다',
+  'errors.pipeline.alreadyRunning': '파이프라인 실행이 이미 진행 중입니다',
+  'errors.recommendation.notFound': '추천을 찾을 수 없습니다',
+  'errors.subscription.notFound': '구독을 찾을 수 없습니다',
+  'errors.playlist.notFound': '재생목록을 찾을 수 없습니다',
+}
