+++ src/core/i18n/messages/overrides/nl.ts
@@ -1,0 +1,242 @@
+import type { MessageCatalog } from '../types'
+
+export const nlOverrides: Partial<MessageCatalog> = {
+  'common.confirm': 'Bevestigen',
+  'discover.filters': 'Filteropties',
+  'discover.filtersActive': 'Filteropties ({0})',
+  'discover.moreActions': 'Meer acties',
+  'discover.decadesLabel': 'Decennia',
+  'analytics.title': 'Analyse',
+  'dashboard.title': 'Overzicht',
+  'discover.title': 'Ontdekken',
+  'setup.title': 'Instellen',
+  'todaysPick.discoverSimilar': 'Vergelijkbare ontdekken',
+  'libraryHealth.fix': 'Repareren',
+  'libraryHealth.fixing': 'Repareren...',
+  'libraryHealth.showLess': 'Minder tonen',
+  'libraryHealth.showAll': 'Alle {0} tonen',
+  'libraryHealth.fixHint.missingMetadata': 'Vernieuwt artiestmetadata vanuit Lidarr',
+  'libraryHealth.fixHint.unmonitored': 'Zet alle vermelde artiesten in Lidarr op gemonitord',
+  'libraryHealth.fixHint.missingAlbums': 'Vernieuwt albums in Lidarr om ontbrekende te vinden',
+  'libraryHealth.fixHint.genreGaps': 'Vernieuwt Lidarr-metadata om genres op te halen',
+  'libraryHealth.fixHint.imageGaps': 'Zoekt artiestafbeeldingen opnieuw',
+  'libraryHealth.fixHint.missingWikidata': 'Haalt biografieën en links op uit Wikidata',
+  'nav.dashboard': 'Overzicht',
+  'setup.model': 'Modelkeuze',
+  'common.importing': 'Importeren...',
+  'common.current': 'Huidig',
+  'common.custom': 'Aangepast',
+  'common.enabled': 'Ingeschakeld',
+  'common.unauthorized': 'Niet gemachtigd',
+  'common.adminAccessRequired': 'Beheerderstoegang vereist',
+  'common.status': 'Toestand',
+  'genres.genrePlural': 'muziekgenres',
+  'genres.genreSingular': 'muziekgenre',
+  'jobHistory.pipeline': 'Ontdekkingsrun',
+  'jobHistory.librarySync': 'Bibliotheeksynchronisatie',
+  'jobHistory.playlistType': 'Afspeellijst',
+  'jobHistory.metadataLabel': 'Metagegevens:',
+  'playlistForm.dialogCreate': 'Afspeellijst maken',
+  'playlistForm.dialogEdit': 'Afspeellijst bewerken',
+  'playlistForm.titleCreate': 'Afspeellijst maken',
+  'playlistForm.titleEdit': 'Afspeellijst bewerken',
+  'playlistForm.namePlaceholder': 'Mijn wekelijkse mix',
+  'playlistForm.strategy': 'Samenstelling',
+  'playlistForm.strategyWeeklyDigestDesc': 'Een zorgvuldige selectie uit je recente ontdekkingen',
+  'playlistForm.strategyGenreFocusDesc': 'Duik diep in één specifiek genre',
+  'playlistForm.strategyMoodMixDesc': 'Tracks die bij een bepaalde sfeer passen',
+  'playlistForm.strategyRediscoverDesc': 'Haal artiesten terug die je misschien was vergeten',
+  'playlistForm.mood': 'Sfeer',
+  'playlistForm.moodPlaceholder': 'nachtelijke rit',
+  'playlistForm.trackCount': 'Aantal tracks',
+  'playlistForm.scheduleAutomaticGeneration': 'Automatisch genereren plannen',
+  'playlistForm.nameRequired': 'Naam is verplicht',
+  'playlistForm.genreRequired': 'Genre is verplicht voor de strategie Genrefocus',
+  'playlistForm.moodRequired': 'Sfeer is verplicht voor de strategie Moodmix',
+  'subscriptionForm.dialogCreate': 'Abonnement maken',
+  'subscriptionForm.dialogEdit': 'Abonnement bewerken',
+  'subscriptionForm.titleCreate': 'Abonnement maken',
+  'subscriptionForm.titleEdit': 'Abonnement bewerken',
+  'subscriptionForm.name': 'Naam',
+  'subscriptionForm.namePlaceholder': 'Nieuwe Shoegaze Weekly',
+  'subscriptionForm.nameRequired': 'Naam is verplicht',
+  'subscriptionForm.source': 'Bron',
+  'subscriptionForm.sources': 'Bronnen',
+  'subscriptionForm.sourceType': 'Brontype',
+  'subscriptionForm.sourceTypeGenre': 'Muziekgenre',
+  'subscriptionForm.sourceTypeGenreDesc': 'Ontdek artiesten binnen een specifiek genre',
+  'subscriptionForm.sourceTypeSimilar': 'Vergelijkbare artiesten',
+  'subscriptionForm.sourceTypeSimilarDesc': 'Vind artiesten die op je favorieten lijken',
+  'subscriptionForm.sourceTypeDiscoveryModeDesc':
+    'Gebruik een ontdekkingsmodus om artiesten te genereren',
+  'subscriptionForm.discoveryMode': 'Ontdekkingsmodus',
+  'subscriptionForm.selectDiscoveryMode': 'Kies een ontdekkingsmodus om in te stellen.',
+  'subscriptionForm.discoveryModeSettingsRequired':
+    'De instellingen voor de ontdekkingsmodus zijn verplicht',
+  'subscriptionForm.genre': 'Muziekgenre',
+  'subscriptionForm.genrePlaceholder': 'shoegaze',
+  'subscriptionForm.seedArtistsPlaceholder': 'Radiohead, Portishead, Massive Attack',
+  'subscriptionForm.genreRequired': 'Genre is verplicht',
+  'subscriptionForm.seedArtists': 'Startartiesten',
+  'subscriptionForm.seedArtistsHelp': 'Artiestnamen gescheiden door komma s',
+  'subscriptionForm.seedArtistsRequired': 'Minstens één startartiest is verplicht',
+  'subscriptionForm.configureInSettings': 'Stel dit in via Instellingen > Verbindingen',
+  'subscriptionForm.notConfigured': 'niet ingesteld',
+  'subscriptionForm.selectAtLeastOneSource': 'Selecteer minstens één bron',
+  'subscriptionForm.sourceLockedHelp':
+    'Het brontype en de configuratie kunnen voor dit abonnement niet worden gewijzigd.',
+  'subscriptionForm.maxArtistsPerRun': 'Maximaal aantal artiesten per uitvoering',
+  'subscriptionForm.scoringPreset': 'Scoreprofiel',
+  'subscriptionForm.scoringPresetDefault': 'Standaard',
+  'subscriptionForm.scoringPresetGenre': 'Geoptimaliseerd voor genres',
+  'subscriptions.spotifyNotConnected': 'Spotify is niet gekoppeld',
+  'subscriptionPresets.getStarted': 'Begin met een voorinstelling',
+  'subscriptionPresets.intro':
+    'Kies een voorinstelling om automatisch abonnementen aan te maken of stel zelf een configuratie samen.',
+  'subscriptionPresets.casualTitle': 'Ontspannen',
+  'subscriptionPresets.casualDescription': 'Wekelijkse aanbevelingen met weinig ruis',
+  'subscriptionPresets.deepDiverTitle': 'Diepgraver',
+  'subscriptionPresets.deepDiverDescription':
+    'Brede ontdekking uit meerdere bronnen met dagelijkse updates',
+  'subscriptionPresets.discoveryOnlyTitle': 'Alleen ontdekken',
+  'subscriptionPresets.discoveryOnlyDescription': 'Alleen de pijplijn, zonder externe feeds',
+  'subscriptionPresets.customTitle': 'Aangepast',
+  'subscriptionPresets.customDescription': 'Stel je eigen combinatie van abonnementen samen',
+  'subscriptionPresets.oneConnectedService': '1 verbonden dienst',
+  'subscriptionPresets.twoConnectedServices': '2 verbonden diensten',
+  'subscriptionPresets.requires': 'Vereist {0}',
+  'subscriptionPresets.setUp': 'Instellen',
+  'subscriptionPresets.applying': 'Bezig met toepassen...',
+  'subscriptionPresets.casualApplied': 'De voorinstelling Ontspannen is toegepast',
+  'subscriptionPresets.deepDiverApplied': 'De voorinstelling Diepgraver is toegepast',
+  'subscriptionPresets.applyFailed': 'De voorinstelling kon niet worden toegepast',
+  'subscriptionPresets.aiOnlyEnabled': 'Alleen-AI-modus is ingeschakeld',
+  'subscriptionPresets.aiOnlyEnableFailed': 'Alleen-AI-modus kon niet worden ingeschakeld',
+  'cronPicker.preset': 'Voorinstelling',
+  'cronPicker.dailyMidnight': 'Dagelijks (middernacht)',
+  'cronPicker.twiceAWeek': 'Twee keer per week (ma + do)',
+  'cronPicker.everyTwoWeeks': 'Om de 2 weken (1 en 15)',
+  'cronPicker.monthlyFirst': 'Maandelijks (dag 1)',
+  'cronPicker.standardHint': 'Standaard cron: minuut uur dag-van-de-maand maand dag-van-de-week',
+  'recommendation.sourceGenre': 'Muziekgenre',
+  'libraryHealth.syncNow': 'Nu synchroniseren',
+  'libraryHealth.snapshotTitle': 'Laatste momentopname',
+  'libraryHealth.lastSynced': 'Laatst gesynchroniseerd',
+  'libraryHealth.neverSynced': 'Nog niet gesynchroniseerd.',
+  'libraryHealth.autoSyncEvery': 'Automatisch synchroniseren elke {0} uur',
+  'settings.capImport': 'Importeren',
+  'systemHealth.pipeline': 'Ontdekkingsrun',
+  'systemHealth.librarySync': 'Bibliotheeksynchronisatie',
+  'discoveryMode.labels.label': 'Etiketten',
+  'discoveryMode.field.feed': 'Feedbron',
+  'discoveryMode.field.tags': 'Trefwoorden',
+  'discoveryMode.option.weeklyJams': 'Weekly Jams',
+  'libraryStats.albums': 'Albumreleases',
+  'settings.tabs.account': 'Gebruikersaccount',
+  'settings.tabs.systemHealth': 'Systeemstatus',
+  'settings.targetUpdated': 'Doel is bijgewerkt',
+  'settings.targetUpdateFailed': 'Doel bijwerken is mislukt',
+  'settings.editTarget': 'Doel bewerken',
+  'settings.saveChanges': 'Wijzigingen opslaan',
+  'settings.linkedTarget': 'Gekoppeld doel:',
+  'settings.type': 'Soort',
+  'settings.consensus': 'Consensusscore',
+  'settings.librarySyncInterval': 'Bibliotheeksynchronisatie',
+  'settings.librarySyncIntervalHours': 'Synchronisatie-interval (uren)',
+  'settings.librarySyncIntervalHelp':
+    'Bepaalt hoe vaak Digarr synchronisatiegegevens van de bibliotheek ververst en gezondheidscontroles opnieuw uitvoert.',
+  'settings.app': 'Applicatie',
+  'settings.oidcSso': 'OIDC / SSO',
+  'settings.fieldModel': 'Modelkeuze',
+  'settings.aiProviderTitle': 'AI-provider',
+  'settings.webhookTitle': 'Webhook-koppeling',
+  'settings.plexToken': 'Plex-token',
+  'settings.serviceConnectedToast': '{0} gekoppeld',
+  'settings.serviceUnreachableToast': '{0} is niet bereikbaar',
+  'settings.serviceSettingsSavedToast': 'Instellingen voor {0} zijn opgeslagen',
+  'settings.serviceSettingsFailedToast': 'Instellingen voor {0} konden niet worden opgeslagen',
+  'settings.spotifyAuthorizationFailed': 'Spotify-autorisatie kon niet worden gestart',
+  'settings.spotifyDisconnectFailed': 'Spotify kon niet worden losgekoppeld',
+  'settings.spotifyLikedSongsFailed':
+    'Het importeren van Spotify-liked songs kon niet worden gestart',
+  'settings.spotifyPlaylistFailed': 'Het importeren van de afspeellijst kon niet worden gestart',
+  'settings.deezerFavoritesFailed': 'Het importeren van Deezer-favorieten kon niet worden gestart',
+  'settings.deezerFollowedFailed':
+    'Het importeren van gevolgde Deezer-artiesten kon niet worden gestart',
+  'settings.deezerDisconnectFailed': 'Deezer kon niet worden losgekoppeld',
+  'settings.deezerAuthorizationFailed': 'Deezer-autorisatie kon niet worden gestart',
+  'settings.reverseProxyConfigHint':
+    'Gebruik PROXY_AUTH_ENABLED en PROXY_AUTH_TRUSTED_PROXIES; je proxy moet X-Forwarded-User meesturen.',
+  'searchResultCard.addToQueue': 'Aan ontdekkingswachtrij toevoegen',
+  'searchResultCard.queuing': 'Toevoegen...',
+  'searchResultCard.queuedSuccess': '{0} is toegevoegd aan de ontdekkingswachtrij',
+  'searchResultCard.queuedFailure': '{0} kon niet aan de wachtrij worden toegevoegd',
+  'searchResultCard.popularity': 'Populariteit {0}{1}',
+  'admin.restoreSummary': '{0} rijen hersteld in {1} tabellen',
+  'admin.restoreSummaryForce':
+    '{0} rijen hersteld. Voer de gegevens voor versleutelde velden opnieuw in.',
+  'admin.encryptionMismatchMessage':
+    'Deze back-up is gemaakt met een andere versleutelingssleutel.',
+  'admin.encryptionMismatchAffected': 'Deze velden moeten opnieuw worden ingevuld: {0}',
+  'admin.encryptionMismatchFallback':
+    'Sommige versleutelde velden moeten mogelijk opnieuw worden ingevuld.',
+  'admin.autoFixInProgress': 'Automatische correctie bezig... ({0}/{1} hersteld)',
+  'errorBoundary.title': 'Er is iets misgegaan',
+  'errorBoundary.retry': 'Opnieuw proberen',
+  'errorBoundary.home': 'Naar startpagina',
+  'connectionSuggestions.title': 'Suggesties voor {0}',
+  'connectionSuggestions.close': 'Suggesties sluiten',
+  'connectionSuggestions.gotIt': 'Begrepen',
+  'connectionSuggestions.skipForNow': 'Nu overslaan',
+  'connectionSuggestions.spotifyDiscoverWeekly': 'Een Discover Weekly-abonnement instellen',
+  'connectionSuggestions.spotifyReleaseRadar': 'Een Release Radar-abonnement inschakelen',
+  'connectionSuggestions.lastfmTopGenreTags': 'Je belangrijkste genrelabels volgen',
+  'connectionSuggestions.listenbrainzWeeklyJams': 'Het Weekly Jams-abonnement inschakelen',
+  'connectionSuggestions.listenbrainzWeeklyExploration': 'Weekly Exploration inschakelen',
+  'connectionSuggestions.weeklyDigestPlaylist': 'Een wekelijkse Digarr Digest-afspeellijst maken',
+  'keyboardShortcuts.title': 'Sneltoetsen',
+  'keyboardShortcuts.close': 'Sneltoetsen sluiten',
+  'keyboardShortcuts.nextCard': 'Volgende kaart',
+  'keyboardShortcuts.previousCard': 'Vorige kaart',
+  'keyboardShortcuts.approveFocusedCard': 'Actieve kaart goedkeuren',
+  'keyboardShortcuts.rejectFocusedCard': 'Actieve kaart afwijzen',
+  'keyboardShortcuts.toggleDetailView': 'Detailweergave openen / sluiten',
+  'keyboardShortcuts.toggleShortcuts': 'Sneltoetsen tonen / verbergen',
+  'keyboardShortcuts.previousCardStack': 'Vorige kaart (stapelweergave)',
+  'keyboardShortcuts.nextCardStack': 'Volgende kaart (stapelweergave)',
+  'keyboardShortcuts.disabledWhileTyping':
+    'Sneltoetsen zijn uitgeschakeld zolang een tekstveld focus heeft.',
+  'settings.fieldClientId': 'Client-ID',
+  'settings.fieldClientSecret': 'Clientgeheim',
+  'settings.fieldScopes': 'Reikwijdten',
+  'analytics.source.consensus': 'Overeenstemming',
+  'albumPicker.title': 'Albums selecteren',
+  'albumPicker.loadFailed': 'Laden van albums mislukt',
+  'albumPicker.selectAll': 'Alle',
+  'albumPicker.selectNone': 'Geen',
+  'albumPicker.selectedCount': '{0} geselecteerd',
+  'albumPicker.aiPick': 'AI-keuze',
+  'albumPicker.monitor': 'Volgen',
+  'albumPicker.monitorSingular': '{0} album volgen',
+  'albumPicker.monitorPlural': '{0} albums volgen',
+  'approveDialog.title': 'Lidarr-instellingen voor deze artiest',
+  'approveDialog.loadFailed':
+    'Laden van Lidarr-profielen mislukt. Controleer je Lidarr-verbinding.',
+  'approveDialog.qualityProfile': 'Kwaliteitsprofiel',
+  'approveDialog.metadataProfile': 'Metadataprofiel',
+  'approveDialog.rootFolder': 'Hoofdmap',
+  'cardStack.counter': '{0} van {1}',
+  'cardStack.swipeHint': 'Veeg naar rechts om goed te keuren, naar links om af te wijzen',
+  'serviceCard.testConnection': 'Verbinding testen',
+  'preview.noPreviewAvailable': 'Geen voorbeeld beschikbaar voor deze artiest',
+  'preview.playbackBlocked': 'Afspelen geblokkeerd door browser - probeer opnieuw te klikken',
+  'errors.auth.invalidCredentials': 'Ongeldige inloggegevens',
+  'errors.auth.notAuthenticated': 'Niet aangemeld',
+  'errors.auth.passwordIncorrect': 'Huidig wachtwoord is onjuist',
+  'errors.auth.usernameTaken': 'Gebruikersnaam al in gebruik',
+  'errors.target.notFound': 'Doel niet gevonden',
+  'errors.pipeline.alreadyRunning': 'Er is al een pipeline-uitvoering bezig',
+  'errors.recommendation.notFound': 'Aanbeveling niet gevonden',
+  'errors.subscription.notFound': 'Abonnement niet gevonden',
+  'errors.playlist.notFound': 'Afspeellijst niet gevonden',
+}
