+++ src/core/i18n/messages/overrides/pl.ts
@@ -1,0 +1,222 @@
+import type { MessageCatalog } from '../types'
+
+export const plOverrides: Partial<MessageCatalog> = {
+  'common.confirm': 'Potwierdź',
+  'discover.filters': 'Filtry',
+  'discover.filtersActive': 'Filtry ({0})',
+  'discover.moreActions': 'Więcej działań',
+  'discover.decadesLabel': 'Dekady',
+  'analytics.title': 'Analityka',
+  'dashboard.title': 'Panel',
+  'discover.title': 'Odkrywaj',
+  'setup.title': 'Konfiguracja',
+  'todaysPick.discoverSimilar': 'Odkryj podobnych',
+  'libraryHealth.fix': 'Napraw',
+  'libraryHealth.fixing': 'Naprawianie...',
+  'libraryHealth.showLess': 'Pokaż mniej',
+  'libraryHealth.showAll': 'Pokaż wszystkie {0}',
+  'libraryHealth.fixHint.missingMetadata': 'Odświeża metadane artystów z Lidarr',
+  'libraryHealth.fixHint.unmonitored':
+    'Ustawia wszystkich wymienionych artystów jako monitorowanych w Lidarr',
+  'libraryHealth.fixHint.missingAlbums': 'Odświeża albumy w Lidarr, aby znaleźć brakujące',
+  'libraryHealth.fixHint.genreGaps': 'Odświeża metadane Lidarr, aby pobrać gatunki',
+  'libraryHealth.fixHint.imageGaps': 'Ponownie wyszukuje obrazy artystów',
+  'libraryHealth.fixHint.missingWikidata': 'Pobiera biografie i linki z Wikidata',
+  'common.unauthorized': 'Brak autoryzacji',
+  'common.adminAccessRequired': 'Wymagany jest dostep administratora',
+  'common.current': 'Biezacy',
+  'common.custom': 'Niestandardowy',
+  'common.enabled': 'Wlaczone',
+  'common.importing': 'Importowanie...',
+  'libraryHealth.syncNow': 'Synchronizuj teraz',
+  'libraryHealth.snapshotTitle': 'Najnowsza migawka',
+  'libraryHealth.lastSynced': 'Ostatnia synchronizacja',
+  'libraryHealth.neverSynced': 'Nigdy nie synchronizowano.',
+  'libraryHealth.autoSyncEvery': 'Automatyczna synchronizacja co {0} h',
+  'jobHistory.librarySync': 'Synchronizacja biblioteki',
+  'errorBoundary.title': 'Cos poszlo nie tak',
+  'errorBoundary.retry': 'Sprobuj ponownie',
+  'errorBoundary.home': 'Przejdz do strony glownej',
+  'connectionSuggestions.title': 'Sugestie dla {0}',
+  'connectionSuggestions.close': 'Zamknij sugestie',
+  'connectionSuggestions.gotIt': 'Rozumiem',
+  'connectionSuggestions.skipForNow': 'Pomin na razie',
+  'connectionSuggestions.spotifyDiscoverWeekly': 'Skonfiguruj subskrypcje Discover Weekly',
+  'connectionSuggestions.spotifyReleaseRadar': 'Wlacz subskrypcje Release Radar',
+  'connectionSuggestions.lastfmTopGenreTags': 'Subskrybuj swoje najwazniejsze tagi gatunkowe',
+  'connectionSuggestions.listenbrainzWeeklyJams': 'Wlacz subskrypcje Weekly Jams',
+  'connectionSuggestions.listenbrainzWeeklyExploration': 'Wlacz Weekly Exploration',
+  'connectionSuggestions.weeklyDigestPlaylist': 'Utworz cotygodniowa playliste Digarr Digest',
+  'keyboardShortcuts.title': 'Skroty klawiaturowe',
+  'keyboardShortcuts.close': 'Zamknij skroty',
+  'keyboardShortcuts.nextCard': 'Nastepna karta',
+  'keyboardShortcuts.previousCard': 'Poprzednia karta',
+  'keyboardShortcuts.approveFocusedCard': 'Zatwierdz aktywna karte',
+  'keyboardShortcuts.rejectFocusedCard': 'Odrzuc aktywna karte',
+  'keyboardShortcuts.toggleDetailView': 'Otworz / zamknij widok szczegolow',
+  'keyboardShortcuts.toggleShortcuts': 'Pokaz / ukryj skroty',
+  'keyboardShortcuts.previousCardStack': 'Poprzednia karta (widok stosu)',
+  'keyboardShortcuts.nextCardStack': 'Nastepna karta (widok stosu)',
+  'keyboardShortcuts.disabledWhileTyping': 'Skroty sa wylaczone, gdy aktywne jest pole tekstowe.',
+  'admin.restoreSummary': 'Przywrocono {0} wierszy w {1} tabelach',
+  'admin.restoreSummaryForce':
+    'Przywrocono {0} wierszy. Wpisz ponownie dane logowania dla zaszyfrowanych pol.',
+  'admin.encryptionMismatchMessage':
+    'Kopia zostala utworzona przy uzyciu innego klucza szyfrowania.',
+  'admin.encryptionMismatchAffected': 'Te pola trzeba bedzie wpisac ponownie: {0}',
+  'admin.encryptionMismatchFallback': 'Niektore zaszyfrowane pola moga wymagac ponownego wpisania.',
+  'subscriptions.spotifyNotConnected': 'Spotify nie jest polaczone',
+  'playlistForm.dialogCreate': 'Utworz playliste',
+  'playlistForm.dialogEdit': 'Edytuj playliste',
+  'playlistForm.titleCreate': 'Utworz playliste',
+  'playlistForm.titleEdit': 'Edytuj playliste',
+  'playlistForm.namePlaceholder': 'Moj tygodniowy miks',
+  'playlistForm.strategy': 'Strategia',
+  'playlistForm.strategyWeeklyDigestDesc': 'Wybor utworow z ostatnich odkryc',
+  'playlistForm.strategyGenreFocusDesc': 'Glebokie zanurzenie w wybrany gatunek',
+  'playlistForm.strategyMoodMixDesc': 'Utwory pasujace do wybranego nastroju',
+  'playlistForm.strategyRediscoverDesc': 'Przypomnij sobie artystów, o ktorych mogles zapomniec',
+  'playlistForm.mood': 'Nastroj',
+  'playlistForm.moodPlaceholder': 'nocna jazda autem',
+  'playlistForm.trackCount': 'Liczba utworow',
+  'playlistForm.scheduleAutomaticGeneration': 'Zaplanuj automatyczne generowanie',
+  'playlistForm.nameRequired': 'Nazwa jest wymagana',
+  'playlistForm.genreRequired': 'Gatunek jest wymagany dla strategii Fokus na gatunek',
+  'playlistForm.moodRequired': 'Nastroj jest wymagany dla strategii Mood Mix',
+  'searchResultCard.addToQueue': 'Dodaj do kolejki odkryc',
+  'searchResultCard.queuing': 'Dodawanie...',
+  'searchResultCard.queuedSuccess': '{0} dodano do kolejki odkryc',
+  'searchResultCard.queuedFailure': 'Nie udalo się dodac {0} do kolejki',
+  'searchResultCard.popularity': '{0}{1} popularnosci',
+  'subscriptionPresets.getStarted': 'Zacznij od gotowego zestawu',
+  'subscriptionPresets.intro':
+    'Wybierz zestaw, aby automatycznie utworzyc subskrypcje, albo zbuduj własna konfiguracje.',
+  'subscriptionPresets.casualTitle': 'Na luzie',
+  'subscriptionPresets.casualDescription': 'Cotygodniowe rekomendacje z niewielka iloscia szumu',
+  'subscriptionPresets.deepDiverTitle': 'Gleboki nurek',
+  'subscriptionPresets.deepDiverDescription':
+    'Szerokie odkrywanie z wielu zrodel i codziennymi aktualizacjami',
+  'subscriptionPresets.discoveryOnlyTitle': 'Tylko odkrywanie',
+  'subscriptionPresets.discoveryOnlyDescription': 'Tylko pipeline, bez zewnetrznych feedow',
+  'subscriptionPresets.customTitle': 'Wlasny',
+  'subscriptionPresets.customDescription': 'Zbuduj własna mieszanke subskrypcji',
+  'subscriptionPresets.oneConnectedService': '1 podlaczona usluga',
+  'subscriptionPresets.twoConnectedServices': '2 podlaczone uslugi',
+  'subscriptionPresets.requires': 'Wymaga: {0}',
+  'subscriptionPresets.setUp': 'Skonfiguruj',
+  'subscriptionPresets.applying': 'Stosowanie...',
+  'subscriptionPresets.casualApplied': 'Zastosowano zestaw Na luzie',
+  'subscriptionPresets.deepDiverApplied': 'Zastosowano zestaw Gleboki nurek',
+  'subscriptionPresets.applyFailed': 'Nie udalo się zastosowac zestawu',
+  'subscriptionPresets.aiOnlyEnabled': 'Wlaczono tryb tylko AI',
+  'subscriptionPresets.aiOnlyEnableFailed': 'Nie udalo się włączyć trybu tylko AI',
+  'jobHistory.pipeline': 'Proces odkrywania',
+  'common.status': 'Stan',
+  'subscriptionForm.dialogCreate': 'Utwórz subskrypcję',
+  'subscriptionForm.dialogEdit': 'Edytuj subskrypcję',
+  'subscriptionForm.titleCreate': 'Utwórz subskrypcję',
+  'subscriptionForm.titleEdit': 'Edytuj subskrypcję',
+  'subscriptionForm.name': 'Nazwa',
+  'subscriptionForm.namePlaceholder': 'Nowe Shoegaze Weekly',
+  'subscriptionForm.nameRequired': 'Nazwa jest wymagana',
+  'subscriptionForm.source': 'Źródło',
+  'subscriptionForm.sources': 'Źródła',
+  'subscriptionForm.sourceType': 'Typ źródła',
+  'subscriptionForm.sourceTypeGenre': 'Gatunek muzyczny',
+  'subscriptionForm.sourceTypeGenreDesc': 'Odkrywaj artystów w określonym gatunku',
+  'subscriptionForm.sourceTypeSimilar': 'Podobni artyści',
+  'subscriptionForm.sourceTypeSimilarDesc': 'Znajdź artystów podobnych do swoich ulubionych',
+  'subscriptionForm.sourceTypeDiscoveryModeDesc': 'Użyj trybu odkrywania, aby generować artystów',
+  'subscriptionForm.discoveryMode': 'Tryb odkrywania',
+  'subscriptionForm.selectDiscoveryMode': 'Wybierz tryb odkrywania do skonfigurowania.',
+  'subscriptionForm.discoveryModeSettingsRequired': 'Ustawienia trybu odkrywania są wymagane',
+  'subscriptionForm.genre': 'Gatunek muzyczny',
+  'subscriptionForm.genrePlaceholder': 'shoegaze',
+  'subscriptionForm.seedArtistsPlaceholder': 'Radiohead, Portishead, Massive Attack',
+  'subscriptionForm.genreRequired': 'Gatunek jest wymagany',
+  'subscriptionForm.seedArtists': 'Artyści startowi',
+  'subscriptionForm.seedArtistsHelp': 'Nazwy artystów oddzielone przecinkami',
+  'subscriptionForm.seedArtistsRequired': 'Wymagany jest co najmniej jeden artysta startowy',
+  'subscriptionForm.configureInSettings': 'Skonfiguruj w Ustawienia > Połączenia',
+  'subscriptionForm.notConfigured': 'nie skonfigurowano',
+  'subscriptionForm.selectAtLeastOneSource': 'Wybierz co najmniej jedno źródło',
+  'subscriptionForm.sourceLockedHelp':
+    'Typ źródła i konfiguracja nie mogą zostać zmienione dla tej subskrypcji.',
+  'subscriptionForm.maxArtistsPerRun': 'Maksymalna liczba artystów na uruchomienie',
+  'subscriptionForm.scoringPreset': 'Preset punktacji',
+  'subscriptionForm.scoringPresetDefault': 'Domyślny',
+  'subscriptionForm.scoringPresetGenre': 'Zoptymalizowany pod gatunki',
+  'cronPicker.preset': 'Ustawienie harmonogramu',
+  'cronPicker.dailyMidnight': 'Codziennie (północ)',
+  'cronPicker.twiceAWeek': 'Dwa razy w tygodniu (pon + czw)',
+  'cronPicker.everyTwoWeeks': 'Co 2 tygodnie (1 i 15)',
+  'cronPicker.monthlyFirst': 'Co miesiąc (1. dzień)',
+  'cronPicker.standardHint':
+    'Standardowy cron: minuta godzina dzień-miesiąca miesiąc dzień-tygodnia',
+  'settings.capImport': 'Importowanie',
+  'systemHealth.pipeline': 'Proces odkrywania',
+  'systemHealth.librarySync': 'Synchronizacja biblioteki',
+  'discoveryMode.field.feed': 'Źródło feedu',
+  'discoveryMode.field.limit': 'Maksymalna liczba',
+  'settings.tabs.systemHealth': 'Stan systemu',
+  'settings.targetUpdated': 'Cel został zaktualizowany',
+  'settings.targetUpdateFailed': 'Nie udało się zaktualizować celu',
+  'settings.editTarget': 'Edytuj cel',
+  'settings.saveChanges': 'Zapisz zmiany',
+  'settings.linkedTarget': 'Powiązany cel:',
+  'settings.librarySyncInterval': 'Synchronizacja biblioteki',
+  'settings.librarySyncIntervalHours': 'Interwał synchronizacji (godziny)',
+  'settings.librarySyncIntervalHelp':
+    'Określa, jak często Digarr odświeża dane synchronizacji biblioteki i automatycznie ponawia kontrole.',
+  'settings.fieldModel': 'Model AI',
+  'settings.fieldClientId': 'Identyfikator klienta',
+  'settings.fieldClientSecret': 'Sekret klienta',
+  'settings.aiProviderTitle': 'Dostawca AI',
+  'settings.webhookTitle': 'Integracja webhook',
+  'settings.plexToken': 'Token Plex',
+  'settings.serviceConnectedToast': '{0} połączono',
+  'settings.serviceUnreachableToast': 'Nie udało się połączyć z {0}',
+  'settings.serviceSettingsSavedToast': 'Ustawienia {0} zostały zapisane',
+  'settings.serviceSettingsFailedToast': 'Nie udało się zapisać ustawień {0}',
+  'settings.spotifyAuthorizationFailed': 'Nie udało się rozpocząć autoryzacji Spotify',
+  'settings.spotifyDisconnectFailed': 'Nie udało się odłączyć Spotify',
+  'settings.spotifyLikedSongsFailed': 'Nie udało się rozpocząć importu Spotify Liked Songs',
+  'settings.spotifyPlaylistFailed': 'Nie udało się rozpocząć importu playlisty',
+  'settings.deezerFavoritesFailed': 'Nie udało się rozpocząć importu ulubionych z Deezer',
+  'settings.deezerFollowedFailed':
+    'Nie udało się rozpocząć importu obserwowanych artystów z Deezer',
+  'settings.deezerDisconnectFailed': 'Nie udało się odłączyć Deezer',
+  'settings.deezerAuthorizationFailed': 'Nie udało się rozpocząć autoryzacji Deezer',
+  'settings.reverseProxyConfigHint':
+    'Użyj PROXY_AUTH_ENABLED oraz PROXY_AUTH_TRUSTED_PROXIES; proxy musi wysyłać X-Forwarded-User.',
+  'admin.autoFixInProgress': 'Automatyczna naprawa trwa... ({0}/{1} naprawiono)',
+  'albumPicker.title': 'Wybierz albumy',
+  'albumPicker.loadFailed': 'Nie udało się wczytać albumów',
+  'albumPicker.selectAll': 'Wszystkie',
+  'albumPicker.selectNone': 'Żadne',
+  'albumPicker.selectedCount': 'Wybrano: {0}',
+  'albumPicker.aiPick': 'Wybór AI',
+  'albumPicker.monitor': 'Monitoruj',
+  'albumPicker.monitorSingular': 'Monitoruj {0} album',
+  'albumPicker.monitorPlural': 'Monitoruj albumy: {0}',
+  'approveDialog.title': 'Ustawienia Lidarr dla tego wykonawcy',
+  'approveDialog.loadFailed': 'Nie udało się wczytać profili Lidarr. Sprawdź połączenie z Lidarr.',
+  'approveDialog.qualityProfile': 'Profil jakości',
+  'approveDialog.metadataProfile': 'Profil metadanych',
+  'approveDialog.rootFolder': 'Folder główny',
+  'cardStack.counter': '{0} z {1}',
+  'cardStack.swipeHint': 'Przesuń w prawo, aby zatwierdzić, w lewo, aby odrzucić',
+  'serviceCard.testConnection': 'Testuj połączenie',
+  'preview.noPreviewAvailable': 'Brak dostępnego podglądu dla tego wykonawcy',
+  'preview.playbackBlocked':
+    'Odtwarzanie zablokowane przez przeglądarkę - spróbuj kliknąć ponownie',
+  'errors.auth.invalidCredentials': 'Nieprawidłowe dane logowania',
+  'errors.auth.notAuthenticated': 'Nie uwierzytelniono',
+  'errors.auth.passwordIncorrect': 'Bieżące hasło jest nieprawidłowe',
+  'errors.auth.usernameTaken': 'Nazwa użytkownika jest już zajęta',
+  'errors.target.notFound': 'Nie znaleziono celu',
+  'errors.pipeline.alreadyRunning': 'Uruchomienie pipeline już trwa',
+  'errors.recommendation.notFound': 'Nie znaleziono rekomendacji',
+  'errors.subscription.notFound': 'Nie znaleziono subskrypcji',
+  'errors.playlist.notFound': 'Nie znaleziono playlisty',
+}
