+++ src/core/i18n/messages/overrides/pt-BR.ts
@@ -1,0 +1,234 @@
+import type { MessageCatalog } from '../types'
+
+export const ptBROverrides: Partial<MessageCatalog> = {
+  'common.confirm': 'Confirmar',
+  'discover.filters': 'Filtros',
+  'discover.filtersActive': 'Filtros ({0})',
+  'discover.moreActions': 'Mais ações',
+  'discover.decadesLabel': 'Décadas',
+  'analytics.title': 'Análises',
+  'dashboard.title': 'Painel',
+  'discover.title': 'Descobrir',
+  'setup.title': 'Configuração',
+  'todaysPick.discoverSimilar': 'Descobrir similares',
+  'libraryHealth.fix': 'Corrigir',
+  'libraryHealth.fixing': 'Corrigindo...',
+  'libraryHealth.showLess': 'Mostrar menos',
+  'libraryHealth.showAll': 'Mostrar todos {0}',
+  'libraryHealth.fixHint.missingMetadata': 'Atualiza metadados de artistas pelo Lidarr',
+  'libraryHealth.fixHint.unmonitored':
+    'Define todos os artistas listados como monitorados no Lidarr',
+  'libraryHealth.fixHint.missingAlbums': 'Atualiza álbuns no Lidarr para encontrar ausentes',
+  'libraryHealth.fixHint.genreGaps': 'Atualiza metadados do Lidarr para buscar gêneros',
+  'libraryHealth.fixHint.imageGaps': 'Busca imagens de artistas novamente',
+  'libraryHealth.fixHint.missingWikidata': 'Busca biografias e links no Wikidata',
+  'common.importing': 'Importando...',
+  'common.current': 'Atual',
+  'common.custom': 'Personalizado',
+  'common.enabled': 'Ativado',
+  'common.unauthorized': 'Não autorizado',
+  'common.adminAccessRequired': 'Acesso de administrador obrigatório',
+  'jobHistory.pipeline': 'Fluxo de descoberta',
+  'jobHistory.librarySync': 'Sincronização da biblioteca',
+  'jobHistory.playlists': 'Listas de reprodução',
+  'playlist.total': 'no total',
+  'playlistForm.dialogCreate': 'Criar playlist',
+  'playlistForm.dialogEdit': 'Editar playlist',
+  'playlistForm.titleCreate': 'Criar Playlist',
+  'playlistForm.titleEdit': 'Editar Playlist',
+  'playlistForm.namePlaceholder': 'Meu Mix Semanal',
+  'playlistForm.strategy': 'Estratégia',
+  'playlistForm.strategyWeeklyDigestDesc': 'Seleção curada das suas descobertas recentes',
+  'playlistForm.strategyGenreFocusDesc': 'Mergulho profundo em um gênero específico',
+  'playlistForm.strategyMoodMixDesc': 'Faixas que combinam com um clima específico',
+  'playlistForm.strategyRediscoverDesc': 'Traga de volta artistas que você pode ter esquecido',
+  'playlistForm.mood': 'Clima',
+  'playlistForm.moodPlaceholder': 'dirigindo tarde da noite',
+  'playlistForm.trackCount': 'Número de faixas',
+  'playlistForm.scheduleAutomaticGeneration': 'Agendar geração automática',
+  'playlistForm.nameRequired': 'O nome é obrigatório',
+  'playlistForm.genreRequired': 'O gênero é obrigatório para a estratégia Foco em gênero',
+  'playlistForm.moodRequired': 'O clima é obrigatório para a estratégia Mix de clima',
+  'subscriptionForm.dialogCreate': 'Criar assinatura',
+  'subscriptionForm.dialogEdit': 'Editar assinatura',
+  'subscriptionForm.titleCreate': 'Criar assinatura',
+  'subscriptionForm.titleEdit': 'Editar assinatura',
+  'subscriptionForm.name': 'Nome',
+  'subscriptionForm.namePlaceholder': 'Nova Shoegaze Weekly',
+  'subscriptionForm.nameRequired': 'O nome é obrigatório',
+  'subscriptionForm.source': 'Fonte',
+  'subscriptionForm.sources': 'Fontes',
+  'subscriptionForm.sourceType': 'Tipo de fonte',
+  'subscriptionForm.sourceTypeGenre': 'Gênero',
+  'subscriptionForm.sourceTypeGenreDesc': 'Descubra artistas em um gênero específico',
+  'subscriptionForm.sourceTypeSimilar': 'Semelhantes',
+  'subscriptionForm.sourceTypeSimilarDesc': 'Encontre artistas parecidos com os seus favoritos',
+  'subscriptionForm.sourceTypeDiscoveryModeDesc': 'Use um modo de descoberta para gerar artistas',
+  'subscriptionForm.discoveryMode': 'Modo de descoberta',
+  'subscriptionForm.selectDiscoveryMode': 'Selecione um modo de descoberta para configurá-lo.',
+  'subscriptionForm.discoveryModeSettingsRequired':
+    'As configurações do modo de descoberta são obrigatórias',
+  'subscriptionForm.genre': 'Gênero',
+  'subscriptionForm.genrePlaceholder': 'shoegaze',
+  'subscriptionForm.genreRequired': 'O gênero é obrigatório',
+  'subscriptionForm.seedArtists': 'Artistas semente',
+  'subscriptionForm.seedArtistsPlaceholder': 'Radiohead, Portishead, Massive Attack',
+  'subscriptionForm.seedArtistsHelp': 'Nomes de artistas separados por vírgulas',
+  'subscriptionForm.seedArtistsRequired': 'É necessário pelo menos um artista semente',
+  'subscriptionForm.configureInSettings': 'Configure em Configurações > Conexões',
+  'subscriptionForm.notConfigured': 'não configurado',
+  'subscriptionForm.selectAtLeastOneSource': 'Selecione pelo menos uma fonte',
+  'subscriptionForm.sourceLockedHelp':
+    'O tipo de fonte e a configuração não podem ser alterados para esta assinatura.',
+  'subscriptionForm.maxArtistsPerRun': 'Máximo de artistas por execução',
+  'subscriptionForm.scoringPreset': 'Preset de pontuação',
+  'subscriptionForm.scoringPresetDefault': 'Padrão',
+  'subscriptionForm.scoringPresetGenre': 'Otimizado para gênero',
+  'subscriptions.spotifyNotConnected': 'O Spotify não está conectado',
+  'subscriptionPresets.getStarted': 'Comece com um preset',
+  'subscriptionPresets.intro':
+    'Escolha um preset para criar assinaturas automaticamente ou monte uma configuração personalizada.',
+  'subscriptionPresets.casualTitle': 'Leve',
+  'subscriptionPresets.casualDescription': 'Recomendações semanais com pouco ruído',
+  'subscriptionPresets.deepDiverTitle': 'Mergulho profundo',
+  'subscriptionPresets.deepDiverDescription':
+    'Descoberta ampla com várias fontes e atualizações diárias',
+  'subscriptionPresets.discoveryOnlyTitle': 'Somente descoberta',
+  'subscriptionPresets.discoveryOnlyDescription': 'Apenas pipeline, sem feeds externos',
+  'subscriptionPresets.customTitle': 'Personalizado',
+  'subscriptionPresets.customDescription': 'Monte sua própria combinação de assinaturas',
+  'subscriptionPresets.oneConnectedService': '1 serviço conectado',
+  'subscriptionPresets.twoConnectedServices': '2 serviços conectados',
+  'subscriptionPresets.requires': 'Requer {0}',
+  'subscriptionPresets.setUp': 'Configurar',
+  'subscriptionPresets.applying': 'Aplicando...',
+  'subscriptionPresets.casualApplied': 'Preset Leve aplicado',
+  'subscriptionPresets.deepDiverApplied': 'Preset Mergulho profundo aplicado',
+  'subscriptionPresets.applyFailed': 'Falha ao aplicar o preset',
+  'subscriptionPresets.aiOnlyEnabled': 'Modo somente IA ativado',
+  'subscriptionPresets.aiOnlyEnableFailed': 'Falha ao ativar o modo somente IA',
+  'cronPicker.preset': 'Predefinição',
+  'cronPicker.dailyMidnight': 'Diário (meia-noite)',
+  'cronPicker.twiceAWeek': 'Duas vezes por semana (seg + qui)',
+  'cronPicker.everyTwoWeeks': 'A cada 2 semanas (1 e 15)',
+  'cronPicker.monthlyFirst': 'Mensal (dia 1)',
+  'cronPicker.standardHint': 'Cron padrão: minuto hora dia-do-mês mês dia-da-semana',
+  'libraryHealth.syncNow': 'Sincronizar agora',
+  'libraryHealth.snapshotTitle': 'Última captura',
+  'libraryHealth.lastSynced': 'Última sincronização',
+  'libraryHealth.neverSynced': 'Nunca sincronizado.',
+  'libraryHealth.autoSyncEvery': 'Sincronização automática a cada {0} h',
+  'systemHealth.pipeline': 'Fluxo de descoberta',
+  'systemHealth.librarySync': 'Sincronização da biblioteca',
+  'settings.tabs.systemHealth': 'Saúde do sistema',
+  'settings.targetUpdated': 'Destino atualizado',
+  'settings.targetUpdateFailed': 'Falha ao atualizar o destino',
+  'settings.editTarget': 'Editar destino',
+  'settings.saveChanges': 'Salvar alterações',
+  'settings.linkedTarget': 'Destino vinculado:',
+  'settings.librarySyncInterval': 'Sincronização da biblioteca',
+  'settings.librarySyncIntervalHours': 'Intervalo de sincronização (horas)',
+  'settings.librarySyncIntervalHelp':
+    'Controla com que frequência o Digarr atualiza os dados de sincronização da biblioteca e executa novamente as verificações automaticamente.',
+  'settings.oidcSso': 'OIDC e SSO',
+  'settings.aiProviderTitle': 'Provedor de IA',
+  'settings.webhookTitle': 'Integração webhook',
+  'settings.plexToken': 'Token do Plex',
+  'settings.serviceConnectedToast': '{0} conectado',
+  'settings.serviceUnreachableToast': '{0} está inacessível',
+  'settings.serviceSettingsSavedToast': 'Configurações de {0} salvas',
+  'settings.serviceSettingsFailedToast': 'Falha ao salvar as configurações de {0}',
+  'settings.spotifyAuthorizationFailed': 'Falha ao iniciar a autorização do Spotify',
+  'settings.spotifyDisconnectFailed': 'Falha ao desconectar o Spotify',
+  'settings.spotifyLikedSongsFailed': 'Falha ao iniciar a importação das curtidas do Spotify',
+  'settings.spotifyPlaylistFailed': 'Falha ao importar playlist do Spotify',
+  'settings.deezerFavoritesFailed': 'Falha ao iniciar a importação dos favoritos do Deezer',
+  'settings.deezerFollowedFailed': 'Falha ao iniciar a importação dos artistas seguidos do Deezer',
+  'settings.deezerDisconnectFailed': 'Falha ao desconectar o Deezer',
+  'settings.deezerAuthorizationFailed': 'Falha ao iniciar a autorização do Deezer',
+  'settings.reverseProxyConfigHint':
+    'Use PROXY_AUTH_ENABLED e PROXY_AUTH_TRUSTED_PROXIES; seu proxy deve enviar X-Forwarded-User.',
+  'searchResultCard.addToQueue': 'Adicionar à fila de descoberta',
+  'searchResultCard.queuing': 'Adicionando...',
+  'searchResultCard.queuedSuccess': '{0} foi adicionado à fila de descoberta',
+  'searchResultCard.queuedFailure': 'Falha ao adicionar {0} à fila',
+  'searchResultCard.popularity': '{0}{1} de popularidade',
+  'admin.restoreSummary': '{0} linhas restauradas em {1} tabelas',
+  'admin.restoreSummaryForce':
+    '{0} linhas restauradas. Digite novamente as credenciais dos campos criptografados.',
+  'admin.encryptionMismatchMessage': 'O backup foi criado com uma chave de criptografia diferente.',
+  'admin.encryptionMismatchAffected': 'Estes campos precisarão ser preenchidos novamente: {0}',
+  'admin.encryptionMismatchFallback':
+    'Alguns campos criptografados podem precisar ser preenchidos novamente.',
+  'admin.autoFixInProgress': 'Correção automática em andamento... ({0}/{1} corrigidos)',
+  'errorBoundary.title': 'Algo deu errado',
+  'errorBoundary.retry': 'Tentar novamente',
+  'errorBoundary.home': 'Ir para o início',
+  'connectionSuggestions.title': 'Sugestões para {0}',
+  'connectionSuggestions.close': 'Fechar sugestões',
+  'connectionSuggestions.gotIt': 'Entendi',
+  'connectionSuggestions.skipForNow': 'Pular por enquanto',
+  'connectionSuggestions.spotifyDiscoverWeekly': 'Configurar uma assinatura do Discover Weekly',
+  'connectionSuggestions.spotifyReleaseRadar': 'Ativar uma assinatura do Release Radar',
+  'connectionSuggestions.lastfmTopGenreTags': 'Assinar suas principais tags de gênero',
+  'connectionSuggestions.listenbrainzWeeklyJams': 'Ativar a assinatura Weekly Jams',
+  'connectionSuggestions.listenbrainzWeeklyExploration': 'Ativar Weekly Exploration',
+  'connectionSuggestions.weeklyDigestPlaylist': 'Criar uma playlist semanal do Digarr Digest',
+  'keyboardShortcuts.title': 'Atalhos de teclado',
+  'keyboardShortcuts.close': 'Fechar atalhos',
+  'keyboardShortcuts.nextCard': 'Próximo cartão',
+  'keyboardShortcuts.previousCard': 'Cartão anterior',
+  'keyboardShortcuts.approveFocusedCard': 'Aprovar o cartão em foco',
+  'keyboardShortcuts.rejectFocusedCard': 'Rejeitar o cartão em foco',
+  'keyboardShortcuts.toggleDetailView': 'Abrir / fechar a visualização de detalhes',
+  'keyboardShortcuts.toggleShortcuts': 'Mostrar / ocultar atalhos',
+  'keyboardShortcuts.previousCardStack': 'Cartão anterior (visão em pilha)',
+  'keyboardShortcuts.nextCardStack': 'Próximo cartão (visão em pilha)',
+  'keyboardShortcuts.disabledWhileTyping':
+    'Os atalhos são desativados quando um campo de texto está em foco.',
+  'app.themeGroupEditor': 'Ferramentas do editor',
+  'common.manual': 'manual',
+  'jobHistory.playlistType': 'Lista de reprodução',
+  'common.status': 'Status atual',
+  'common.total': 'Total acumulado',
+  'settings.capDiscovery': 'Descoberta',
+  'settings.capImport': 'Importação',
+  'settings.capLibrarySync': 'Sincronização da biblioteca',
+  'settings.capPlaylistExport': 'Exportação de playlists',
+  'settings.capService': 'Serviço conectado',
+  'settings.capSubscriptions': 'Assinaturas',
+  'search.experimental': 'Em teste',
+  'discoveryMode.field.feed': 'Origem do feed',
+  'discoveryMode.field.tags': 'Etiquetas',
+  'settings.totalWeight': 'peso total:',
+  'integration.playlist': 'Lista de reprodução',
+  'albumPicker.title': 'Selecionar álbuns',
+  'albumPicker.loadFailed': 'Falha ao carregar álbuns',
+  'albumPicker.selectAll': 'Todos',
+  'albumPicker.selectNone': 'Nenhum',
+  'albumPicker.selectedCount': '{0} selecionados',
+  'albumPicker.aiPick': 'Escolha da IA',
+  'albumPicker.monitor': 'Monitorar',
+  'albumPicker.monitorSingular': 'Monitorar {0} álbum',
+  'albumPicker.monitorPlural': 'Monitorar {0} álbuns',
+  'approveDialog.title': 'Configurações do Lidarr para este artista',
+  'approveDialog.loadFailed':
+    'Falha ao carregar os perfis do Lidarr. Verifique sua conexão com o Lidarr.',
+  'approveDialog.qualityProfile': 'Perfil de qualidade',
+  'approveDialog.metadataProfile': 'Perfil de metadados',
+  'approveDialog.rootFolder': 'Pasta raiz',
+  'cardStack.counter': '{0} de {1}',
+  'cardStack.swipeHint': 'Deslize para a direita para aprovar, para a esquerda para rejeitar',
+  'serviceCard.testConnection': 'Testar conexão',
+  'preview.noPreviewAvailable': 'Sem prévia disponível para este artista',
+  'preview.playbackBlocked': 'Reprodução bloqueada pelo navegador - tente clicar novamente',
+  'errors.auth.invalidCredentials': 'Credenciais inválidas',
+  'errors.auth.notAuthenticated': 'Não autenticado',
+  'errors.auth.passwordIncorrect': 'A senha atual está incorreta',
+  'errors.auth.usernameTaken': 'Nome de usuário já em uso',
+  'errors.target.notFound': 'Destino não encontrado',
+  'errors.pipeline.alreadyRunning': 'Já há uma execução do pipeline em andamento',
+  'errors.recommendation.notFound': 'Recomendação não encontrada',
+  'errors.subscription.notFound': 'Assinatura não encontrada',
+  'errors.playlist.notFound': 'Playlist não encontrada',
+}
