+++ src/core/i18n/messages/overrides/ro.ts
@@ -1,0 +1,239 @@
+import type { MessageCatalog } from '../types'
+
+export const roOverrides: Partial<MessageCatalog> = {
+  'common.confirm': 'Confirmă',
+  'discover.filters': 'Filtre',
+  'discover.filtersActive': 'Filtre ({0})',
+  'discover.moreActions': 'Mai multe acțiuni',
+  'discover.decadesLabel': 'Decenii',
+  'analytics.title': 'Analize',
+  'dashboard.title': 'Tablou de bord',
+  'discover.title': 'Descoperire',
+  'setup.title': 'Configurare',
+  'todaysPick.discoverSimilar': 'Descoperă similari',
+  'libraryHealth.fix': 'Remediază',
+  'libraryHealth.fixing': 'Se remediază...',
+  'libraryHealth.showLess': 'Afișează mai puțin',
+  'libraryHealth.showAll': 'Afișează toate {0}',
+  'libraryHealth.fixHint.missingMetadata': 'Actualizează metadatele artiștilor din Lidarr',
+  'libraryHealth.fixHint.unmonitored': 'Setează toți artiștii listați ca monitorizați în Lidarr',
+  'libraryHealth.fixHint.missingAlbums': 'Actualizează albumele în Lidarr pentru a găsi lipsurile',
+  'libraryHealth.fixHint.genreGaps': 'Actualizează metadatele Lidarr pentru genuri',
+  'libraryHealth.fixHint.imageGaps': 'Caută din nou imagini pentru artiști',
+  'libraryHealth.fixHint.missingWikidata': 'Preia biografii și linkuri din Wikidata',
+  'app.themeGroupEditor': 'Editare',
+  'app.themeGroupStreaming': 'Servicii de streaming',
+  'setup.lidarrUrl': 'Adresa Lidarr',
+  'common.importing': 'Importare...',
+  'common.manual': 'manual',
+  'common.current': 'Curent',
+  'common.custom': 'Personalizat',
+  'common.enabled': 'Activat',
+  'common.unauthorized': 'Neautorizat',
+  'common.adminAccessRequired': 'Este necesar accesul de administrator',
+  'common.total': 'În total',
+  'genreDetail.artistSingular': 'muzician',
+  'jobHistory.pipeline': 'Rulare de descoperire',
+  'jobHistory.librarySync': 'Sincronizarea bibliotecii',
+  'jobHistory.playlistType': 'Listă de redare',
+  'playlist.total': 'în total',
+  'playlistForm.dialogCreate': 'Creează lista de redare',
+  'playlistForm.dialogEdit': 'Editează lista de redare',
+  'playlistForm.titleCreate': 'Creează lista de redare',
+  'playlistForm.titleEdit': 'Editează lista de redare',
+  'playlistForm.namePlaceholder': 'Mixul meu săptămânal',
+  'playlistForm.strategy': 'Strategie',
+  'playlistForm.strategyWeeklyDigestDesc': 'Selecție atentă din descoperirile tale recente',
+  'playlistForm.strategyGenreFocusDesc': 'Explorează în profunzime un anumit gen',
+  'playlistForm.strategyMoodMixDesc': 'Piese potrivite pentru o anumită stare',
+  'playlistForm.strategyRediscoverDesc': 'Adu înapoi artiști pe care poate i-ai uitat',
+  'playlistForm.mood': 'Stare',
+  'playlistForm.moodPlaceholder': 'drum de noapte',
+  'playlistForm.trackCount': 'Număr de piese',
+  'playlistForm.scheduleAutomaticGeneration': 'Planifică generarea automată',
+  'playlistForm.nameRequired': 'Numele este obligatoriu',
+  'playlistForm.genreRequired': 'Genul este obligatoriu pentru strategia Focus pe gen',
+  'playlistForm.moodRequired': 'Starea este obligatorie pentru strategia Mood Mix',
+  'subscriptionForm.dialogCreate': 'Creează abonament',
+  'subscriptionForm.dialogEdit': 'Editează abonamentul',
+  'subscriptionForm.titleCreate': 'Creează abonament',
+  'subscriptionForm.titleEdit': 'Editează abonamentul',
+  'subscriptionForm.name': 'Nume',
+  'subscriptionForm.namePlaceholder': 'Shoegaze Weekly nou',
+  'subscriptionForm.nameRequired': 'Numele este obligatoriu',
+  'subscriptionForm.source': 'Sursă',
+  'subscriptionForm.sources': 'Surse',
+  'subscriptionForm.sourceType': 'Tip de sursă',
+  'subscriptionForm.sourceTypeGenre': 'Gen muzical',
+  'subscriptionForm.sourceTypeGenreDesc': 'Descoperă artiști într-un gen anume',
+  'subscriptionForm.sourceTypeSimilar': 'Artiști similari',
+  'subscriptionForm.sourceTypeSimilarDesc': 'Găsește artiști asemănători cu favoriții tăi',
+  'subscriptionForm.sourceTypeDiscoveryModeDesc':
+    'Folosește un mod de descoperire pentru a genera artiști',
+  'subscriptionForm.discoveryMode': 'Mod de descoperire',
+  'subscriptionForm.selectDiscoveryMode': 'Alege un mod de descoperire pe care să îl configurezi.',
+  'subscriptionForm.discoveryModeSettingsRequired':
+    'Setările pentru modul de descoperire sunt obligatorii',
+  'subscriptionForm.genre': 'Gen muzical',
+  'subscriptionForm.genrePlaceholder': 'shoegaze',
+  'subscriptionForm.seedArtistsPlaceholder': 'Radiohead, Portishead, Massive Attack',
+  'subscriptionForm.genreRequired': 'Genul este obligatoriu',
+  'subscriptionForm.seedArtists': 'Artiști de pornire',
+  'subscriptionForm.seedArtistsHelp': 'Nume de artiști separate prin virgulă',
+  'subscriptionForm.seedArtistsRequired': 'Este necesar cel puțin un artist de pornire',
+  'subscriptionForm.configureInSettings': 'Configurează în Setări > Conexiuni',
+  'subscriptionForm.notConfigured': 'neconfigurat',
+  'subscriptionForm.selectAtLeastOneSource': 'Selectează cel puțin o sursă',
+  'subscriptionForm.sourceLockedHelp':
+    'Tipul sursei și configurația nu pot fi schimbate pentru acest abonament.',
+  'subscriptionForm.maxArtistsPerRun': 'Număr maxim de artiști pe rulare',
+  'subscriptionForm.scoringPreset': 'Preset de scor',
+  'subscriptionForm.scoringPresetDefault': 'Implicit',
+  'subscriptionForm.scoringPresetGenre': 'Optimizat pentru gen',
+  'subscriptions.spotifyNotConnected': 'Spotify nu este conectat',
+  'subscriptionPresets.getStarted': 'Începe cu un preset',
+  'subscriptionPresets.intro':
+    'Alege un preset pentru a crea abonamente automat sau construiește-ți propria configurație.',
+  'subscriptionPresets.casualTitle': 'Lejer',
+  'subscriptionPresets.casualDescription': 'Recomandări săptămânale cu puțin zgomot',
+  'subscriptionPresets.deepDiverTitle': 'Explorare profundă',
+  'subscriptionPresets.deepDiverDescription':
+    'Descoperire amplă din mai multe surse și actualizări zilnice',
+  'subscriptionPresets.discoveryOnlyTitle': 'Doar descoperire',
+  'subscriptionPresets.discoveryOnlyDescription': 'Doar pipeline, fără feeduri externe',
+  'subscriptionPresets.customTitle': 'Personalizat',
+  'subscriptionPresets.customDescription': 'Construiește-ți propria combinație de abonamente',
+  'subscriptionPresets.oneConnectedService': '1 serviciu conectat',
+  'subscriptionPresets.twoConnectedServices': '2 servicii conectate',
+  'subscriptionPresets.requires': 'Necesită {0}',
+  'subscriptionPresets.setUp': 'Configurează',
+  'subscriptionPresets.applying': 'Se aplică...',
+  'subscriptionPresets.casualApplied': 'Presetul Lejer a fost aplicat',
+  'subscriptionPresets.deepDiverApplied': 'Presetul Explorare profundă a fost aplicat',
+  'subscriptionPresets.applyFailed': 'Presetul nu a putut fi aplicat',
+  'subscriptionPresets.aiOnlyEnabled': 'Modul doar AI a fost activat',
+  'subscriptionPresets.aiOnlyEnableFailed': 'Modul doar AI nu a putut fi activat',
+  'cronPicker.preset': 'Programare predefinită',
+  'cronPicker.dailyMidnight': 'Zilnic (miezul nopții)',
+  'cronPicker.twiceAWeek': 'De două ori pe săptămână (lun + joi)',
+  'cronPicker.everyTwoWeeks': 'La fiecare 2 săptămâni (1 și 15)',
+  'cronPicker.monthlyFirst': 'Lunar (ziua 1)',
+  'cronPicker.standardHint': 'Cron standard: minut oră zi-din-lună lună zi-din-săptămână',
+  'recommendation.sourceSimilar': 'Asemănător',
+  'recommendation.streaming': 'Serviciu de streaming',
+  'libraryHealth.syncNow': 'Sincronizează acum',
+  'libraryHealth.snapshotTitle': 'Ultimul instantaneu',
+  'libraryHealth.lastSynced': 'Ultima sincronizare',
+  'libraryHealth.neverSynced': 'Nu a fost sincronizat niciodată.',
+  'libraryHealth.autoSyncEvery': 'Sincronizare automată la fiecare {0} h',
+  'settings.capDiscovery': 'Descoperire',
+  'settings.capImport': 'Importare',
+  'settings.capLibrarySync': 'Sync bibliotecă',
+  'settings.capPlaylistExport': 'Export listă',
+  'settings.capService': 'Serviciu',
+  'settings.capSubscriptions': 'Abonamente',
+  'systemHealth.pipeline': 'Rulare de descoperire',
+  'systemHealth.librarySync': 'Sincronizarea bibliotecii',
+  'search.experimental': 'În test',
+  'discoveryMode.field.artist': 'Nume artist',
+  'settings.tabs.systemHealth': 'Starea sistemului',
+  'settings.targetUpdated': 'Ținta a fost actualizată',
+  'settings.targetUpdateFailed': 'Actualizarea țintei a eșuat',
+  'settings.editTarget': 'Editează ținta',
+  'settings.saveChanges': 'Salvează modificările',
+  'settings.linkedTarget': 'Țintă asociată:',
+  'settings.totalWeight': 'în total:',
+  'settings.librarySyncInterval': 'Sync bibliotecă',
+  'settings.librarySyncIntervalHours': 'Interval de sincronizare (ore)',
+  'settings.librarySyncIntervalHelp':
+    'Controlează cât de des Digarr reîmprospătează datele de sincronizare ale bibliotecii și rerulează verificările automat.',
+  'settings.oidcSso': 'OIDC / SSO',
+  'settings.fieldModel': 'Model AI',
+  'settings.aiProviderTitle': 'Furnizor AI',
+  'settings.webhookTitle': 'Integrare webhook',
+  'settings.plexToken': 'Token Plex',
+  'settings.serviceConnectedToast': '{0} conectat',
+  'settings.serviceUnreachableToast': '{0} nu poate fi accesat',
+  'settings.serviceSettingsSavedToast': 'Setările pentru {0} au fost salvate',
+  'settings.serviceSettingsFailedToast': 'Setările pentru {0} nu au putut fi salvate',
+  'settings.spotifyAuthorizationFailed': 'Pornirea autorizării Spotify a eșuat',
+  'settings.spotifyDisconnectFailed': 'Deconectarea Spotify a eșuat',
+  'settings.spotifyLikedSongsFailed': 'Pornirea importului Spotify Liked Songs a eșuat',
+  'settings.spotifyPlaylistFailed': 'Pornirea importului playlistului a eșuat',
+  'settings.deezerFavoritesFailed': 'Pornirea importului favorite Deezer a eșuat',
+  'settings.deezerFollowedFailed': 'Pornirea importului artiștilor urmăriți pe Deezer a eșuat',
+  'settings.deezerDisconnectFailed': 'Deconectarea Deezer a eșuat',
+  'settings.deezerAuthorizationFailed': 'Pornirea autorizării Deezer a eșuat',
+  'settings.reverseProxyConfigHint':
+    'Folosește PROXY_AUTH_ENABLED și PROXY_AUTH_TRUSTED_PROXIES; proxy-ul tău trebuie să trimită X-Forwarded-User.',
+  'searchResultCard.addToQueue': 'Adaugă în coada de descoperire',
+  'searchResultCard.queuing': 'Se adaugă...',
+  'searchResultCard.queuedSuccess': '{0} a fost adăugat în coada de descoperire',
+  'searchResultCard.queuedFailure': '{0} nu a putut fi adăugat în coadă',
+  'searchResultCard.popularity': 'Popularitate {0}{1}',
+  'genres.artistCountSingular': 'artist',
+  'admin.restoreSummary': '{0} rânduri restaurate în {1} tabele',
+  'admin.restoreSummaryForce':
+    '{0} rânduri restaurate. Reintrodu credențialele pentru câmpurile criptate.',
+  'admin.encryptionMismatchMessage':
+    'Această copie de siguranță a fost creată cu o altă cheie de criptare.',
+  'admin.encryptionMismatchAffected': 'Aceste câmpuri trebuie completate din nou: {0}',
+  'admin.encryptionMismatchFallback': 'Unele câmpuri criptate pot necesita completare din nou.',
+  'admin.autoFixInProgress': 'Corectare automată în curs... ({0}/{1} corectate)',
+  'integration.playlist': 'Listă de redare',
+  'errorBoundary.title': 'A apărut o problemă',
+  'errorBoundary.retry': 'Încearcă din nou',
+  'errorBoundary.home': 'Mergi la început',
+  'connectionSuggestions.title': 'Sugestii pentru {0}',
+  'connectionSuggestions.close': 'Închide sugestiile',
+  'connectionSuggestions.gotIt': 'Am înțeles',
+  'connectionSuggestions.skipForNow': 'Sari peste momentan',
+  'connectionSuggestions.spotifyDiscoverWeekly': 'Configurează un abonament Discover Weekly',
+  'connectionSuggestions.spotifyReleaseRadar': 'Activează un abonament Release Radar',
+  'connectionSuggestions.lastfmTopGenreTags': 'Abonează-te la principalele tale etichete de gen',
+  'connectionSuggestions.listenbrainzWeeklyJams': 'Activează abonamentul Weekly Jams',
+  'connectionSuggestions.listenbrainzWeeklyExploration': 'Activează Weekly Exploration',
+  'connectionSuggestions.weeklyDigestPlaylist': 'Creează o listă săptămânală Digarr Digest',
+  'keyboardShortcuts.title': 'Scurtături de tastatură',
+  'keyboardShortcuts.close': 'Închide scurtăturile',
+  'keyboardShortcuts.nextCard': 'Cardul următor',
+  'keyboardShortcuts.previousCard': 'Cardul anterior',
+  'keyboardShortcuts.approveFocusedCard': 'Aprobă cardul activ',
+  'keyboardShortcuts.rejectFocusedCard': 'Respinge cardul activ',
+  'keyboardShortcuts.toggleDetailView': 'Deschide / închide vizualizarea detaliată',
+  'keyboardShortcuts.toggleShortcuts': 'Arată / ascunde scurtăturile',
+  'keyboardShortcuts.previousCardStack': 'Cardul anterior (vizualizare stivă)',
+  'keyboardShortcuts.nextCardStack': 'Cardul următor (vizualizare stivă)',
+  'keyboardShortcuts.disabledWhileTyping':
+    'Scurtăturile sunt dezactivate cât timp un câmp text are focus.',
+  'setup.model': 'Model AI',
+  'albumPicker.title': 'Selectează albume',
+  'albumPicker.loadFailed': 'Încărcarea albumelor a eșuat',
+  'albumPicker.selectAll': 'Toate',
+  'albumPicker.selectNone': 'Niciunul',
+  'albumPicker.selectedCount': '{0} selectate',
+  'albumPicker.aiPick': 'Alegere AI',
+  'albumPicker.monitor': 'Monitorizează',
+  'albumPicker.monitorSingular': 'Monitorizează {0} album',
+  'albumPicker.monitorPlural': 'Monitorizează {0} albume',
+  'approveDialog.title': 'Setări Lidarr pentru acest artist',
+  'approveDialog.loadFailed':
+    'Încărcarea profilurilor Lidarr a eșuat. Verifică conexiunea la Lidarr.',
+  'approveDialog.qualityProfile': 'Profil de calitate',
+  'approveDialog.metadataProfile': 'Profil de metadate',
+  'approveDialog.rootFolder': 'Dosar rădăcină',
+  'cardStack.counter': '{0} din {1}',
+  'cardStack.swipeHint': 'Glisează dreapta pentru a aproba, stânga pentru a respinge',
+  'serviceCard.testConnection': 'Testează conexiunea',
+  'preview.noPreviewAvailable': 'Nu există previzualizare pentru acest artist',
+  'preview.playbackBlocked': 'Redare blocată de browser - încearcă să dai clic din nou',
+  'errors.auth.invalidCredentials': 'Credențiale nevalide',
+  'errors.auth.notAuthenticated': 'Neautentificat',
+  'errors.auth.passwordIncorrect': 'Parola actuală este incorectă',
+  'errors.auth.usernameTaken': 'Numele de utilizator este deja folosit',
+  'errors.target.notFound': 'Destinația nu a fost găsită',
+  'errors.pipeline.alreadyRunning': 'O execuție a pipeline-ului este deja în desfășurare',
+  'errors.recommendation.notFound': 'Recomandarea nu a fost găsită',
+  'errors.subscription.notFound': 'Abonamentul nu a fost găsit',
+  'errors.playlist.notFound': 'Lista de redare nu a fost găsită',
+}
