+++ src/core/i18n/messages/overrides/ru.ts
@@ -1,0 +1,229 @@
+import type { MessageCatalog } from '../types'
+
+export const ruOverrides: Partial<MessageCatalog> = {
+  'common.confirm': 'Подтвердить',
+  'discover.filters': 'Фильтры',
+  'discover.filtersActive': 'Фильтры ({0})',
+  'discover.moreActions': 'Больше действий',
+  'discover.decadesLabel': 'Десятилетия',
+  'analytics.title': 'Аналитика',
+  'dashboard.title': 'Панель',
+  'discover.title': 'Поиск',
+  'setup.title': 'Настройка',
+  'todaysPick.discoverSimilar': 'Найти похожих',
+  'libraryHealth.fix': 'Исправить',
+  'libraryHealth.fixing': 'Исправление...',
+  'libraryHealth.showLess': 'Показать меньше',
+  'libraryHealth.showAll': 'Показать все {0}',
+  'libraryHealth.fixHint.missingMetadata': 'Обновляет метаданные артистов из Lidarr',
+  'libraryHealth.fixHint.unmonitored':
+    'Помечает всех перечисленных артистов как отслеживаемых в Lidarr',
+  'libraryHealth.fixHint.missingAlbums': 'Обновляет альбомы в Lidarr, чтобы найти отсутствующие',
+  'libraryHealth.fixHint.genreGaps': 'Обновляет метаданные Lidarr для получения жанров',
+  'libraryHealth.fixHint.imageGaps': 'Повторно ищет изображения артистов',
+  'libraryHealth.fixHint.missingWikidata': 'Получает биографии и ссылки из Wikidata',
+  'common.importing': 'Импорт...',
+  'common.current': 'Текущий',
+  'common.custom': 'Пользовательский',
+  'common.enabled': 'Включено',
+  'common.unauthorized': 'Нет авторизации',
+  'common.adminAccessRequired': 'Требуется доступ администратора',
+  'jobHistory.pipeline': 'Пайплайн',
+  'jobHistory.librarySync': 'Синхронизация библиотеки',
+  'playlistForm.dialogCreate': 'Создать плейлист',
+  'playlistForm.dialogEdit': 'Изменить плейлист',
+  'playlistForm.titleCreate': 'Создать плейлист',
+  'playlistForm.titleEdit': 'Изменить плейлист',
+  'playlistForm.namePlaceholder': 'Мой недельный микс',
+  'playlistForm.strategy': 'Стратегия',
+  'playlistForm.strategyWeeklyDigestDesc': 'Подборка из ваших недавних открытий',
+  'playlistForm.strategyGenreFocusDesc': 'Глубокое погружение в конкретный жанр',
+  'playlistForm.strategyMoodMixDesc': 'Треки под определенное настроение',
+  'playlistForm.strategyRediscoverDesc': 'Верните артистов, о которых могли забыть',
+  'playlistForm.mood': 'Настроение',
+  'playlistForm.moodPlaceholder': 'ночная поездка',
+  'playlistForm.trackCount': 'Количество треков',
+  'playlistForm.scheduleAutomaticGeneration': 'Запланировать автоматическую генерацию',
+  'playlistForm.nameRequired': 'Название обязательно',
+  'playlistForm.genreRequired': 'Жанр обязателен для стратегии Фокус на жанре',
+  'playlistForm.moodRequired': 'Настроение обязательно для стратегии Mood Mix',
+  'subscriptionForm.dialogCreate': 'Создать подписку',
+  'subscriptionForm.dialogEdit': 'Изменить подписку',
+  'subscriptionForm.titleCreate': 'Создать подписку',
+  'subscriptionForm.titleEdit': 'Изменить подписку',
+  'subscriptionForm.name': 'Название',
+  'subscriptionForm.namePlaceholder': 'Новая Shoegaze Weekly',
+  'subscriptionForm.nameRequired': 'Название обязательно',
+  'subscriptionForm.source': 'Источник',
+  'subscriptionForm.sources': 'Источники',
+  'subscriptionForm.sourceType': 'Тип источника',
+  'subscriptionForm.sourceTypeGenre': 'Жанр',
+  'subscriptionForm.sourceTypeGenreDesc': 'Открывайте артистов в определенном жанре',
+  'subscriptionForm.sourceTypeSimilar': 'Похожие',
+  'subscriptionForm.sourceTypeSimilarDesc': 'Находите артистов, похожих на ваших любимых',
+  'subscriptionForm.sourceTypeDiscoveryModeDesc':
+    'Используйте режим открытия для генерации артистов',
+  'subscriptionForm.discoveryMode': 'Режим открытия',
+  'subscriptionForm.selectDiscoveryMode': 'Выберите режим открытия, чтобы настроить его.',
+  'subscriptionForm.discoveryModeSettingsRequired': 'Настройки режима открытия обязательны',
+  'subscriptionForm.genre': 'Жанр',
+  'subscriptionForm.genrePlaceholder': 'shoegaze',
+  'subscriptionForm.genreRequired': 'Жанр обязателен',
+  'subscriptionForm.seedArtists': 'Базовые артисты',
+  'subscriptionForm.seedArtistsPlaceholder': 'Radiohead, Portishead, Massive Attack',
+  'subscriptionForm.seedArtistsHelp': 'Имена артистов через запятую',
+  'subscriptionForm.seedArtistsRequired': 'Нужен хотя бы один базовый артист',
+  'subscriptionForm.configureInSettings': 'Настройте в Настройках > Подключения',
+  'subscriptionForm.notConfigured': 'не настроено',
+  'subscriptionForm.selectAtLeastOneSource': 'Выберите хотя бы один источник',
+  'subscriptionForm.sourceLockedHelp':
+    'Тип источника и конфигурацию нельзя изменить для этой подписки.',
+  'subscriptionForm.maxArtistsPerRun': 'Максимум артистов за запуск',
+  'subscriptionForm.scoringPreset': 'Пресет оценки',
+  'subscriptionForm.scoringPresetDefault': 'По умолчанию',
+  'subscriptionForm.scoringPresetGenre': 'Оптимизировано под жанр',
+  'subscriptions.spotifyNotConnected': 'Spotify не подключен',
+  'subscriptionPresets.getStarted': 'Начните с пресета',
+  'subscriptionPresets.intro':
+    'Выберите пресет, чтобы автоматически создать подписки, или соберите собственную конфигурацию.',
+  'subscriptionPresets.casualTitle': 'Легкий',
+  'subscriptionPresets.casualDescription': 'Еженедельные рекомендации с минимумом шума',
+  'subscriptionPresets.deepDiverTitle': 'Глубокое погружение',
+  'subscriptionPresets.deepDiverDescription':
+    'Широкое открытие из нескольких источников с ежедневными обновлениями',
+  'subscriptionPresets.discoveryOnlyTitle': 'Только открытие',
+  'subscriptionPresets.discoveryOnlyDescription': 'Только пайплайн, без внешних фидов',
+  'subscriptionPresets.customTitle': 'Свой вариант',
+  'subscriptionPresets.customDescription': 'Соберите свою комбинацию подписок',
+  'subscriptionPresets.oneConnectedService': '1 подключенный сервис',
+  'subscriptionPresets.twoConnectedServices': '2 подключенных сервиса',
+  'subscriptionPresets.requires': 'Требуется {0}',
+  'subscriptionPresets.setUp': 'Настроить',
+  'subscriptionPresets.applying': 'Применение...',
+  'subscriptionPresets.casualApplied': 'Пресет Легкий применен',
+  'subscriptionPresets.deepDiverApplied': 'Пресет Глубокое погружение применен',
+  'subscriptionPresets.applyFailed': 'Не удалось применить пресет',
+  'subscriptionPresets.aiOnlyEnabled': 'Режим только AI включен',
+  'subscriptionPresets.aiOnlyEnableFailed': 'Не удалось включить режим только AI',
+  'cronPicker.preset': 'Пресет',
+  'cronPicker.dailyMidnight': 'Ежедневно (полночь)',
+  'cronPicker.twiceAWeek': 'Два раза в неделю (пн + чт)',
+  'cronPicker.everyTwoWeeks': 'Каждые 2 недели (1 и 15)',
+  'cronPicker.monthlyFirst': 'Ежемесячно (1-е число)',
+  'cronPicker.standardHint': 'Стандартный cron: минута час день-месяца месяц день-недели',
+  'libraryHealth.syncNow': 'Синхронизировать сейчас',
+  'libraryHealth.snapshotTitle': 'Последний снимок',
+  'libraryHealth.lastSynced': 'Последняя синхронизация',
+  'libraryHealth.neverSynced': 'Никогда не синхронизировалось.',
+  'libraryHealth.autoSyncEvery': 'Автосинхронизация каждые {0} ч',
+  'systemHealth.librarySync': 'Синхронизация библиотеки',
+  'settings.capDiscovery': 'Открытие',
+  'settings.capImport': 'Импорт',
+  'settings.capLibrarySync': 'Синхронизация библиотеки',
+  'settings.capPlaylistExport': 'Экспорт плейлистов',
+  'settings.capService': 'Сервис',
+  'settings.capSubscriptions': 'Подписки',
+  'settings.connectDeezer': 'Подключить Deezer',
+  'settings.deezerConnectHelp':
+    'Подключите Deezer, чтобы импортировать избранное, подписки и плейлисты.',
+  'settings.importFavorites': 'Импортировать избранное',
+  'settings.importFollowed': 'Импортировать подписки',
+  'settings.integrationCapabilities': 'Возможности интеграций',
+  'settings.tabs.systemHealth': 'Состояние системы',
+  'settings.targetUpdated': 'Цель обновлена',
+  'settings.targetUpdateFailed': 'Не удалось обновить цель',
+  'settings.editTarget': 'Изменить цель',
+  'settings.saveChanges': 'Сохранить изменения',
+  'settings.linkedTarget': 'Связанная цель:',
+  'settings.librarySyncInterval': 'Синхронизация библиотеки',
+  'settings.librarySyncIntervalHours': 'Интервал синхронизации (часы)',
+  'settings.librarySyncIntervalHelp':
+    'Определяет, как часто Digarr обновляет данные синхронизации библиотеки и автоматически заново запускает проверки.',
+  'settings.oidcSso': 'OIDC и SSO',
+  'settings.fieldClientId': 'ID клиента',
+  'settings.fieldClientSecret': 'Секрет клиента',
+  'settings.aiProviderTitle': 'Провайдер AI',
+  'settings.webhookTitle': 'Вебхук',
+  'settings.plexToken': 'Токен Plex',
+  'settings.serviceConnectedToast': '{0} подключен',
+  'settings.serviceUnreachableToast': '{0} недоступен',
+  'settings.serviceSettingsSavedToast': 'Настройки {0} сохранены',
+  'settings.serviceSettingsFailedToast': 'Не удалось сохранить настройки {0}',
+  'settings.spotifyAuthorizationFailed': 'Не удалось начать авторизацию Spotify',
+  'settings.spotifyDisconnectFailed': 'Не удалось отключить Spotify',
+  'settings.spotifyLikedSongsFailed': 'Не удалось запустить импорт понравившихся треков Spotify',
+  'settings.spotifyPlaylistFailed': 'Не удалось импортировать плейлист Spotify',
+  'settings.deezerFavoritesFailed': 'Не удалось запустить импорт избранного Deezer',
+  'settings.deezerFollowedFailed': 'Не удалось запустить импорт подписок Deezer',
+  'settings.deezerDisconnectFailed': 'Не удалось отключить Deezer',
+  'settings.deezerAuthorizationFailed': 'Не удалось начать авторизацию Deezer',
+  'settings.reverseProxyConfigHint':
+    'Используйте PROXY_AUTH_ENABLED и PROXY_AUTH_TRUSTED_PROXIES; ваш прокси должен передавать X-Forwarded-User.',
+  'searchResultCard.addToQueue': 'Добавить в очередь открытий',
+  'searchResultCard.queuing': 'Добавление...',
+  'searchResultCard.queuedSuccess': '{0} добавлен в очередь открытий',
+  'searchResultCard.queuedFailure': 'Не удалось поставить {0} в очередь',
+  'searchResultCard.popularity': '{0}{1} популярности',
+  'admin.restoreSummary': 'Восстановлено {0} строк в {1} таблицах',
+  'admin.restoreSummaryForce':
+    'Восстановлено {0} строк. Повторно введите учетные данные для зашифрованных полей.',
+  'admin.encryptionMismatchMessage': 'Резервная копия создана с другим ключом шифрования.',
+  'admin.encryptionMismatchAffected': 'Эти поля потребуется ввести заново: {0}',
+  'admin.encryptionMismatchFallback':
+    'Некоторые зашифрованные поля, возможно, потребуется ввести заново.',
+  'admin.autoFixInProgress': 'Автоисправление выполняется... ({0}/{1} исправлено)',
+  'errorBoundary.title': 'Что-то пошло не так',
+  'errorBoundary.retry': 'Повторить',
+  'errorBoundary.home': 'На главную',
+  'connectionSuggestions.title': 'Подсказки для {0}',
+  'connectionSuggestions.close': 'Закрыть подсказки',
+  'connectionSuggestions.gotIt': 'Понятно',
+  'connectionSuggestions.skipForNow': 'Пока пропустить',
+  'connectionSuggestions.spotifyDiscoverWeekly': 'Настроить подписку Discover Weekly',
+  'connectionSuggestions.spotifyReleaseRadar': 'Включить подписку Release Radar',
+  'connectionSuggestions.lastfmTopGenreTags': 'Подписаться на ваши основные жанровые теги',
+  'connectionSuggestions.listenbrainzWeeklyJams': 'Включить подписку Weekly Jams',
+  'connectionSuggestions.listenbrainzWeeklyExploration': 'Включить Weekly Exploration',
+  'connectionSuggestions.weeklyDigestPlaylist': 'Создать еженедельный плейлист Digarr Digest',
+  'keyboardShortcuts.title': 'Сочетания клавиш',
+  'keyboardShortcuts.close': 'Закрыть сочетания клавиш',
+  'keyboardShortcuts.nextCard': 'Следующая карточка',
+  'keyboardShortcuts.previousCard': 'Предыдущая карточка',
+  'keyboardShortcuts.approveFocusedCard': 'Одобрить активную карточку',
+  'keyboardShortcuts.rejectFocusedCard': 'Отклонить активную карточку',
+  'keyboardShortcuts.toggleDetailView': 'Открыть / закрыть подробный вид',
+  'keyboardShortcuts.toggleShortcuts': 'Показать / скрыть сочетания клавиш',
+  'keyboardShortcuts.previousCardStack': 'Предыдущая карточка (вид стопкой)',
+  'keyboardShortcuts.nextCardStack': 'Следующая карточка (вид стопкой)',
+  'keyboardShortcuts.disabledWhileTyping':
+    'Сочетания клавиш отключены, когда в фокусе текстовое поле.',
+  'albumPicker.title': 'Выбрать альбомы',
+  'albumPicker.loadFailed': 'Не удалось загрузить альбомы',
+  'albumPicker.selectAll': 'Все',
+  'albumPicker.selectNone': 'Ничего',
+  'albumPicker.selectedCount': 'Выбрано: {0}',
+  'albumPicker.aiPick': 'Выбор AI',
+  'albumPicker.monitor': 'Отслеживать',
+  'albumPicker.monitorSingular': 'Отслеживать {0} альбом',
+  'albumPicker.monitorPlural': 'Отслеживать альбомов: {0}',
+  'approveDialog.title': 'Настройки Lidarr для этого исполнителя',
+  'approveDialog.loadFailed':
+    'Не удалось загрузить профили Lidarr. Проверьте подключение к Lidarr.',
+  'approveDialog.qualityProfile': 'Профиль качества',
+  'approveDialog.metadataProfile': 'Профиль метаданных',
+  'approveDialog.rootFolder': 'Корневая папка',
+  'cardStack.counter': '{0} из {1}',
+  'cardStack.swipeHint': 'Свайп вправо - одобрить, влево - отклонить',
+  'serviceCard.testConnection': 'Проверить подключение',
+  'preview.noPreviewAvailable': 'Предпросмотр для этого исполнителя недоступен',
+  'preview.playbackBlocked': 'Воспроизведение заблокировано браузером - попробуйте нажать ещё раз',
+  'errors.auth.invalidCredentials': 'Неверные учётные данные',
+  'errors.auth.notAuthenticated': 'Не аутентифицирован',
+  'errors.auth.passwordIncorrect': 'Текущий пароль неверный',
+  'errors.auth.usernameTaken': 'Имя пользователя уже занято',
+  'errors.target.notFound': 'Цель не найдена',
+  'errors.pipeline.alreadyRunning': 'Запуск пайплайна уже выполняется',
+  'errors.recommendation.notFound': 'Рекомендация не найдена',
+  'errors.subscription.notFound': 'Подписка не найдена',
+  'errors.playlist.notFound': 'Плейлист не найден',
+}
