+++ src/core/i18n/messages/overrides/tr.ts
@@ -1,0 +1,217 @@
+import type { MessageCatalog } from '../types'
+
+export const trOverrides: Partial<MessageCatalog> = {
+  'common.confirm': 'Onayla',
+  'discover.filters': 'Filtreler',
+  'discover.filtersActive': 'Filtreler ({0})',
+  'discover.moreActions': 'Daha fazla işlem',
+  'discover.decadesLabel': 'On yıllar',
+  'analytics.title': 'Analizler',
+  'dashboard.title': 'Pano',
+  'discover.title': 'Keşfet',
+  'setup.title': 'Kurulum',
+  'todaysPick.discoverSimilar': 'Benzerleri keşfet',
+  'libraryHealth.fix': 'Düzelt',
+  'libraryHealth.fixing': 'Düzeltiliyor...',
+  'libraryHealth.showLess': 'Daha az göster',
+  'libraryHealth.showAll': 'Tümünü göster {0}',
+  'libraryHealth.fixHint.missingMetadata': 'Lidarr üzerinden sanatçı meta verilerini yeniler',
+  'libraryHealth.fixHint.unmonitored': 'Listelenen tüm sanatçıları Lidarr içinde izleniyor yapar',
+  'libraryHealth.fixHint.missingAlbums': 'Eksikleri bulmak için Lidarr albümlerini yeniler',
+  'libraryHealth.fixHint.genreGaps': 'Türleri almak için Lidarr meta verilerini yeniler',
+  'libraryHealth.fixHint.imageGaps': 'Sanatçı görsellerini yeniden arar',
+  'libraryHealth.fixHint.missingWikidata': 'Wikidata üzerinden biyografi ve bağlantıları alır',
+  'common.importing': 'İçe aktarılıyor...',
+  'common.current': 'Geçerli',
+  'common.custom': 'Özel',
+  'common.enabled': 'Etkin',
+  'common.unauthorized': 'Yetkisiz',
+  'common.adminAccessRequired': 'Yönetici erişimi gerekiyor',
+  'jobHistory.pipeline': 'Keşif çalıştırması',
+  'jobHistory.librarySync': 'Kütüphane eşitleme',
+  'playlistForm.dialogCreate': 'Çalma listesi oluştur',
+  'playlistForm.dialogEdit': 'Çalma listesini düzenle',
+  'playlistForm.titleCreate': 'Çalma listesi oluştur',
+  'playlistForm.titleEdit': 'Çalma listesini düzenle',
+  'playlistForm.namePlaceholder': 'Haftalık karışımım',
+  'playlistForm.strategy': 'Strateji',
+  'playlistForm.strategyWeeklyDigestDesc': 'Son keşiflerinden özenle seçilmiş parçalar',
+  'playlistForm.strategyGenreFocusDesc': 'Belirli bir türe derinlemesine dal',
+  'playlistForm.strategyMoodMixDesc': 'Belirli bir havaya uyan parçalar',
+  'playlistForm.strategyRediscoverDesc': 'Unutmuş olabileceğin sanatçıları yeniden ortaya çıkar',
+  'playlistForm.mood': 'Ruh hali',
+  'playlistForm.moodPlaceholder': 'gece yolculuğu',
+  'playlistForm.trackCount': 'Parça sayısı',
+  'playlistForm.scheduleAutomaticGeneration': 'Otomatik oluşturmayı zamanla',
+  'playlistForm.nameRequired': 'Ad zorunludur',
+  'playlistForm.genreRequired': 'Tür Odağı stratejisi için tür zorunludur',
+  'playlistForm.moodRequired': 'Mood Mix stratejisi için ruh hali zorunludur',
+  'subscriptionForm.dialogCreate': 'Abonelik oluştur',
+  'subscriptionForm.dialogEdit': 'Aboneliği düzenle',
+  'subscriptionForm.titleCreate': 'Abonelik oluştur',
+  'subscriptionForm.titleEdit': 'Aboneliği düzenle',
+  'subscriptionForm.name': 'Ad',
+  'subscriptionForm.namePlaceholder': 'Yeni Shoegaze Weekly',
+  'subscriptionForm.nameRequired': 'Ad zorunludur',
+  'subscriptionForm.source': 'Kaynak',
+  'subscriptionForm.sources': 'Kaynaklar',
+  'subscriptionForm.sourceType': 'Kaynak türü',
+  'subscriptionForm.sourceTypeGenre': 'Müzik türü',
+  'subscriptionForm.sourceTypeGenreDesc': 'Belirli bir tür içinde sanatçılar keşfet',
+  'subscriptionForm.sourceTypeSimilar': 'Benzer sanatçılar',
+  'subscriptionForm.sourceTypeSimilarDesc': 'Favorilerine benzeyen sanatçıları bul',
+  'subscriptionForm.sourceTypeDiscoveryModeDesc': 'Sanatçı üretmek için bir keşif modu kullan',
+  'subscriptionForm.discoveryMode': 'Keşif modu',
+  'subscriptionForm.selectDiscoveryMode': 'Yapılandırmak için bir keşif modu seç.',
+  'subscriptionForm.discoveryModeSettingsRequired': 'Keşif modu ayarları zorunludur',
+  'subscriptionForm.genre': 'Müzik türü',
+  'subscriptionForm.genrePlaceholder': 'shoegaze',
+  'subscriptionForm.seedArtistsPlaceholder': 'Radiohead, Portishead, Massive Attack',
+  'subscriptionForm.genreRequired': 'Tür zorunludur',
+  'subscriptionForm.seedArtists': 'Başlangıç sanatçıları',
+  'subscriptionForm.seedArtistsHelp': 'Virgülle ayrılmış sanatçı adları',
+  'subscriptionForm.seedArtistsRequired': 'En az bir başlangıç sanatçısı gereklidir',
+  'subscriptionForm.configureInSettings': 'Ayarlar > Bağlantılar bölümünden yapılandır',
+  'subscriptionForm.notConfigured': 'yapılandırılmadı',
+  'subscriptionForm.selectAtLeastOneSource': 'En az bir kaynak seç',
+  'subscriptionForm.sourceLockedHelp':
+    'Bu abonelik için kaynak türü ve yapılandırma değiştirilemez.',
+  'subscriptionForm.maxArtistsPerRun': 'Çalıştırma başına en fazla sanatçı',
+  'subscriptionForm.scoringPreset': 'Puanlama ön ayarı',
+  'subscriptionForm.scoringPresetDefault': 'Varsayılan',
+  'subscriptionForm.scoringPresetGenre': 'Türe göre optimize edildi',
+  'subscriptions.spotifyNotConnected': 'Spotify bağlı değil',
+  'subscriptionPresets.getStarted': 'Bir ön ayarla başla',
+  'subscriptionPresets.intro':
+    'Abonelikleri otomatik oluşturmak için bir ön ayar seç ya da kendi yapılandırmanı kur.',
+  'subscriptionPresets.casualTitle': 'Rahat',
+  'subscriptionPresets.casualDescription': 'Az gürültülü haftalık öneriler',
+  'subscriptionPresets.deepDiverTitle': 'Derin keşif',
+  'subscriptionPresets.deepDiverDescription':
+    'Birden fazla kaynaktan geniş keşif ve günlük güncellemeler',
+  'subscriptionPresets.discoveryOnlyTitle': 'Sadece keşif',
+  'subscriptionPresets.discoveryOnlyDescription': 'Yalnızca pipeline, harici akış yok',
+  'subscriptionPresets.customTitle': 'Özel',
+  'subscriptionPresets.customDescription': 'Kendi abonelik karışımını oluştur',
+  'subscriptionPresets.oneConnectedService': '1 bağlı hizmet',
+  'subscriptionPresets.twoConnectedServices': '2 bağlı hizmet',
+  'subscriptionPresets.requires': '{0} gerekir',
+  'subscriptionPresets.setUp': 'Kur',
+  'subscriptionPresets.applying': 'Uygulanıyor...',
+  'subscriptionPresets.casualApplied': 'Rahat ön ayarı uygulandı',
+  'subscriptionPresets.deepDiverApplied': 'Derin keşif ön ayarı uygulandı',
+  'subscriptionPresets.applyFailed': 'Ön ayar uygulanamadı',
+  'subscriptionPresets.aiOnlyEnabled': 'Yalnızca AI modu etkinleştirildi',
+  'subscriptionPresets.aiOnlyEnableFailed': 'Yalnızca AI modu etkinleştirilemedi',
+  'cronPicker.preset': 'Ön ayar',
+  'cronPicker.dailyMidnight': 'Her gün (gece yarısı)',
+  'cronPicker.twiceAWeek': 'Haftada iki kez (Pzt + Per)',
+  'cronPicker.everyTwoWeeks': 'Her 2 haftada bir (1 ve 15)',
+  'cronPicker.monthlyFirst': 'Aylık (1. gün)',
+  'cronPicker.standardHint': 'Standart cron: dakika saat ayın-günü ay haftanın-günü',
+  'libraryHealth.syncNow': 'Şimdi eşitle',
+  'libraryHealth.snapshotTitle': 'Son anlık görüntü',
+  'libraryHealth.lastSynced': 'Son eşitleme',
+  'libraryHealth.neverSynced': 'Henüz hiç eşitlenmedi.',
+  'libraryHealth.autoSyncEvery': 'Her {0} saatte otomatik eşitleme',
+  'systemHealth.pipeline': 'Keşif çalıştırması',
+  'systemHealth.librarySync': 'Kütüphane eşitleme',
+  'discoveryMode.field.feed': 'Besleme kaynağı',
+  'settings.tabs.systemHealth': 'Sistem durumu',
+  'settings.targetUpdated': 'Hedef güncellendi',
+  'settings.targetUpdateFailed': 'Hedef güncellenemedi',
+  'settings.editTarget': 'Hedefi düzenle',
+  'settings.saveChanges': 'Değişiklikleri kaydet',
+  'settings.linkedTarget': 'Bağlı hedef:',
+  'settings.librarySyncInterval': 'Kütüphane eşitleme',
+  'settings.librarySyncIntervalHours': 'Eşitleme aralığı (saat)',
+  'settings.librarySyncIntervalHelp':
+    'Digarr ın kütüphane eşitleme verilerini ne sıklıkla yenileyeceğini ve denetimleri otomatik olarak yeniden çalıştıracağını belirler.',
+  'settings.oidcSso': 'OIDC / SSO',
+  'settings.fieldModel': 'Model seçimi',
+  'settings.fieldClientId': 'İstemci kimliği',
+  'settings.fieldClientSecret': 'İstemci gizlisi',
+  'settings.fieldWebhookUrl': 'Webhook adresi',
+  'settings.aiProviderTitle': 'AI sağlayıcısı',
+  'settings.webhookTitle': 'Webhook entegrasyonu',
+  'settings.plexToken': 'Plex belirteci',
+  'settings.serviceConnectedToast': '{0} bağlandı',
+  'settings.serviceUnreachableToast': '{0} erişilemiyor',
+  'settings.serviceSettingsSavedToast': '{0} ayarları kaydedildi',
+  'settings.serviceSettingsFailedToast': '{0} ayarları kaydedilemedi',
+  'settings.spotifyAuthorizationFailed': 'Spotify yetkilendirmesi başlatılamadı',
+  'settings.spotifyDisconnectFailed': 'Spotify bağlantısı kesilemedi',
+  'settings.spotifyLikedSongsFailed': 'Spotify Liked Songs içe aktarma başlatılamadı',
+  'settings.spotifyPlaylistFailed': 'Çalma listesi içe aktarma başlatılamadı',
+  'settings.deezerFavoritesFailed': 'Deezer favorileri içe aktarma başlatılamadı',
+  'settings.deezerFollowedFailed': 'Takip edilen Deezer sanatçıları içe aktarma başlatılamadı',
+  'settings.deezerDisconnectFailed': 'Deezer bağlantısı kesilemedi',
+  'settings.deezerAuthorizationFailed': 'Deezer yetkilendirmesi başlatılamadı',
+  'settings.reverseProxyConfigHint':
+    'PROXY_AUTH_ENABLED ve PROXY_AUTH_TRUSTED_PROXIES kullanın; proxy nizin X-Forwarded-User göndermesi gerekir.',
+  'searchResultCard.addToQueue': 'Keşif kuyruğuna ekle',
+  'searchResultCard.queuing': 'Ekleniyor...',
+  'searchResultCard.queuedSuccess': '{0} keşif kuyruğuna eklendi',
+  'searchResultCard.queuedFailure': '{0} kuyruğa eklenemedi',
+  'searchResultCard.popularity': 'Popülerlik {0}{1}',
+  'admin.restoreSummary': '{1} tabloda {0} satır geri yüklendi',
+  'admin.restoreSummaryForce':
+    '{0} satır geri yüklendi. Şifrelenmiş alanlar için kimlik bilgilerini yeniden girin.',
+  'admin.encryptionMismatchMessage': 'Bu yedek farklı bir şifreleme anahtarıyla oluşturulmuş.',
+  'admin.encryptionMismatchAffected': 'Bu alanların yeniden girilmesi gerekecek: {0}',
+  'admin.encryptionMismatchFallback': 'Bazı şifrelenmiş alanların yeniden girilmesi gerekebilir.',
+  'admin.autoFixInProgress': 'Otomatik düzeltme sürüyor... ({0}/{1} düzeltildi)',
+  'errorBoundary.title': 'Bir sorun oluştu',
+  'errorBoundary.retry': 'Tekrar dene',
+  'errorBoundary.home': 'Ana sayfaya git',
+  'connectionSuggestions.title': '{0} için öneriler',
+  'connectionSuggestions.close': 'Önerileri kapat',
+  'connectionSuggestions.gotIt': 'Tamam',
+  'connectionSuggestions.skipForNow': 'Şimdilik geç',
+  'connectionSuggestions.spotifyDiscoverWeekly': 'Bir Discover Weekly aboneliği ayarla',
+  'connectionSuggestions.spotifyReleaseRadar': 'Bir Release Radar aboneliğini etkinleştir',
+  'connectionSuggestions.lastfmTopGenreTags': 'En önemli tür etiketlerine abone ol',
+  'connectionSuggestions.listenbrainzWeeklyJams': 'Weekly Jams aboneliğini etkinleştir',
+  'connectionSuggestions.listenbrainzWeeklyExploration': 'Weekly Exploration ı etkinleştir',
+  'connectionSuggestions.weeklyDigestPlaylist': 'Haftalık bir Digarr Digest çalma listesi oluştur',
+  'keyboardShortcuts.title': 'Klavye kısayolları',
+  'keyboardShortcuts.close': 'Kısayolları kapat',
+  'keyboardShortcuts.nextCard': 'Sonraki kart',
+  'keyboardShortcuts.previousCard': 'Önceki kart',
+  'keyboardShortcuts.approveFocusedCard': 'Etkin kartı onayla',
+  'keyboardShortcuts.rejectFocusedCard': 'Etkin kartı reddet',
+  'keyboardShortcuts.toggleDetailView': 'Detay görünümünü aç / kapat',
+  'keyboardShortcuts.toggleShortcuts': 'Kısayolları göster / gizle',
+  'keyboardShortcuts.previousCardStack': 'Önceki kart (yığın görünümü)',
+  'keyboardShortcuts.nextCardStack': 'Sonraki kart (yığın görünümü)',
+  'keyboardShortcuts.disabledWhileTyping':
+    'Bir metin alanı odaktayken kısayollar devre dışı kalır.',
+  'albumPicker.title': 'Albümleri seç',
+  'albumPicker.loadFailed': 'Albümler yüklenemedi',
+  'albumPicker.selectAll': 'Tümü',
+  'albumPicker.selectNone': 'Hiçbiri',
+  'albumPicker.selectedCount': '{0} seçili',
+  'albumPicker.aiPick': 'AI seçimi',
+  'albumPicker.monitor': 'İzle',
+  'albumPicker.monitorSingular': '{0} albümü izle',
+  'albumPicker.monitorPlural': '{0} albümü izle',
+  'approveDialog.title': 'Bu sanatçı için Lidarr ayarları',
+  'approveDialog.loadFailed': 'Lidarr profilleri yüklenemedi. Lidarr bağlantınızı kontrol edin.',
+  'approveDialog.qualityProfile': 'Kalite Profili',
+  'approveDialog.metadataProfile': 'Meta Veri Profili',
+  'approveDialog.rootFolder': 'Kök Klasör',
+  'cardStack.counter': '{0} / {1}',
+  'cardStack.swipeHint': 'Onaylamak için sağa, reddetmek için sola kaydırın',
+  'serviceCard.testConnection': 'Bağlantıyı Test Et',
+  'preview.noPreviewAvailable': 'Bu sanatçı için önizleme yok',
+  'preview.playbackBlocked': 'Oynatma tarayıcı tarafından engellendi - tekrar tıklamayı deneyin',
+  'errors.auth.invalidCredentials': 'Geçersiz kimlik bilgileri',
+  'errors.auth.notAuthenticated': 'Kimlik doğrulanmadı',
+  'errors.auth.passwordIncorrect': 'Mevcut parola yanlış',
+  'errors.auth.usernameTaken': 'Kullanıcı adı zaten alınmış',
+  'errors.target.notFound': 'Hedef bulunamadı',
+  'errors.pipeline.alreadyRunning': 'Zaten bir pipeline çalışması devam ediyor',
+  'errors.recommendation.notFound': 'Öneri bulunamadı',
+  'errors.subscription.notFound': 'Abonelik bulunamadı',
+  'errors.playlist.notFound': 'Oynatma listesi bulunamadı',
+}
