+++ src/core/i18n/messages/overrides/uk.ts
@@ -1,0 +1,228 @@
+import type { MessageCatalog } from '../types'
+
+export const ukOverrides: Partial<MessageCatalog> = {
+  'common.confirm': 'Підтвердити',
+  'discover.filters': 'Фільтри',
+  'discover.filtersActive': 'Фільтри ({0})',
+  'discover.moreActions': 'Більше дій',
+  'discover.decadesLabel': 'Десятиліття',
+  'analytics.title': 'Аналітика',
+  'dashboard.title': 'Панель',
+  'discover.title': 'Відкриття',
+  'setup.title': 'Налаштування',
+  'todaysPick.discoverSimilar': 'Знайти схожих',
+  'libraryHealth.fix': 'Виправити',
+  'libraryHealth.fixing': 'Виправлення...',
+  'libraryHealth.showLess': 'Показати менше',
+  'libraryHealth.showAll': 'Показати всі {0}',
+  'libraryHealth.fixHint.missingMetadata': 'Оновлює метадані артистів із Lidarr',
+  'libraryHealth.fixHint.unmonitored':
+    'Позначає всіх перелічених артистів як відстежуваних у Lidarr',
+  'libraryHealth.fixHint.missingAlbums': 'Оновлює альбоми в Lidarr, щоб знайти відсутні',
+  'libraryHealth.fixHint.genreGaps': 'Оновлює метадані Lidarr, щоб отримати жанри',
+  'libraryHealth.fixHint.imageGaps': 'Повторно шукає зображення артистів',
+  'libraryHealth.fixHint.missingWikidata': 'Отримує біографії та посилання з Wikidata',
+  'setup.embyUrl': 'Адреса Emby',
+  'common.importing': 'Імпорт...',
+  'common.current': 'Поточний',
+  'common.custom': 'Власний',
+  'common.enabled': 'Увімкнено',
+  'common.unauthorized': 'Немає авторизації',
+  'common.adminAccessRequired': 'Потрібен доступ адміністратора',
+  'jobHistory.pipeline': 'Запуск пошуку',
+  'jobHistory.librarySync': 'Синхронізація бібліотеки',
+  'playlistForm.dialogCreate': 'Створити плейлист',
+  'playlistForm.dialogEdit': 'Редагувати плейлист',
+  'playlistForm.titleCreate': 'Створити плейлист',
+  'playlistForm.titleEdit': 'Редагувати плейлист',
+  'playlistForm.namePlaceholder': 'Мій тижневий мікс',
+  'playlistForm.strategy': 'Стратегія',
+  'playlistForm.strategyWeeklyDigestDesc': 'Добірка з ваших недавніх відкриттів',
+  'playlistForm.strategyGenreFocusDesc': 'Глибоке занурення в конкретний жанр',
+  'playlistForm.strategyMoodMixDesc': 'Треки під певний настрій',
+  'playlistForm.strategyRediscoverDesc': 'Поверніть артистів, про яких могли забути',
+  'playlistForm.mood': 'Настрій',
+  'playlistForm.moodPlaceholder': 'нічна поїздка',
+  'playlistForm.trackCount': 'Кількість треків',
+  'playlistForm.scheduleAutomaticGeneration': 'Запланувати автоматичне створення',
+  'playlistForm.nameRequired': 'Назва обов язкова',
+  'playlistForm.genreRequired': 'Жанр обов язковий для стратегії Фокус на жанрі',
+  'playlistForm.moodRequired': 'Настрій обов язковий для стратегії Mood Mix',
+  'subscriptionForm.dialogCreate': 'Створити підписку',
+  'subscriptionForm.dialogEdit': 'Редагувати підписку',
+  'subscriptionForm.titleCreate': 'Створити підписку',
+  'subscriptionForm.titleEdit': 'Редагувати підписку',
+  'subscriptionForm.name': 'Назва',
+  'subscriptionForm.namePlaceholder': 'Нова Shoegaze Weekly',
+  'subscriptionForm.nameRequired': 'Назва обов язкова',
+  'subscriptionForm.source': 'Джерело',
+  'subscriptionForm.sources': 'Джерела',
+  'subscriptionForm.sourceType': 'Тип джерела',
+  'subscriptionForm.sourceTypeGenre': 'Музичний жанр',
+  'subscriptionForm.sourceTypeGenreDesc': 'Відкривайте артистів у певному жанрі',
+  'subscriptionForm.sourceTypeSimilar': 'Схожі артисти',
+  'subscriptionForm.sourceTypeSimilarDesc': 'Знайдіть артистів, схожих на ваших улюблених',
+  'subscriptionForm.sourceTypeDiscoveryModeDesc':
+    'Використовуйте режим пошуку, щоб генерувати артистів',
+  'subscriptionForm.discoveryMode': 'Режим пошуку',
+  'subscriptionForm.selectDiscoveryMode': 'Виберіть режим пошуку для налаштування.',
+  'subscriptionForm.discoveryModeSettingsRequired': 'Налаштування режиму пошуку є обов язковими',
+  'subscriptionForm.genre': 'Музичний жанр',
+  'subscriptionForm.genrePlaceholder': 'shoegaze',
+  'subscriptionForm.seedArtistsPlaceholder': 'Radiohead, Portishead, Massive Attack',
+  'subscriptionForm.genreRequired': 'Жанр обов язковий',
+  'subscriptionForm.seedArtists': 'Початкові артисти',
+  'subscriptionForm.seedArtistsHelp': 'Імена артистів, розділені комами',
+  'subscriptionForm.seedArtistsRequired': 'Потрібен щонайменше один початковий артист',
+  'subscriptionForm.configureInSettings': 'Налаштуйте в Розділі Налаштування > Підключення',
+  'subscriptionForm.notConfigured': 'не налаштовано',
+  'subscriptionForm.selectAtLeastOneSource': 'Виберіть принаймні одне джерело',
+  'subscriptionForm.sourceLockedHelp':
+    'Тип джерела та конфігурацію не можна змінити для цієї підписки.',
+  'subscriptionForm.maxArtistsPerRun': 'Максимум артистів за запуск',
+  'subscriptionForm.scoringPreset': 'Шаблон оцінювання',
+  'subscriptionForm.scoringPresetDefault': 'Типовий',
+  'subscriptionForm.scoringPresetGenre': 'Оптимізовано для жанру',
+  'subscriptions.spotifyNotConnected': 'Spotify не підключено',
+  'subscriptionPresets.getStarted': 'Почніть із пресета',
+  'subscriptionPresets.intro':
+    'Виберіть пресет, щоб автоматично створювати підписки, або зберіть власну конфігурацію.',
+  'subscriptionPresets.casualTitle': 'Легкий режим',
+  'subscriptionPresets.casualDescription': 'Щотижневі рекомендації з мінімумом шуму',
+  'subscriptionPresets.deepDiverTitle': 'Глибоке занурення',
+  'subscriptionPresets.deepDiverDescription':
+    'Широке відкриття з кількох джерел і щоденними оновленнями',
+  'subscriptionPresets.discoveryOnlyTitle': 'Лише пошук',
+  'subscriptionPresets.discoveryOnlyDescription': 'Лише pipeline, без зовнішніх стрічок',
+  'subscriptionPresets.customTitle': 'Власний',
+  'subscriptionPresets.customDescription': 'Створіть власну комбінацію підписок',
+  'subscriptionPresets.oneConnectedService': '1 підключений сервіс',
+  'subscriptionPresets.twoConnectedServices': '2 підключені сервіси',
+  'subscriptionPresets.requires': 'Потрібно: {0}',
+  'subscriptionPresets.setUp': 'Налаштувати',
+  'subscriptionPresets.applying': 'Застосування...',
+  'subscriptionPresets.casualApplied': 'Пресет Легкий режим застосовано',
+  'subscriptionPresets.deepDiverApplied': 'Пресет Глибоке занурення застосовано',
+  'subscriptionPresets.applyFailed': 'Не вдалося застосувати пресет',
+  'subscriptionPresets.aiOnlyEnabled': 'Режим лише AI увімкнено',
+  'subscriptionPresets.aiOnlyEnableFailed': 'Не вдалося увімкнути режим лише AI',
+  'cronPicker.preset': 'Пресет',
+  'cronPicker.dailyMidnight': 'Щодня (опівночі)',
+  'cronPicker.twiceAWeek': 'Двічі на тиждень (пн + чт)',
+  'cronPicker.everyTwoWeeks': 'Кожні 2 тижні (1 і 15)',
+  'cronPicker.monthlyFirst': 'Щомісяця (день 1)',
+  'cronPicker.standardHint': 'Стандартний cron: хвилина година день-місяця місяць день-тижня',
+  'libraryHealth.syncNow': 'Синхронізувати зараз',
+  'libraryHealth.snapshotTitle': 'Останній знімок',
+  'libraryHealth.lastSynced': 'Остання синхронізація',
+  'libraryHealth.neverSynced': 'Ще не синхронізовано.',
+  'libraryHealth.autoSyncEvery': 'Автосинхронізація кожні {0} год',
+  'settings.capDiscovery': 'Пошук',
+  'settings.capImport': 'Імпорт',
+  'settings.capLibrarySync': 'Синхронізація бібліотеки',
+  'settings.capPlaylistExport': 'Експорт плейлистів',
+  'settings.capService': 'Сервіс',
+  'settings.capSubscriptions': 'Підписки',
+  'settings.connectDeezer': 'Підключити Deezer',
+  'settings.deezerConnectHelp':
+    'Підключіть свій обліковий запис Deezer, щоб імпортувати улюблене, артистів, на яких ви підписані, і дані плейлистів.',
+  'settings.importFavorites': 'Імпортувати улюблене',
+  'settings.importFollowed': 'Імпортувати підписки',
+  'settings.integrationCapabilities': 'Можливості інтеграції',
+  'systemHealth.librarySync': 'Синхронізація бібліотеки',
+  'settings.tabs.systemHealth': 'Стан системи',
+  'settings.targetUpdated': 'Ціль оновлено',
+  'settings.targetUpdateFailed': 'Не вдалося оновити ціль',
+  'settings.editTarget': 'Редагувати ціль',
+  'settings.saveChanges': 'Зберегти зміни',
+  'settings.linkedTarget': 'Пов язана ціль:',
+  'settings.librarySyncInterval': 'Синхронізація бібліотеки',
+  'settings.librarySyncIntervalHours': 'Інтервал синхронізації (години)',
+  'settings.librarySyncIntervalHelp':
+    'Визначає, як часто Digarr оновлює дані синхронізації бібліотеки та автоматично повторно запускає перевірки.',
+  'settings.oidcSso': 'OIDC / SSO',
+  'settings.fieldClientId': 'ID клієнта',
+  'settings.fieldClientSecret': 'Секрет клієнта',
+  'settings.aiProviderTitle': 'AI-провайдер',
+  'settings.webhookTitle': 'Webhook-інтеграція',
+  'settings.plexToken': 'Токен Plex',
+  'settings.serviceConnectedToast': '{0} підключено',
+  'settings.serviceUnreachableToast': 'Не вдалося дістатися до {0}',
+  'settings.serviceSettingsSavedToast': 'Налаштування {0} збережено',
+  'settings.serviceSettingsFailedToast': 'Не вдалося зберегти налаштування {0}',
+  'settings.spotifyAuthorizationFailed': 'Не вдалося запустити авторизацію Spotify',
+  'settings.spotifyDisconnectFailed': 'Не вдалося відключити Spotify',
+  'settings.spotifyLikedSongsFailed': 'Не вдалося запустити імпорт Spotify Liked Songs',
+  'settings.spotifyPlaylistFailed': 'Не вдалося запустити імпорт плейлиста',
+  'settings.deezerFavoritesFailed': 'Не вдалося запустити імпорт Deezer Favorites',
+  'settings.deezerFollowedFailed': 'Не вдалося запустити імпорт Deezer Followed',
+  'settings.deezerDisconnectFailed': 'Не вдалося відключити Deezer',
+  'settings.deezerAuthorizationFailed': 'Не вдалося запустити авторизацію Deezer',
+  'settings.reverseProxyConfigHint':
+    'Використовуйте PROXY_AUTH_ENABLED і PROXY_AUTH_TRUSTED_PROXIES; ваш proxy має надсилати X-Forwarded-User.',
+  'searchResultCard.addToQueue': 'Додати до черги пошуку',
+  'searchResultCard.queuing': 'Додається...',
+  'searchResultCard.queuedSuccess': '{0} додано до черги пошуку',
+  'searchResultCard.queuedFailure': '{0} не вдалося додати до черги',
+  'searchResultCard.popularity': 'Популярність {0}{1}',
+  'admin.restoreSummary': '{0} рядків відновлено у {1} таблицях',
+  'admin.restoreSummaryForce':
+    '{0} рядків відновлено. Повторно введіть облікові дані для зашифрованих полів.',
+  'admin.encryptionMismatchMessage': 'Цю резервну копію створено з іншим ключем шифрування.',
+  'admin.encryptionMismatchAffected': 'Ці поля потрібно буде ввести знову: {0}',
+  'admin.encryptionMismatchFallback': 'Деякі зашифровані поля можуть вимагати повторного введення.',
+  'admin.autoFixInProgress': 'Автовиправлення триває... ({0}/{1} виправлено)',
+  'errorBoundary.title': 'Сталася проблема',
+  'errorBoundary.retry': 'Спробувати знову',
+  'errorBoundary.home': 'Перейти на головну',
+  'connectionSuggestions.title': 'Пропозиції для {0}',
+  'connectionSuggestions.close': 'Закрити пропозиції',
+  'connectionSuggestions.gotIt': 'Зрозуміло',
+  'connectionSuggestions.skipForNow': 'Пропустити поки що',
+  'connectionSuggestions.spotifyDiscoverWeekly': 'Налаштувати підписку Discover Weekly',
+  'connectionSuggestions.spotifyReleaseRadar': 'Увімкнути підписку Release Radar',
+  'connectionSuggestions.lastfmTopGenreTags': 'Підписатися на ваші основні теги жанрів',
+  'connectionSuggestions.listenbrainzWeeklyJams': 'Увімкнути підписку Weekly Jams',
+  'connectionSuggestions.listenbrainzWeeklyExploration': 'Увімкнути Weekly Exploration',
+  'connectionSuggestions.weeklyDigestPlaylist': 'Створити щотижневий плейлист Digarr Digest',
+  'keyboardShortcuts.title': 'Клавіатурні скорочення',
+  'keyboardShortcuts.close': 'Закрити скорочення',
+  'keyboardShortcuts.nextCard': 'Наступна картка',
+  'keyboardShortcuts.previousCard': 'Попередня картка',
+  'keyboardShortcuts.approveFocusedCard': 'Схвалити активну картку',
+  'keyboardShortcuts.rejectFocusedCard': 'Відхилити активну картку',
+  'keyboardShortcuts.toggleDetailView': 'Відкрити / закрити детальний перегляд',
+  'keyboardShortcuts.toggleShortcuts': 'Показати / приховати скорочення',
+  'keyboardShortcuts.previousCardStack': 'Попередня картка (вигляд стосу)',
+  'keyboardShortcuts.nextCardStack': 'Наступна картка (вигляд стосу)',
+  'keyboardShortcuts.disabledWhileTyping': 'Скорочення вимикаються, коли у фокусі текстове поле.',
+  'albumPicker.title': 'Вибрати альбоми',
+  'albumPicker.loadFailed': 'Не вдалося завантажити альбоми',
+  'albumPicker.selectAll': 'Усі',
+  'albumPicker.selectNone': 'Жодного',
+  'albumPicker.selectedCount': 'Вибрано: {0}',
+  'albumPicker.aiPick': 'Вибір AI',
+  'albumPicker.monitor': 'Відстежувати',
+  'albumPicker.monitorSingular': 'Відстежувати {0} альбом',
+  'albumPicker.monitorPlural': 'Відстежувати {0} альбомів',
+  'approveDialog.title': 'Налаштування Lidarr для цього виконавця',
+  'approveDialog.loadFailed':
+    'Не вдалося завантажити профілі Lidarr. Перевірте підключення до Lidarr.',
+  'approveDialog.qualityProfile': 'Профіль якості',
+  'approveDialog.metadataProfile': 'Профіль метаданих',
+  'approveDialog.rootFolder': 'Коренева папка',
+  'cardStack.counter': '{0} з {1}',
+  'cardStack.swipeHint': 'Проведіть праворуч, щоб схвалити, ліворуч - щоб відхилити',
+  'serviceCard.testConnection': 'Перевірити зʼєднання',
+  'preview.noPreviewAvailable': 'Попередній перегляд для цього виконавця недоступний',
+  'preview.playbackBlocked': 'Відтворення заблоковано браузером - спробуйте натиснути ще раз',
+  'errors.auth.invalidCredentials': 'Недійсні облікові дані',
+  'errors.auth.notAuthenticated': 'Не автентифіковано',
+  'errors.auth.passwordIncorrect': 'Поточний пароль неправильний',
+  'errors.auth.usernameTaken': 'Імʼя користувача вже зайняте',
+  'errors.target.notFound': 'Ціль не знайдено',
+  'errors.pipeline.alreadyRunning': 'Запуск конвеєра вже триває',
+  'errors.recommendation.notFound': 'Рекомендацію не знайдено',
+  'errors.subscription.notFound': 'Підписку не знайдено',
+  'errors.playlist.notFound': 'Список відтворення не знайдено',
+}
