+++ src/core/i18n/messages/overrides/zh-CN.ts
@@ -1,0 +1,208 @@
+import type { MessageCatalog } from '../types'
+
+export const zhCNOverrides: Partial<MessageCatalog> = {
+  'common.confirm': '确认',
+  'discover.filters': '筛选',
+  'discover.filtersActive': '筛选 ({0})',
+  'discover.moreActions': '更多操作',
+  'discover.decadesLabel': '年代',
+  'analytics.title': '分析',
+  'dashboard.title': '仪表盘',
+  'discover.title': '发现',
+  'setup.title': '设置',
+  'todaysPick.discoverSimilar': '发现相似艺人',
+  'libraryHealth.fix': '修复',
+  'libraryHealth.fixing': '修复中...',
+  'libraryHealth.showLess': '收起',
+  'libraryHealth.showAll': '显示全部 {0}',
+  'libraryHealth.fixHint.missingMetadata': '从 Lidarr 刷新艺人元数据',
+  'libraryHealth.fixHint.unmonitored': '将所有列出的艺人在 Lidarr 中设为监控',
+  'libraryHealth.fixHint.missingAlbums': '刷新 Lidarr 专辑以查找缺失项',
+  'libraryHealth.fixHint.genreGaps': '刷新 Lidarr 元数据以获取流派标签',
+  'libraryHealth.fixHint.imageGaps': '重新查找艺人图片',
+  'libraryHealth.fixHint.missingWikidata': '从 Wikidata 获取简介和链接',
+  'common.importing': '导入中...',
+  'common.current': '当前',
+  'common.custom': '自定义',
+  'common.enabled': '已启用',
+  'common.unauthorized': '未授权',
+  'common.adminAccessRequired': '需要管理员权限',
+  'jobHistory.pipeline': '流水线',
+  'jobHistory.librarySync': '资料库同步',
+  'playlistForm.dialogCreate': '创建播放列表',
+  'playlistForm.dialogEdit': '编辑播放列表',
+  'playlistForm.titleCreate': '创建播放列表',
+  'playlistForm.titleEdit': '编辑播放列表',
+  'playlistForm.namePlaceholder': '我的每周混音',
+  'playlistForm.strategy': '策略',
+  'playlistForm.strategyWeeklyDigestDesc': '从你最近的发现中精选整理',
+  'playlistForm.strategyGenreFocusDesc': '深入探索某个特定流派',
+  'playlistForm.strategyMoodMixDesc': '符合特定氛围的曲目集合',
+  'playlistForm.strategyRediscoverDesc': '重新挖掘你可能已经淡忘的艺人',
+  'playlistForm.mood': '氛围',
+  'playlistForm.moodPlaceholder': '深夜开车',
+  'playlistForm.trackCount': '曲目数量',
+  'playlistForm.scheduleAutomaticGeneration': '安排自动生成',
+  'playlistForm.nameRequired': '名称为必填项',
+  'playlistForm.genreRequired': '流派聚焦策略必须填写流派',
+  'playlistForm.moodRequired': 'Mood Mix 策略必须填写氛围',
+  'subscriptionForm.dialogCreate': '创建订阅',
+  'subscriptionForm.dialogEdit': '编辑订阅',
+  'subscriptionForm.titleCreate': '创建订阅',
+  'subscriptionForm.titleEdit': '编辑订阅',
+  'subscriptionForm.name': '名称',
+  'subscriptionForm.namePlaceholder': '新的 Shoegaze Weekly',
+  'subscriptionForm.nameRequired': '名称为必填项',
+  'subscriptionForm.source': '来源',
+  'subscriptionForm.sources': '来源',
+  'subscriptionForm.sourceType': '来源类型',
+  'subscriptionForm.sourceTypeGenre': '流派',
+  'subscriptionForm.sourceTypeGenreDesc': '在特定流派中发现艺人',
+  'subscriptionForm.sourceTypeSimilar': '相似艺人',
+  'subscriptionForm.sourceTypeSimilarDesc': '查找与你喜欢的艺人相似的艺人',
+  'subscriptionForm.sourceTypeDiscoveryModeDesc': '使用发现模式来生成艺人',
+  'subscriptionForm.discoveryMode': '发现模式',
+  'subscriptionForm.selectDiscoveryMode': '选择一个发现模式进行配置。',
+  'subscriptionForm.discoveryModeSettingsRequired': '发现模式设置为必填项',
+  'subscriptionForm.genre': '流派',
+  'subscriptionForm.genrePlaceholder': 'shoegaze',
+  'subscriptionForm.genreRequired': '流派为必填项',
+  'subscriptionForm.seedArtists': '种子艺人',
+  'subscriptionForm.seedArtistsPlaceholder': 'Radiohead, Portishead, Massive Attack',
+  'subscriptionForm.seedArtistsHelp': '用逗号分隔艺人名称',
+  'subscriptionForm.seedArtistsRequired': '至少需要一位种子艺人',
+  'subscriptionForm.configureInSettings': '请在 设置 > 连接 中配置',
+  'subscriptionForm.notConfigured': '未配置',
+  'subscriptionForm.selectAtLeastOneSource': '至少选择一个来源',
+  'subscriptionForm.sourceLockedHelp': '此订阅的来源类型和配置无法更改。',
+  'subscriptionForm.maxArtistsPerRun': '每次运行的最大艺人数',
+  'subscriptionForm.scoringPreset': '评分预设',
+  'subscriptionForm.scoringPresetDefault': '默认',
+  'subscriptionForm.scoringPresetGenre': '流派优化',
+  'subscriptions.spotifyNotConnected': 'Spotify 尚未连接',
+  'subscriptionPresets.getStarted': '从预设开始',
+  'subscriptionPresets.intro': '选择一个预设来自动创建订阅，或自行搭建自定义配置。',
+  'subscriptionPresets.casualTitle': '轻松模式',
+  'subscriptionPresets.casualDescription': '低噪音的每周推荐',
+  'subscriptionPresets.deepDiverTitle': '深度探索',
+  'subscriptionPresets.deepDiverDescription': '多来源广泛发现，并提供每日更新',
+  'subscriptionPresets.discoveryOnlyTitle': '仅发现',
+  'subscriptionPresets.discoveryOnlyDescription': '仅运行流水线，不使用外部订阅源',
+  'subscriptionPresets.customTitle': '自定义',
+  'subscriptionPresets.customDescription': '构建你自己的订阅组合',
+  'subscriptionPresets.oneConnectedService': '1 个已连接服务',
+  'subscriptionPresets.twoConnectedServices': '2 个已连接服务',
+  'subscriptionPresets.requires': '需要 {0}',
+  'subscriptionPresets.setUp': '设置',
+  'subscriptionPresets.applying': '应用中...',
+  'subscriptionPresets.casualApplied': '已应用轻松模式预设',
+  'subscriptionPresets.deepDiverApplied': '已应用深度探索预设',
+  'subscriptionPresets.applyFailed': '无法应用预设',
+  'subscriptionPresets.aiOnlyEnabled': '已启用仅 AI 模式',
+  'subscriptionPresets.aiOnlyEnableFailed': '无法启用仅 AI 模式',
+  'cronPicker.preset': '预设',
+  'cronPicker.dailyMidnight': '每天（午夜）',
+  'cronPicker.twiceAWeek': '每周两次（周一 + 周四）',
+  'cronPicker.everyTwoWeeks': '每两周（1日 + 15日）',
+  'cronPicker.monthlyFirst': '每月（1日）',
+  'cronPicker.standardHint': '标准 cron：分钟 小时 月内日期 月 星期',
+  'libraryHealth.syncNow': '立即同步',
+  'libraryHealth.snapshotTitle': '最新快照',
+  'libraryHealth.lastSynced': '上次同步',
+  'libraryHealth.neverSynced': '从未同步。',
+  'libraryHealth.autoSyncEvery': '每 {0} 小时自动同步',
+  'systemHealth.librarySync': '资料库同步',
+  'settings.tabs.systemHealth': '系统健康',
+  'settings.targetUpdated': '目标已更新',
+  'settings.targetUpdateFailed': '无法更新目标',
+  'settings.editTarget': '编辑目标',
+  'settings.saveChanges': '保存更改',
+  'settings.linkedTarget': '已关联目标：',
+  'settings.librarySyncInterval': '资料库同步',
+  'settings.librarySyncIntervalHours': '同步间隔（小时）',
+  'settings.librarySyncIntervalHelp': '控制 Digarr 刷新资料库同步数据并自动重新运行检查的频率。',
+  'settings.oidcSso': 'OIDC 与 SSO',
+  'settings.fieldClientId': '客户端 ID',
+  'settings.fieldClientSecret': '客户端密钥',
+  'settings.fieldWebhookUrl': 'Webhook 地址',
+  'settings.aiProviderTitle': 'AI 提供商',
+  'settings.webhookTitle': 'Webhook 集成',
+  'settings.plexToken': 'Plex 令牌',
+  'settings.serviceConnectedToast': '{0} 已连接',
+  'settings.serviceUnreachableToast': '无法访问 {0}',
+  'settings.serviceSettingsSavedToast': '{0} 设置已保存',
+  'settings.serviceSettingsFailedToast': '无法保存 {0} 设置',
+  'settings.spotifyAuthorizationFailed': '无法开始 Spotify 授权',
+  'settings.spotifyDisconnectFailed': '无法断开 Spotify 连接',
+  'settings.spotifyLikedSongsFailed': '无法开始导入 Spotify 点赞歌曲',
+  'settings.spotifyPlaylistFailed': '无法导入 Spotify 播放列表',
+  'settings.deezerFavoritesFailed': '无法开始导入 Deezer 收藏',
+  'settings.deezerFollowedFailed': '无法开始导入 Deezer 已关注艺人',
+  'settings.deezerDisconnectFailed': '无法断开 Deezer 连接',
+  'settings.deezerAuthorizationFailed': '无法开始 Deezer 授权',
+  'settings.reverseProxyConfigHint':
+    '使用 PROXY_AUTH_ENABLED 和 PROXY_AUTH_TRUSTED_PROXIES；你的代理必须发送 X-Forwarded-User。',
+  'searchResultCard.addToQueue': '加入发现队列',
+  'searchResultCard.queuing': '加入中...',
+  'searchResultCard.queuedSuccess': '已将 {0} 加入发现队列',
+  'searchResultCard.queuedFailure': '无法将 {0} 加入队列',
+  'searchResultCard.popularity': '热度 {0}{1}',
+  'admin.restoreSummary': '已在 {1} 张表中恢复 {0} 行',
+  'admin.restoreSummaryForce': '已恢复 {0} 行。请重新输入加密字段的凭据。',
+  'admin.encryptionMismatchMessage': '该备份使用了不同的加密密钥创建。',
+  'admin.encryptionMismatchAffected': '这些字段需要重新输入：{0}',
+  'admin.encryptionMismatchFallback': '部分加密字段可能需要重新输入。',
+  'admin.autoFixInProgress': '自动修复进行中...（已修复 {0}/{1}）',
+  'errorBoundary.title': '出了点问题',
+  'errorBoundary.retry': '重试',
+  'errorBoundary.home': '返回首页',
+  'connectionSuggestions.title': '{0} 的建议',
+  'connectionSuggestions.close': '关闭建议',
+  'connectionSuggestions.gotIt': '知道了',
+  'connectionSuggestions.skipForNow': '暂时跳过',
+  'connectionSuggestions.spotifyDiscoverWeekly': '设置 Discover Weekly 订阅',
+  'connectionSuggestions.spotifyReleaseRadar': '启用 Release Radar 订阅',
+  'connectionSuggestions.lastfmTopGenreTags': '订阅你最常用的流派标签',
+  'connectionSuggestions.listenbrainzWeeklyJams': '启用 Weekly Jams 订阅',
+  'connectionSuggestions.listenbrainzWeeklyExploration': '启用 Weekly Exploration',
+  'connectionSuggestions.weeklyDigestPlaylist': '创建每周 Digarr Digest 播放列表',
+  'keyboardShortcuts.title': '键盘快捷键',
+  'keyboardShortcuts.close': '关闭快捷键',
+  'keyboardShortcuts.nextCard': '下一张卡片',
+  'keyboardShortcuts.previousCard': '上一张卡片',
+  'keyboardShortcuts.approveFocusedCard': '批准当前卡片',
+  'keyboardShortcuts.rejectFocusedCard': '拒绝当前卡片',
+  'keyboardShortcuts.toggleDetailView': '打开 / 关闭详情视图',
+  'keyboardShortcuts.toggleShortcuts': '显示 / 隐藏快捷键',
+  'keyboardShortcuts.previousCardStack': '上一张卡片（堆叠视图）',
+  'keyboardShortcuts.nextCardStack': '下一张卡片（堆叠视图）',
+  'keyboardShortcuts.disabledWhileTyping': '当文本输入框处于焦点状态时，快捷键会被禁用。',
+  'albumPicker.title': '选择专辑',
+  'albumPicker.loadFailed': '加载专辑失败',
+  'albumPicker.selectAll': '全部',
+  'albumPicker.selectNone': '无',
+  'albumPicker.selectedCount': '已选择 {0} 个',
+  'albumPicker.aiPick': 'AI 精选',
+  'albumPicker.monitor': '监控',
+  'albumPicker.monitorSingular': '监控 {0} 张专辑',
+  'albumPicker.monitorPlural': '监控 {0} 张专辑',
+  'approveDialog.title': '此艺术家的 Lidarr 设置',
+  'approveDialog.loadFailed': '加载 Lidarr 配置文件失败。请检查 Lidarr 连接。',
+  'approveDialog.qualityProfile': '质量配置文件',
+  'approveDialog.metadataProfile': '元数据配置文件',
+  'approveDialog.rootFolder': '根目录',
+  'cardStack.counter': '第 {0} / 共 {1}',
+  'cardStack.swipeHint': '向右滑动批准，向左滑动拒绝',
+  'serviceCard.testConnection': '测试连接',
+  'preview.noPreviewAvailable': '该艺术家没有可用的预览',
+  'preview.playbackBlocked': '浏览器阻止了播放 - 请再次点击',
+  'errors.auth.invalidCredentials': '凭据无效',
+  'errors.auth.notAuthenticated': '未认证',
+  'errors.auth.passwordIncorrect': '当前密码不正确',
+  'errors.auth.usernameTaken': '用户名已被占用',
+  'errors.target.notFound': '未找到目标',
+  'errors.pipeline.alreadyRunning': '已有流水线正在运行',
+  'errors.recommendation.notFound': '未找到推荐',
+  'errors.subscription.notFound': '未找到订阅',
+  'errors.playlist.notFound': '未找到播放列表',
+}
